% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_bot}
\alias{connect_associate_bot}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_associate_bot(InstanceId, LexBot = NULL, LexV2Bot = NULL)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{LexBot}{}

\item{LexV2Bot}{The Amazon Lex V2 bot to associate with the instance.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/associate_bot.html} for full documentation.
}
\keyword{internal}
