% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_contact_with_user}
\alias{connect_associate_contact_with_user}
\title{Associates a queued contact with an agent}
\usage{
connect_associate_contact_with_user(InstanceId, ContactId, UserId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact in this instance of Amazon Connect.}

\item{UserId}{[required] The identifier for the user. This can be the ID or the ARN of the user.}
}
\description{
Associates a queued contact with an agent.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_contact_with_user/} for full documentation.
}
\keyword{internal}
