% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_default_vocabulary}
\alias{connect_associate_default_vocabulary}
\title{Associates an existing vocabulary as the default}
\usage{
connect_associate_default_vocabulary(
  InstanceId,
  LanguageCode,
  VocabularyId = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{LanguageCode}{[required] The language code of the vocabulary entries. For a list of languages and
their corresponding language codes, see \href{https://docs.aws.amazon.com/transcribe/latest/dg/what-is.html}{What is Amazon Transcribe?}}

\item{VocabularyId}{The identifier of the custom vocabulary. If this is empty, the default
is set to none.}
}
\description{
Associates an existing vocabulary as the default. Contact Lens for Amazon Connect uses the vocabulary in post-call and real-time analysis sessions for the given language.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_default_vocabulary/} for full documentation.
}
\keyword{internal}
