% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_email_address_alias}
\alias{connect_associate_email_address_alias}
\title{Associates an email address alias with an existing email address in an
Amazon Connect instance}
\usage{
connect_associate_email_address_alias(
  EmailAddressId,
  InstanceId,
  AliasConfiguration,
  ClientToken = NULL
)
}
\arguments{
\item{EmailAddressId}{[required] The identifier of the email address.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{AliasConfiguration}{[required] Configuration object that specifies which email address will serve as
the alias. The specified email address must already exist in the Amazon
Connect instance and cannot already be configured as an alias or have an
alias of its own.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}
}
\description{
Associates an email address alias with an existing email address in an Amazon Connect instance. This creates a forwarding relationship where emails sent to the alias email address are automatically forwarded to the primary email address.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_email_address_alias/} for full documentation.
}
\keyword{internal}
