% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_flow}
\alias{connect_associate_flow}
\title{Associates a connect resource to a flow}
\usage{
connect_associate_flow(InstanceId, ResourceId, FlowId, ResourceType)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ResourceId}{[required] The identifier of the resource.}

\item{FlowId}{[required] The identifier of the flow.}

\item{ResourceType}{[required] A valid resource type.}
}
\description{
Associates a connect resource to a flow.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_flow/} for full documentation.
}
\keyword{internal}
