% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_hours_of_operations}
\alias{connect_associate_hours_of_operations}
\title{Associates a set of hours of operations with another hours of operation}
\usage{
connect_associate_hours_of_operations(
  InstanceId,
  HoursOfOperationId,
  ParentHoursOfOperationConfigs
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{HoursOfOperationId}{[required] The identifier of the child hours of operation.}

\item{ParentHoursOfOperationConfigs}{[required] The Amazon Resource Names (ARNs) of the parent hours of operation resources to associate with the child hours of operation resource.}
}
\description{
Associates a set of hours of operations with another hours of operation. For more information about inheriting overrides from parent hours of operation, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/hours-of-operation-overrides.html}{Hours of operation overrides} in the Administrator Guide.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_hours_of_operations/} for full documentation.
}
\keyword{internal}
