% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_lambda_function}
\alias{connect_associate_lambda_function}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_associate_lambda_function(InstanceId, FunctionArn)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{FunctionArn}{[required] The Amazon Resource Name (ARN) for the Lambda function being associated.
Maximum number of characters allowed is 140.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/associate_lambda_function.html} for full documentation.
}
\keyword{internal}
