% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_queue_email_addresses}
\alias{connect_associate_queue_email_addresses}
\title{Associates a set of email addresses with a queue to enable agents to
select different "From" (system) email addresses when replying to
inbound email contacts or initiating outbound email contacts}
\usage{
connect_associate_queue_email_addresses(
  InstanceId,
  QueueId,
  EmailAddressesConfig,
  ClientToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{EmailAddressesConfig}{[required] Configuration list containing the email addresses to associate with the
queue. Each configuration specifies an email address ID that should be
linked to this queue for routing purposes.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}
}
\description{
Associates a set of email addresses with a queue to enable agents to select different "From" (system) email addresses when replying to inbound email contacts or initiating outbound email contacts. This allows agents to handle email contacts across different brands and business units within the same queue.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_queue_email_addresses/} for full documentation.
}
\keyword{internal}
