% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_queue_quick_connects}
\alias{connect_associate_queue_quick_connects}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_associate_queue_quick_connects(InstanceId, QueueId,
  QuickConnectIds)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{QuickConnectIds}{[required] The quick connects to associate with this queue.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Associates a set of quick connects with a queue.
}
\section{Request syntax}{
\preformatted{svc$associate_queue_quick_connects(
  InstanceId = "string",
  QueueId = "string",
  QuickConnectIds = list(
    "string"
  )
)
}
}

\keyword{internal}
