% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_routing_profile_queues}
\alias{connect_associate_routing_profile_queues}
\title{Associates a set of queues with a routing profile}
\usage{
connect_associate_routing_profile_queues(
  InstanceId,
  RoutingProfileId,
  QueueConfigs = NULL,
  ManualAssignmentQueueConfigs = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{QueueConfigs}{The queues to associate with this routing profile.}

\item{ManualAssignmentQueueConfigs}{The manual assignment queues to associate with this routing profile.

For voice contacts, manual assignment supports only agent-first callback contacts. Chat, email, and task contacts are fully supported.}
}
\description{
Associates a set of queues with a routing profile.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_routing_profile_queues/} for full documentation.
}
\keyword{internal}
