% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_security_profiles}
\alias{connect_associate_security_profiles}
\title{Associate security profiles with an Entity in an Amazon Connect instance}
\usage{
connect_associate_security_profiles(
  InstanceId,
  SecurityProfiles,
  EntityType,
  EntityArn
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the instance
ID in the Amazon Resource Name (ARN) of the instance.}

\item{SecurityProfiles}{[required] List of Security Profile Object.}

\item{EntityType}{[required] Only supported type is AI_AGENT.}

\item{EntityArn}{[required] Arn of a Q in Connect AI Agent.}
}
\description{
Associate security profiles with an Entity in an Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_security_profiles/} for full documentation.
}
\keyword{internal}
