% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_traffic_distribution_group_user}
\alias{connect_associate_traffic_distribution_group_user}
\title{Associates an agent with a traffic distribution group}
\usage{
connect_associate_traffic_distribution_group_user(
  TrafficDistributionGroupId,
  UserId,
  InstanceId
)
}
\arguments{
\item{TrafficDistributionGroupId}{[required] The identifier of the traffic distribution group. This can be the ID or
the ARN if the API is being called in the Region where the traffic
distribution group was created. The ARN must be provided if the call is
from the replicated Region.}

\item{UserId}{[required] The identifier of the user account. This can be the ID or the ARN of the
user.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}
}
\description{
Associates an agent with a traffic distribution group.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_traffic_distribution_group_user/} for full documentation.
}
\keyword{internal}
