% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_user_proficiencies}
\alias{connect_associate_user_proficiencies}
\title{Associates a set of proficiencies with a user}
\usage{
connect_associate_user_proficiencies(InstanceId, UserId, UserProficiencies)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the instance
ID in the Amazon Resource Name (ARN of the instance).}

\item{UserId}{[required] The identifier of the user account.}

\item{UserProficiencies}{[required] The proficiencies to associate with the user.}
}
\description{
\>Associates a set of proficiencies with a user.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_user_proficiencies/} for full documentation.
}
\keyword{internal}
