% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_workspace}
\alias{connect_associate_workspace}
\title{Associates a workspace with one or more users or routing profiles,
allowing them to access the workspace's configured views and pages}
\usage{
connect_associate_workspace(InstanceId, WorkspaceId, ResourceArns)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}

\item{ResourceArns}{[required] The Amazon Resource Names (ARNs) of the resources to associate with the
workspace. Valid resource types are users and routing profiles.}
}
\description{
Associates a workspace with one or more users or routing profiles, allowing them to access the workspace's configured views and pages.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_workspace/} for full documentation.
}
\keyword{internal}
