% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_batch_associate_analytics_data_set}
\alias{connect_batch_associate_analytics_data_set}
\title{Associates a list of analytics datasets for a given Amazon Connect
instance to a target account}
\usage{
connect_batch_associate_analytics_data_set(
  InstanceId,
  DataSetIds,
  TargetAccountId = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{DataSetIds}{[required] An array of dataset identifiers to associate.}

\item{TargetAccountId}{The identifier of the target account. Use to associate a dataset to a
different account than the one containing the Amazon Connect instance.
If not specified, by default this value is the Amazon Web Services
account that has the Amazon Connect instance.}
}
\description{
Associates a list of analytics datasets for a given Amazon Connect instance to a target account. You can associate multiple datasets in a single call.

See \url{https://www.paws-r-sdk.com/docs/connect_batch_associate_analytics_data_set/} for full documentation.
}
\keyword{internal}
