% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_batch_delete_data_table_value}
\alias{connect_batch_delete_data_table_value}
\title{Deletes multiple values from a data table}
\usage{
connect_batch_delete_data_table_value(InstanceId, DataTableId, Values)
}
\arguments{
\item{InstanceId}{[required] The unique identifier for the Amazon Connect instance.}

\item{DataTableId}{[required] The unique identifier for the data table. Must also accept the table ARN
with or without a version alias.}

\item{Values}{[required] A list of value identifiers to delete, each specifying primary values,
attribute name, and lock version information.}
}
\description{
Deletes multiple values from a data table. API users may delete values at any time. When deletion is requested from the admin website, a warning is shown alerting the user of the most recent time the attribute and its values were accessed. System managed values are not deletable by customers.

See \url{https://www.paws-r-sdk.com/docs/connect_batch_delete_data_table_value/} for full documentation.
}
\keyword{internal}
