% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_batch_get_flow_association}
\alias{connect_batch_get_flow_association}
\title{Retrieve the flow associations for the given resources}
\usage{
connect_batch_get_flow_association(
  InstanceId,
  ResourceIds,
  ResourceType = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ResourceIds}{[required] A list of resource identifiers to retrieve flow associations.}

\item{ResourceType}{The type of resource association.}
}
\description{
Retrieve the flow associations for the given resources.

See \url{https://www.paws-r-sdk.com/docs/connect_batch_get_flow_association/} for full documentation.
}
\keyword{internal}
