% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_batch_put_contact}
\alias{connect_batch_put_contact}
\title{Only the Amazon Connect outbound campaigns service principal is allowed
to assume a role in your account and call this API}
\usage{
connect_batch_put_contact(
  ClientToken = NULL,
  InstanceId,
  ContactDataRequestList
)
}
\arguments{
\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactDataRequestList}{[required] List of individual contact requests.}
}
\description{
Only the Amazon Connect outbound campaigns service principal is allowed to assume a role in your account and call this API.

See \url{https://www.paws-r-sdk.com/docs/connect_batch_put_contact/} for full documentation.
}
\keyword{internal}
