% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_batch_update_data_table_value}
\alias{connect_batch_update_data_table_value}
\title{Updates multiple data table values using all properties from
BatchCreateDataTableValue}
\usage{
connect_batch_update_data_table_value(InstanceId, DataTableId, Values)
}
\arguments{
\item{InstanceId}{[required] The unique identifier for the Amazon Connect instance.}

\item{DataTableId}{[required] The unique identifier for the data table. Must also accept the table ARN
with or without a version alias.}

\item{Values}{[required] A list of values to update, each including the current lock version to
ensure optimistic locking.}
}
\description{
Updates multiple data table values using all properties from BatchCreateDataTableValue. System managed values are not modifiable by customers. The operation requires proper lock versions to prevent concurrent modification conflicts.

See \url{https://www.paws-r-sdk.com/docs/connect_batch_update_data_table_value/} for full documentation.
}
\keyword{internal}
