% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_attached_file}
\alias{connect_create_attached_file}
\title{Creates an attached file for a completed voice contact by copying a
recording from a source S3 URI into Connect Customer managed storage}
\usage{
connect_create_attached_file(
  ClientToken = NULL,
  InstanceId,
  FileUseCaseType,
  FileSourceUri,
  AssociatedResourceArn,
  Tags = NULL
)
}
\arguments{
\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see \href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{FileUseCaseType}{[required] The use case for the file.

Only \code{VOICE_RECORDING} is supported.}

\item{FileSourceUri}{[required] The S3 URI of the file to be attached. Only S3 source URIs are supported.}

\item{AssociatedResourceArn}{[required] The ARN of the completed voice contact to attach the file to. Only voice contacts with Telephony subtype are supported.

This value must be a valid ARN.}

\item{Tags}{The tags used to organize, track, or control access for this resource. For example, \verb{\{ "Tags": \{"key1":"value1", "key2":"value2"\} \}}.}
}
\description{
Creates an attached file for a completed voice contact by copying a recording from a source S3 URI into Connect Customer managed storage. Use this API to attach voice recordings to contacts for downstream processing such as conversational analytics.

See \url{https://www.paws-r-sdk.com/docs/connect_create_attached_file/} for full documentation.
}
\keyword{internal}
