% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_auth_code}
\alias{connect_create_auth_code}
\title{Creates an authorization code for the specified Connect Customer
instance}
\usage{
connect_create_auth_code(
  InstanceId,
  Scope,
  MaxSessionDurationMinutes = NULL,
  SessionInactivityDurationMinutes
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{Scope}{[required] The scope for the authorization code. Defines the permissions and access boundaries for the session.}

\item{MaxSessionDurationMinutes}{The maximum duration of the session, in minutes. Minimum value of 1440 (24 hours). Maximum value of 43200 (30 days). If no value is provided, the session will expire after 400 days.}

\item{SessionInactivityDurationMinutes}{[required] The duration of inactivity, in minutes, after which the session expires. Minimum value of 1440 (24 hours). Maximum value of 20160 (14 days).}
}
\description{
Creates an authorization code for the specified Connect Customer instance. The authorization code can be used to establish a session with scoped permissions defined by the specified scope parameters.

See \url{https://www.paws-r-sdk.com/docs/connect_create_auth_code/} for full documentation.
}
\keyword{internal}
