% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_contact_flow}
\alias{connect_create_contact_flow}
\title{Creates a contact flow for the specified Amazon Connect instance}
\usage{
connect_create_contact_flow(
  InstanceId,
  Name,
  Type,
  Description = NULL,
  Content,
  Tags = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{Name}{[required] The name of the contact flow.}

\item{Type}{[required] The type of the contact flow. For descriptions of the available types,
see \href{https://docs.aws.amazon.com/connect/latest/adminguide/create-contact-flow.html#contact-flow-types}{Choose a Contact Flow Type}
in the \emph{Amazon Connect Administrator Guide}.}

\item{Description}{The description of the contact flow.}

\item{Content}{[required] The content of the contact flow.}

\item{Tags}{One or more tags.}
}
\description{
Creates a contact flow for the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/create_contact_flow.html} for full documentation.
}
\keyword{internal}
