% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_contact_flow_module}
\alias{connect_create_contact_flow_module}
\title{Creates a contact flow module for the specified Amazon Connect instance}
\usage{
connect_create_contact_flow_module(
  InstanceId,
  Name,
  Description = NULL,
  Content,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{Name}{[required] The name of the contact flow module.}

\item{Description}{The description of the contact flow module.}

\item{Content}{[required] The content of the contact flow module.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Creates a contact flow module for the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/create_contact_flow_module.html} for full documentation.
}
\keyword{internal}
