% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_contact_flow_module_alias}
\alias{connect_create_contact_flow_module_alias}
\title{Creates a named alias that points to a specific version of a contact
flow module}
\usage{
connect_create_contact_flow_module_alias(
  InstanceId,
  Description = NULL,
  ContactFlowModuleId,
  ContactFlowModuleVersion,
  AliasName
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{Description}{The description of the alias.}

\item{ContactFlowModuleId}{[required] The identifier of the flow module.}

\item{ContactFlowModuleVersion}{[required] The version of the flow module.}

\item{AliasName}{[required] The name of the alias.}
}
\description{
Creates a named alias that points to a specific version of a contact flow module.

See \url{https://www.paws-r-sdk.com/docs/connect_create_contact_flow_module_alias/} for full documentation.
}
\keyword{internal}
