% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_contact_flow_module_version}
\alias{connect_create_contact_flow_module_version}
\title{Creates an immutable snapshot of a contact flow module, preserving its
content and settings at a specific point in time for version control and
rollback capabilities}
\usage{
connect_create_contact_flow_module_version(
  InstanceId,
  Description = NULL,
  ContactFlowModuleId,
  FlowModuleContentSha256 = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{Description}{The description of the flow module version.}

\item{ContactFlowModuleId}{[required] The identifier of the flow module.}

\item{FlowModuleContentSha256}{Indicates the checksum value of the flow module content.}
}
\description{
Creates an immutable snapshot of a contact flow module, preserving its content and settings at a specific point in time for version control and rollback capabilities.

See \url{https://www.paws-r-sdk.com/docs/connect_create_contact_flow_module_version/} for full documentation.
}
\keyword{internal}
