% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_data_table}
\alias{connect_create_data_table}
\title{Creates a new data table with the specified properties}
\usage{
connect_create_data_table(
  InstanceId,
  Name,
  Description = NULL,
  TimeZone,
  ValueLockLevel,
  Status,
  Tags = NULL
)
}
\arguments{
\item{InstanceId}{[required] The unique identifier for the Amazon Connect instance where the data
table will be created.}

\item{Name}{[required] The name for the data table. Must conform to Connect human readable
string specification and have 1-127 characters. Whitespace must be
trimmed first. Must not start with the reserved case insensitive values
'connect:' and 'aws:'. Must be unique for the instance using
case-insensitive comparison.}

\item{Description}{An optional description for the data table. Must conform to Connect
human readable string specification and have 0-250 characters.
Whitespace must be trimmed first.}

\item{TimeZone}{[required] The IANA timezone identifier to use when resolving time based dynamic
values. Required even if no time slices are specified.}

\item{ValueLockLevel}{[required] The data level that concurrent value edits are locked on. One of
DATA_TABLE, PRIMARY_VALUE, ATTRIBUTE, VALUE, and NONE. NONE is the
default if unspecified. This determines how concurrent edits are handled
when multiple users attempt to modify values simultaneously.}

\item{Status}{[required] The status of the data table. One of PUBLISHED or SAVED. Required
parameter that determines the initial state of the table.}

\item{Tags}{Key value pairs for attribute based access control (TBAC or ABAC).
Optional tags to apply to the data table for organization and access
control purposes.}
}
\description{
Creates a new data table with the specified properties. Supports the creation of all table properties except for attributes and values. A table with no attributes and values is a valid state for a table. The number of tables per instance is limited to 100 per instance. Customers can request an increase by using Amazon Web Services Service Quotas.

See \url{https://www.paws-r-sdk.com/docs/connect_create_data_table/} for full documentation.
}
\keyword{internal}
