% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_data_table_attribute}
\alias{connect_create_data_table_attribute}
\title{Adds an attribute to an existing data table}
\usage{
connect_create_data_table_attribute(
  InstanceId,
  DataTableId,
  Name,
  ValueType,
  Description = NULL,
  Primary = NULL,
  Validation = NULL
)
}
\arguments{
\item{InstanceId}{[required] The unique identifier for the Amazon Connect instance.}

\item{DataTableId}{[required] The unique identifier for the data table. Must also accept the table ARN
with or without a version alias. If the version is provided as part of
the identifier or ARN, the version must be one of the two available
system managed aliases, $SAVED or $LATEST.}

\item{Name}{[required] The name for the attribute. Must conform to Connect human readable
string specification and have 1-127 characters. Must not start with the
reserved case insensitive values 'connect:' and 'aws:'. Whitespace
trimmed before persisting. Must be unique for the data table using
case-insensitive comparison.}

\item{ValueType}{[required] The type of value allowed or the resultant type after the value's
expression is evaluated. Must be one of TEXT, TEXT_LIST, NUMBER,
NUMBER_LIST, and BOOLEAN.}

\item{Description}{An optional description for the attribute. Must conform to Connect human
readable string specification and have 0-250 characters. Whitespace
trimmed before persisting.}

\item{Primary}{Optional boolean that defaults to false. Determines if the value is used
to identify a record in the table. Values for primary attributes must
not be expressions.}

\item{Validation}{Optional validation rules for the attribute. Borrows heavily from JSON
Schema - Draft 2020-12. The maximum length of arrays within validations
and depth of validations is 5. There are default limits that apply to
all types. Customer specified limits in excess of the default limits are
not permitted.}
}
\description{
Adds an attribute to an existing data table. Creating a new primary attribute uses the empty value for the specified value type for all existing records. This should not affect uniqueness of published data tables since the existing primary values will already be unique. Creating attributes does not create any values. System managed tables may not allow customers to create new attributes.

See \url{https://www.paws-r-sdk.com/docs/connect_create_data_table_attribute/} for full documentation.
}
\keyword{internal}
