% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_predefined_attribute}
\alias{connect_create_predefined_attribute}
\title{Creates a new predefined attribute for the specified Amazon Connect
instance}
\usage{
connect_create_predefined_attribute(
  InstanceId,
  Name,
  Values = NULL,
  Purposes = NULL,
  AttributeConfiguration = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the instance
ID in the Amazon Resource Name (ARN) of the instance.}

\item{Name}{[required] The name of the predefined attribute.}

\item{Values}{The values of the predefined attribute.}

\item{Purposes}{Values that enable you to categorize your predefined attributes. You can
use them in custom UI elements across the Amazon Connect admin website.}

\item{AttributeConfiguration}{Custom metadata that is associated to predefined attributes to control
behavior in upstream services, such as controlling how a predefined
attribute should be displayed in the Amazon Connect admin website.}
}
\description{
Creates a new predefined attribute for the specified Amazon Connect instance. A \emph{predefined attribute} is made up of a name and a value.

See \url{https://www.paws-r-sdk.com/docs/connect_create_predefined_attribute/} for full documentation.
}
\keyword{internal}
