% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_queue}
\alias{connect_create_queue}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_create_queue(InstanceId, Name, Description,
  OutboundCallerConfig, HoursOfOperationId, MaxContacts, QuickConnectIds,
  Tags)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{Name}{[required] The name of the queue.}

\item{Description}{The description of the queue.}

\item{OutboundCallerConfig}{The outbound caller ID name, number, and outbound whisper flow.}

\item{HoursOfOperationId}{[required] The identifier for the hours of operation.}

\item{MaxContacts}{The maximum number of contacts that can be in the queue before it is
considered full.}

\item{QuickConnectIds}{The quick connects available to agents who are working the queue.}

\item{Tags}{One or more tags.}
}
\value{
A list with the following syntax:\preformatted{list(
  QueueArn = "string",
  QueueId = "string"
)
}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Creates a new queue for the specified Amazon Connect instance.
}
\section{Request syntax}{
\preformatted{svc$create_queue(
  InstanceId = "string",
  Name = "string",
  Description = "string",
  OutboundCallerConfig = list(
    OutboundCallerIdName = "string",
    OutboundCallerIdNumberId = "string",
    OutboundFlowId = "string"
  ),
  HoursOfOperationId = "string",
  MaxContacts = 123,
  QuickConnectIds = list(
    "string"
  ),
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
