% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_quick_connect}
\alias{connect_create_quick_connect}
\title{Creates a quick connect for the specified Amazon Connect instance}
\usage{
connect_create_quick_connect(
  InstanceId,
  Name,
  Description = NULL,
  QuickConnectConfig,
  Tags = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{Name}{[required] The name of the quick connect.}

\item{Description}{The description of the quick connect.}

\item{QuickConnectConfig}{[required] Configuration settings for the quick connect.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Creates a quick connect for the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/create_quick_connect.html} for full documentation.
}
\keyword{internal}
