% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_rule}
\alias{connect_create_rule}
\title{Creates a rule for the specified Connect Customer instance}
\usage{
connect_create_rule(
  InstanceId,
  Name,
  TriggerEventSource,
  Function,
  Actions,
  PublishStatus,
  PreEvaluationFilters = NULL,
  ClientToken = NULL,
  Tags = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{Name}{[required] A unique name for the rule.}

\item{TriggerEventSource}{[required] The event source to trigger the rule.}

\item{Function}{[required] The conditions of the rule.}

\item{Actions}{[required] A list of actions to be run when the rule is triggered.}

\item{PublishStatus}{[required] The publish status of the rule.}

\item{PreEvaluationFilters}{The pre-evaluation filters for the rule, that restrict the rule to be applied to only certain resources based on the resource's attributes, such as tags assigned to a contact. The pre-evaluation filters are applied even before rule conditions are evaluated and are used to enforce tag-based-access-control while applying rules.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see \href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{Tags}{The tags used to organize, track, or control access for this resource. For example, \{ "Tags": \{"key1":"value1", "key2":"value2"\} \}.}
}
\description{
Creates a rule for the specified Connect Customer instance.

See \url{https://www.paws-r-sdk.com/docs/connect_create_rule/} for full documentation.
}
\keyword{internal}
