% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_test_case}
\alias{connect_create_test_case}
\title{Creates a test case with its content and metadata for the specified
Amazon Connect instance}
\usage{
connect_create_test_case(
  InstanceId,
  Name,
  Description = NULL,
  Content,
  EntryPoint = NULL,
  InitializationData = NULL,
  Status = NULL,
  TestCaseId = NULL,
  Tags = NULL,
  LastModifiedTime = NULL,
  LastModifiedRegion = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{Name}{[required] The name of the test.}

\item{Description}{The description of the test.}

\item{Content}{[required] The JSON string that represents the content of the test.}

\item{EntryPoint}{Defines the starting point for your test.}

\item{InitializationData}{Defines the initial custom attributes for your test.}

\item{Status}{Indicates the test status as either SAVED or PUBLISHED. The PUBLISHED
status will initiate validation on the content. The SAVED status does
not initiate validation of the content.}

\item{TestCaseId}{Id of the test case if you want to create it in a replica region using
Amazon Connect Global Resiliency}

\item{Tags}{The tags used to organize, track, or control access for this resource.}

\item{LastModifiedTime}{The time at which the resource was last modified.}

\item{LastModifiedRegion}{The region in which the resource was last modified}
}
\description{
Creates a test case with its content and metadata for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_create_test_case/} for full documentation.
}
\keyword{internal}
