% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_user_hierarchy_group}
\alias{connect_create_user_hierarchy_group}
\title{Creates a new user hierarchy group}
\usage{
connect_create_user_hierarchy_group(
  Name,
  ParentGroupId = NULL,
  InstanceId,
  Tags = NULL
)
}
\arguments{
\item{Name}{[required] The name of the user hierarchy group. Must not be more than 100
characters.}

\item{ParentGroupId}{The identifier for the parent hierarchy group. The user hierarchy is
created at level one if the parent group ID is null.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{Tags}{The tags used to organize, track, or control access for this resource.
For example, \{ "tags": \{"key1":"value1", "key2":"value2"\} \}.}
}
\description{
Creates a new user hierarchy group.

See \url{https://paws-r.github.io/docs/connect/create_user_hierarchy_group.html} for full documentation.
}
\keyword{internal}
