% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_workspace}
\alias{connect_create_workspace}
\title{Creates a workspace that defines the user experience by mapping views to
pages}
\usage{
connect_create_workspace(
  InstanceId,
  Name,
  Description = NULL,
  Theme = NULL,
  Title = NULL,
  Tags = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{Name}{[required] The name of the workspace. Must be unique within the instance and can
contain 1-127 characters.}

\item{Description}{The description of the workspace. Maximum length is 250 characters.}

\item{Theme}{The theme configuration for the workspace, including colors and styling.}

\item{Title}{The title displayed for the workspace.}

\item{Tags}{The tags used to organize, track, or control access for this resource.
For example, \verb{\{ "Tags": \{"key1":"value1", "key2":"value2"\} \}}.}
}
\description{
Creates a workspace that defines the user experience by mapping views to pages. Workspaces can be assigned to users or routing profiles.

See \url{https://www.paws-r-sdk.com/docs/connect_create_workspace/} for full documentation.
}
\keyword{internal}
