% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_workspace_page}
\alias{connect_create_workspace_page}
\title{Associates a view with a page in a workspace, defining what users see
when they navigate to that page}
\usage{
connect_create_workspace_page(
  InstanceId,
  WorkspaceId,
  ResourceArn,
  Page,
  Slug = NULL,
  InputData = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}

\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the view to associate with the page.}

\item{Page}{[required] The page identifier. Valid system pages include \code{HOME} and
\code{AGENT_EXPERIENCE}. Custom pages cannot use the \verb{aws:} or \verb{connect:}
prefixes.}

\item{Slug}{The URL-friendly identifier for the page.}

\item{InputData}{A JSON string containing input parameters for the view, validated
against the view's input schema.}
}
\description{
Associates a view with a page in a workspace, defining what users see when they navigate to that page.

See \url{https://www.paws-r-sdk.com/docs/connect_create_workspace_page/} for full documentation.
}
\keyword{internal}
