% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_attached_file}
\alias{connect_delete_attached_file}
\title{Deletes an attached file along with the underlying S3 Object}
\usage{
connect_delete_attached_file(InstanceId, FileId, AssociatedResourceArn)
}
\arguments{
\item{InstanceId}{[required] The unique identifier of the Connect instance.}

\item{FileId}{[required] The unique identifier of the attached file resource.}

\item{AssociatedResourceArn}{[required] The resource to which the attached file is (being) uploaded to. The supported resources are \href{https://docs.aws.amazon.com/connect/latest/adminguide/cases.html}{Cases}, \href{https://docs.aws.amazon.com/connect/latest/adminguide/setup-email-channel.html}{Email}, and \href{https://docs.aws.amazon.com/connect/latest/adminguide/concepts-getting-started-tasks.html}{Task}.

This value must be a valid ARN.}
}
\description{
Deletes an attached file along with the underlying S3 Object.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_attached_file/} for full documentation.
}
\keyword{internal}
