% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_contact_data}
\alias{connect_delete_contact_data}
\title{Deletes the specified fields containing personally identifiable
information (PII) from a contact in the specified Connect Customer
instance}
\usage{
connect_delete_contact_data(InstanceId, ContactId, ContactFields)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact. You can delete PII only from a contact that has been disconnected (is in a terminated state).}

\item{ContactFields}{[required] The categories of PII to redact from the contact. Specify one or more of the following values:
\itemize{
\item \code{CUSTOMER_ENDPOINT} – The customer's contact endpoint.
\item \code{ADDITIONAL_EMAIL_RECIPIENTS} – Additional recipients on an email contact (email channel only).
\item \code{EMAIL_SUBJECT} – The subject line of an email contact (email channel only).
}}
}
\description{
Deletes the specified fields containing personally identifiable information (PII) from a contact in the specified Connect Customer instance. We redact PII (such as customer endpoints, additional email recipients, and the email subject) from the contact and its associated contact trace record (CTR). The contact must be in a terminated state.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_contact_data/} for full documentation.
}
\keyword{internal}
