% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_contact_flow_module_alias}
\alias{connect_delete_contact_flow_module_alias}
\title{Removes an alias reference, breaking the named connection to the
underlying module version without affecting the version itself}
\usage{
connect_delete_contact_flow_module_alias(
  InstanceId,
  ContactFlowModuleId,
  AliasId
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowModuleId}{[required] The identifier of the flow module.}

\item{AliasId}{[required] The identifier of the alias.}
}
\description{
Removes an alias reference, breaking the named connection to the underlying module version without affecting the version itself.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_contact_flow_module_alias/} for full documentation.
}
\keyword{internal}
