% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_contact_flow_module_version}
\alias{connect_delete_contact_flow_module_version}
\title{Removes a specific version of a contact flow module}
\usage{
connect_delete_contact_flow_module_version(
  InstanceId,
  ContactFlowModuleId,
  ContactFlowModuleVersion
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowModuleId}{[required] The identifier of the flow module.}

\item{ContactFlowModuleVersion}{[required] The version of the flow module to delete.}
}
\description{
Removes a specific version of a contact flow module.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_contact_flow_module_version/} for full documentation.
}
\keyword{internal}
