% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_data_table_attribute}
\alias{connect_delete_data_table_attribute}
\title{Deletes an attribute and all its values from a data table}
\usage{
connect_delete_data_table_attribute(InstanceId, DataTableId, AttributeName)
}
\arguments{
\item{InstanceId}{[required] The unique identifier for the Amazon Connect instance.}

\item{DataTableId}{[required] The unique identifier for the data table.}

\item{AttributeName}{[required] The name of the attribute to delete.}
}
\description{
Deletes an attribute and all its values from a data table.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_data_table_attribute/} for full documentation.
}
\keyword{internal}
