% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_email_address}
\alias{connect_delete_email_address}
\title{Deletes email address from the specified Amazon Connect instance}
\usage{
connect_delete_email_address(InstanceId, EmailAddressId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{EmailAddressId}{[required] The identifier of the email address.}
}
\description{
Deletes email address from the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_email_address/} for full documentation.
}
\keyword{internal}
