% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_extraction_definition}
\alias{connect_delete_extraction_definition}
\title{Deletes an extraction definition from the specified Connect Customer
instance}
\usage{
connect_delete_extraction_definition(InstanceId, ExtractionDefinitionId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{ExtractionDefinitionId}{[required] The identifier of the extraction definition to delete.}
}
\description{
Deletes an extraction definition from the specified Connect Customer instance.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_extraction_definition/} for full documentation.
}
\keyword{internal}
