% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_instance}
\alias{connect_delete_instance}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_delete_instance(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/delete_instance.html} for full documentation.
}
\keyword{internal}
