% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_integration_association}
\alias{connect_delete_integration_association}
\title{Deletes an Amazon Web Services resource association from an Amazon
Connect instance}
\usage{
connect_delete_integration_association(InstanceId, IntegrationAssociationId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{IntegrationAssociationId}{[required] The identifier for the integration association.}
}
\description{
Deletes an Amazon Web Services resource association from an Amazon Connect instance. The association must not have any use cases associated with it.

See \url{https://paws-r.github.io/docs/connect/delete_integration_association.html} for full documentation.
}
\keyword{internal}
