% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_metric}
\alias{connect_delete_metric}
\title{Deletes an existing metric from the specified Connect Customer instance}
\usage{
connect_delete_metric(InstanceId, MetricId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{MetricId}{[required] The identifier of the metric to delete.}
}
\description{
Deletes an existing metric from the specified Connect Customer instance. This operation fails with \code{ResourceConflictException} if the metric is currently in use in a dashboard.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_metric/} for full documentation.
}
\keyword{internal}
