% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_notification}
\alias{connect_delete_notification}
\title{Deletes a notification}
\usage{
connect_delete_notification(InstanceId, NotificationId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{NotificationId}{[required] The unique identifier for the notification to delete.}
}
\description{
Deletes a notification. Once deleted, the notification is no longer visible to all users and cannot be managed through the Admin Website or APIs.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_notification/} for full documentation.
}
\keyword{internal}
