% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_prompt}
\alias{connect_delete_prompt}
\title{Deletes a prompt}
\usage{
connect_delete_prompt(InstanceId, PromptId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{PromptId}{[required] A unique identifier for the prompt.}
}
\description{
Deletes a prompt.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_prompt/} for full documentation.
}
\keyword{internal}
