% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_push_notification_registration}
\alias{connect_delete_push_notification_registration}
\title{Deletes registration for a device token and a chat contact}
\usage{
connect_delete_push_notification_registration(
  InstanceId,
  RegistrationId,
  ContactId
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{RegistrationId}{[required] The identifier for the registration.}

\item{ContactId}{[required] The identifier of the contact within the Amazon Connect instance.}
}
\description{
Deletes registration for a device token and a chat contact.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_push_notification_registration/} for full documentation.
}
\keyword{internal}
