% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_security_profile}
\alias{connect_delete_security_profile}
\title{Deletes a security profile}
\usage{
connect_delete_security_profile(InstanceId, SecurityProfileId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{SecurityProfileId}{[required] The identifier for the security profle.}
}
\description{
Deletes a security profile.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_security_profile/} for full documentation.
}
\keyword{internal}
