% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_session}
\alias{connect_delete_session}
\title{Deletes a session for the specified Connect Customer instance}
\usage{
connect_delete_session(InstanceId, SessionId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{SessionId}{[required] The identifier of the session to delete.}
}
\description{
Deletes a session for the specified Connect Customer instance.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_session/} for full documentation.
}
\keyword{internal}
