% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_task_template}
\alias{connect_delete_task_template}
\title{Deletes the task template}
\usage{
connect_delete_task_template(InstanceId, TaskTemplateId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{TaskTemplateId}{[required] A unique identifier for the task template.}
}
\description{
Deletes the task template.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_task_template/} for full documentation.
}
\keyword{internal}
