% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_test_case}
\alias{connect_delete_test_case}
\title{Deletes the test case that has already been created for the specified
Amazon Connect instance}
\usage{
connect_delete_test_case(InstanceId, TestCaseId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{TestCaseId}{[required] The identifier of the test case to delete.}
}
\description{
Deletes the test case that has already been created for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_test_case/} for full documentation.
}
\keyword{internal}
