% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_use_case}
\alias{connect_delete_use_case}
\title{Deletes a use case from an integration association}
\usage{
connect_delete_use_case(InstanceId, IntegrationAssociationId, UseCaseId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{IntegrationAssociationId}{[required] The identifier for the integration association.}

\item{UseCaseId}{[required] The identifier for the use case.}
}
\description{
Deletes a use case from an integration association.

See \url{https://paws-r.github.io/docs/connect/delete_use_case.html} for full documentation.
}
\keyword{internal}
