% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_view}
\alias{connect_delete_view}
\title{Deletes the view entirely}
\usage{
connect_delete_view(InstanceId, ViewId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ViewId}{[required] The identifier of the view. Both \code{ViewArn} and \code{ViewId} can be used.}
}
\description{
Deletes the view entirely. It deletes the view and all associated qualifiers (versions and aliases).

See \url{https://www.paws-r-sdk.com/docs/connect_delete_view/} for full documentation.
}
\keyword{internal}
