% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_view_version}
\alias{connect_delete_view_version}
\title{Deletes the particular version specified in ViewVersion identifier}
\usage{
connect_delete_view_version(InstanceId, ViewId, ViewVersion)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ViewId}{[required] The identifier of the view. Both \code{ViewArn} and \code{ViewId} can be used.}

\item{ViewVersion}{[required] The version number of the view.}
}
\description{
Deletes the particular version specified in \code{ViewVersion} identifier.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_view_version/} for full documentation.
}
\keyword{internal}
