% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_vocabulary}
\alias{connect_delete_vocabulary}
\title{Deletes the vocabulary that has the given identifier}
\usage{
connect_delete_vocabulary(InstanceId, VocabularyId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{VocabularyId}{[required] The identifier of the custom vocabulary.}
}
\description{
Deletes the vocabulary that has the given identifier.

See \url{https://paws-r.github.io/docs/connect/delete_vocabulary.html} for full documentation.
}
\keyword{internal}
