% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_workspace}
\alias{connect_delete_workspace}
\title{Deletes a workspace and removes all associated view and resource
assignments}
\usage{
connect_delete_workspace(InstanceId, WorkspaceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}
}
\description{
Deletes a workspace and removes all associated view and resource assignments.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_workspace/} for full documentation.
}
\keyword{internal}
