% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_workspace_media}
\alias{connect_delete_workspace_media}
\title{Deletes a media asset (such as a logo) from a workspace}
\usage{
connect_delete_workspace_media(InstanceId, WorkspaceId, MediaType)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}

\item{MediaType}{[required] The type of media to delete. Valid values are: \code{IMAGE_LOGO_FAVICON} and
\code{IMAGE_LOGO_HORIZONTAL}.}
}
\description{
Deletes a media asset (such as a logo) from a workspace.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_workspace_media/} for full documentation.
}
\keyword{internal}
