% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_workspace_page}
\alias{connect_delete_workspace_page}
\title{Removes the association between a view and a page in a workspace}
\usage{
connect_delete_workspace_page(InstanceId, WorkspaceId, Page)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}

\item{Page}{[required] The page identifier.}
}
\description{
Removes the association between a view and a page in a workspace. The page will display the default view after deletion.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_workspace_page/} for full documentation.
}
\keyword{internal}
