% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_authentication_profile}
\alias{connect_describe_authentication_profile}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_authentication_profile(AuthenticationProfileId, InstanceId)
}
\arguments{
\item{AuthenticationProfileId}{[required] A unique identifier for the authentication profile.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change. To request access to this API, contact Amazon Web Services Support.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_authentication_profile/} for full documentation.
}
\keyword{internal}
