% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_contact_evaluation}
\alias{connect_describe_contact_evaluation}
\title{Describes a contact evaluation in the specified Amazon Connect instance}
\usage{
connect_describe_contact_evaluation(InstanceId, EvaluationId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{EvaluationId}{[required] A unique identifier for the contact evaluation.}
}
\description{
Describes a contact evaluation in the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/describe_contact_evaluation.html} for full documentation.
}
\keyword{internal}
