% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_contact_flow_module}
\alias{connect_describe_contact_flow_module}
\title{Describes the specified flow module}
\usage{
connect_describe_contact_flow_module(InstanceId, ContactFlowModuleId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowModuleId}{[required] The identifier of the flow module.}
}
\description{
Describes the specified flow module.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_contact_flow_module/} for full documentation.
}
\keyword{internal}
