% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_contact_flow_module_alias}
\alias{connect_describe_contact_flow_module_alias}
\title{Retrieves detailed information about a specific alias, including which
version it currently points to and its metadata}
\usage{
connect_describe_contact_flow_module_alias(
  InstanceId,
  ContactFlowModuleId,
  AliasId
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowModuleId}{[required] The identifier of the flow module.}

\item{AliasId}{[required] The identifier of the alias.}
}
\description{
Retrieves detailed information about a specific alias, including which version it currently points to and its metadata.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_contact_flow_module_alias/} for full documentation.
}
\keyword{internal}
