% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_hours_of_operation}
\alias{connect_describe_hours_of_operation}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_hours_of_operation(InstanceId, HoursOfOperationId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{HoursOfOperationId}{[required] The identifier for the hours of operation.}
}
\value{
A list with the following syntax:\preformatted{list(
  HoursOfOperation = list(
    HoursOfOperationId = "string",
    HoursOfOperationArn = "string",
    Name = "string",
    Description = "string",
    TimeZone = "string",
    Config = list(
      list(
        Day = "SUNDAY"|"MONDAY"|"TUESDAY"|"WEDNESDAY"|"THURSDAY"|"FRIDAY"|"SATURDAY",
        StartTime = list(
          Hours = 123,
          Minutes = 123
        ),
        EndTime = list(
          Hours = 123,
          Minutes = 123
        )
      )
    ),
    Tags = list(
      "string"
    )
  )
)
}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Describes the hours of operation.
}
\section{Request syntax}{
\preformatted{svc$describe_hours_of_operation(
  InstanceId = "string",
  HoursOfOperationId = "string"
)
}
}

\keyword{internal}
