% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_hours_of_operation_override}
\alias{connect_describe_hours_of_operation_override}
\title{Describes the hours of operation override}
\usage{
connect_describe_hours_of_operation_override(
  InstanceId,
  HoursOfOperationId,
  HoursOfOperationOverrideId
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{HoursOfOperationId}{[required] The identifier for the hours of operation.}

\item{HoursOfOperationOverrideId}{[required] The identifier for the hours of operation override.}
}
\description{
Describes the hours of operation override.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_hours_of_operation_override/} for full documentation.
}
\keyword{internal}
