% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_instance_attribute}
\alias{connect_describe_instance_attribute}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_instance_attribute(InstanceId, AttributeType)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{AttributeType}{[required] The type of attribute.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/describe_instance_attribute.html} for full documentation.
}
\keyword{internal}
