% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_metric}
\alias{connect_describe_metric}
\title{Retrieves the full definition of an existing metric from the specified
Connect Customer instance}
\usage{
connect_describe_metric(InstanceId, MetricId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{MetricId}{[required] The identifier of the metric to describe. Adding the \verb{$SAVED} qualifier will describe the saved version of the metric. Adding \verb{$LATEST} or omitting a qualifier will describe the published version.}
}
\description{
Retrieves the full definition of an existing metric from the specified Connect Customer instance.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_metric/} for full documentation.
}
\keyword{internal}
