% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_notification}
\alias{connect_describe_notification}
\title{Retrieves detailed information about a specific notification, including
its content, priority, recipients, and metadata}
\usage{
connect_describe_notification(InstanceId, NotificationId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{NotificationId}{[required] The unique identifier for the notification.}
}
\description{
Retrieves detailed information about a specific notification, including its content, priority, recipients, and metadata.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_notification/} for full documentation.
}
\keyword{internal}
