% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_phone_number}
\alias{connect_describe_phone_number}
\title{Gets details and status of a phone number that’s claimed to your Amazon
Connect instance or traffic distribution group}
\usage{
connect_describe_phone_number(PhoneNumberId)
}
\arguments{
\item{PhoneNumberId}{[required] A unique identifier for the phone number.}
}
\description{
Gets details and status of a phone number that’s claimed to your Amazon Connect instance or traffic distribution group.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_phone_number/} for full documentation.
}
\keyword{internal}
