% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_prompt}
\alias{connect_describe_prompt}
\title{Describes the prompt}
\usage{
connect_describe_prompt(InstanceId, PromptId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{PromptId}{[required] A unique identifier for the prompt.}
}
\description{
Describes the prompt.

See \url{https://paws-r.github.io/docs/connect/describe_prompt.html} for full documentation.
}
\keyword{internal}
