% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_queue}
\alias{connect_describe_queue}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_queue(InstanceId, QueueId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{QueueId}{[required] The identifier for the queue.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/describe_queue.html} for full documentation.
}
\keyword{internal}
