% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_quick_connect}
\alias{connect_describe_quick_connect}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_quick_connect(InstanceId, QuickConnectId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QuickConnectId}{[required] The identifier for the quick connect.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  QuickConnect = list(
    QuickConnectARN = "string",
    QuickConnectId = "string",
    Name = "string",
    Description = "string",
    QuickConnectConfig = list(
      QuickConnectType = "USER"|"QUEUE"|"PHONE_NUMBER",
      UserConfig = list(
        UserId = "string",
        ContactFlowId = "string"
      ),
      QueueConfig = list(
        QueueId = "string",
        ContactFlowId = "string"
      ),
      PhoneConfig = list(
        PhoneNumber = "string"
      )
    ),
    Tags = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Describes the quick connect.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_quick_connect(
  InstanceId = "string",
  QuickConnectId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
