% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_routing_profile}
\alias{connect_describe_routing_profile}
\title{Describes the specified routing profile}
\usage{
connect_describe_routing_profile(InstanceId, RoutingProfileId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RoutingProfile = list(
    InstanceId = "string",
    Name = "string",
    RoutingProfileArn = "string",
    RoutingProfileId = "string",
    Description = "string",
    MediaConcurrencies = list(
      list(
        Channel = "VOICE"|"CHAT"|"TASK",
        Concurrency = 123
      )
    ),
    DefaultOutboundQueueId = "string",
    Tags = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the specified routing profile.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_routing_profile(
  InstanceId = "string",
  RoutingProfileId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
