% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_security_profile}
\alias{connect_describe_security_profile}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_security_profile(SecurityProfileId, InstanceId)
}
\arguments{
\item{SecurityProfileId}{[required] The identifier for the security profle.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_security_profile/} for full documentation.
}
\keyword{internal}
