% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_test_case}
\alias{connect_describe_test_case}
\title{Describes the specified test case and allows you to get the content and
metadata of the test case for the specified Amazon Connect instance}
\usage{
connect_describe_test_case(InstanceId, TestCaseId, Status = NULL)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{TestCaseId}{[required] The identifier of the test case.}

\item{Status}{The status of the test case version to retrieve. If not specified,
returns the published version if available, otherwise returns the saved
version.}
}
\description{
Describes the specified test case and allows you to get the content and metadata of the test case for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_test_case/} for full documentation.
}
\keyword{internal}
