% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_traffic_distribution_group}
\alias{connect_describe_traffic_distribution_group}
\title{Gets details and status of a traffic distribution group}
\usage{
connect_describe_traffic_distribution_group(TrafficDistributionGroupId)
}
\arguments{
\item{TrafficDistributionGroupId}{[required] The identifier of the traffic distribution group. This can be the ID or
the ARN if the API is being called in the Region where the traffic
distribution group was created. The ARN must be provided if the call is
from the replicated Region.}
}
\description{
Gets details and status of a traffic distribution group.

See \url{https://paws-r.github.io/docs/connect/describe_traffic_distribution_group.html} for full documentation.
}
\keyword{internal}
