% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_user_hierarchy_group}
\alias{connect_describe_user_hierarchy_group}
\title{Describes the specified hierarchy group}
\usage{
connect_describe_user_hierarchy_group(HierarchyGroupId, InstanceId)
}
\arguments{
\item{HierarchyGroupId}{[required] The identifier of the hierarchy group.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  HierarchyGroup = list(
    Id = "string",
    Arn = "string",
    Name = "string",
    LevelId = "string",
    HierarchyPath = list(
      LevelOne = list(
        Id = "string",
        Arn = "string",
        Name = "string"
      ),
      LevelTwo = list(
        Id = "string",
        Arn = "string",
        Name = "string"
      ),
      LevelThree = list(
        Id = "string",
        Arn = "string",
        Name = "string"
      ),
      LevelFour = list(
        Id = "string",
        Arn = "string",
        Name = "string"
      ),
      LevelFive = list(
        Id = "string",
        Arn = "string",
        Name = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the specified hierarchy group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_user_hierarchy_group(
  HierarchyGroupId = "string",
  InstanceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
