% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_vocabulary}
\alias{connect_describe_vocabulary}
\title{Describes the specified vocabulary}
\usage{
connect_describe_vocabulary(InstanceId, VocabularyId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{VocabularyId}{[required] The identifier of the custom vocabulary.}
}
\description{
Describes the specified vocabulary.

See \url{https://paws-r.github.io/docs/connect/describe_vocabulary.html} for full documentation.
}
\keyword{internal}
