% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_workspace}
\alias{connect_describe_workspace}
\title{Retrieves details about a workspace, including its configuration and
metadata}
\usage{
connect_describe_workspace(InstanceId, WorkspaceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}
}
\description{
Retrieves details about a workspace, including its configuration and metadata.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_workspace/} for full documentation.
}
\keyword{internal}
