% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_approved_origin}
\alias{connect_disassociate_approved_origin}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_disassociate_approved_origin(InstanceId, Origin)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{Origin}{[required] The domain URL of the integrated application.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Revokes access to integrated applications from Amazon Connect.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disassociate_approved_origin(
  InstanceId = "string",
  Origin = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
