% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_bot}
\alias{connect_disassociate_bot}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_disassociate_bot(InstanceId, LexBot = NULL, LexV2Bot = NULL)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{LexBot}{}

\item{LexV2Bot}{The Amazon Lex V2 bot to disassociate from the instance.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/disassociate_bot.html} for full documentation.
}
\keyword{internal}
