% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_email_address_alias}
\alias{connect_disassociate_email_address_alias}
\title{Removes the alias association between two email addresses in an Amazon
Connect instance}
\usage{
connect_disassociate_email_address_alias(
  EmailAddressId,
  InstanceId,
  AliasConfiguration,
  ClientToken = NULL
)
}
\arguments{
\item{EmailAddressId}{[required] The identifier of the email address.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{AliasConfiguration}{[required] Configuration object that specifies which alias relationship to remove.
The alias association must currently exist between the primary email
address and the specified alias email address.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}
}
\description{
Removes the alias association between two email addresses in an Amazon Connect instance. After disassociation, emails sent to the former alias email address are no longer forwarded to the primary email address. Both email addresses continue to exist independently and can receive emails directly.

See \url{https://www.paws-r-sdk.com/docs/connect_disassociate_email_address_alias/} for full documentation.
}
\keyword{internal}
