% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_instance_storage_config}
\alias{connect_disassociate_instance_storage_config}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_disassociate_instance_storage_config(InstanceId, AssociationId,
  ResourceType)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{AssociationId}{[required] The existing association identifier that uniquely identifies the
resource type and storage config for the given instance ID.}

\item{ResourceType}{[required] A valid resource type.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Removes the storage type configurations for the specified resource type
and association ID.
}
\section{Request syntax}{
\preformatted{svc$disassociate_instance_storage_config(
  InstanceId = "string",
  AssociationId = "string",
  ResourceType = "CHAT_TRANSCRIPTS"|"CALL_RECORDINGS"|"SCHEDULED_REPORTS"|"MEDIA_STREAMS"|"CONTACT_TRACE_RECORDS"|"AGENT_EVENTS"
)
}
}

\keyword{internal}
