% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_lambda_function}
\alias{connect_disassociate_lambda_function}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_disassociate_lambda_function(InstanceId, FunctionArn)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance..}

\item{FunctionArn}{[required] The Amazon Resource Name (ARN) of the Lambda function being
disassociated.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/disassociate_lambda_function.html} for full documentation.
}
\keyword{internal}
