% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_queue_email_addresses}
\alias{connect_disassociate_queue_email_addresses}
\title{Removes the association between a set of email addresses and a queue}
\usage{
connect_disassociate_queue_email_addresses(
  InstanceId,
  QueueId,
  EmailAddressesId,
  ClientToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{EmailAddressesId}{[required] List of email address identifiers to disassociate from the queue. These
are the unique identifiers of email addresses that should no longer be
routed to this queue.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}
}
\description{
Removes the association between a set of email addresses and a queue. After disassociation, agents will no longer be able to select these email addresses as "From" addresses when replying to inbound email contacts or initiating outbound email contacts in this queue.

See \url{https://www.paws-r-sdk.com/docs/connect_disassociate_queue_email_addresses/} for full documentation.
}
\keyword{internal}
