% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_workspace}
\alias{connect_disassociate_workspace}
\title{Removes the association between a workspace and one or more users or
routing profiles}
\usage{
connect_disassociate_workspace(InstanceId, WorkspaceId, ResourceArns)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}

\item{ResourceArns}{[required] The Amazon Resource Names (ARNs) of the resources to disassociate from
the workspace.}
}
\description{
Removes the association between a workspace and one or more users or routing profiles.

See \url{https://www.paws-r-sdk.com/docs/connect_disassociate_workspace/} for full documentation.
}
\keyword{internal}
