% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_contact_attributes}
\alias{connect_get_contact_attributes}
\title{Retrieves the contact attributes for the specified contact}
\usage{
connect_get_contact_attributes(InstanceId, InitialContactId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{InitialContactId}{[required] The identifier of the initial contact.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Attributes = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the contact attributes for the specified contact.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_contact_attributes(
  InstanceId = "string",
  InitialContactId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
