% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_cross_region_routing}
\alias{connect_get_cross_region_routing}
\title{Retrieves the current cross-region routing configuration for an Amazon
Connect Global Resiliency instance enabled for global routing}
\usage{
connect_get_cross_region_routing(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}
}
\description{
Retrieves the current cross-region routing configuration for an Amazon Connect Global Resiliency instance enabled for global routing. This operation returns whether cross-region routing is currently enabled or disabled (isolated) for the instance.

See \url{https://www.paws-r-sdk.com/docs/connect_get_cross_region_routing/} for full documentation.
}
\keyword{internal}
