% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_evaluation_form_validation}
\alias{connect_get_evaluation_form_validation}
\title{Retrieves the status and results of a validation process started by
StartEvaluationFormValidation}
\usage{
connect_get_evaluation_form_validation(
  InstanceId,
  EvaluationFormId,
  EvaluationFormVersion = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{EvaluationFormId}{[required] The unique identifier for the evaluation form.}

\item{EvaluationFormVersion}{The version of the evaluation form to retrieve validation results for.}
}
\description{
Retrieves the status and results of a validation process started by \code{\link[=connect_start_evaluation_form_validation]{start_evaluation_form_validation}}. Returns the current execution status (\code{IN_PROGRESS}, \code{COMPLETED}, or \code{FAILED}), the validated form version, and when completed, a list of findings that identify structural issues and quality improvements for the evaluation form, and may include suggested fixes. If the validation failed, a reason is provided indicating the cause of the failure.

See \url{https://www.paws-r-sdk.com/docs/connect_get_evaluation_form_validation/} for full documentation.
}
\keyword{internal}
