% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_test_case_execution_summary}
\alias{connect_get_test_case_execution_summary}
\title{Retrieves an overview of a test execution that includes the status of
the execution, start and end time, and observation summary}
\usage{
connect_get_test_case_execution_summary(
  InstanceId,
  TestCaseId,
  TestCaseExecutionId
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{TestCaseId}{[required] The identifier of the test case.}

\item{TestCaseExecutionId}{[required] The identifier of the test case execution.}
}
\description{
Retrieves an overview of a test execution that includes the status of the execution, start and end time, and observation summary.

See \url{https://www.paws-r-sdk.com/docs/connect_get_test_case_execution_summary/} for full documentation.
}
\keyword{internal}
