% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_traffic_distribution}
\alias{connect_get_traffic_distribution}
\title{Retrieves the current traffic distribution for a given traffic
distribution group}
\usage{
connect_get_traffic_distribution(Id)
}
\arguments{
\item{Id}{[required] The identifier of the traffic distribution group.}
}
\description{
Retrieves the current traffic distribution for a given traffic distribution group.

See \url{https://www.paws-r-sdk.com/docs/connect_get_traffic_distribution/} for full documentation.
}
\keyword{internal}
