% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_approved_origins}
\alias{connect_list_approved_origins}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_list_approved_origins(InstanceId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/list_approved_origins.html} for full documentation.
}
\keyword{internal}
