% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_associated_contacts}
\alias{connect_list_associated_contacts}
\title{Provides information about contact tree, a list of associated contacts
with a unique identifier}
\usage{
connect_list_associated_contacts(
  InstanceId,
  ContactId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact in this instance of Amazon Connect.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Provides information about contact tree, a list of associated contacts with a unique identifier.

See \url{https://www.paws-r-sdk.com/docs/connect_list_associated_contacts/} for full documentation.
}
\keyword{internal}
