% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_authentication_profiles}
\alias{connect_list_authentication_profiles}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_list_authentication_profiles(
  InstanceId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change. To request access to this API, contact Amazon Web Services Support.

See \url{https://www.paws-r-sdk.com/docs/connect_list_authentication_profiles/} for full documentation.
}
\keyword{internal}
