% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_contact_flow_module_aliases}
\alias{connect_list_contact_flow_module_aliases}
\title{Lists all aliases associated with a contact flow module, showing their
current version mappings and metadata}
\usage{
connect_list_contact_flow_module_aliases(
  InstanceId,
  ContactFlowModuleId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowModuleId}{[required] The identifier of the flow module.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Lists all aliases associated with a contact flow module, showing their current version mappings and metadata.

See \url{https://www.paws-r-sdk.com/docs/connect_list_contact_flow_module_aliases/} for full documentation.
}
\keyword{internal}
