% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_contact_flow_module_versions}
\alias{connect_list_contact_flow_module_versions}
\title{Retrieves a paginated list of all versions for a specific contact flow
module}
\usage{
connect_list_contact_flow_module_versions(
  InstanceId,
  ContactFlowModuleId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowModuleId}{[required] The identifier of the flow module.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Retrieves a paginated list of all versions for a specific contact flow module.

See \url{https://www.paws-r-sdk.com/docs/connect_list_contact_flow_module_versions/} for full documentation.
}
\keyword{internal}
