% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_contact_flows}
\alias{connect_list_contact_flows}
\title{Provides information about the contact flows for the specified Amazon
Connect instance}
\usage{
connect_list_contact_flows(
  InstanceId,
  ContactFlowTypes = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ContactFlowTypes}{The type of contact flow.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Provides information about the contact flows for the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/list_contact_flows.html} for full documentation.
}
\keyword{internal}
