% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_data_tables}
\alias{connect_list_data_tables}
\title{Lists all data tables for the specified Amazon Connect instance}
\usage{
connect_list_data_tables(InstanceId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{InstanceId}{[required] The unique identifier for the Amazon Connect instance whose data tables
should be listed.}

\item{NextToken}{Specify the pagination token from a previous request to retrieve the
next page of results.}

\item{MaxResults}{The maximum number of data tables to return in one page of results.}
}
\description{
Lists all data tables for the specified Amazon Connect instance. Returns summary information for each table including basic metadata and modification details.

See \url{https://www.paws-r-sdk.com/docs/connect_list_data_tables/} for full documentation.
}
\keyword{internal}
