% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_entity_security_profiles}
\alias{connect_list_entity_security_profiles}
\title{Lists all security profiles attached to a Q in Connect AIAgent Entity in
an Amazon Connect instance}
\usage{
connect_list_entity_security_profiles(
  InstanceId,
  EntityType,
  EntityArn,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the instance
ID in the Amazon Resource Name (ARN) of the instance.}

\item{EntityType}{[required] Only supported type is AI_AGENT.}

\item{EntityArn}{[required] ARN of a Q in Connect AI Agent.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page. The default MaxResult
size is 100.}
}
\description{
Lists all security profiles attached to a Q in Connect AIAgent Entity in an Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_list_entity_security_profiles/} for full documentation.
}
\keyword{internal}
