% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_evaluation_form_ai_versions}
\alias{connect_list_evaluation_form_ai_versions}
\title{Lists the available AI versions for evaluation forms in the specified
Connect Customer instance}
\usage{
connect_list_evaluation_form_ai_versions(
  InstanceId,
  ContactInteractionType,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{ContactInteractionType}{[required] The contact interaction type for the evaluation form.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{NextToken}{The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.}
}
\description{
Lists the available AI versions for evaluation forms in the specified Connect Customer instance.

See \url{https://www.paws-r-sdk.com/docs/connect_list_evaluation_form_ai_versions/} for full documentation.
}
\keyword{internal}
