% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_evaluation_form_versions}
\alias{connect_list_evaluation_form_versions}
\title{Lists versions of an evaluation form in the specified Amazon Connect
instance}
\usage{
connect_list_evaluation_form_versions(
  InstanceId,
  EvaluationFormId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{EvaluationFormId}{[required] The unique identifier for the evaluation form.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Lists versions of an evaluation form in the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/list_evaluation_form_versions.html} for full documentation.
}
\keyword{internal}
