% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_extraction_definitions}
\alias{connect_list_extraction_definitions}
\title{Lists extraction definitions in the specified Connect Customer instance}
\usage{
connect_list_extraction_definitions(
  InstanceId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{MaxResults}{The maximum number of results to return per page. The default MaxResult size is 100.}

\item{NextToken}{The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.}
}
\description{
Lists extraction definitions in the specified Connect Customer instance.

See \url{https://www.paws-r-sdk.com/docs/connect_list_extraction_definitions/} for full documentation.
}
\keyword{internal}
