% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_hours_of_operation_overrides}
\alias{connect_list_hours_of_operation_overrides}
\title{List the hours of operation overrides}
\usage{
connect_list_hours_of_operation_overrides(
  InstanceId,
  HoursOfOperationId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{HoursOfOperationId}{[required] The identifier for the hours of operation.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
List the hours of operation overrides.

See \url{https://www.paws-r-sdk.com/docs/connect_list_hours_of_operation_overrides/} for full documentation.
}
\keyword{internal}
