% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_hours_of_operations}
\alias{connect_list_hours_of_operations}
\title{Provides information about the hours of operation for the specified
Amazon Connect instance}
\usage{
connect_list_hours_of_operations(InstanceId, NextToken, MaxResults)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximimum number of results to return per page.}
}
\description{
Provides information about the hours of operation for the specified
Amazon Connect instance.

For more information about hours of operation, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/set-hours-operation.html}{Set the Hours of Operation for a Queue}
in the \emph{Amazon Connect Administrator Guide}.
}
\section{Request syntax}{
\preformatted{svc$list_hours_of_operations(
  InstanceId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
