% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_instances}
\alias{connect_list_instances}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_list_instances(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximimum number of results to return per page.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Return a list of instances which are in active state,
creation-in-progress state, and failed state. Instances that aren't
successfully created (they are in a failed state) are returned only for
24 hours after the CreateInstance API was invoked.
}
\section{Request syntax}{
\preformatted{svc$list_instances(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
