% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_integration_associations}
\alias{connect_list_integration_associations}
\title{Provides summary information about the AppIntegration associations for
the specified Amazon Connect instance}
\usage{
connect_list_integration_associations(InstanceId, NextToken, MaxResults)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\value{
A list with the following syntax:\preformatted{list(
  IntegrationAssociationSummaryList = list(
    list(
      IntegrationAssociationId = "string",
      IntegrationAssociationArn = "string",
      InstanceId = "string",
      IntegrationType = "EVENT",
      IntegrationArn = "string",
      SourceApplicationUrl = "string",
      SourceApplicationName = "string",
      SourceType = "SALESFORCE"|"ZENDESK"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Provides summary information about the AppIntegration associations for
the specified Amazon Connect instance.
}
\section{Request syntax}{
\preformatted{svc$list_integration_associations(
  InstanceId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
