% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_metrics}
\alias{connect_list_metrics}
\title{Retrieves a paginated list of metric summaries for the specified Connect
Customer instance}
\usage{
connect_list_metrics(
  InstanceId,
  Type = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{Type}{The type of metrics to list. Valid values: \code{AWS_MANAGED} | \code{CUSTOMER_MANAGED}.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{NextToken}{The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.}
}
\description{
Retrieves a paginated list of metric summaries for the specified Connect Customer instance. Use pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/connect_list_metrics/} for full documentation.
}
\keyword{internal}
