% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_notifications}
\alias{connect_list_notifications}
\title{Retrieves a paginated list of all notifications in the Amazon Connect
instance}
\usage{
connect_list_notifications(InstanceId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response to retrieve the next page of results.}

\item{MaxResults}{The maximum number of results to return per page. Valid range is 1-100.}
}
\description{
Retrieves a paginated list of all notifications in the Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_list_notifications/} for full documentation.
}
\keyword{internal}
