% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_phone_numbers}
\alias{connect_list_phone_numbers}
\title{Provides information about the phone numbers for the specified Amazon
Connect instance}
\usage{
connect_list_phone_numbers(
  InstanceId,
  PhoneNumberTypes = NULL,
  PhoneNumberCountryCodes = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{PhoneNumberTypes}{The type of phone number.}

\item{PhoneNumberCountryCodes}{The ISO country code.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Provides information about the phone numbers for the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/list_phone_numbers.html} for full documentation.
}
\keyword{internal}
