% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_phone_numbers_v2}
\alias{connect_list_phone_numbers_v2}
\title{Lists phone numbers claimed to your Amazon Connect instance}
\usage{
connect_list_phone_numbers_v2(
  TargetArn = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  PhoneNumberCountryCodes = NULL,
  PhoneNumberTypes = NULL,
  PhoneNumberPrefix = NULL
)
}
\arguments{
\item{TargetArn}{The Amazon Resource Name (ARN) for Amazon Connect instances that phone
numbers are claimed to. If \code{TargetArn} input is not provided, this API
lists numbers claimed to all the Amazon Connect instances belonging to
your account.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{PhoneNumberCountryCodes}{The ISO country code.}

\item{PhoneNumberTypes}{The type of phone number.}

\item{PhoneNumberPrefix}{The prefix of the phone number. If provided, it must contain \code{+} as part
of the country code.}
}
\description{
Lists phone numbers claimed to your Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/list_phone_numbers_v2.html} for full documentation.
}
\keyword{internal}
