% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_queue_email_addresses}
\alias{connect_list_queue_email_addresses}
\title{Lists all email addresses that are currently associated with a specific
queue, providing details about which "From" email addresses agents can
select when handling email contacts}
\usage{
connect_list_queue_email_addresses(
  InstanceId,
  QueueId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Lists all email addresses that are currently associated with a specific queue, providing details about which "From" email addresses agents can select when handling email contacts. This helps administrators manage agent email address options and understand the available choices for different brands and business units.

See \url{https://www.paws-r-sdk.com/docs/connect_list_queue_email_addresses/} for full documentation.
}
\keyword{internal}
