% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_queue_quick_connects}
\alias{connect_list_queue_quick_connects}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_list_queue_quick_connects(InstanceId, QueueId, NextToken,
  MaxResults)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  QuickConnectSummaryList = list(
    list(
      Id = "string",
      Arn = "string",
      Name = "string",
      QuickConnectType = "USER"|"QUEUE"|"PHONE_NUMBER"
    )
  )
)
}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Lists the quick connects associated with a queue.
}
\section{Request syntax}{
\preformatted{svc$list_queue_quick_connects(
  InstanceId = "string",
  QueueId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
