% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_routing_profile_manual_assignment_queues}
\alias{connect_list_routing_profile_manual_assignment_queues}
\title{Lists the manual assignment queues associated with a routing profile}
\usage{
connect_list_routing_profile_manual_assignment_queues(
  InstanceId,
  RoutingProfileId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Lists the manual assignment queues associated with a routing profile.

See \url{https://www.paws-r-sdk.com/docs/connect_list_routing_profile_manual_assignment_queues/} for full documentation.
}
\keyword{internal}
