% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_rules}
\alias{connect_list_rules}
\title{List all rules for the specified Amazon Connect instance}
\usage{
connect_list_rules(
  InstanceId,
  PublishStatus = NULL,
  EventSourceName = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{PublishStatus}{The publish status of the rule.}

\item{EventSourceName}{The name of the event source.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
List all rules for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_list_rules/} for full documentation.
}
\keyword{internal}
