% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_security_profile_applications}
\alias{connect_list_security_profile_applications}
\title{Returns a list of third-party applications or MCP Servers in a specific
security profile}
\usage{
connect_list_security_profile_applications(
  SecurityProfileId,
  InstanceId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{SecurityProfileId}{[required] The identifier for the security profle.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Returns a list of third-party applications or MCP Servers in a specific security profile.

See \url{https://www.paws-r-sdk.com/docs/connect_list_security_profile_applications/} for full documentation.
}
\keyword{internal}
