% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_security_profile_flow_modules}
\alias{connect_list_security_profile_flow_modules}
\title{A list of Flow Modules an AI Agent can invoke as a tool}
\usage{
connect_list_security_profile_flow_modules(
  SecurityProfileId,
  InstanceId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{SecurityProfileId}{[required] The identifier for the security profile.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page. The default MaxResult
size is 100.}
}
\description{
A list of Flow Modules an AI Agent can invoke as a tool

See \url{https://www.paws-r-sdk.com/docs/connect_list_security_profile_flow_modules/} for full documentation.
}
\keyword{internal}
