% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_security_profiles}
\alias{connect_list_security_profiles}
\title{Provides summary information about the security profiles for the
specified Amazon Connect instance}
\usage{
connect_list_security_profiles(InstanceId, NextToken, MaxResults)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximimum number of results to return per page.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SecurityProfileSummaryList = list(
    list(
      Id = "string",
      Arn = "string",
      Name = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Provides summary information about the security profiles for the
specified Amazon Connect instance.

For more information about security profiles, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/connect-security-profiles.html}{Security Profiles}
in the \emph{Amazon Connect Administrator Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_security_profiles(
  InstanceId = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
