% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_tags_for_resource}
\alias{connect_list_tags_for_resource}
\title{Lists the tags for the specified resource}
\usage{
connect_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Lists the tags for the specified resource.

See \url{https://paws-r.github.io/docs/connect/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
