% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_test_case_execution_records}
\alias{connect_list_test_case_execution_records}
\title{Lists detailed steps of test case execution that includes all
observations along with actions taken and data associated in the
specified Amazon Connect instance}
\usage{
connect_list_test_case_execution_records(
  InstanceId,
  TestCaseId,
  TestCaseExecutionId,
  Status = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{TestCaseId}{[required] The identifier of the test case.}

\item{TestCaseExecutionId}{[required] The identifier of the test case execution.}

\item{Status}{Filter execution records by status.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Lists detailed steps of test case execution that includes all observations along with actions taken and data associated in the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_list_test_case_execution_records/} for full documentation.
}
\keyword{internal}
