% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_test_case_executions}
\alias{connect_list_test_case_executions}
\title{Lists all test case executions and allows filtering by test case id,
test case name, start time, end time or status of the execution for the
specified Amazon Connect instance}
\usage{
connect_list_test_case_executions(
  InstanceId,
  TestCaseId = NULL,
  TestCaseName = NULL,
  StartTime = NULL,
  EndTime = NULL,
  Status = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{TestCaseId}{Filter executions by test case identifier.}

\item{TestCaseName}{Filter executions by test case name.}

\item{StartTime}{Filter executions that started after this time.}

\item{EndTime}{Filter executions that started before this time.}

\item{Status}{Filter executions by status.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Lists all test case executions and allows filtering by test case id, test case name, start time, end time or status of the execution for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_list_test_case_executions/} for full documentation.
}
\keyword{internal}
