% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_use_cases}
\alias{connect_list_use_cases}
\title{Lists the use cases for the integration association}
\usage{
connect_list_use_cases(
  InstanceId,
  IntegrationAssociationId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{IntegrationAssociationId}{[required] The identifier for the integration association.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Lists the use cases for the integration association.

See \url{https://www.paws-r-sdk.com/docs/connect_list_use_cases/} for full documentation.
}
\keyword{internal}
