% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_views}
\alias{connect_list_views}
\title{Returns views in the given instance}
\usage{
connect_list_views(
  InstanceId,
  Type = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{Type}{The type of the view.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page. The default MaxResult
size is 100.}
}
\description{
Returns views in the given instance.

See \url{https://www.paws-r-sdk.com/docs/connect_list_views/} for full documentation.
}
\keyword{internal}
