% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_workspace_media}
\alias{connect_list_workspace_media}
\title{Lists media assets (such as logos) associated with a workspace}
\usage{
connect_list_workspace_media(InstanceId, WorkspaceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}
}
\description{
Lists media assets (such as logos) associated with a workspace.

See \url{https://www.paws-r-sdk.com/docs/connect_list_workspace_media/} for full documentation.
}
\keyword{internal}
