% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_workspace_pages}
\alias{connect_list_workspace_pages}
\title{Lists the page configurations in a workspace, including the views
assigned to each page}
\usage{
connect_list_workspace_pages(
  InstanceId,
  WorkspaceId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Lists the page configurations in a workspace, including the views assigned to each page.

See \url{https://www.paws-r-sdk.com/docs/connect_list_workspace_pages/} for full documentation.
}
\keyword{internal}
