% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_put_user_status}
\alias{connect_put_user_status}
\title{Changes the current status of a user or agent in Connect Customer}
\usage{
connect_put_user_status(UserId, InstanceId, AgentStatusId)
}
\arguments{
\item{UserId}{[required] The identifier of the user.}

\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{AgentStatusId}{[required] The identifier of the agent status.}
}
\description{
Changes the current status of a user or agent in Connect Customer. If the agent is currently handling a contact, this sets the agent's next status.

See \url{https://www.paws-r-sdk.com/docs/connect_put_user_status/} for full documentation.
}
\keyword{internal}
