% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_release_phone_number}
\alias{connect_release_phone_number}
\title{Releases a phone number previously claimed to an Amazon Connect instance}
\usage{
connect_release_phone_number(PhoneNumberId, ClientToken = NULL)
}
\arguments{
\item{PhoneNumberId}{[required] A unique identifier for the phone number.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Releases a phone number previously claimed to an Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/release_phone_number.html} for full documentation.
}
\keyword{internal}
