% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_resume_contact}
\alias{connect_resume_contact}
\title{Allows resuming a task contact in a paused state}
\usage{
connect_resume_contact(ContactId, InstanceId, ContactFlowId = NULL)
}
\arguments{
\item{ContactId}{[required] The identifier of the contact.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
\code{instanceId} in the ARN of the instance.}

\item{ContactFlowId}{The identifier of the flow.}
}
\description{
Allows resuming a task contact in a paused state.

See \url{https://www.paws-r-sdk.com/docs/connect_resume_contact/} for full documentation.
}
\keyword{internal}
