% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_search_contacts}
\alias{connect_search_contacts}
\title{Searches contacts in an Amazon Connect instance}
\usage{
connect_search_contacts(
  InstanceId,
  TimeRange,
  SearchCriteria = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  Sort = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of Amazon Connect instance. You can find the instance ID
in the Amazon Resource Name (ARN) of the instance.}

\item{TimeRange}{[required] Time range that you want to search results.}

\item{SearchCriteria}{The search criteria to be used to return contacts.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{Sort}{Specifies a field to sort by and a sort order.}
}
\description{
Searches contacts in an Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_search_contacts/} for full documentation.
}
\keyword{internal}
