% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_search_email_addresses}
\alias{connect_search_email_addresses}
\title{Searches email address in an instance, with optional filtering}
\usage{
connect_search_email_addresses(
  InstanceId,
  MaxResults = NULL,
  NextToken = NULL,
  SearchCriteria = NULL,
  SearchFilter = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{SearchCriteria}{The search criteria to be used to return email addresses.}

\item{SearchFilter}{Filters to be applied to search results.}
}
\description{
Searches email address in an instance, with optional filtering.

See \url{https://www.paws-r-sdk.com/docs/connect_search_email_addresses/} for full documentation.
}
\keyword{internal}
