% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_search_hours_of_operation_overrides}
\alias{connect_search_hours_of_operation_overrides}
\title{Searches the hours of operation overrides}
\usage{
connect_search_hours_of_operation_overrides(
  InstanceId,
  NextToken = NULL,
  MaxResults = NULL,
  SearchFilter = NULL,
  SearchCriteria = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{SearchFilter}{}

\item{SearchCriteria}{The search criteria to be used to return hours of operations overrides.}
}
\description{
Searches the hours of operation overrides.

See \url{https://www.paws-r-sdk.com/docs/connect_search_hours_of_operation_overrides/} for full documentation.
}
\keyword{internal}
