% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_search_metrics}
\alias{connect_search_metrics}
\title{Searches for metrics in the specified Connect Customer instance using
search criteria and optional tag-based filters}
\usage{
connect_search_metrics(
  InstanceId,
  NextToken = NULL,
  MaxResults = NULL,
  SearchFilter = NULL,
  SearchCriteria = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{SearchFilter}{Filters to be applied to search results.}

\item{SearchCriteria}{The search criteria to filter the metrics.}
}
\description{
Searches for metrics in the specified Connect Customer instance using search criteria and optional tag-based filters. Use pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/connect_search_metrics/} for full documentation.
}
\keyword{internal}
