% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_search_resource_tags}
\alias{connect_search_resource_tags}
\title{Searches tags used in an Connect Customer instance using optional search
criteria}
\usage{
connect_search_resource_tags(
  InstanceId,
  ResourceTypes = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  SearchCriteria = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can find the instanceId in the Amazon Resource Name (ARN) of the instance.}

\item{ResourceTypes}{The list of resource types to be used to search tags from. If not provided or if any empty list is provided, this API will search from all supported resource types. Note that lowercase and - are required.

\strong{Supported resource types}
\itemize{
\item agent
\item agent-state
\item routing-profile
\item standard-queue
\item security-profile
\item operating-hours
\item prompt
\item contact-flow
\item flow- module
\item transfer-destination (also known as quick connect)
\item metric
}}

\item{NextToken}{The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{SearchCriteria}{The search criteria to be used to return tags.}
}
\description{
Searches tags used in an Connect Customer instance using optional search criteria.

See \url{https://www.paws-r-sdk.com/docs/connect_search_resource_tags/} for full documentation.
}
\keyword{internal}
