% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_search_test_cases}
\alias{connect_search_test_cases}
\title{Searches for test cases in the specified Amazon Connect instance, with
optional filtering}
\usage{
connect_search_test_cases(
  InstanceId,
  NextToken = NULL,
  MaxResults = NULL,
  SearchFilter = NULL,
  SearchCriteria = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the instance
ID in the Amazon Resource Name (ARN) of the instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{SearchFilter}{Filters to be applied to search results.}

\item{SearchCriteria}{The search criteria to be used to return test cases.}
}
\description{
Searches for test cases in the specified Amazon Connect instance, with optional filtering.

See \url{https://www.paws-r-sdk.com/docs/connect_search_test_cases/} for full documentation.
}
\keyword{internal}
