% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_search_views}
\alias{connect_search_views}
\title{Searches views based on name, description, or tags}
\usage{
connect_search_views(
  InstanceId,
  NextToken = NULL,
  MaxResults = NULL,
  SearchFilter = NULL,
  SearchCriteria = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{SearchFilter}{Filters to apply to the search, such as tag-based filters.}

\item{SearchCriteria}{The search criteria, including field names and comparison types.}
}
\description{
Searches views based on name, description, or tags.

See \url{https://www.paws-r-sdk.com/docs/connect_search_views/} for full documentation.
}
\keyword{internal}
