% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_send_outbound_web_notification}
\alias{connect_send_outbound_web_notification}
\title{Sends an outbound web notification to a customer's web browser for
outbound campaigns}
\usage{
connect_send_outbound_web_notification(
  InstanceId,
  ClientToken = NULL,
  BrowserId,
  SessionId,
  ExpiresAt,
  Source,
  Destination,
  Content
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see \href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{BrowserId}{[required] A unique identifier for the customer's web browser instance to which the notification is being sent.}

\item{SessionId}{[required] A unique identifier for the customer's web session to which the notification is being sent.}

\item{ExpiresAt}{[required] The timestamp, in Unix epoch time format, at which the web notification expires. After this time, the notification is no longer delivered to the customer's browser.}

\item{Source}{[required] The source of the web notification. A \code{SourceCampaign} object identifies the campaign and outbound request that triggered this notification.}

\item{Destination}{[required] The destination for the web notification, specifying the communication widget that delivers the notification and the customer profile of the recipient.}

\item{Content}{[required] The content of the web notification, including the notification type, the view to render, and any optional attributes used to populate it.}
}
\description{
Sends an outbound web notification to a customer's web browser for outbound campaigns. For more information about outbound campaigns, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/enable-outbound-campaigns.html}{Set up Connect Customer outbound campaigns}.

See \url{https://www.paws-r-sdk.com/docs/connect_send_outbound_web_notification/} for full documentation.
}
\keyword{internal}
