% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_start_assistant_contact}
\alias{connect_start_assistant_contact}
\title{Starts a chat contact with an AI agent}
\usage{
connect_start_assistant_contact(
  InstanceId,
  AiAgent,
  ParticipantDetails,
  InitialMessage = NULL,
  Attributes = NULL,
  ClientToken = NULL,
  PersistentChat = NULL,
  RelatedContactId = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{AiAgent}{[required] The AI agent configuration for this contact.}

\item{ParticipantDetails}{[required] The display name and other details that identify the chat participant.}

\item{InitialMessage}{The initial message to send to the newly created chat.}

\item{Attributes}{A map of key-value pairs to associate with the contact. We make these attributes available to flows as standard contact attributes.

You can provide up to 32,768 UTF-8 bytes across all key-value pairs for each contact.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see \href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{PersistentChat}{The configuration that enables persistent chat. For more information about persistent chat and its use cases, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/chat-persistence.html}{Enable persistent chat}.}

\item{RelatedContactId}{The identifier of an Connect Customer contact related to the new assistant contact.

You cannot provide both \code{RelatedContactId} and \code{PersistentChat}.}
}
\description{
Starts a chat contact with an AI agent.

See \url{https://www.paws-r-sdk.com/docs/connect_start_assistant_contact/} for full documentation.
}
\keyword{internal}
