% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_start_chat_contact}
\alias{connect_start_chat_contact}
\title{Initiates a flow to start a new chat for the customer}
\usage{
connect_start_chat_contact(
  InstanceId,
  ContactFlowId,
  Attributes = NULL,
  ParticipantDetails,
  ParticipantConfiguration = NULL,
  InitialMessage = NULL,
  ClientToken = NULL,
  ChatDurationInMinutes = NULL,
  SupportedMessagingContentTypes = NULL,
  PersistentChat = NULL,
  RelatedContactId = NULL,
  SegmentAttributes = NULL,
  CustomerId = NULL,
  DisconnectOnCustomerExit = NULL,
  ConnectionTypes = NULL,
  ChatStreamingConfiguration = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowId}{[required] The identifier of the flow for initiating the chat. To see the ContactFlowId in the Connect Customer admin website, on the navigation menu go to \strong{Routing}, \strong{Flows}. Choose the flow. On the flow page, under the name of the flow, choose \strong{Show additional flow information}. The ContactFlowId is the last part of the ARN, shown here in bold:

arn:aws:connect:us-west-2:xxxxxxxxxxxx:instance/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/contact-flow/\strong{846ec553-a005-41c0-8341-xxxxxxxxxxxx}}

\item{Attributes}{A custom key-value pair using an attribute map. The attributes are standard Connect Customer attributes. They can be accessed in flows just like any other contact attributes.

There can be up to 32,768 UTF-8 bytes across all key-value pairs per contact. Attribute keys can include only alphanumeric, dash, and underscore characters.}

\item{ParticipantDetails}{[required] Information identifying the participant.}

\item{ParticipantConfiguration}{The configuration of the participant.}

\item{InitialMessage}{The initial message to be sent to the newly created chat.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see \href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{ChatDurationInMinutes}{The total duration of the newly started chat session. If not specified, the chat session duration defaults to 25 hour. The minimum configurable time is 60 minutes. The maximum configurable time is 10,080 minutes (7 days).}

\item{SupportedMessagingContentTypes}{The supported chat message content types. Supported types are \code{text/plain}, \code{text/markdown}, \code{application/json}, \code{application/vnd.amazonaws.connect.message.interactive}, and \code{application/vnd.amazonaws.connect.message.interactive.response}.

Content types must always contain \code{text/plain}. You can then put any other supported type in the list. For example, all the following lists are valid because they contain \code{text/plain}: \verb{[text/plain, text/markdown, application/json]}, \verb{[text/markdown, text/plain]}, \verb{[text/plain, application/json, application/vnd.amazonaws.connect.message.interactive.response]}.

The type \code{application/vnd.amazonaws.connect.message.interactive} is required to use the \href{https://docs.aws.amazon.com/connect/latest/adminguide/show-view-block.html}{Show view} flow block.}

\item{PersistentChat}{Enable persistent chats. For more information about enabling persistent chat, and for example use cases and how to configure for them, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/chat-persistence.html}{Enable persistent chat}.}

\item{RelatedContactId}{The unique identifier for an Connect Customer contact. This identifier is related to the chat starting.

You cannot provide data for both RelatedContactId and PersistentChat.}

\item{SegmentAttributes}{A set of system defined key-value pairs stored on individual contact segments using an attribute map. The attributes are standard Connect Customer attributes. They can be accessed in flows.

Attribute keys can include only alphanumeric, -, and _.

This field can be used to show channel subtype, such as \code{connect:Guide}.

The types \code{application/vnd.amazonaws.connect.message.interactive} and \code{application/vnd.amazonaws.connect.message.interactive.response} must be present in the SupportedMessagingContentTypes field of this API in order to set \code{SegmentAttributes} as \{\verb{ "connect:Subtype": \\\{"valueString" : "connect:Guide" \\\}\\\}}.}

\item{CustomerId}{The customer's identification number. For example, the \code{CustomerId} may be a customer number from your CRM.}

\item{DisconnectOnCustomerExit}{A list of participant types to automatically disconnect when the end customer ends the chat session, allowing them to continue through disconnect flows such as surveys or feedback forms.}

\item{ConnectionTypes}{The types of connection information to return in the response. This parameter is optional.

To receive connection information, specify one or both of the following values:
\itemize{
\item \code{CONNECTION_CREDENTIALS}: Returns a connection token.
\item \code{WEBSOCKET}: Returns a websocket URL.
}

\code{WEBSOCKET} and \code{CONNECTION_CREDENTIALS} are the values this operation acts on. No other value returns connection information.

Request \code{WEBSOCKET} to get a URL the participant connects to directly. You do not need to call \href{https://docs.aws.amazon.com/connect/latest/APIReference/API_connect-participant_CreateParticipantConnection.html}{CreateParticipantConnection} for it. Request \code{CONNECTION_CREDENTIALS} on its own and the response returns a connection token but no websocket URL.

If you omit this parameter, the response has no connection information.

When you start a new chat contact and the information you request cannot be returned, StartChatContact returns an error rather than a response that omits it. When you retry a request with the same \code{ClientToken}, the response repeats the original contact and can omit a websocket URL if the chat has already ended.}

\item{ChatStreamingConfiguration}{The streaming configuration, such as the Amazon SNS streaming endpoint. Use it to initiate real-time message streaming when the chat is created. This parameter is optional.

Setting this parameter returns a \code{StreamingId} in the response, and you do not need to call \code{\link[=connect_start_contact_streaming]{start_contact_streaming}}.

This parameter starts message streaming only. The response does not include connection information, and setting this parameter does not remove the need to call \href{https://docs.aws.amazon.com/connect/latest/APIReference/API_connect-participant_CreateParticipantConnection.html}{CreateParticipantConnection}.}
}
\description{
Initiates a flow to start a new chat for the customer. Response of this API provides a token required to obtain credentials from the \href{https://docs.aws.amazon.com/connect/latest/APIReference/API_connect-participant_CreateParticipantConnection.html}{CreateParticipantConnection} API in the Connect Customer Participant Service.

See \url{https://www.paws-r-sdk.com/docs/connect_start_chat_contact/} for full documentation.
}
\keyword{internal}
