% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_start_contact_conversational_analytics_job}
\alias{connect_start_contact_conversational_analytics_job}
\title{Starts a Contact Lens post-call analytics job for the specified contact}
\usage{
connect_start_contact_conversational_analytics_job(
  InstanceId,
  ContactId,
  AnalyticsModes,
  AnalyticsConfiguration,
  ClientToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact in this instance of Connect Customer.}

\item{AnalyticsModes}{[required] The analytics modes to run for the contact. Valid values: \code{PostContact}.}

\item{AnalyticsConfiguration}{[required] The configuration for the conversational analytics job.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see \href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}
}
\description{
Starts a Contact Lens post-call analytics job for the specified contact. This API runs Conversational Analytics post-contact analysis on a voice recording that is already attached to the contact, generating transcription, sentiment analysis, redaction, and summarization results based on the provided configuration.

See \url{https://www.paws-r-sdk.com/docs/connect_start_contact_conversational_analytics_job/} for full documentation.
}
\keyword{internal}
