% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_start_contact_media_processing}
\alias{connect_start_contact_media_processing}
\title{Enables in-flight message processing for an ongoing chat session}
\usage{
connect_start_contact_media_processing(
  InstanceId = NULL,
  ContactId = NULL,
  ProcessorArn = NULL,
  FailureMode = NULL
)
}
\arguments{
\item{InstanceId}{The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{The identifier of the contact.}

\item{ProcessorArn}{The Amazon Resource Name (ARN) of the Lambda processor. You can find the
Amazon Resource Name of the lambda in the lambda console.}

\item{FailureMode}{The desired behavior for failed message processing.}
}
\description{
Enables in-flight message processing for an ongoing chat session. Message processing will stay active for the rest of the chat, even if an individual contact segment ends.

See \url{https://www.paws-r-sdk.com/docs/connect_start_contact_media_processing/} for full documentation.
}
\keyword{internal}
