% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_start_contact_recording}
\alias{connect_start_contact_recording}
\title{Starts recording the contact:}
\usage{
connect_start_contact_recording(
  InstanceId,
  ContactId,
  InitialContactId,
  VoiceRecordingConfiguration
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact.}

\item{InitialContactId}{[required] The identifier of the contact. This is the identifier of the contact
associated with the first interaction with the contact center.}

\item{VoiceRecordingConfiguration}{[required] The person being recorded.}
}
\description{
Starts recording the contact:

See \url{https://www.paws-r-sdk.com/docs/connect_start_contact_recording/} for full documentation.
}
\keyword{internal}
