% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_start_screen_sharing}
\alias{connect_start_screen_sharing}
\title{Starts screen sharing for a contact}
\usage{
connect_start_screen_sharing(ClientToken = NULL, InstanceId, ContactId)
}
\arguments{
\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact in this instance of Amazon Connect.}
}
\description{
Starts screen sharing for a contact. For more information about screen sharing, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/inapp-calling.html}{Set up in-app, web, video calling, and screen sharing capabilities} in the \emph{Amazon Connect Administrator Guide}.

See \url{https://www.paws-r-sdk.com/docs/connect_start_screen_sharing/} for full documentation.
}
\keyword{internal}
