% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_stop_contact_media_processing}
\alias{connect_stop_contact_media_processing}
\title{Stops in-flight message processing for an ongoing chat session}
\usage{
connect_stop_contact_media_processing(InstanceId = NULL, ContactId = NULL)
}
\arguments{
\item{InstanceId}{The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{The identifier of the contact.}
}
\description{
Stops in-flight message processing for an ongoing chat session.

See \url{https://www.paws-r-sdk.com/docs/connect_stop_contact_media_processing/} for full documentation.
}
\keyword{internal}
