% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_stop_contact_streaming}
\alias{connect_stop_contact_streaming}
\title{Ends message streaming on a specified contact}
\usage{
connect_stop_contact_streaming(InstanceId, ContactId, StreamingId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ContactId}{[required] The identifier of the contact. This is the identifier of the contact
that is associated with the first interaction with the contact center.}

\item{StreamingId}{[required] The identifier of the streaming configuration enabled.}
}
\description{
Ends message streaming on a specified contact. To restart message streaming on that contact, call the \code{\link[=connect_start_contact_streaming]{start_contact_streaming}} API.

See \url{https://paws-r.github.io/docs/connect/stop_contact_streaming.html} for full documentation.
}
\keyword{internal}
