% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_stop_test_case_execution}
\alias{connect_stop_test_case_execution}
\title{Stops a running test execution}
\usage{
connect_stop_test_case_execution(
  InstanceId,
  TestCaseExecutionId,
  TestCaseId,
  ClientToken = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{TestCaseExecutionId}{[required] The identifier of the test case execution to stop.}

\item{TestCaseId}{[required] The identifier of the test case.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}
}
\description{
Stops a running test execution.

See \url{https://www.paws-r-sdk.com/docs/connect_stop_test_case_execution/} for full documentation.
}
\keyword{internal}
