% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_submit_contact_evaluation}
\alias{connect_submit_contact_evaluation}
\title{Submits a contact evaluation in the specified Amazon Connect instance}
\usage{
connect_submit_contact_evaluation(
  InstanceId,
  EvaluationId,
  Answers = NULL,
  Notes = NULL,
  SubmittedBy = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{EvaluationId}{[required] A unique identifier for the contact evaluation.}

\item{Answers}{A map of question identifiers to answer value.}

\item{Notes}{A map of question identifiers to note value.}

\item{SubmittedBy}{The ID of the user who submitted the contact evaluation.}
}
\description{
Submits a contact evaluation in the specified Amazon Connect instance. Answers included in the request are merged with existing answers for the given evaluation. If no answers or notes are passed, the evaluation is submitted with the existing answers and notes. You can delete an answer or note by passing an empty object (\code{{}}) to the question identifier.

See \url{https://www.paws-r-sdk.com/docs/connect_submit_contact_evaluation/} for full documentation.
}
\keyword{internal}
