% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_untag_contact}
\alias{connect_untag_contact}
\title{Removes the specified tags from the contact resource}
\usage{
connect_untag_contact(ContactId, InstanceId, TagKeys)
}
\arguments{
\item{ContactId}{[required] The identifier of the contact in this instance of Amazon Connect.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{TagKeys}{[required] A list of tag keys. Existing tags on the contact whose keys are members
of this list will be removed.}
}
\description{
Removes the specified tags from the contact resource. For more information about this API is used, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/granular-billing.html}{Set up granular billing for a detailed view of your Amazon Connect usage}.

See \url{https://www.paws-r-sdk.com/docs/connect_untag_contact/} for full documentation.
}
\keyword{internal}
