% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_agent_status}
\alias{connect_update_agent_status}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_agent_status(
  InstanceId,
  AgentStatusId,
  Name = NULL,
  Description = NULL,
  State = NULL,
  DisplayOrder = NULL,
  ResetOrderNumber = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{AgentStatusId}{[required] The identifier of the agent status.}

\item{Name}{The name of the agent status.}

\item{Description}{The description of the agent status.}

\item{State}{The state of the agent status.}

\item{DisplayOrder}{The display order of the agent status.}

\item{ResetOrderNumber}{A number indicating the reset order of the agent status.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/update_agent_status.html} for full documentation.
}
\keyword{internal}
