% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact}
\alias{connect_update_contact}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_contact(
  InstanceId,
  ContactId,
  Name = NULL,
  Description = NULL,
  References = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ContactId}{[required] The identifier of the contact. This is the identifier of the contact
associated with the first interaction with your contact center.}

\item{Name}{The name of the contact.}

\item{Description}{The description of the contact.}

\item{References}{Well-formed data on contact, shown to agents on Contact Control Panel
(CCP).}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/update_contact.html} for full documentation.
}
\keyword{internal}
