% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_flow_metadata}
\alias{connect_update_contact_flow_metadata}
\title{Updates metadata about specified contact flow}
\usage{
connect_update_contact_flow_metadata(
  InstanceId,
  ContactFlowId,
  Name = NULL,
  Description = NULL,
  ContactFlowState = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ContactFlowId}{[required] The identifier of the contact flow.}

\item{Name}{TThe name of the contact flow.}

\item{Description}{The description of the contact flow.}

\item{ContactFlowState}{The state of contact flow.}
}
\description{
Updates metadata about specified contact flow.

See \url{https://paws-r.github.io/docs/connect/update_contact_flow_metadata.html} for full documentation.
}
\keyword{internal}
