% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_flow_module_alias}
\alias{connect_update_contact_flow_module_alias}
\title{Updates a specific Aliases metadata, including the version it’s tied to,
it’s name, and description}
\usage{
connect_update_contact_flow_module_alias(
  InstanceId,
  ContactFlowModuleId,
  AliasId,
  Name = NULL,
  Description = NULL,
  ContactFlowModuleVersion = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowModuleId}{[required] The identifier of the flow module.}

\item{AliasId}{[required] The identifier of the alias.}

\item{Name}{The name of the alias.}

\item{Description}{The description of the alias.}

\item{ContactFlowModuleVersion}{The version of the flow module.}
}
\description{
Updates a specific Aliases metadata, including the version it’s tied to, it’s name, and description.

See \url{https://www.paws-r-sdk.com/docs/connect_update_contact_flow_module_alias/} for full documentation.
}
\keyword{internal}
