% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_flow_module_metadata}
\alias{connect_update_contact_flow_module_metadata}
\title{Updates metadata about specified contact flow module}
\usage{
connect_update_contact_flow_module_metadata(
  InstanceId,
  ContactFlowModuleId,
  Name = NULL,
  Description = NULL,
  State = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ContactFlowModuleId}{[required] The identifier of the contact flow module.}

\item{Name}{The name of the contact flow module.}

\item{Description}{The description of the contact flow module.}

\item{State}{The state of contact flow module.}
}
\description{
Updates metadata about specified contact flow module.

See \url{https://paws-r.github.io/docs/connect/update_contact_flow_module_metadata.html} for full documentation.
}
\keyword{internal}
