% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_flow_name}
\alias{connect_update_contact_flow_name}
\title{The name of the contact flow}
\usage{
connect_update_contact_flow_name(InstanceId, ContactFlowId, Name,
  Description)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{ContactFlowId}{[required] The identifier of the contact flow.}

\item{Name}{The name of the contact flow.}

\item{Description}{The description of the contact flow.}
}
\value{
An empty list.
}
\description{
The name of the contact flow.

You can also create and update contact flows using the \href{https://docs.aws.amazon.com/connect/latest/adminguide/flow-language.html}{Amazon Connect Flow language}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_contact_flow_name(
  InstanceId = "string",
  ContactFlowId = "string",
  Name = "string",
  Description = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
