% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_schedule}
\alias{connect_update_contact_schedule}
\title{Updates the scheduled time of a task contact that is already scheduled}
\usage{
connect_update_contact_schedule(InstanceId, ContactId, ScheduledTime)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact.}

\item{ScheduledTime}{[required] The timestamp, in Unix Epoch seconds format, at which to start running
the inbound flow. The scheduled time cannot be in the past. It must be
within up to 6 days in future.}
}
\description{
Updates the scheduled time of a task contact that is already scheduled.

See \url{https://paws-r.github.io/docs/connect/update_contact_schedule.html} for full documentation.
}
\keyword{internal}
