% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_task_template}
\alias{connect_update_contact_task_template}
\title{Updates the task template association on an existing task contact}
\usage{
connect_update_contact_task_template(InstanceId, TaskTemplateId, ContactId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{TaskTemplateId}{[required] A unique identifier for the task template. For more information about task templates, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/task-templates.html}{Task templates} in the \emph{Connect Customer Administrator Guide}.}

\item{ContactId}{[required] The identifier of the contact in this instance of Connect Customer.}
}
\description{
Updates the task template association on an existing task contact. You can update the task template on a contact before assignment to support tasks that are created without a template (for example \href{https://docs.aws.amazon.com/connect/latest/adminguide/connect-rules.html}{Rules} or \href{https://docs.aws.amazon.com/connect/latest/adminguide/set-disconnect-flow.html}{disconnect flows}) or change the agent interaction form to represent the latest task data (for example an initial request that was submitted as a refund gets updated to an account cancellation and requires a new template).

See \url{https://www.paws-r-sdk.com/docs/connect_update_contact_task_template/} for full documentation.
}
\keyword{internal}
