% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_cross_region_routing}
\alias{connect_update_cross_region_routing}
\title{Updates the cross-region routing configuration for an Amazon Connect
Global Resiliency instance enabled for global routing}
\usage{
connect_update_cross_region_routing(InstanceId, IsolatedAll)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{IsolatedAll}{[required] Set to \code{true} to disable cross-region routing for all Regions associated with this instance. Set to \code{false} to re-enable cross-region routing.}
}
\description{
Updates the cross-region routing configuration for an Amazon Connect Global Resiliency instance enabled for global routing. When invoked with \code{IsolatedAll} set to \code{true}, this operation disables cross-region routing, meaning contacts originating in one Region will no longer be routed to agents in another Region.

See \url{https://www.paws-r-sdk.com/docs/connect_update_cross_region_routing/} for full documentation.
}
\keyword{internal}
