% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_data_table_attribute}
\alias{connect_update_data_table_attribute}
\title{Updates all properties for an attribute using all properties from
CreateDataTableAttribute}
\usage{
connect_update_data_table_attribute(
  InstanceId,
  DataTableId,
  AttributeName,
  Name,
  ValueType,
  Description = NULL,
  Primary = NULL,
  Validation = NULL
)
}
\arguments{
\item{InstanceId}{[required] The unique identifier for the Amazon Connect instance.}

\item{DataTableId}{[required] The unique identifier for the data table. Must also accept the table ARN
with or without a version alias.}

\item{AttributeName}{[required] The current name of the attribute to update. Used as an identifier since
attribute names can be changed.}

\item{Name}{[required] The new name for the attribute. Must conform to Connect human readable
string specification and be unique within the data table.}

\item{ValueType}{[required] The updated value type for the attribute. When changing value types,
existing values are not deleted but may return default values if
incompatible.}

\item{Description}{The updated description for the attribute.}

\item{Primary}{Whether the attribute should be treated as a primary key. Converting to
primary attribute requires existing values to maintain uniqueness.}

\item{Validation}{The updated validation rules for the attribute. Changes do not affect
existing values until they are modified.}
}
\description{
Updates all properties for an attribute using all properties from CreateDataTableAttribute. There are no other granular update endpoints. It does not act as a patch operation - all properties must be provided. System managed attributes are not mutable by customers. Changing an attribute's validation does not invalidate existing values since validation only runs when values are created or updated.

See \url{https://www.paws-r-sdk.com/docs/connect_update_data_table_attribute/} for full documentation.
}
\keyword{internal}
