% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_data_table_metadata}
\alias{connect_update_data_table_metadata}
\title{Updates the metadata properties of a data table}
\usage{
connect_update_data_table_metadata(
  InstanceId,
  DataTableId,
  Name,
  Description = NULL,
  ValueLockLevel,
  TimeZone
)
}
\arguments{
\item{InstanceId}{[required] The unique identifier for the Amazon Connect instance.}

\item{DataTableId}{[required] The unique identifier for the data table. Must also accept the table ARN
with or without a version alias. If the version is provided as part of
the identifier or ARN, the version must be $LATEST. Providing any other
alias fails with an error.}

\item{Name}{[required] The updated name for the data table. Must conform to Connect human
readable string specification and have 1-127 characters. Must be unique
for the instance using case-insensitive comparison.}

\item{Description}{The updated description for the data table. Must conform to Connect
human readable string specification and have 0-250 characters.}

\item{ValueLockLevel}{[required] The updated value lock level for the data table. One of DATA_TABLE,
PRIMARY_VALUE, ATTRIBUTE, VALUE, and NONE.}

\item{TimeZone}{[required] The updated IANA timezone identifier to use when resolving time based
dynamic values.}
}
\description{
Updates the metadata properties of a data table. Accepts all fields similar to CreateDataTable, except for fields and tags. There are no other granular update endpoints. It does not act as a patch operation - all properties must be provided or defaults will be used. Fields follow the same requirements as CreateDataTable.

See \url{https://www.paws-r-sdk.com/docs/connect_update_data_table_metadata/} for full documentation.
}
\keyword{internal}
