% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_extraction_definition}
\alias{connect_update_extraction_definition}
\title{Updates an extraction definition in the specified Connect Customer
instance}
\usage{
connect_update_extraction_definition(
  ClientToken = NULL,
  ExtractionDefinitionId,
  InstanceId,
  Name,
  ExtractionConfiguration,
  Display = NULL
)
}
\arguments{
\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field.}

\item{ExtractionDefinitionId}{[required] The identifier of the extraction definition to update.}

\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{Name}{[required] The name of the extraction definition.}

\item{ExtractionConfiguration}{[required] The configuration that defines how data is extracted, including the prompt hint and not-found behavior.}

\item{Display}{The display settings for the extraction definition.}
}
\description{
Updates an extraction definition in the specified Connect Customer instance.

See \url{https://www.paws-r-sdk.com/docs/connect_update_extraction_definition/} for full documentation.
}
\keyword{internal}
