% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_metric_content}
\alias{connect_update_metric_content}
\title{Updates the calculation, unit, and/or trend indicator of an existing
metric in the specified Connect Customer instance}
\usage{
connect_update_metric_content(
  InstanceId,
  MetricId,
  MetricCalculation = NULL,
  Unit = NULL,
  PositiveTrendIndicator = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{MetricId}{[required] The identifier of the metric to update. Adding the \verb{$SAVED} qualifier will update the saved version of the metric. Adding \verb{$LATEST} or omitting a qualifier will update the published version.}

\item{MetricCalculation}{The updated calculation definition for the metric.}

\item{Unit}{The updated display unit for the metric.}

\item{PositiveTrendIndicator}{How an increase in the metric value should be interpreted. Valid values: \code{POSITIVE}, \code{NEUTRAL}, \code{NEGATIVE}.}
}
\description{
Updates the calculation, unit, and/or trend indicator of an existing metric in the specified Connect Customer instance.

See \url{https://www.paws-r-sdk.com/docs/connect_update_metric_content/} for full documentation.
}
\keyword{internal}
