% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_metric_metadata}
\alias{connect_update_metric_metadata}
\title{Updates the name and/or description of an existing metric in the
specified Connect Customer instance}
\usage{
connect_update_metric_metadata(
  InstanceId,
  MetricId,
  Name = NULL,
  Description = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{MetricId}{[required] The identifier of the metric to update. Adding the \verb{$SAVED} qualifier will update the saved version of the metric. Adding \verb{$LATEST} or omitting a qualifier will update the published version.}

\item{Name}{The updated name of the metric.}

\item{Description}{The updated description of the metric.}
}
\description{
Updates the name and/or description of an existing metric in the specified Connect Customer instance.

See \url{https://www.paws-r-sdk.com/docs/connect_update_metric_metadata/} for full documentation.
}
\keyword{internal}
