% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_notification_content}
\alias{connect_update_notification_content}
\title{Updates the localized content of an existing notification}
\usage{
connect_update_notification_content(InstanceId, NotificationId, Content)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{NotificationId}{[required] The unique identifier for the notification to update.}

\item{Content}{[required] The updated localized content of the notification. A map of locale codes and values. Maximum 500 visible characters per locale.}
}
\description{
Updates the localized content of an existing notification. This operation applies to all users for whom the notification was sent.

See \url{https://www.paws-r-sdk.com/docs/connect_update_notification_content/} for full documentation.
}
\keyword{internal}
