% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_participant_role_config}
\alias{connect_update_participant_role_config}
\title{Updates timeouts for when human chat participants are to be considered
idle, and when agents are automatically disconnected from a chat due to
idleness}
\usage{
connect_update_participant_role_config(
  InstanceId,
  ContactId,
  ChannelConfiguration
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact in this instance of Amazon Connect.}

\item{ChannelConfiguration}{[required] The Amazon Connect channel you want to configure.}
}
\description{
Updates timeouts for when human chat participants are to be considered idle, and when agents are automatically disconnected from a chat due to idleness. You can set four timers:

See \url{https://www.paws-r-sdk.com/docs/connect_update_participant_role_config/} for full documentation.
}
\keyword{internal}
