% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_phone_number_metadata}
\alias{connect_update_phone_number_metadata}
\title{Updates a phone number’s metadata}
\usage{
connect_update_phone_number_metadata(
  PhoneNumberId,
  PhoneNumberDescription = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{PhoneNumberId}{[required] The Amazon Resource Name (ARN) or resource ID of the phone number.}

\item{PhoneNumberDescription}{The description of the phone number.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}
}
\description{
Updates a phone number’s metadata.

See \url{https://www.paws-r-sdk.com/docs/connect_update_phone_number_metadata/} for full documentation.
}
\keyword{internal}
