% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_predefined_attribute}
\alias{connect_update_predefined_attribute}
\title{Updates a predefined attribute for the specified Amazon Connect instance}
\usage{
connect_update_predefined_attribute(InstanceId, Name, Values = NULL)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the instance
ID in the Amazon Resource Name (ARN) of the instance.}

\item{Name}{[required] The name of the predefined attribute.}

\item{Values}{The values of the predefined attribute.}
}
\description{
Updates a predefined attribute for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_update_predefined_attribute/} for full documentation.
}
\keyword{internal}
