% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_prompt}
\alias{connect_update_prompt}
\title{Updates a prompt}
\usage{
connect_update_prompt(
  InstanceId,
  PromptId,
  Name = NULL,
  Description = NULL,
  S3Uri = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{PromptId}{[required] A unique identifier for the prompt.}

\item{Name}{The name of the prompt.}

\item{Description}{A description of the prompt.}

\item{S3Uri}{The URI for the S3 bucket where the prompt is stored.}
}
\description{
Updates a prompt.

See \url{https://www.paws-r-sdk.com/docs/connect_update_prompt/} for full documentation.
}
\keyword{internal}
