% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_queue_max_contacts}
\alias{connect_update_queue_max_contacts}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_queue_max_contacts(InstanceId, QueueId, MaxContacts)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{MaxContacts}{The maximum number of contacts that can be in the queue before it is
considered full.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Updates the maximum number of contacts allowed in a queue before it is
considered full.
}
\section{Request syntax}{
\preformatted{svc$update_queue_max_contacts(
  InstanceId = "string",
  QueueId = "string",
  MaxContacts = 123
)
}
}

\keyword{internal}
