% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_queue_outbound_caller_config}
\alias{connect_update_queue_outbound_caller_config}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_queue_outbound_caller_config(
  InstanceId,
  QueueId,
  OutboundCallerConfig
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{OutboundCallerConfig}{[required] The outbound caller ID name, number, and outbound whisper flow.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/update_queue_outbound_caller_config.html} for full documentation.
}
\keyword{internal}
