% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_queue_outbound_email_config}
\alias{connect_update_queue_outbound_email_config}
\title{Updates the outbound email address Id for a specified queue}
\usage{
connect_update_queue_outbound_email_config(
  InstanceId,
  QueueId,
  OutboundEmailConfig
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{OutboundEmailConfig}{[required] The outbound email address ID for a specified queue.}
}
\description{
Updates the outbound email address Id for a specified queue.

See \url{https://www.paws-r-sdk.com/docs/connect_update_queue_outbound_email_config/} for full documentation.
}
\keyword{internal}
