% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_queue_status}
\alias{connect_update_queue_status}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_queue_status(InstanceId, QueueId, Status)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{Status}{[required] The status of the queue.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Updates the status of the queue.
}
\section{Request syntax}{
\preformatted{svc$update_queue_status(
  InstanceId = "string",
  QueueId = "string",
  Status = "ENABLED"|"DISABLED"
)
}
}

\keyword{internal}
