% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_quick_connect_config}
\alias{connect_update_quick_connect_config}
\title{Updates the configuration settings for the specified quick connect}
\usage{
connect_update_quick_connect_config(
  InstanceId,
  QuickConnectId,
  QuickConnectConfig
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QuickConnectId}{[required] The identifier for the quick connect.}

\item{QuickConnectConfig}{[required] Information about the configuration settings for the quick connect.}
}
\description{
Updates the configuration settings for the specified quick connect.

See \url{https://paws-r.github.io/docs/connect/update_quick_connect_config.html} for full documentation.
}
\keyword{internal}
