% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_quick_connect_name}
\alias{connect_update_quick_connect_name}
\title{Updates the name and description of a quick connect}
\usage{
connect_update_quick_connect_name(
  InstanceId,
  QuickConnectId,
  Name = NULL,
  Description = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QuickConnectId}{[required] The identifier for the quick connect.}

\item{Name}{The name of the quick connect.}

\item{Description}{The description of the quick connect.}
}
\description{
Updates the name and description of a quick connect. The request accepts the following data in JSON format. At least \code{Name} or \code{Description} must be provided.

See \url{https://paws-r.github.io/docs/connect/update_quick_connect_name.html} for full documentation.
}
\keyword{internal}
