% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_routing_profile_agent_availability_timer}
\alias{connect_update_routing_profile_agent_availability_timer}
\title{Whether agents with this routing profile will have their routing order
calculated based on time since their last inbound contact or longest
idle time}
\usage{
connect_update_routing_profile_agent_availability_timer(
  InstanceId,
  RoutingProfileId,
  AgentAvailabilityTimer
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{AgentAvailabilityTimer}{[required] Whether agents with this routing profile will have their routing order
calculated based on \emph{time since their last inbound contact} or \emph{longest
idle time}.}
}
\description{
Whether agents with this routing profile will have their routing order calculated based on \emph{time since their last inbound contact} or \emph{longest idle time}.

See \url{https://www.paws-r-sdk.com/docs/connect_update_routing_profile_agent_availability_timer/} for full documentation.
}
\keyword{internal}
