% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_routing_profile_concurrency}
\alias{connect_update_routing_profile_concurrency}
\title{Updates the channels that agents can handle in the Contact Control Panel
(CCP) for a routing profile}
\usage{
connect_update_routing_profile_concurrency(
  InstanceId,
  RoutingProfileId,
  MediaConcurrencies
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{MediaConcurrencies}{[required] The channels that agents can handle in the Contact Control Panel (CCP).}
}
\description{
Updates the channels that agents can handle in the Contact Control Panel (CCP) for a routing profile.

See \url{https://paws-r.github.io/docs/connect/update_routing_profile_concurrency.html} for full documentation.
}
\keyword{internal}
