% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_routing_profile_default_outbound_queue}
\alias{connect_update_routing_profile_default_outbound_queue}
\title{Updates the default outbound queue of a routing profile}
\usage{
connect_update_routing_profile_default_outbound_queue(
  InstanceId,
  RoutingProfileId,
  DefaultOutboundQueueId
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{DefaultOutboundQueueId}{[required] The identifier for the default outbound queue.}
}
\description{
Updates the default outbound queue of a routing profile.

See \url{https://www.paws-r-sdk.com/docs/connect_update_routing_profile_default_outbound_queue/} for full documentation.
}
\keyword{internal}
