% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_routing_profile_queues}
\alias{connect_update_routing_profile_queues}
\title{Updates the properties associated with a set of queues for a routing
profile}
\usage{
connect_update_routing_profile_queues(
  InstanceId,
  RoutingProfileId,
  QueueConfigs
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{QueueConfigs}{[required] The queues to be updated for this routing profile. Queues must first be
associated to the routing profile. You can do this using
AssociateRoutingProfileQueues.}
}
\description{
Updates the properties associated with a set of queues for a routing profile.

See \url{https://paws-r.github.io/docs/connect/update_routing_profile_queues.html} for full documentation.
}
\keyword{internal}
