% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_rule}
\alias{connect_update_rule}
\title{Updates a rule for the specified Connect Customer instance}
\usage{
connect_update_rule(
  RuleId,
  InstanceId,
  Name,
  Function,
  Actions,
  PublishStatus,
  PreEvaluationFilters = NULL
)
}
\arguments{
\item{RuleId}{[required] A unique identifier for the rule.}

\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{Name}{[required] The name of the rule. You can change the name only if \code{TriggerEventSource} is one of the following values: \code{OnZendeskTicketCreate} | \code{OnZendeskTicketStatusUpdate} | \code{OnSalesforceCaseCreate}}

\item{Function}{[required] The conditions of the rule.}

\item{Actions}{[required] A list of actions to be run when the rule is triggered.}

\item{PublishStatus}{[required] The publish status of the rule.}

\item{PreEvaluationFilters}{The pre-evaluation filters for the rule, that restrict the rule to be applied to only certain resources based on the resource's attributes, such as tags assigned to a contact. The pre-evaluation filters are applied even before rule conditions are evaluated and are used to enforce tag-based-access-control while applying rules.}
}
\description{
Updates a rule for the specified Connect Customer instance.

See \url{https://www.paws-r-sdk.com/docs/connect_update_rule/} for full documentation.
}
\keyword{internal}
