% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_security_profile}
\alias{connect_update_security_profile}
\title{Updates a security profile}
\usage{
connect_update_security_profile(
  Description = NULL,
  Permissions = NULL,
  SecurityProfileId,
  InstanceId,
  AllowedAccessControlTags = NULL,
  TagRestrictedResources = NULL,
  Applications = NULL,
  HierarchyRestrictedResources = NULL,
  AllowedAccessControlHierarchyGroupId = NULL,
  AllowedFlowModules = NULL,
  AllowedAIAgents = NULL,
  GranularAccessControlConfiguration = NULL
)
}
\arguments{
\item{Description}{The description of the security profile.}

\item{Permissions}{The permissions granted to a security profile. For a list of valid permissions, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/security-profile-list.html}{List of security profile permissions}.}

\item{SecurityProfileId}{[required] The identifier for the security profle.}

\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{AllowedAccessControlTags}{The list of tags that a security profile uses to restrict access to resources in Connect Customer.}

\item{TagRestrictedResources}{The list of resources that a security profile applies tag restrictions to in Connect Customer.}

\item{Applications}{A list of the third-party application's metadata.}

\item{HierarchyRestrictedResources}{The list of resources that a security profile applies hierarchy restrictions to in Connect Customer. Following are acceptable ResourceNames: \code{User}.}

\item{AllowedAccessControlHierarchyGroupId}{The identifier of the hierarchy group that a security profile uses to restrict access to resources in Connect Customer.}

\item{AllowedFlowModules}{A list of Flow Modules an AI Agent can invoke as a tool}

\item{AllowedAIAgents}{A list of AI agents that the security profile will give access to.}

\item{GranularAccessControlConfiguration}{The granular access control configuration for the security profile, including data table permissions.}
}
\description{
Updates a security profile.

See \url{https://www.paws-r-sdk.com/docs/connect_update_security_profile/} for full documentation.
}
\keyword{internal}
