% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_task_template}
\alias{connect_update_task_template}
\title{Updates details about a specific task template in the specified Connect
Customer instance}
\usage{
connect_update_task_template(
  TaskTemplateId,
  InstanceId,
  Name = NULL,
  Description = NULL,
  ContactFlowId = NULL,
  SelfAssignFlowId = NULL,
  Constraints = NULL,
  Defaults = NULL,
  Status = NULL,
  Fields = NULL
)
}
\arguments{
\item{TaskTemplateId}{[required] A unique identifier for the task template.}

\item{InstanceId}{[required] The identifier of the Connect Customer instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID} in the Amazon Resource Name (ARN) of the instance.}

\item{Name}{The name of the task template.}

\item{Description}{The description of the task template.}

\item{ContactFlowId}{The identifier of the flow that runs by default when a task is created by referencing this template.

Although this parameter is marked as optional, the request must contain either a \code{ContactFlowId} or a field of type \code{QUICK_CONNECT}.}

\item{SelfAssignFlowId}{The ContactFlowId for the flow that will be run if this template is used to create a self-assigned task.}

\item{Constraints}{Constraints that are applicable to the fields listed. Although this parameter is marked as optional in the API model, the service requires it when calling \code{\link[=connect_create_task_template]{create_task_template}} or \code{\link[=connect_update_task_template]{update_task_template}}. The \code{RequiredFields} array must contain at least one element, and the field of type \code{NAME} must be included in \code{RequiredFields}.}

\item{Defaults}{The default values for fields when a task is created by referencing this template.}

\item{Status}{Marks a template as \code{ACTIVE} or \code{INACTIVE} for a task to refer to it. Tasks can only be created from \code{ACTIVE} templates. If a template is marked as \code{INACTIVE}, then a task that refers to this template cannot be created.

Although this parameter is marked as optional, the service requires it when calling \code{\link[=connect_update_task_template]{update_task_template}}.}

\item{Fields}{Fields that are part of the template.

The request must contain exactly one field of type \code{NAME}. This field must also be listed in the \code{RequiredFields} array within the \code{Constraints} parameter.}
}
\description{
Updates details about a specific task template in the specified Connect Customer instance. This operation does not support partial updates. Instead it does a full update of template content.

See \url{https://www.paws-r-sdk.com/docs/connect_update_task_template/} for full documentation.
}
\keyword{internal}
