% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_test_case}
\alias{connect_update_test_case}
\title{Updates any of the metadata for a test case, such as the name,
description, and status or content of an existing test case}
\usage{
connect_update_test_case(
  InstanceId,
  TestCaseId,
  Content = NULL,
  EntryPoint = NULL,
  InitializationData = NULL,
  Name = NULL,
  Description = NULL,
  Status = NULL,
  LastModifiedTime = NULL,
  LastModifiedRegion = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{TestCaseId}{[required] The identifier of the test case to update.}

\item{Content}{The JSON string that represents the content of the test.}

\item{EntryPoint}{Defines the starting point for your test.}

\item{InitializationData}{Defines the test attributes for precise data representation.}

\item{Name}{The name of the test case.}

\item{Description}{The description of the test case.}

\item{Status}{Indicates the test status as either SAVED or PUBLISHED. The PUBLISHED
status will initiate validation on the content. The SAVED status does
not initiate validation of the content.}

\item{LastModifiedTime}{The time at which the resource was last modified.}

\item{LastModifiedRegion}{The region in which the resource was last modified}
}
\description{
Updates any of the metadata for a test case, such as the name, description, and status or content of an existing test case. This API doesn't allow customers to update the tags of the test case resource for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_update_test_case/} for full documentation.
}
\keyword{internal}
