% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_traffic_distribution}
\alias{connect_update_traffic_distribution}
\title{Updates the traffic distribution for a given traffic distribution group}
\usage{
connect_update_traffic_distribution(Id, TelephonyConfig = NULL)
}
\arguments{
\item{Id}{[required] The identifier of the traffic distribution group. This can be the ID or
the ARN if the API is being called in the Region where the traffic
distribution group was created. The ARN must be provided if the call is
from the replicated Region.}

\item{TelephonyConfig}{The distribution of traffic between the instance and its replica(s).}
}
\description{
Updates the traffic distribution for a given traffic distribution group.

See \url{https://www.paws-r-sdk.com/docs/connect_update_traffic_distribution/} for full documentation.
}
\keyword{internal}
