% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_user_hierarchy}
\alias{connect_update_user_hierarchy}
\title{Assigns the specified hierarchy group to the specified user}
\usage{
connect_update_user_hierarchy(HierarchyGroupId = NULL, UserId, InstanceId)
}
\arguments{
\item{HierarchyGroupId}{The identifier of the hierarchy group.}

\item{UserId}{[required] The identifier of the user account.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}
}
\description{
Assigns the specified hierarchy group to the specified user.

See \url{https://www.paws-r-sdk.com/docs/connect_update_user_hierarchy/} for full documentation.
}
\keyword{internal}
