% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_user_hierarchy_group_name}
\alias{connect_update_user_hierarchy_group_name}
\title{Updates the name of the user hierarchy group}
\usage{
connect_update_user_hierarchy_group_name(Name, HierarchyGroupId, InstanceId)
}
\arguments{
\item{Name}{[required] The name of the hierarchy group. Must not be more than 100 characters.}

\item{HierarchyGroupId}{[required] The identifier of the hierarchy group.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}
}
\description{
Updates the name of the user hierarchy group.

See \url{https://www.paws-r-sdk.com/docs/connect_update_user_hierarchy_group_name/} for full documentation.
}
\keyword{internal}
