% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_user_phone_config}
\alias{connect_update_user_phone_config}
\title{Updates the phone configuration settings for the specified user}
\usage{
connect_update_user_phone_config(PhoneConfig, UserId, InstanceId)
}
\arguments{
\item{PhoneConfig}{[required] Information about phone configuration settings for the user.}

\item{UserId}{[required] The identifier of the user account.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\value{
An empty list.
}
\description{
Updates the phone configuration settings for the specified user.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_user_phone_config(
  PhoneConfig = list(
    PhoneType = "SOFT_PHONE"|"DESK_PHONE",
    AutoAccept = TRUE|FALSE,
    AfterContactWorkTimeLimit = 123,
    DeskPhoneNumber = "string"
  ),
  UserId = "string",
  InstanceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
