% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_user_routing_profile}
\alias{connect_update_user_routing_profile}
\title{Assigns the specified routing profile to the specified user}
\usage{
connect_update_user_routing_profile(RoutingProfileId, UserId, InstanceId)
}
\arguments{
\item{RoutingProfileId}{[required] The identifier of the routing profile for the user.}

\item{UserId}{[required] The identifier of the user account.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}
}
\description{
Assigns the specified routing profile to the specified user.

See \url{https://paws-r.github.io/docs/connect/update_user_routing_profile.html} for full documentation.
}
\keyword{internal}
