% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_view_content}
\alias{connect_update_view_content}
\title{Updates the view content of the given view identifier in the specified
Amazon Connect instance}
\usage{
connect_update_view_content(InstanceId, ViewId, Status, Content)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ViewId}{[required] The identifier of the view. Both \code{ViewArn} and \code{ViewId} can be used.}

\item{Status}{[required] Indicates the view status as either \code{SAVED} or \code{PUBLISHED}. The
\code{PUBLISHED} status will initiate validation on the content.}

\item{Content}{[required] View content containing all content necessary to render a view except
for runtime input data and the runtime input schema, which is
auto-generated by this operation.

The total uncompressed content has a maximum file size of 400kB.}
}
\description{
Updates the view content of the given view identifier in the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_update_view_content/} for full documentation.
}
\keyword{internal}
