% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_view_metadata}
\alias{connect_update_view_metadata}
\title{Updates the view metadata}
\usage{
connect_update_view_metadata(
  InstanceId,
  ViewId,
  Name = NULL,
  Description = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ViewId}{[required] The identifier of the view. Both \code{ViewArn} and \code{ViewId} can be used.}

\item{Name}{The name of the view.}

\item{Description}{The description of the view.}
}
\description{
Updates the view metadata. Note that either \code{Name} or \code{Description} must be provided.

See \url{https://www.paws-r-sdk.com/docs/connect_update_view_metadata/} for full documentation.
}
\keyword{internal}
