% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_workspace_metadata}
\alias{connect_update_workspace_metadata}
\title{Updates the metadata of a workspace, such as its name and description}
\usage{
connect_update_workspace_metadata(
  InstanceId,
  WorkspaceId,
  Name = NULL,
  Description = NULL,
  Title = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}

\item{Name}{The name of the workspace.}

\item{Description}{The description of the workspace.}

\item{Title}{The title displayed for the workspace.}
}
\description{
Updates the metadata of a workspace, such as its name and description.

See \url{https://www.paws-r-sdk.com/docs/connect_update_workspace_metadata/} for full documentation.
}
\keyword{internal}
