% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_workspace_page}
\alias{connect_update_workspace_page}
\title{Updates the configuration of a page in a workspace, including the
associated view and input data}
\usage{
connect_update_workspace_page(
  InstanceId,
  WorkspaceId,
  Page,
  NewPage = NULL,
  ResourceArn = NULL,
  Slug = NULL,
  InputData = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}

\item{Page}{[required] The current page identifier.}

\item{NewPage}{The new page identifier, if changing the page name.}

\item{ResourceArn}{The Amazon Resource Name (ARN) of the view to associate with the page.}

\item{Slug}{The URL-friendly identifier for the page.}

\item{InputData}{A JSON string containing input parameters for the view.}
}
\description{
Updates the configuration of a page in a workspace, including the associated view and input data.

See \url{https://www.paws-r-sdk.com/docs/connect_update_workspace_page/} for full documentation.
}
\keyword{internal}
