% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_workspace_theme}
\alias{connect_update_workspace_theme}
\title{Updates the theme configuration for a workspace, including colors and
styling}
\usage{
connect_update_workspace_theme(InstanceId, WorkspaceId, Theme = NULL)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{WorkspaceId}{[required] The identifier of the workspace.}

\item{Theme}{The theme configuration, including color schemes and visual styles.}
}
\description{
Updates the theme configuration for a workspace, including colors and styling.

See \url{https://www.paws-r-sdk.com/docs/connect_update_workspace_theme/} for full documentation.
}
\keyword{internal}
