% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_create_campaign}
\alias{connectcampaignservice_create_campaign}
\title{Creates a campaign for the specified Amazon Connect account}
\usage{
connectcampaignservice_create_campaign(
  name,
  connectInstanceId,
  dialerConfig,
  outboundCallConfig,
  tags = NULL
)
}
\arguments{
\item{name}{[required]}

\item{connectInstanceId}{[required]}

\item{dialerConfig}{[required]}

\item{outboundCallConfig}{[required]}

\item{tags}{}
}
\description{
Creates a campaign for the specified Amazon Connect account. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_create_campaign/} for full documentation.
}
\keyword{internal}
