% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_delete_campaign}
\alias{connectcampaignservice_delete_campaign}
\title{Deletes a campaign from the specified Amazon Connect account}
\usage{
connectcampaignservice_delete_campaign(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Deletes a campaign from the specified Amazon Connect account.

See \url{https://paws-r.github.io/docs/connectcampaignservice/delete_campaign.html} for full documentation.
}
\keyword{internal}
