% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_delete_connect_instance_config}
\alias{connectcampaignservice_delete_connect_instance_config}
\title{Deletes a connect instance config from the specified AWS account}
\usage{
connectcampaignservice_delete_connect_instance_config(connectInstanceId)
}
\arguments{
\item{connectInstanceId}{[required]}
}
\description{
Deletes a connect instance config from the specified AWS account.

See \url{https://paws-r.github.io/docs/connectcampaignservice/delete_connect_instance_config.html} for full documentation.
}
\keyword{internal}
