% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_describe_campaign}
\alias{connectcampaignservice_describe_campaign}
\title{Describes the specific campaign}
\usage{
connectcampaignservice_describe_campaign(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Describes the specific campaign.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_describe_campaign/} for full documentation.
}
\keyword{internal}
