% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_get_instance_onboarding_job_status}
\alias{connectcampaignservice_get_instance_onboarding_job_status}
\title{Get the specific instance onboarding job status}
\usage{
connectcampaignservice_get_instance_onboarding_job_status(connectInstanceId)
}
\arguments{
\item{connectInstanceId}{[required]}
}
\description{
Get the specific instance onboarding job status.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_get_instance_onboarding_job_status/} for full documentation.
}
\keyword{internal}
