% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_list_campaigns}
\alias{connectcampaignservice_list_campaigns}
\title{Provides summary information about the campaigns under the specified
Amazon Connect account}
\usage{
connectcampaignservice_list_campaigns(
  filters = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filters}{}

\item{maxResults}{}

\item{nextToken}{}
}
\description{
Provides summary information about the campaigns under the specified Amazon Connect account.

See \url{https://paws-r.github.io/docs/connectcampaignservice/list_campaigns.html} for full documentation.
}
\keyword{internal}
