% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_pause_campaign}
\alias{connectcampaignservice_pause_campaign}
\title{Pauses a campaign for the specified Amazon Connect account}
\usage{
connectcampaignservice_pause_campaign(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Pauses a campaign for the specified Amazon Connect account.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_pause_campaign/} for full documentation.
}
\keyword{internal}
