% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_put_dial_request_batch}
\alias{connectcampaignservice_put_dial_request_batch}
\title{Creates dials requests for the specified campaign Amazon Connect account}
\usage{
connectcampaignservice_put_dial_request_batch(dialRequests, id)
}
\arguments{
\item{dialRequests}{[required]}

\item{id}{[required]}
}
\description{
Creates dials requests for the specified campaign Amazon Connect account. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_put_dial_request_batch/} for full documentation.
}
\keyword{internal}
