% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_resume_campaign}
\alias{connectcampaignservice_resume_campaign}
\title{Stops a campaign for the specified Amazon Connect account}
\usage{
connectcampaignservice_resume_campaign(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Stops a campaign for the specified Amazon Connect account.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_resume_campaign/} for full documentation.
}
\keyword{internal}
