% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_start_campaign}
\alias{connectcampaignservice_start_campaign}
\title{Starts a campaign for the specified Amazon Connect account}
\usage{
connectcampaignservice_start_campaign(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Starts a campaign for the specified Amazon Connect account.

See \url{https://paws-r.github.io/docs/connectcampaignservice/start_campaign.html} for full documentation.
}
\keyword{internal}
