% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_stop_campaign}
\alias{connectcampaignservice_stop_campaign}
\title{Stops a campaign for the specified Amazon Connect account}
\usage{
connectcampaignservice_stop_campaign(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Stops a campaign for the specified Amazon Connect account.

See \url{https://paws-r.github.io/docs/connectcampaignservice/stop_campaign.html} for full documentation.
}
\keyword{internal}
