% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_tag_resource}
\alias{connectcampaignservice_tag_resource}
\title{Tag a resource}
\usage{
connectcampaignservice_tag_resource(arn, tags)
}
\arguments{
\item{arn}{[required] Arn}

\item{tags}{[required] Tag map with key and value.}
}
\description{
Tag a resource.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_tag_resource/} for full documentation.
}
\keyword{internal}
