% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_untag_resource}
\alias{connectcampaignservice_untag_resource}
\title{Untag a resource}
\usage{
connectcampaignservice_untag_resource(arn, tagKeys)
}
\arguments{
\item{arn}{[required] Arn}

\item{tagKeys}{[required] List of tag keys.}
}
\description{
Untag a resource.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_untag_resource/} for full documentation.
}
\keyword{internal}
