% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_update_campaign_name}
\alias{connectcampaignservice_update_campaign_name}
\title{Updates the name of a campaign}
\usage{
connectcampaignservice_update_campaign_name(id, name)
}
\arguments{
\item{id}{[required]}

\item{name}{[required]}
}
\description{
Updates the name of a campaign. This API is idempotent.

See \url{https://paws-r.github.io/docs/connectcampaignservice/update_campaign_name.html} for full documentation.
}
\keyword{internal}
