% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_update_campaign_outbound_call_config}
\alias{connectcampaignservice_update_campaign_outbound_call_config}
\title{Updates the outbound call config of a campaign}
\usage{
connectcampaignservice_update_campaign_outbound_call_config(
  id,
  connectContactFlowId = NULL,
  connectSourcePhoneNumber = NULL,
  answerMachineDetectionConfig = NULL
)
}
\arguments{
\item{id}{[required] Identifier representing a Campaign}

\item{connectContactFlowId}{The identifier of the contact flow for the outbound call.}

\item{connectSourcePhoneNumber}{The phone number associated with the Amazon Connect instance, in E.164 format. If you do not specify a source phone number, you must specify a queue.}

\item{answerMachineDetectionConfig}{Answering Machine Detection config}
}
\description{
Updates the outbound call config of a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_update_campaign_outbound_call_config/} for full documentation.
}
\keyword{internal}
