% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_create_campaign}
\alias{connectcampaignservicev2_create_campaign}
\title{Creates a campaign for the specified Amazon Connect account}
\usage{
connectcampaignservicev2_create_campaign(
  name,
  connectInstanceId,
  channelSubtypeConfig = NULL,
  type = NULL,
  source = NULL,
  connectCampaignFlowArn = NULL,
  schedule = NULL,
  entryLimitsConfig = NULL,
  communicationTimeConfig = NULL,
  communicationLimitsOverride = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of an Amazon Connect Campaign name.}

\item{connectInstanceId}{[required] Amazon Connect Instance Id}

\item{channelSubtypeConfig}{Campaign Channel Subtype config}

\item{type}{The type of campaign externally exposed in APIs.}

\item{source}{Source of the campaign}

\item{connectCampaignFlowArn}{Amazon Resource Names(ARN)}

\item{schedule}{Campaign schedule}

\item{entryLimitsConfig}{Campaign entry limits config}

\item{communicationTimeConfig}{Campaign communication time config}

\item{communicationLimitsOverride}{Communication limits config}

\item{tags}{Tag map with key and value.}
}
\description{
Creates a campaign for the specified Amazon Connect account. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_create_campaign/} for full documentation.
}
\keyword{internal}
