% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_delete_campaign_channel_subtype_config}
\alias{connectcampaignservicev2_delete_campaign_channel_subtype_config}
\title{Deletes the channel subtype config of a campaign}
\usage{
connectcampaignservicev2_delete_campaign_channel_subtype_config(
  id,
  channelSubtype
)
}
\arguments{
\item{id}{[required]}

\item{channelSubtype}{[required]}
}
\description{
Deletes the channel subtype config of a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_delete_campaign_channel_subtype_config/} for full documentation.
}
\keyword{internal}
