% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_delete_campaign_communication_time}
\alias{connectcampaignservicev2_delete_campaign_communication_time}
\title{Deletes the communication time config for a campaign}
\usage{
connectcampaignservicev2_delete_campaign_communication_time(id, config)
}
\arguments{
\item{id}{[required]}

\item{config}{[required]}
}
\description{
Deletes the communication time config for a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_delete_campaign_communication_time/} for full documentation.
}
\keyword{internal}
