% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_delete_campaign_entry_limits}
\alias{connectcampaignservicev2_delete_campaign_entry_limits}
\title{Deletes the entry limits config for a campaign}
\usage{
connectcampaignservicev2_delete_campaign_entry_limits(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Deletes the entry limits config for a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_delete_campaign_entry_limits/} for full documentation.
}
\keyword{internal}
