% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_delete_connect_instance_config}
\alias{connectcampaignservicev2_delete_connect_instance_config}
\title{Deletes a connect instance config from the specified AWS account}
\usage{
connectcampaignservicev2_delete_connect_instance_config(
  connectInstanceId,
  campaignDeletionPolicy = NULL
)
}
\arguments{
\item{connectInstanceId}{[required] Amazon Connect Instance Id}

\item{campaignDeletionPolicy}{Enumeration of the policies to enact on existing campaigns during instance config deletion}
}
\description{
Deletes a connect instance config from the specified AWS account.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_delete_connect_instance_config/} for full documentation.
}
\keyword{internal}
