% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_delete_connect_instance_integration}
\alias{connectcampaignservicev2_delete_connect_instance_integration}
\title{Delete the integration for the specified Amazon Connect instance}
\usage{
connectcampaignservicev2_delete_connect_instance_integration(
  connectInstanceId,
  integrationIdentifier
)
}
\arguments{
\item{connectInstanceId}{[required]}

\item{integrationIdentifier}{[required]}
}
\description{
Delete the integration for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_delete_connect_instance_integration/} for full documentation.
}
\keyword{internal}
