% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_describe_campaign}
\alias{connectcampaignservicev2_describe_campaign}
\title{Describes the specific campaign}
\usage{
connectcampaignservicev2_describe_campaign(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Describes the specific campaign.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_describe_campaign/} for full documentation.
}
\keyword{internal}
