% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_get_connect_instance_config}
\alias{connectcampaignservicev2_get_connect_instance_config}
\title{Get the specific Connect instance config}
\usage{
connectcampaignservicev2_get_connect_instance_config(connectInstanceId)
}
\arguments{
\item{connectInstanceId}{[required]}
}
\description{
Get the specific Connect instance config.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_get_connect_instance_config/} for full documentation.
}
\keyword{internal}
