% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_get_instance_communication_limits}
\alias{connectcampaignservicev2_get_instance_communication_limits}
\title{Get the instance communication limits}
\usage{
connectcampaignservicev2_get_instance_communication_limits(connectInstanceId)
}
\arguments{
\item{connectInstanceId}{[required]}
}
\description{
Get the instance communication limits.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_get_instance_communication_limits/} for full documentation.
}
\keyword{internal}
