% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_list_campaigns}
\alias{connectcampaignservicev2_list_campaigns}
\title{Provides summary information about the campaigns under the specified
Amazon Connect account}
\usage{
connectcampaignservicev2_list_campaigns(
  maxResults = NULL,
  nextToken = NULL,
  filters = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results.}

\item{filters}{Filter model by type}
}
\description{
Provides summary information about the campaigns under the specified Amazon Connect account.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_list_campaigns/} for full documentation.
}
\keyword{internal}
