% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_list_connect_instance_integrations}
\alias{connectcampaignservicev2_list_connect_instance_integrations}
\title{Provides summary information about the integration under the specified
Connect instance}
\usage{
connectcampaignservicev2_list_connect_instance_integrations(
  connectInstanceId,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{connectInstanceId}{[required]}

\item{maxResults}{}

\item{nextToken}{}
}
\description{
Provides summary information about the integration under the specified Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_list_connect_instance_integrations/} for full documentation.
}
\keyword{internal}
