% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_list_tags_for_resource}
\alias{connectcampaignservicev2_list_tags_for_resource}
\title{List tags for a resource}
\usage{
connectcampaignservicev2_list_tags_for_resource(arn)
}
\arguments{
\item{arn}{[required]}
}
\description{
List tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
