% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_put_connect_instance_integration}
\alias{connectcampaignservicev2_put_connect_instance_integration}
\title{Put or update the integration for the specified Amazon Connect instance}
\usage{
connectcampaignservicev2_put_connect_instance_integration(
  connectInstanceId,
  integrationConfig
)
}
\arguments{
\item{connectInstanceId}{[required]}

\item{integrationConfig}{[required]}
}
\description{
Put or update the integration for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_put_connect_instance_integration/} for full documentation.
}
\keyword{internal}
