% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_put_outbound_request_batch}
\alias{connectcampaignservicev2_put_outbound_request_batch}
\title{Creates outbound requests for the specified campaign Amazon Connect
account}
\usage{
connectcampaignservicev2_put_outbound_request_batch(id, outboundRequests)
}
\arguments{
\item{id}{[required]}

\item{outboundRequests}{[required]}
}
\description{
Creates outbound requests for the specified campaign Amazon Connect account. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_put_outbound_request_batch/} for full documentation.
}
\keyword{internal}
