% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_put_profile_outbound_request_batch}
\alias{connectcampaignservicev2_put_profile_outbound_request_batch}
\title{Takes in a list of profile outbound requests to be placed as part of an
outbound campaign}
\usage{
connectcampaignservicev2_put_profile_outbound_request_batch(
  id,
  profileOutboundRequests
)
}
\arguments{
\item{id}{[required]}

\item{profileOutboundRequests}{[required]}
}
\description{
Takes in a list of profile outbound requests to be placed as part of an outbound campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_put_profile_outbound_request_batch/} for full documentation.
}
\keyword{internal}
