% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_start_campaign}
\alias{connectcampaignservicev2_start_campaign}
\title{Starts a campaign for the specified Amazon Connect account}
\usage{
connectcampaignservicev2_start_campaign(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Starts a campaign for the specified Amazon Connect account.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_start_campaign/} for full documentation.
}
\keyword{internal}
