% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_start_instance_onboarding_job}
\alias{connectcampaignservicev2_start_instance_onboarding_job}
\title{Onboard the specific Amazon Connect instance to Connect Campaigns}
\usage{
connectcampaignservicev2_start_instance_onboarding_job(
  connectInstanceId,
  encryptionConfig
)
}
\arguments{
\item{connectInstanceId}{[required] Amazon Connect Instance Id}

\item{encryptionConfig}{[required] Encryption config for Connect Instance. Note that sensitive data will always be encrypted. If disabled, service will perform encryption with its own key. If enabled, a KMS key id needs to be provided and KMS charges will apply. KMS is only type supported}
}
\description{
Onboard the specific Amazon Connect instance to Connect Campaigns.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_start_instance_onboarding_job/} for full documentation.
}
\keyword{internal}
