% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_stop_campaign}
\alias{connectcampaignservicev2_stop_campaign}
\title{Stops a campaign for the specified Amazon Connect account}
\usage{
connectcampaignservicev2_stop_campaign(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Stops a campaign for the specified Amazon Connect account.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_stop_campaign/} for full documentation.
}
\keyword{internal}
