% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_tag_resource}
\alias{connectcampaignservicev2_tag_resource}
\title{Tag a resource}
\usage{
connectcampaignservicev2_tag_resource(arn, tags)
}
\arguments{
\item{arn}{[required]}

\item{tags}{[required]}
}
\description{
Tag a resource.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_tag_resource/} for full documentation.
}
\keyword{internal}
