% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_untag_resource}
\alias{connectcampaignservicev2_untag_resource}
\title{Untag a resource}
\usage{
connectcampaignservicev2_untag_resource(arn, tagKeys)
}
\arguments{
\item{arn}{[required]}

\item{tagKeys}{[required]}
}
\description{
Untag a resource.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_untag_resource/} for full documentation.
}
\keyword{internal}
