% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_update_campaign_channel_subtype_config}
\alias{connectcampaignservicev2_update_campaign_channel_subtype_config}
\title{Updates the channel subtype config of a campaign}
\usage{
connectcampaignservicev2_update_campaign_channel_subtype_config(
  id,
  channelSubtypeConfig
)
}
\arguments{
\item{id}{[required]}

\item{channelSubtypeConfig}{[required]}
}
\description{
Updates the channel subtype config of a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_update_campaign_channel_subtype_config/} for full documentation.
}
\keyword{internal}
