% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_update_campaign_communication_limits}
\alias{connectcampaignservicev2_update_campaign_communication_limits}
\title{Updates the communication limits config for a campaign}
\usage{
connectcampaignservicev2_update_campaign_communication_limits(
  id,
  communicationLimitsOverride
)
}
\arguments{
\item{id}{[required]}

\item{communicationLimitsOverride}{[required]}
}
\description{
Updates the communication limits config for a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_update_campaign_communication_limits/} for full documentation.
}
\keyword{internal}
