% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_update_campaign_communication_time}
\alias{connectcampaignservicev2_update_campaign_communication_time}
\title{Updates the communication time config for a campaign}
\usage{
connectcampaignservicev2_update_campaign_communication_time(
  id,
  communicationTimeConfig
)
}
\arguments{
\item{id}{[required]}

\item{communicationTimeConfig}{[required]}
}
\description{
Updates the communication time config for a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_update_campaign_communication_time/} for full documentation.
}
\keyword{internal}
