% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_update_campaign_entry_limits}
\alias{connectcampaignservicev2_update_campaign_entry_limits}
\title{Updates the entry limits config for a campaign}
\usage{
connectcampaignservicev2_update_campaign_entry_limits(id, entryLimitsConfig)
}
\arguments{
\item{id}{[required]}

\item{entryLimitsConfig}{[required]}
}
\description{
Updates the entry limits config for a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_update_campaign_entry_limits/} for full documentation.
}
\keyword{internal}
