% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_update_campaign_flow_association}
\alias{connectcampaignservicev2_update_campaign_flow_association}
\title{Updates the campaign flow associated with a campaign}
\usage{
connectcampaignservicev2_update_campaign_flow_association(
  id,
  connectCampaignFlowArn
)
}
\arguments{
\item{id}{[required]}

\item{connectCampaignFlowArn}{[required]}
}
\description{
Updates the campaign flow associated with a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_update_campaign_flow_association/} for full documentation.
}
\keyword{internal}
