% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_update_campaign_name}
\alias{connectcampaignservicev2_update_campaign_name}
\title{Updates the name of a campaign}
\usage{
connectcampaignservicev2_update_campaign_name(id, name)
}
\arguments{
\item{id}{[required]}

\item{name}{[required]}
}
\description{
Updates the name of a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_update_campaign_name/} for full documentation.
}
\keyword{internal}
