% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_update_campaign_schedule}
\alias{connectcampaignservicev2_update_campaign_schedule}
\title{Updates the schedule for a campaign}
\usage{
connectcampaignservicev2_update_campaign_schedule(id, schedule)
}
\arguments{
\item{id}{[required]}

\item{schedule}{[required]}
}
\description{
Updates the schedule for a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_update_campaign_schedule/} for full documentation.
}
\keyword{internal}
