% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_update_campaign_source}
\alias{connectcampaignservicev2_update_campaign_source}
\title{Updates the campaign source with a campaign}
\usage{
connectcampaignservicev2_update_campaign_source(id, source)
}
\arguments{
\item{id}{[required]}

\item{source}{[required]}
}
\description{
Updates the campaign source with a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_update_campaign_source/} for full documentation.
}
\keyword{internal}
