% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_batch_get_field}
\alias{connectcases_batch_get_field}
\title{Returns the description for the list of fields in the request parameters}
\usage{
connectcases_batch_get_field(domainId, fields)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{fields}{[required] A list of unique field identifiers.}
}
\description{
Returns the description for the list of fields in the request parameters.

See \url{https://www.paws-r-sdk.com/docs/connectcases_batch_get_field/} for full documentation.
}
\keyword{internal}
