% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_batch_put_field_options}
\alias{connectcases_batch_put_field_options}
\title{Creates and updates a set of field options for a single select field in
a Cases domain}
\usage{
connectcases_batch_put_field_options(domainId, fieldId, options)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{fieldId}{[required] The unique identifier of a field.}

\item{options}{[required] A list of \code{FieldOption} objects.}
}
\description{
Creates and updates a set of field options for a single select field in a Cases domain.

See \url{https://www.paws-r-sdk.com/docs/connectcases_batch_put_field_options/} for full documentation.
}
\keyword{internal}
