% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_create_case_rule}
\alias{connectcases_create_case_rule}
\title{Creates a new case rule}
\usage{
connectcases_create_case_rule(domainId, name, description = NULL, rule)
}
\arguments{
\item{domainId}{[required] Unique identifier of a Cases domain.}

\item{name}{[required] Name of the case rule.}

\item{description}{The description of a case rule.}

\item{rule}{[required] Represents what rule type should take place, under what conditions.}
}
\description{
Creates a new case rule. In the Amazon Connect admin website, case rules are known as \emph{case field conditions}. For more information about case field conditions, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/case-field-conditions.html}{Add case field conditions to a case template}.

See \url{https://www.paws-r-sdk.com/docs/connectcases_create_case_rule/} for full documentation.
}
\keyword{internal}
