% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_create_domain}
\alias{connectcases_create_domain}
\title{Creates a domain, which is a container for all case data, such as cases,
fields, templates and layouts}
\usage{
connectcases_create_domain(name)
}
\arguments{
\item{name}{[required] The name for your Cases domain. It must be unique for your Amazon Web
Services account.}
}
\description{
Creates a domain, which is a container for all case data, such as cases, fields, templates and layouts. Each Amazon Connect instance can be associated with only one Cases domain.

See \url{https://www.paws-r-sdk.com/docs/connectcases_create_domain/} for full documentation.
}
\keyword{internal}
