% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_create_field}
\alias{connectcases_create_field}
\title{Creates a field in the Cases domain}
\usage{
connectcases_create_field(domainId, name, type, description = NULL)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{name}{[required] The name of the field.}

\item{type}{[required] Defines the data type, some system constraints, and default display of
the field.}

\item{description}{The description of the field.}
}
\description{
Creates a field in the Cases domain. This field is used to define the case object model (that is, defines what data can be captured on cases) in a Cases domain.

See \url{https://www.paws-r-sdk.com/docs/connectcases_create_field/} for full documentation.
}
\keyword{internal}
