% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_create_layout}
\alias{connectcases_create_layout}
\title{Creates a layout in the Cases domain}
\usage{
connectcases_create_layout(domainId, name, content)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{name}{[required] The name of the layout. It must be unique for the Cases domain.}

\item{content}{[required] Information about which fields will be present in the layout, and
information about the order of the fields.}
}
\description{
Creates a layout in the Cases domain. Layouts define the following configuration in the top section and More Info tab of the Cases user interface:

See \url{https://www.paws-r-sdk.com/docs/connectcases_create_layout/} for full documentation.
}
\keyword{internal}
