% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_delete_case}
\alias{connectcases_delete_case}
\title{The DeleteCase API permanently deletes a case and all its associated
resources from the cases data store}
\usage{
connectcases_delete_case(domainId, caseId)
}
\arguments{
\item{domainId}{[required] A unique identifier of the Cases domain.}

\item{caseId}{[required] A unique identifier of the case.}
}
\description{
The DeleteCase API permanently deletes a case and all its associated resources from the cases data store. After a successful deletion, you cannot:

See \url{https://www.paws-r-sdk.com/docs/connectcases_delete_case/} for full documentation.
}
\keyword{internal}
