% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_delete_case_rule}
\alias{connectcases_delete_case_rule}
\title{Deletes a case rule}
\usage{
connectcases_delete_case_rule(domainId, caseRuleId)
}
\arguments{
\item{domainId}{[required] Unique identifier of a Cases domain.}

\item{caseRuleId}{[required] Unique identifier of a case rule.}
}
\description{
Deletes a case rule. In the Amazon Connect admin website, case rules are known as \emph{case field conditions}. For more information about case field conditions, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/case-field-conditions.html}{Add case field conditions to a case template}.

See \url{https://www.paws-r-sdk.com/docs/connectcases_delete_case_rule/} for full documentation.
}
\keyword{internal}
