% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_delete_domain}
\alias{connectcases_delete_domain}
\title{Deletes a Cases domain}
\usage{
connectcases_delete_domain(domainId)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}
}
\description{
Deletes a Cases domain.

See \url{https://www.paws-r-sdk.com/docs/connectcases_delete_domain/} for full documentation.
}
\keyword{internal}
