% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_delete_field}
\alias{connectcases_delete_field}
\title{Deletes a field from a cases template}
\usage{
connectcases_delete_field(domainId, fieldId)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{fieldId}{[required] Unique identifier of the field.}
}
\description{
Deletes a field from a cases template. You can delete up to 100 fields per domain.

See \url{https://www.paws-r-sdk.com/docs/connectcases_delete_field/} for full documentation.
}
\keyword{internal}
