% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_delete_layout}
\alias{connectcases_delete_layout}
\title{Deletes a layout from a cases template}
\usage{
connectcases_delete_layout(domainId, layoutId)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{layoutId}{[required] The unique identifier of the layout.}
}
\description{
Deletes a layout from a cases template. You can delete up to 100 layouts per domain.

See \url{https://www.paws-r-sdk.com/docs/connectcases_delete_layout/} for full documentation.
}
\keyword{internal}
