% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_delete_related_item}
\alias{connectcases_delete_related_item}
\title{Deletes the related item resource under a case}
\usage{
connectcases_delete_related_item(domainId, caseId, relatedItemId)
}
\arguments{
\item{domainId}{[required] A unique identifier of the Cases domain.}

\item{caseId}{[required] A unique identifier of the case.}

\item{relatedItemId}{[required] A unique identifier of a related item.}
}
\description{
Deletes the related item resource under a case.

See \url{https://www.paws-r-sdk.com/docs/connectcases_delete_related_item/} for full documentation.
}
\keyword{internal}
