% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_delete_template}
\alias{connectcases_delete_template}
\title{Deletes a cases template}
\usage{
connectcases_delete_template(domainId, templateId)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{templateId}{[required] A unique identifier of a template.}
}
\description{
Deletes a cases template. You can delete up to 100 templates per domain.

See \url{https://www.paws-r-sdk.com/docs/connectcases_delete_template/} for full documentation.
}
\keyword{internal}
