% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_get_case}
\alias{connectcases_get_case}
\title{Returns information about a specific case if it exists}
\usage{
connectcases_get_case(caseId, domainId, fields, nextToken = NULL)
}
\arguments{
\item{caseId}{[required] A unique identifier of the case.}

\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{fields}{[required] A list of unique field identifiers.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Returns information about a specific case if it exists.

See \url{https://www.paws-r-sdk.com/docs/connectcases_get_case/} for full documentation.
}
\keyword{internal}
