% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_get_case_audit_events}
\alias{connectcases_get_case_audit_events}
\title{Returns the audit history about a specific case if it exists}
\usage{
connectcases_get_case_audit_events(
  caseId,
  domainId,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{caseId}{[required] A unique identifier of the case.}

\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{maxResults}{The maximum number of audit events to return. The current maximum
supported value is 25. This is also the default when no other value is
provided.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Returns the audit history about a specific case if it exists.

See \url{https://www.paws-r-sdk.com/docs/connectcases_get_case_audit_events/} for full documentation.
}
\keyword{internal}
