% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_get_case_event_configuration}
\alias{connectcases_get_case_event_configuration}
\title{Returns the case event publishing configuration}
\usage{
connectcases_get_case_event_configuration(domainId)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}
}
\description{
Returns the case event publishing configuration.

See \url{https://www.paws-r-sdk.com/docs/connectcases_get_case_event_configuration/} for full documentation.
}
\keyword{internal}
