% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_get_domain}
\alias{connectcases_get_domain}
\title{Returns information about a specific domain if it exists}
\usage{
connectcases_get_domain(domainId)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}
}
\description{
Returns information about a specific domain if it exists.

See \url{https://www.paws-r-sdk.com/docs/connectcases_get_domain/} for full documentation.
}
\keyword{internal}
