% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_get_layout}
\alias{connectcases_get_layout}
\title{Returns the details for the requested layout}
\usage{
connectcases_get_layout(domainId, layoutId)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{layoutId}{[required] The unique identifier of the layout.}
}
\description{
Returns the details for the requested layout.

See \url{https://www.paws-r-sdk.com/docs/connectcases_get_layout/} for full documentation.
}
\keyword{internal}
