% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_get_template}
\alias{connectcases_get_template}
\title{Returns the details for the requested template}
\usage{
connectcases_get_template(domainId, templateId)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{templateId}{[required] A unique identifier of a template.}
}
\description{
Returns the details for the requested template.

See \url{https://www.paws-r-sdk.com/docs/connectcases_get_template/} for full documentation.
}
\keyword{internal}
