% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_list_case_rules}
\alias{connectcases_list_case_rules}
\title{Lists all case rules in a Cases domain}
\usage{
connectcases_list_case_rules(domainId, maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{domainId}{[required] Unique identifier of a Cases domain.}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Lists all case rules in a Cases domain. In the Amazon Connect admin website, case rules are known as \emph{case field conditions}. For more information about case field conditions, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/case-field-conditions.html}{Add case field conditions to a case template}.

See \url{https://www.paws-r-sdk.com/docs/connectcases_list_case_rules/} for full documentation.
}
\keyword{internal}
