% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_list_domains}
\alias{connectcases_list_domains}
\title{Lists all cases domains in the Amazon Web Services account}
\usage{
connectcases_list_domains(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Lists all cases domains in the Amazon Web Services account. Each list item is a condensed summary object of the domain.

See \url{https://www.paws-r-sdk.com/docs/connectcases_list_domains/} for full documentation.
}
\keyword{internal}
