% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_list_field_options}
\alias{connectcases_list_field_options}
\title{Lists all of the field options for a field identifier in the domain}
\usage{
connectcases_list_field_options(
  domainId,
  fieldId,
  maxResults = NULL,
  nextToken = NULL,
  values = NULL
)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{fieldId}{[required] The unique identifier of a field.}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{values}{A list of \code{FieldOption} values to filter on for
\code{\link[=connectcases_list_field_options]{list_field_options}}.}
}
\description{
Lists all of the field options for a field identifier in the domain.

See \url{https://www.paws-r-sdk.com/docs/connectcases_list_field_options/} for full documentation.
}
\keyword{internal}
