% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_list_tags_for_resource}
\alias{connectcases_list_tags_for_resource}
\title{Lists tags for a resource}
\usage{
connectcases_list_tags_for_resource(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN)}
}
\description{
Lists tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/connectcases_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
