% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_put_case_event_configuration}
\alias{connectcases_put_case_event_configuration}
\title{Adds case event publishing configuration}
\usage{
connectcases_put_case_event_configuration(domainId, eventBridge)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{eventBridge}{[required] Configuration to enable EventBridge case event delivery and determine
what data is delivered.}
}
\description{
Adds case event publishing configuration. For a complete list of fields you can add to the event message, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/case-fields.html}{Create case fields} in the \emph{Amazon Connect Administrator Guide}

See \url{https://www.paws-r-sdk.com/docs/connectcases_put_case_event_configuration/} for full documentation.
}
\keyword{internal}
