% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_tag_resource}
\alias{connectcases_tag_resource}
\title{Adds tags to a resource}
\usage{
connectcases_tag_resource(arn, tags)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN)}

\item{tags}{[required] A map of of key-value pairs that represent tags on a resource. Tags are
used to organize, track, or control access for this resource.}
}
\description{
Adds tags to a resource.

See \url{https://www.paws-r-sdk.com/docs/connectcases_tag_resource/} for full documentation.
}
\keyword{internal}
