% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_untag_resource}
\alias{connectcases_untag_resource}
\title{Untags a resource}
\usage{
connectcases_untag_resource(arn, tagKeys)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN)}

\item{tagKeys}{[required] List of tag keys.}
}
\description{
Untags a resource.

See \url{https://www.paws-r-sdk.com/docs/connectcases_untag_resource/} for full documentation.
}
\keyword{internal}
