% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_update_case}
\alias{connectcases_update_case}
\title{Updates the values of fields on a case}
\usage{
connectcases_update_case(caseId, domainId, fields)
}
\arguments{
\item{caseId}{[required] A unique identifier of the case.}

\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{fields}{[required] An array of objects with \code{fieldId} (matching ListFields/DescribeField)
and value union data, structured identical to
\code{\link[=connectcases_create_case]{create_case}}.}
}
\description{
Updates the values of fields on a case. Fields to be updated are received as an array of id/value pairs identical to the \code{\link[=connectcases_create_case]{create_case}} input .

See \url{https://www.paws-r-sdk.com/docs/connectcases_update_case/} for full documentation.
}
\keyword{internal}
