% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_update_case_rule}
\alias{connectcases_update_case_rule}
\title{Updates a case rule}
\usage{
connectcases_update_case_rule(
  domainId,
  caseRuleId,
  name = NULL,
  description = NULL,
  rule = NULL
)
}
\arguments{
\item{domainId}{[required] Unique identifier of a Cases domain.}

\item{caseRuleId}{[required] Unique identifier of a case rule.}

\item{name}{Name of the case rule.}

\item{description}{Description of a case rule.}

\item{rule}{Represents what rule type should take place, under what conditions.}
}
\description{
Updates a case rule. In the Amazon Connect admin website, case rules are known as \emph{case field conditions}. For more information about case field conditions, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/case-field-conditions.html}{Add case field conditions to a case template}.

See \url{https://www.paws-r-sdk.com/docs/connectcases_update_case_rule/} for full documentation.
}
\keyword{internal}
