% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_update_field}
\alias{connectcases_update_field}
\title{Updates the properties of an existing field}
\usage{
connectcases_update_field(description = NULL, domainId, fieldId, name = NULL)
}
\arguments{
\item{description}{The description of a field.}

\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{fieldId}{[required] The unique identifier of a field.}

\item{name}{The name of the field.}
}
\description{
Updates the properties of an existing field.

See \url{https://www.paws-r-sdk.com/docs/connectcases_update_field/} for full documentation.
}
\keyword{internal}
