% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_update_related_item}
\alias{connectcases_update_related_item}
\title{Updates the content of a related item associated with a case}
\usage{
connectcases_update_related_item(
  domainId,
  caseId,
  relatedItemId,
  content,
  performedBy = NULL
)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{caseId}{[required] A unique identifier of the case.}

\item{relatedItemId}{[required] Unique identifier of a related item.}

\item{content}{[required] The content of a related item to be updated.}

\item{performedBy}{Represents the user who performed the update of the related item.}
}
\description{
Updates the content of a related item associated with a case. The following related item types are supported:

See \url{https://www.paws-r-sdk.com/docs/connectcases_update_related_item/} for full documentation.
}
\keyword{internal}
