% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_cancel_participant_authentication}
\alias{connectparticipant_cancel_participant_authentication}
\title{Cancels the authentication session}
\usage{
connectparticipant_cancel_participant_authentication(
  SessionId,
  ConnectionToken
)
}
\arguments{
\item{SessionId}{[required] The \code{sessionId} provided in the \code{authenticationInitiated} event.}

\item{ConnectionToken}{[required] The authentication token associated with the participant's connection.}
}
\description{
Cancels the authentication session. The opted out branch of the Authenticate Customer flow block will be taken.

See \url{https://www.paws-r-sdk.com/docs/connectparticipant_cancel_participant_authentication/} for full documentation.
}
\keyword{internal}
