% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_complete_attachment_upload}
\alias{connectparticipant_complete_attachment_upload}
\title{Allows you to confirm that the attachment has been uploaded using the
pre-signed URL provided in StartAttachmentUpload API}
\usage{
connectparticipant_complete_attachment_upload(
  AttachmentIds,
  ClientToken,
  ConnectionToken
)
}
\arguments{
\item{AttachmentIds}{[required] A list of unique identifiers for the attachments.}

\item{ClientToken}{[required] A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{ConnectionToken}{[required] The authentication token associated with the participant's connection.}
}
\description{
Allows you to confirm that the attachment has been uploaded using the pre-signed URL provided in StartAttachmentUpload API.

See \url{https://paws-r.github.io/docs/connectparticipant/complete_attachment_upload.html} for full documentation.
}
\keyword{internal}
