% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_create_participant_connection}
\alias{connectparticipant_create_participant_connection}
\title{Creates the participant's connection}
\usage{
connectparticipant_create_participant_connection(
  Type = NULL,
  ParticipantToken,
  ConnectParticipant = NULL
)
}
\arguments{
\item{Type}{Type of connection information required. If you need
\code{CONNECTION_CREDENTIALS} along with marking participant as connected,
pass \code{CONNECTION_CREDENTIALS} in \code{Type}.}

\item{ParticipantToken}{[required] This is a header parameter.

The ParticipantToken as obtained from
\href{https://docs.aws.amazon.com/connect/latest/APIReference/API_StartChatContact.html}{StartChatContact}
API response.}

\item{ConnectParticipant}{Amazon Connect Participant is used to mark the participant as connected
for customer participant in message streaming, as well as for agent or
manager participant in non-streaming chats.}
}
\description{
Creates the participant's connection.

See \url{https://www.paws-r-sdk.com/docs/connectparticipant_create_participant_connection/} for full documentation.
}
\keyword{internal}
