% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_describe_view}
\alias{connectparticipant_describe_view}
\title{Retrieves the view for the specified view token}
\usage{
connectparticipant_describe_view(ViewToken, ConnectionToken)
}
\arguments{
\item{ViewToken}{[required] An encrypted token originating from the interactive message of a
ShowView block operation. Represents the desired view.}

\item{ConnectionToken}{[required] The connection token.}
}
\description{
Retrieves the view for the specified view token.

See \url{https://www.paws-r-sdk.com/docs/connectparticipant_describe_view/} for full documentation.
}
\keyword{internal}
