% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_disconnect_participant}
\alias{connectparticipant_disconnect_participant}
\title{Disconnects a participant}
\usage{
connectparticipant_disconnect_participant(ClientToken = NULL, ConnectionToken)
}
\arguments{
\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{ConnectionToken}{[required] The authentication token associated with the participant's connection.}
}
\description{
Disconnects a participant. Note that ConnectionToken is used for invoking this API instead of ParticipantToken.

See \url{https://paws-r.github.io/docs/connectparticipant/disconnect_participant.html} for full documentation.
}
\keyword{internal}
