% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_get_authentication_url}
\alias{connectparticipant_get_authentication_url}
\title{Retrieves the AuthenticationUrl for the current authentication session
for the AuthenticateCustomer flow block}
\usage{
connectparticipant_get_authentication_url(
  SessionId,
  RedirectUri,
  ConnectionToken
)
}
\arguments{
\item{SessionId}{[required] The sessionId provided in the authenticationInitiated event.}

\item{RedirectUri}{[required] The URL where the customer will be redirected after Amazon Cognito
authorizes the user.}

\item{ConnectionToken}{[required] The authentication token associated with the participant's connection.}
}
\description{
Retrieves the AuthenticationUrl for the current authentication session for the AuthenticateCustomer flow block.

See \url{https://www.paws-r-sdk.com/docs/connectparticipant_get_authentication_url/} for full documentation.
}
\keyword{internal}
