% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_send_event}
\alias{connectparticipant_send_event}
\title{Sends an event}
\usage{
connectparticipant_send_event(
  ContentType,
  Content = NULL,
  ClientToken = NULL,
  ConnectionToken
)
}
\arguments{
\item{ContentType}{[required] The content type of the request. Supported types are:
\itemize{
\item application/vnd.amazonaws.connect.event.typing
\item application/vnd.amazonaws.connect.event.connection.acknowledged
}}

\item{Content}{The content of the event to be sent (for example, message text). This is
not yet supported.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{ConnectionToken}{[required] The authentication token associated with the participant's connection.}
}
\description{
Sends an event. Note that ConnectionToken is used for invoking this API instead of ParticipantToken.

See \url{https://paws-r.github.io/docs/connectparticipant/send_event.html} for full documentation.
}
\keyword{internal}
