% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_send_message}
\alias{connectparticipant_send_message}
\title{Sends a message}
\usage{
connectparticipant_send_message(
  ContentType,
  Content,
  ClientToken = NULL,
  ConnectionToken
)
}
\arguments{
\item{ContentType}{[required] The type of the content. Supported types are text/plain.}

\item{Content}{[required] The content of the message.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{ConnectionToken}{[required] The authentication token associated with the connection.}
}
\description{
Sends a message. Note that ConnectionToken is used for invoking this API instead of ParticipantToken.

See \url{https://paws-r.github.io/docs/connectparticipant/send_message.html} for full documentation.
}
\keyword{internal}
