% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_create_assistant}
\alias{connectwisdomservice_create_assistant}
\title{Creates an Amazon Connect Wisdom assistant}
\usage{
connectwisdomservice_create_assistant(
  clientToken = NULL,
  description = NULL,
  name,
  serverSideEncryptionConfiguration = NULL,
  tags = NULL,
  type
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{description}{The description of the assistant.}

\item{name}{[required] The name of the assistant.}

\item{serverSideEncryptionConfiguration}{The KMS key used for encryption.}

\item{tags}{The tags used to organize, track, or control access for this resource.}

\item{type}{[required] The type of assistant.}
}
\description{
Creates an Amazon Connect Wisdom assistant.

See \url{https://paws-r.github.io/docs/connectwisdomservice/create_assistant.html} for full documentation.
}
\keyword{internal}
