% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_delete_assistant}
\alias{connectwisdomservice_delete_assistant}
\title{Deletes an assistant}
\usage{
connectwisdomservice_delete_assistant(assistantId)
}
\arguments{
\item{assistantId}{[required] The identifier of the Wisdom assistant. Can be either the ID or the ARN.
URLs cannot contain the ARN.}
}
\description{
Deletes an assistant.

See \url{https://paws-r.github.io/docs/connectwisdomservice/delete_assistant.html} for full documentation.
}
\keyword{internal}
