% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_delete_assistant_association}
\alias{connectwisdomservice_delete_assistant_association}
\title{Deletes an assistant association}
\usage{
connectwisdomservice_delete_assistant_association(
  assistantAssociationId,
  assistantId
)
}
\arguments{
\item{assistantAssociationId}{[required] The identifier of the assistant association. Can be either the ID or the
ARN. URLs cannot contain the ARN.}

\item{assistantId}{[required] The identifier of the Wisdom assistant. Can be either the ID or the ARN.
URLs cannot contain the ARN.}
}
\description{
Deletes an assistant association.

See \url{https://paws-r.github.io/docs/connectwisdomservice/delete_assistant_association.html} for full documentation.
}
\keyword{internal}
