% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_delete_content}
\alias{connectwisdomservice_delete_content}
\title{Deletes the content}
\usage{
connectwisdomservice_delete_content(contentId, knowledgeBaseId)
}
\arguments{
\item{contentId}{[required] The identifier of the content. Can be either the ID or the ARN. URLs
cannot contain the ARN.}

\item{knowledgeBaseId}{[required] The identifier of the knowledge base. Can be either the ID or the ARN.
URLs cannot contain the ARN.}
}
\description{
Deletes the content.

See \url{https://paws-r.github.io/docs/connectwisdomservice/delete_content.html} for full documentation.
}
\keyword{internal}
