% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_delete_import_job}
\alias{connectwisdomservice_delete_import_job}
\title{Deletes the quick response import job}
\usage{
connectwisdomservice_delete_import_job(importJobId, knowledgeBaseId)
}
\arguments{
\item{importJobId}{[required] The identifier of the import job to be deleted.}

\item{knowledgeBaseId}{[required] The identifier of the knowledge base. This should not be a
QUICK_RESPONSES type knowledge base if you're storing Wisdom Content
resource to it.}
}
\description{
Deletes the quick response import job.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_delete_import_job/} for full documentation.
}
\keyword{internal}
