% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_delete_knowledge_base}
\alias{connectwisdomservice_delete_knowledge_base}
\title{Deletes the knowledge base}
\usage{
connectwisdomservice_delete_knowledge_base(knowledgeBaseId)
}
\arguments{
\item{knowledgeBaseId}{[required] The knowledge base to delete content from. Can be either the ID or the
ARN. URLs cannot contain the ARN.}
}
\description{
Deletes the knowledge base.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_delete_knowledge_base/} for full documentation.
}
\keyword{internal}
