% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_get_content_summary}
\alias{connectwisdomservice_get_content_summary}
\title{Retrieves summary information about the content}
\usage{
connectwisdomservice_get_content_summary(contentId, knowledgeBaseId)
}
\arguments{
\item{contentId}{[required] The identifier of the content. Can be either the ID or the ARN. URLs
cannot contain the ARN.}

\item{knowledgeBaseId}{[required] The identifier of the knowledge base. This should not be a
QUICK_RESPONSES type knowledge base if you're storing Wisdom Content
resource to it. Can be either the ID or the ARN. URLs cannot contain the
ARN.}
}
\description{
Retrieves summary information about the content.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_get_content_summary/} for full documentation.
}
\keyword{internal}
