% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_get_import_job}
\alias{connectwisdomservice_get_import_job}
\title{Retrieves the started import job}
\usage{
connectwisdomservice_get_import_job(importJobId, knowledgeBaseId)
}
\arguments{
\item{importJobId}{[required] The identifier of the import job to retrieve.}

\item{knowledgeBaseId}{[required] The identifier of the knowledge base that the import job belongs to.}
}
\description{
Retrieves the started import job.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_get_import_job/} for full documentation.
}
\keyword{internal}
