% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_get_knowledge_base}
\alias{connectwisdomservice_get_knowledge_base}
\title{Retrieves information about the knowledge base}
\usage{
connectwisdomservice_get_knowledge_base(knowledgeBaseId)
}
\arguments{
\item{knowledgeBaseId}{[required] The identifier of the knowledge base. Can be either the ID or the ARN.
URLs cannot contain the ARN.}
}
\description{
Retrieves information about the knowledge base.

See \url{https://paws-r.github.io/docs/connectwisdomservice/get_knowledge_base.html} for full documentation.
}
\keyword{internal}
