% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_get_quick_response}
\alias{connectwisdomservice_get_quick_response}
\title{Retrieves the quick response}
\usage{
connectwisdomservice_get_quick_response(knowledgeBaseId, quickResponseId)
}
\arguments{
\item{knowledgeBaseId}{[required] The identifier of the knowledge base. This should be a QUICK_RESPONSES
type knowledge base.}

\item{quickResponseId}{[required] The identifier of the quick response.}
}
\description{
Retrieves the quick response.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_get_quick_response/} for full documentation.
}
\keyword{internal}
