% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_get_session}
\alias{connectwisdomservice_get_session}
\title{Retrieves information for a specified session}
\usage{
connectwisdomservice_get_session(assistantId, sessionId)
}
\arguments{
\item{assistantId}{[required] The identifier of the Wisdom assistant. Can be either the ID or the ARN.
URLs cannot contain the ARN.}

\item{sessionId}{[required] The identifier of the session. Can be either the ID or the ARN. URLs
cannot contain the ARN.}
}
\description{
Retrieves information for a specified session.

See \url{https://paws-r.github.io/docs/connectwisdomservice/get_session.html} for full documentation.
}
\keyword{internal}
