% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_list_assistants}
\alias{connectwisdomservice_list_assistants}
\title{Lists information about assistants}
\usage{
connectwisdomservice_list_assistants(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Lists information about assistants.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_list_assistants/} for full documentation.
}
\keyword{internal}
