% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_list_contents}
\alias{connectwisdomservice_list_contents}
\title{Lists the content}
\usage{
connectwisdomservice_list_contents(
  knowledgeBaseId,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{knowledgeBaseId}{[required] The identifier of the knowledge base. Can be either the ID or the ARN.
URLs cannot contain the ARN.}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Lists the content.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_list_contents/} for full documentation.
}
\keyword{internal}
