% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_list_knowledge_bases}
\alias{connectwisdomservice_list_knowledge_bases}
\title{Lists the knowledge bases}
\usage{
connectwisdomservice_list_knowledge_bases(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}
}
\description{
Lists the knowledge bases.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_list_knowledge_bases/} for full documentation.
}
\keyword{internal}
