% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_notify_recommendations_received}
\alias{connectwisdomservice_notify_recommendations_received}
\title{Removes the specified recommendations from the specified assistant's
queue of newly available recommendations}
\usage{
connectwisdomservice_notify_recommendations_received(
  assistantId,
  recommendationIds,
  sessionId
)
}
\arguments{
\item{assistantId}{[required] The identifier of the Wisdom assistant. Can be either the ID or the ARN.
URLs cannot contain the ARN.}

\item{recommendationIds}{[required] The identifiers of the recommendations.}

\item{sessionId}{[required] The identifier of the session. Can be either the ID or the ARN. URLs
cannot contain the ARN.}
}
\description{
Removes the specified recommendations from the specified assistant's queue of newly available recommendations. You can use this API in conjunction with \code{\link[=connectwisdomservice_get_recommendations]{get_recommendations}} and a \code{waitTimeSeconds} input for long-polling behavior and avoiding duplicate recommendations.

See \url{https://paws-r.github.io/docs/connectwisdomservice/notify_recommendations_received.html} for full documentation.
}
\keyword{internal}
