% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_query_assistant}
\alias{connectwisdomservice_query_assistant}
\title{Performs a manual search against the specified assistant}
\usage{
connectwisdomservice_query_assistant(
  assistantId,
  maxResults = NULL,
  nextToken = NULL,
  queryText
)
}
\arguments{
\item{assistantId}{[required] The identifier of the Wisdom assistant. Can be either the ID or the ARN.
URLs cannot contain the ARN.}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{queryText}{[required] The text to search for.}
}
\description{
Performs a manual search against the specified assistant. To retrieve recommendations for an assistant, use \code{\link[=connectwisdomservice_get_recommendations]{get_recommendations}}.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_query_assistant/} for full documentation.
}
\keyword{internal}
