% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_remove_knowledge_base_template_uri}
\alias{connectwisdomservice_remove_knowledge_base_template_uri}
\title{Removes a URI template from a knowledge base}
\usage{
connectwisdomservice_remove_knowledge_base_template_uri(knowledgeBaseId)
}
\arguments{
\item{knowledgeBaseId}{[required] The identifier of the knowledge base. Can be either the ID or the ARN.
URLs cannot contain the ARN.}
}
\description{
Removes a URI template from a knowledge base.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_remove_knowledge_base_template_uri/} for full documentation.
}
\keyword{internal}
