% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_search_sessions}
\alias{connectwisdomservice_search_sessions}
\title{Searches for sessions}
\usage{
connectwisdomservice_search_sessions(
  assistantId,
  maxResults = NULL,
  nextToken = NULL,
  searchExpression
)
}
\arguments{
\item{assistantId}{[required] The identifier of the Wisdom assistant. Can be either the ID or the ARN.
URLs cannot contain the ARN.}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{searchExpression}{[required] The search expression to filter results.}
}
\description{
Searches for sessions.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_search_sessions/} for full documentation.
}
\keyword{internal}
