% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_tag_resource}
\alias{connectwisdomservice_tag_resource}
\title{Adds the specified tags to the specified resource}
\usage{
connectwisdomservice_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tags}{[required] The tags used to organize, track, or control access for this resource.}
}
\description{
Adds the specified tags to the specified resource.

See \url{https://paws-r.github.io/docs/connectwisdomservice/tag_resource.html} for full documentation.
}
\keyword{internal}
