% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_untag_resource}
\alias{connectwisdomservice_untag_resource}
\title{Removes the specified tags from the specified resource}
\usage{
connectwisdomservice_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tagKeys}{[required] The tag keys.}
}
\description{
Removes the specified tags from the specified resource.

See \url{https://paws-r.github.io/docs/connectwisdomservice/untag_resource.html} for full documentation.
}
\keyword{internal}
