% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_add_profile_key}
\alias{customerprofiles_add_profile_key}
\title{Associates a new key value with a specific profile, such as a Contact
Trace Record (CTR) ContactId}
\usage{
customerprofiles_add_profile_key(ProfileId, KeyName, Values, DomainName)
}
\arguments{
\item{ProfileId}{[required] The unique identifier of a customer profile.}

\item{KeyName}{[required] A searchable identifier of a customer profile. The predefined keys you
can use include: _account, _profileId, _assetId, _caseId, _orderId,
_fullName, _phone, _email, _ctrContactId, _marketoLeadId,
_salesforceAccountId, _salesforceContactId, _salesforceAssetId,
_zendeskUserId, _zendeskExternalId, _zendeskTicketId,
_serviceNowSystemId, _serviceNowIncidentId, _segmentUserId,
_shopifyCustomerId, _shopifyOrderId.}

\item{Values}{[required] A list of key values.}

\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Associates a new key value with a specific profile, such as a Contact Trace Record (CTR) ContactId.

See \url{https://paws-r.github.io/docs/customerprofiles/add_profile_key.html} for full documentation.
}
\keyword{internal}
