% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_associate_stream_for_segments}
\alias{customerprofiles_associate_stream_for_segments}
\title{Associates an Amazon Kinesis data stream to receive segment membership
events for a given domain}
\usage{
customerprofiles_associate_stream_for_segments(
  DomainName,
  DestinationArn,
  DestinationRoleArn
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{DestinationArn}{[required] The Amazon Resource Name (ARN) of the Amazon Kinesis data stream to deliver segment membership events to. For example, \code{arn:aws:kinesis:region:account-id:stream/stream-name}.}

\item{DestinationRoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role that allows Customer Profiles service principal to assume the role for conducting AWS Key Management Service (KMS) and Amazon Kinesis operations. The role must grant the following Amazon Kinesis permissions to deliver segment membership events to the stream:
\itemize{
\item \code{kinesis:PutRecord}
\item \code{kinesis:PutRecords}
\item \code{kinesis:DescribeStream}
}}
}
\description{
Associates an Amazon Kinesis data stream to receive segment membership events for a given domain. This is a domain-level configuration that applies to all segment subscriptions within the domain. A domain can have only one associated stream at a time.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_associate_stream_for_segments/} for full documentation.
}
\keyword{internal}
