% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_batch_get_calculated_attribute_for_profile}
\alias{customerprofiles_batch_get_calculated_attribute_for_profile}
\title{Fetch the possible attribute values given the attribute name}
\usage{
customerprofiles_batch_get_calculated_attribute_for_profile(
  CalculatedAttributeName,
  DomainName,
  ProfileIds,
  ConditionOverrides = NULL
)
}
\arguments{
\item{CalculatedAttributeName}{[required] The unique name of the calculated attribute.}

\item{DomainName}{[required] The unique name of the domain.}

\item{ProfileIds}{[required] List of unique identifiers for customer profiles to retrieve.}

\item{ConditionOverrides}{Overrides the condition block within the original calculated attribute
definition.}
}
\description{
Fetch the possible attribute values given the attribute name.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_batch_get_calculated_attribute_for_profile/} for full documentation.
}
\keyword{internal}
