% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_batch_get_profile}
\alias{customerprofiles_batch_get_profile}
\title{Get a batch of profiles}
\usage{
customerprofiles_batch_get_profile(DomainName, ProfileIds)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ProfileIds}{[required] List of unique identifiers for customer profiles to retrieve.}
}
\description{
Get a batch of profiles.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_batch_get_profile/} for full documentation.
}
\keyword{internal}
