% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_batch_put_profile_object}
\alias{customerprofiles_batch_put_profile_object}
\title{Adds multiple profile objects to a domain of a given ObjectType in a
single API call}
\usage{
customerprofiles_batch_put_profile_object(DomainName, ObjectTypeName, Items)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ObjectTypeName}{[required] The name of the profile object type.}

\item{Items}{[required] A list of items to add to the domain.}
}
\description{
Adds multiple profile objects to a domain of a given ObjectType in a single API call.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_batch_put_profile_object/} for full documentation.
}
\keyword{internal}
