% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_create_domain_layout}
\alias{customerprofiles_create_domain_layout}
\title{Creates the layout to view data for a specific domain}
\usage{
customerprofiles_create_domain_layout(
  DomainName,
  LayoutDefinitionName,
  Description,
  DisplayName,
  IsDefault = NULL,
  LayoutType,
  Layout,
  Tags = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{LayoutDefinitionName}{[required] The unique name of the layout.}

\item{Description}{[required] The description of the layout}

\item{DisplayName}{[required] The display name of the layout}

\item{IsDefault}{If set to true for a layout, this layout will be used by default to view
data. If set to false, then the layout will not be used by default, but
it can be used to view data by explicitly selecting it in the console.}

\item{LayoutType}{[required] The type of layout that can be used to view data under a Customer
Profiles domain.}

\item{Layout}{[required] A customizable layout that can be used to view data under a Customer
Profiles domain.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Creates the layout to view data for a specific domain. This API can only be invoked from the Amazon Connect admin website.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_create_domain_layout/} for full documentation.
}
\keyword{internal}
