% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_create_recommender}
\alias{customerprofiles_create_recommender}
\title{Creates a recommender}
\usage{
customerprofiles_create_recommender(
  DomainName,
  RecommenderName,
  RecommenderRecipeName,
  RecommenderConfig = NULL,
  Description = NULL,
  Tags = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderName}{[required] The name of the recommender.}

\item{RecommenderRecipeName}{[required] The name of the recommeder recipe.}

\item{RecommenderConfig}{The recommender configuration.}

\item{Description}{The description of the domain object type.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Creates a recommender

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_create_recommender/} for full documentation.
}
\keyword{internal}
