% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_create_recommender_filter}
\alias{customerprofiles_create_recommender_filter}
\title{Creates a recommender filter}
\usage{
customerprofiles_create_recommender_filter(
  DomainName,
  RecommenderFilterName,
  RecommenderFilterExpression,
  RecommenderSchemaName = NULL,
  Description = NULL,
  Tags = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderFilterName}{[required] The name of the recommender filter. The name must be unique within the
domain.}

\item{RecommenderFilterExpression}{[required] The filter expression that defines which items to include or exclude
from recommendations.}

\item{RecommenderSchemaName}{The name of the recommender schema to use for this recommender filter.
If not specified, the default schema is used.}

\item{Description}{A description of the recommender filter.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Creates a recommender filter. A recommender filter specifies which items to include or exclude from recommendations.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_create_recommender_filter/} for full documentation.
}
\keyword{internal}
