% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_create_recommender_schema}
\alias{customerprofiles_create_recommender_schema}
\title{Creates a recommender schema}
\usage{
customerprofiles_create_recommender_schema(
  DomainName,
  RecommenderSchemaName,
  Fields,
  Tags = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderSchemaName}{[required] The name of the recommender schema. The name must be unique within the
domain.}

\item{Fields}{[required] A map of dataset type to column definitions that specifies which data
columns to include in the schema. Currently only the \verb{_webAnalytics} key
is supported.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Creates a recommender schema. A recommender schema defines the set of data columns available for training recommenders and filters under a domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_create_recommender_schema/} for full documentation.
}
\keyword{internal}
