% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_create_segment_definition}
\alias{customerprofiles_create_segment_definition}
\title{Creates a segment definition associated to the given domain}
\usage{
customerprofiles_create_segment_definition(
  DomainName,
  SegmentDefinitionName,
  DisplayName,
  Description = NULL,
  SegmentGroups = NULL,
  SegmentSqlQuery = NULL,
  Tags = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{SegmentDefinitionName}{[required] The unique name of the segment definition.}

\item{DisplayName}{[required] The display name of the segment definition.}

\item{Description}{The description of the segment definition.}

\item{SegmentGroups}{Specifies the base segments and dimensions for a segment definition
along with their respective relationship.}

\item{SegmentSqlQuery}{The segment SQL query.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Creates a segment definition associated to the given domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_create_segment_definition/} for full documentation.
}
\keyword{internal}
