% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_create_segment_estimate}
\alias{customerprofiles_create_segment_estimate}
\title{Creates a segment estimate query}
\usage{
customerprofiles_create_segment_estimate(
  DomainName,
  SegmentQuery = NULL,
  SegmentSqlQuery = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{SegmentQuery}{The segment query for calculating a segment estimate.}

\item{SegmentSqlQuery}{The segment SQL query.}
}
\description{
Creates a segment estimate query.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_create_segment_estimate/} for full documentation.
}
\keyword{internal}
