% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_create_upload_job}
\alias{customerprofiles_create_upload_job}
\title{Creates an Upload job to ingest data for segment imports}
\usage{
customerprofiles_create_upload_job(
  DomainName,
  DisplayName,
  Fields,
  UniqueKey,
  DataExpiry = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain. Domain should be exists for the upload
job to be created.}

\item{DisplayName}{[required] The unique name of the upload job. Could be a file name to identify the
upload job.}

\item{Fields}{[required] The mapping between CSV Columns and Profile Object attributes. A map of
the name and ObjectType field.}

\item{UniqueKey}{[required] The unique key columns for de-duping the profiles used to map data to
the profile.}

\item{DataExpiry}{The expiry duration for the profiles ingested with the job. If not
provided, the system default of 2 weeks is used.}
}
\description{
Creates an Upload job to ingest data for segment imports. The metadata is created for the job with the provided field mapping and unique key.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_create_upload_job/} for full documentation.
}
\keyword{internal}
