% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_calculated_attribute_definition}
\alias{customerprofiles_delete_calculated_attribute_definition}
\title{Deletes an existing calculated attribute definition}
\usage{
customerprofiles_delete_calculated_attribute_definition(
  DomainName,
  CalculatedAttributeName
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{CalculatedAttributeName}{[required] The unique name of the calculated attribute.}
}
\description{
Deletes an existing calculated attribute definition. Note that deleting a default calculated attribute is possible, however once deleted, you will be unable to undo that action and will need to recreate it on your own using the CreateCalculatedAttributeDefinition API if you want it back.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_calculated_attribute_definition/} for full documentation.
}
\keyword{internal}
