% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_domain_layout}
\alias{customerprofiles_delete_domain_layout}
\title{Deletes the layout used to view data for a specific domain}
\usage{
customerprofiles_delete_domain_layout(DomainName, LayoutDefinitionName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{LayoutDefinitionName}{[required] The unique name of the layout.}
}
\description{
Deletes the layout used to view data for a specific domain. This API can only be invoked from the Amazon Connect admin website.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_domain_layout/} for full documentation.
}
\keyword{internal}
