% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_domain_object_type}
\alias{customerprofiles_delete_domain_object_type}
\title{Delete a DomainObjectType for the given Domain and ObjectType name}
\usage{
customerprofiles_delete_domain_object_type(DomainName, ObjectTypeName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ObjectTypeName}{[required] The unique name of the domain object type.}
}
\description{
Delete a DomainObjectType for the given Domain and ObjectType name.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_domain_object_type/} for full documentation.
}
\keyword{internal}
