% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_event_stream}
\alias{customerprofiles_delete_event_stream}
\title{Disables and deletes the specified event stream}
\usage{
customerprofiles_delete_event_stream(DomainName, EventStreamName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{EventStreamName}{[required] The name of the event stream}
}
\description{
Disables and deletes the specified event stream.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_event_stream/} for full documentation.
}
\keyword{internal}
