% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_event_trigger}
\alias{customerprofiles_delete_event_trigger}
\title{Disable and deletes the Event Trigger}
\usage{
customerprofiles_delete_event_trigger(DomainName, EventTriggerName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{EventTriggerName}{[required] The unique name of the event trigger.}
}
\description{
Disable and deletes the Event Trigger.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_event_trigger/} for full documentation.
}
\keyword{internal}
