% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_profile}
\alias{customerprofiles_delete_profile}
\title{Deletes the standard customer profile and all data pertaining to the
profile}
\usage{
customerprofiles_delete_profile(ProfileId, DomainName)
}
\arguments{
\item{ProfileId}{[required] The unique identifier of a customer profile.}

\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Deletes the standard customer profile and all data pertaining to the profile.

See \url{https://paws-r.github.io/docs/customerprofiles/delete_profile.html} for full documentation.
}
\keyword{internal}
