% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_profile_object}
\alias{customerprofiles_delete_profile_object}
\title{Removes an object associated with a profile of a given ProfileObjectType}
\usage{
customerprofiles_delete_profile_object(
  ProfileId,
  ProfileObjectUniqueKey,
  ObjectTypeName,
  DomainName
)
}
\arguments{
\item{ProfileId}{[required] The unique identifier of a customer profile.}

\item{ProfileObjectUniqueKey}{[required] The unique identifier of the profile object generated by the service.}

\item{ObjectTypeName}{[required] The name of the profile object type.}

\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Removes an object associated with a profile of a given ProfileObjectType.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_profile_object/} for full documentation.
}
\keyword{internal}
