% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_recommender}
\alias{customerprofiles_delete_recommender}
\title{Deletes a recommender}
\usage{
customerprofiles_delete_recommender(DomainName, RecommenderName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderName}{[required] The recommender name.}
}
\description{
Deletes a recommender.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_recommender/} for full documentation.
}
\keyword{internal}
