% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_recommender_filter}
\alias{customerprofiles_delete_recommender_filter}
\title{Deletes a recommender filter from a domain}
\usage{
customerprofiles_delete_recommender_filter(DomainName, RecommenderFilterName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderFilterName}{[required] The name of the recommender filter to delete.}
}
\description{
Deletes a recommender filter from a domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_recommender_filter/} for full documentation.
}
\keyword{internal}
