% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_recommender_schema}
\alias{customerprofiles_delete_recommender_schema}
\title{Deletes a recommender schema from a domain}
\usage{
customerprofiles_delete_recommender_schema(DomainName, RecommenderSchemaName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderSchemaName}{[required] The name of the recommender schema to delete.}
}
\description{
Deletes a recommender schema from a domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_recommender_schema/} for full documentation.
}
\keyword{internal}
