% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_segment_subscription}
\alias{customerprofiles_delete_segment_subscription}
\title{Deletes a segment subscription for membership events}
\usage{
customerprofiles_delete_segment_subscription(DomainName, SegmentDefinitionName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{SegmentDefinitionName}{[required] The unique name of the segment definition.}
}
\description{
Deletes a segment subscription for membership events. All active event notifications for this segment are stopped.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_segment_subscription/} for full documentation.
}
\keyword{internal}
