% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_delete_workflow}
\alias{customerprofiles_delete_workflow}
\title{Deletes the specified workflow and all its corresponding resources}
\usage{
customerprofiles_delete_workflow(DomainName, WorkflowId)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{WorkflowId}{[required] Unique identifier for the workflow.}
}
\description{
Deletes the specified workflow and all its corresponding resources. This is an async process.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_delete_workflow/} for full documentation.
}
\keyword{internal}
