% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_detect_profile_object_type}
\alias{customerprofiles_detect_profile_object_type}
\title{The process of detecting profile object type mapping by using given
objects}
\usage{
customerprofiles_detect_profile_object_type(Objects, DomainName)
}
\arguments{
\item{Objects}{[required] A string that is serialized from a JSON object.}

\item{DomainName}{[required] The unique name of the domain.}
}
\description{
The process of detecting profile object type mapping by using given objects.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_detect_profile_object_type/} for full documentation.
}
\keyword{internal}
