% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_disassociate_stream_for_segments}
\alias{customerprofiles_disassociate_stream_for_segments}
\title{Disassociates the Amazon Kinesis data stream configured for segment
membership events}
\usage{
customerprofiles_disassociate_stream_for_segments(DomainName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Disassociates the Amazon Kinesis data stream configured for segment membership events. All active segment subscriptions delivering events to this stream are eventually stopped.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_disassociate_stream_for_segments/} for full documentation.
}
\keyword{internal}
