% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_calculated_attribute_definition}
\alias{customerprofiles_get_calculated_attribute_definition}
\title{Provides more information on a calculated attribute definition for
Customer Profiles}
\usage{
customerprofiles_get_calculated_attribute_definition(
  DomainName,
  CalculatedAttributeName
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{CalculatedAttributeName}{[required] The unique name of the calculated attribute.}
}
\description{
Provides more information on a calculated attribute definition for Customer Profiles.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_calculated_attribute_definition/} for full documentation.
}
\keyword{internal}
