% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_calculated_attribute_for_profile}
\alias{customerprofiles_get_calculated_attribute_for_profile}
\title{Retrieve a calculated attribute for a customer profile}
\usage{
customerprofiles_get_calculated_attribute_for_profile(
  DomainName,
  ProfileId,
  CalculatedAttributeName
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ProfileId}{[required] The unique identifier of a customer profile.}

\item{CalculatedAttributeName}{[required] The unique name of the calculated attribute.}
}
\description{
Retrieve a calculated attribute for a customer profile.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_calculated_attribute_for_profile/} for full documentation.
}
\keyword{internal}
