% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_domain_object_type}
\alias{customerprofiles_get_domain_object_type}
\title{Return a DomainObjectType for the input Domain and ObjectType names}
\usage{
customerprofiles_get_domain_object_type(DomainName, ObjectTypeName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ObjectTypeName}{[required] The unique name of the domain object type.}
}
\description{
Return a DomainObjectType for the input Domain and ObjectType names.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_domain_object_type/} for full documentation.
}
\keyword{internal}
