% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_event_stream}
\alias{customerprofiles_get_event_stream}
\title{Returns information about the specified event stream in a specific
domain}
\usage{
customerprofiles_get_event_stream(DomainName, EventStreamName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{EventStreamName}{[required] The name of the event stream provided during create operations.}
}
\description{
Returns information about the specified event stream in a specific domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_event_stream/} for full documentation.
}
\keyword{internal}
