% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_event_trigger}
\alias{customerprofiles_get_event_trigger}
\title{Get a specific Event Trigger from the domain}
\usage{
customerprofiles_get_event_trigger(DomainName, EventTriggerName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{EventTriggerName}{[required] The unique name of the event trigger.}
}
\description{
Get a specific Event Trigger from the domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_event_trigger/} for full documentation.
}
\keyword{internal}
