% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_identity_resolution_job}
\alias{customerprofiles_get_identity_resolution_job}
\title{Returns information about an Identity Resolution Job in a specific
domain}
\usage{
customerprofiles_get_identity_resolution_job(DomainName, JobId)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{JobId}{[required] The unique identifier of the Identity Resolution Job.}
}
\description{
Returns information about an Identity Resolution Job in a specific domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_identity_resolution_job/} for full documentation.
}
\keyword{internal}
