% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_object_type_attribute_statistics}
\alias{customerprofiles_get_object_type_attribute_statistics}
\title{The GetObjectTypeAttributeValues API delivers statistical insights about
attributes within a specific object type, but is exclusively available
for domains with data store enabled}
\usage{
customerprofiles_get_object_type_attribute_statistics(
  DomainName,
  ObjectTypeName,
  AttributeName
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ObjectTypeName}{[required] The unique name of the domain object type.}

\item{AttributeName}{[required] The attribute name.}
}
\description{
The GetObjectTypeAttributeValues API delivers statistical insights about attributes within a specific object type, but is exclusively available for domains with data store enabled. This API performs daily calculations to provide statistical information about your attribute values, helping you understand patterns and trends in your data. The statistical calculations are performed once per day, providing a consistent snapshot of your attribute data characteristics.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_object_type_attribute_statistics/} for full documentation.
}
\keyword{internal}
