% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_profile_history_record}
\alias{customerprofiles_get_profile_history_record}
\title{Returns a history record for a specific profile, for a specific domain}
\usage{
customerprofiles_get_profile_history_record(DomainName, ProfileId, Id)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain for which to return a profile history
record.}

\item{ProfileId}{[required] The unique identifier of the profile for which to return a history
record.}

\item{Id}{[required] The unique identifier of the profile history record to return.}
}
\description{
Returns a history record for a specific profile, for a specific domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_profile_history_record/} for full documentation.
}
\keyword{internal}
