% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_profile_object_type_template}
\alias{customerprofiles_get_profile_object_type_template}
\title{Returns the template information for a specific object type}
\usage{
customerprofiles_get_profile_object_type_template(TemplateId)
}
\arguments{
\item{TemplateId}{[required] A unique identifier for the object template.}
}
\description{
Returns the template information for a specific object type.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_profile_object_type_template/} for full documentation.
}
\keyword{internal}
