% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_recommender}
\alias{customerprofiles_get_recommender}
\title{Retrieves a recommender}
\usage{
customerprofiles_get_recommender(
  DomainName,
  RecommenderName,
  TrainingMetricsCount = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderName}{[required] The name of the recommender.}

\item{TrainingMetricsCount}{The number of training metrics to retrieve for the recommender.}
}
\description{
Retrieves a recommender.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_recommender/} for full documentation.
}
\keyword{internal}
