% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_recommender_filter}
\alias{customerprofiles_get_recommender_filter}
\title{Retrieves information about a specific recommender filter in a domain}
\usage{
customerprofiles_get_recommender_filter(DomainName, RecommenderFilterName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderFilterName}{[required] The name of the recommender filter to retrieve.}
}
\description{
Retrieves information about a specific recommender filter in a domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_recommender_filter/} for full documentation.
}
\keyword{internal}
