% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_recommender_schema}
\alias{customerprofiles_get_recommender_schema}
\title{Retrieves information about a specific recommender schema in a domain}
\usage{
customerprofiles_get_recommender_schema(DomainName, RecommenderSchemaName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderSchemaName}{[required] The name of the recommender schema to retrieve.}
}
\description{
Retrieves information about a specific recommender schema in a domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_recommender_schema/} for full documentation.
}
\keyword{internal}
