% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_segment_definition}
\alias{customerprofiles_get_segment_definition}
\title{Gets a segment definition from the domain}
\usage{
customerprofiles_get_segment_definition(DomainName, SegmentDefinitionName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{SegmentDefinitionName}{[required] The unique name of the segment definition.}
}
\description{
Gets a segment definition from the domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_segment_definition/} for full documentation.
}
\keyword{internal}
