% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_segment_estimate}
\alias{customerprofiles_get_segment_estimate}
\title{Gets the result of a segment estimate query}
\usage{
customerprofiles_get_segment_estimate(DomainName, EstimateId)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{EstimateId}{[required] The query Id passed by a previous
\code{\link[=customerprofiles_create_segment_estimate]{create_segment_estimate}}
operation.}
}
\description{
Gets the result of a segment estimate query.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_segment_estimate/} for full documentation.
}
\keyword{internal}
