% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_segment_membership}
\alias{customerprofiles_get_segment_membership}
\title{Determines if the given profiles are within a segment}
\usage{
customerprofiles_get_segment_membership(
  DomainName,
  SegmentDefinitionName,
  ProfileIds
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{SegmentDefinitionName}{[required] The Id of the wanted segment. Needs to be a valid, and existing segment
Id.}

\item{ProfileIds}{[required] The list of profile IDs to query for.}
}
\description{
Determines if the given profiles are within a segment.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_segment_membership/} for full documentation.
}
\keyword{internal}
