% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_segment_snapshot}
\alias{customerprofiles_get_segment_snapshot}
\title{Retrieve the latest status of a segment snapshot}
\usage{
customerprofiles_get_segment_snapshot(
  DomainName,
  SegmentDefinitionName,
  SnapshotId
)
}
\arguments{
\item{DomainName}{[required] The unique identifier of the domain.}

\item{SegmentDefinitionName}{[required] The unique name of the segment definition.}

\item{SnapshotId}{[required] The unique identifier of the segment snapshot.}
}
\description{
Retrieve the latest status of a segment snapshot.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_segment_snapshot/} for full documentation.
}
\keyword{internal}
