% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_segment_subscription}
\alias{customerprofiles_get_segment_subscription}
\title{Returns the current subscription configuration, execution schedule, and
status for segment membership events}
\usage{
customerprofiles_get_segment_subscription(DomainName, SegmentDefinitionName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{SegmentDefinitionName}{[required] The unique name of the segment definition.}
}
\description{
Returns the current subscription configuration, execution schedule, and status for segment membership events.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_segment_subscription/} for full documentation.
}
\keyword{internal}
