% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_stream_for_segments}
\alias{customerprofiles_get_stream_for_segments}
\title{Returns information about the segment membership event stream configured
for a specific domain, including the stream state and associated
segments}
\usage{
customerprofiles_get_stream_for_segments(DomainName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Returns information about the segment membership event stream configured for a specific domain, including the stream state and associated segments.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_stream_for_segments/} for full documentation.
}
\keyword{internal}
