% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_upload_job}
\alias{customerprofiles_get_upload_job}
\title{This API retrieves the details of a specific upload job}
\usage{
customerprofiles_get_upload_job(DomainName, JobId)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain containing the upload job.}

\item{JobId}{[required] The unique identifier of the upload job to retrieve.}
}
\description{
This API retrieves the details of a specific upload job.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_upload_job/} for full documentation.
}
\keyword{internal}
