% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_upload_job_path}
\alias{customerprofiles_get_upload_job_path}
\title{This API retrieves the pre-signed URL and client token for uploading the
file associated with the upload job}
\usage{
customerprofiles_get_upload_job_path(DomainName, JobId)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain containing the upload job.}

\item{JobId}{[required] The unique identifier of the upload job to retrieve the upload path for.
This is generated from the CreateUploadJob API.}
}
\description{
This API retrieves the pre-signed URL and client token for uploading the file associated with the upload job.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_upload_job_path/} for full documentation.
}
\keyword{internal}
