% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_workflow}
\alias{customerprofiles_get_workflow}
\title{Get details of specified workflow}
\usage{
customerprofiles_get_workflow(DomainName, WorkflowId)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{WorkflowId}{[required] Unique identifier for the workflow.}
}
\description{
Get details of specified workflow.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_workflow/} for full documentation.
}
\keyword{internal}
