% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_workflow_steps}
\alias{customerprofiles_get_workflow_steps}
\title{Get granular list of steps in workflow}
\usage{
customerprofiles_get_workflow_steps(
  DomainName,
  WorkflowId,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{WorkflowId}{[required] Unique identifier for the workflow.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Get granular list of steps in workflow.

See \url{https://paws-r.github.io/docs/customerprofiles/get_workflow_steps.html} for full documentation.
}
\keyword{internal}
