% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_account_integrations}
\alias{customerprofiles_list_account_integrations}
\title{Lists all of the integrations associated to a specific URI in the AWS
account}
\usage{
customerprofiles_list_account_integrations(
  Uri,
  NextToken = NULL,
  MaxResults = NULL,
  IncludeHidden = NULL
)
}
\arguments{
\item{Uri}{[required] The URI of the S3 bucket or any other type of data source.}

\item{NextToken}{The pagination token from the previous ListAccountIntegrations API call.}

\item{MaxResults}{The maximum number of objects returned per page.}

\item{IncludeHidden}{Boolean to indicate if hidden integration should be returned. Defaults
to \code{False}.}
}
\description{
Lists all of the integrations associated to a specific URI in the AWS account.

See \url{https://paws-r.github.io/docs/customerprofiles/list_account_integrations.html} for full documentation.
}
\keyword{internal}
