% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_calculated_attribute_definitions}
\alias{customerprofiles_list_calculated_attribute_definitions}
\title{Lists calculated attribute definitions for Customer Profiles}
\usage{
customerprofiles_list_calculated_attribute_definitions(
  DomainName,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{NextToken}{The pagination token from the previous call to
ListCalculatedAttributeDefinitions.}

\item{MaxResults}{The maximum number of calculated attribute definitions returned per
page.}
}
\description{
Lists calculated attribute definitions for Customer Profiles

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_calculated_attribute_definitions/} for full documentation.
}
\keyword{internal}
