% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_calculated_attributes_for_profile}
\alias{customerprofiles_list_calculated_attributes_for_profile}
\title{Retrieve a list of calculated attributes for a customer profile}
\usage{
customerprofiles_list_calculated_attributes_for_profile(
  NextToken = NULL,
  MaxResults = NULL,
  DomainName,
  ProfileId
)
}
\arguments{
\item{NextToken}{The pagination token from the previous call to
ListCalculatedAttributesForProfile.}

\item{MaxResults}{The maximum number of calculated attributes returned per page.}

\item{DomainName}{[required] The unique name of the domain.}

\item{ProfileId}{[required] The unique identifier of a customer profile.}
}
\description{
Retrieve a list of calculated attributes for a customer profile.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_calculated_attributes_for_profile/} for full documentation.
}
\keyword{internal}
