% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_domain_layouts}
\alias{customerprofiles_list_domain_layouts}
\title{Lists the existing layouts that can be used to view data for a specific
domain}
\usage{
customerprofiles_list_domain_layouts(
  DomainName,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{The maximum number of objects returned per page.}
}
\description{
Lists the existing layouts that can be used to view data for a specific domain. This API can only be invoked from the Amazon Connect admin website.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_domain_layouts/} for full documentation.
}
\keyword{internal}
