% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_domain_object_types}
\alias{customerprofiles_list_domain_object_types}
\title{List all DomainObjectType(s) in a Customer Profiles domain}
\usage{
customerprofiles_list_domain_object_types(
  DomainName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{MaxResults}{The maximum number of domain object types returned per page.}

\item{NextToken}{The pagination token from the previous call to ListDomainObjectTypes.}
}
\description{
List all DomainObjectType(s) in a Customer Profiles domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_domain_object_types/} for full documentation.
}
\keyword{internal}
