% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_domains}
\alias{customerprofiles_list_domains}
\title{Returns a list of all the domains for an AWS account that have been
created}
\usage{
customerprofiles_list_domains(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The pagination token from the previous ListDomain API call.}

\item{MaxResults}{The maximum number of objects returned per page.}
}
\description{
Returns a list of all the domains for an AWS account that have been created.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_domains/} for full documentation.
}
\keyword{internal}
