% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_event_triggers}
\alias{customerprofiles_list_event_triggers}
\title{List all Event Triggers under a domain}
\usage{
customerprofiles_list_event_triggers(
  DomainName,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{NextToken}{The pagination token to use with ListEventTriggers.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
List all Event Triggers under a domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_event_triggers/} for full documentation.
}
\keyword{internal}
