% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_identity_resolution_jobs}
\alias{customerprofiles_list_identity_resolution_jobs}
\title{Lists all of the Identity Resolution Jobs in your domain}
\usage{
customerprofiles_list_identity_resolution_jobs(
  DomainName,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Lists all of the Identity Resolution Jobs in your domain. The response sorts the list by \code{JobStartTime}.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_identity_resolution_jobs/} for full documentation.
}
\keyword{internal}
