% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_object_type_attribute_values}
\alias{customerprofiles_list_object_type_attribute_values}
\title{The ListObjectTypeAttributeValues API provides access to the most recent
distinct values for any specified attribute, making it valuable for
real-time data validation and consistency checks within your object
types}
\usage{
customerprofiles_list_object_type_attribute_values(
  NextToken = NULL,
  MaxResults = NULL,
  DomainName,
  ObjectTypeName,
  AttributeName
)
}
\arguments{
\item{NextToken}{The pagination token from the previous call.}

\item{MaxResults}{The maximum number of objects returned per page. Valid Range: Minimum
value of 1. Maximum value of 100. If not provided default as 100.}

\item{DomainName}{[required] The unique name of the domain.}

\item{ObjectTypeName}{[required] The unique name of the domain object type.}

\item{AttributeName}{[required] The attribute name.}
}
\description{
The ListObjectTypeAttributeValues API provides access to the most recent distinct values for any specified attribute, making it valuable for real-time data validation and consistency checks within your object types. This API works across domain, supporting both custom and standard object types. The API accepts the object type name, attribute name, and domain name as input parameters and returns values up to the storage limit of approximately 350KB.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_object_type_attribute_values/} for full documentation.
}
\keyword{internal}
