% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_object_type_attributes}
\alias{customerprofiles_list_object_type_attributes}
\title{Fetch the possible attribute values given the attribute name}
\usage{
customerprofiles_list_object_type_attributes(
  NextToken = NULL,
  MaxResults = NULL,
  DomainName,
  ObjectTypeName
)
}
\arguments{
\item{NextToken}{The pagination token from the previous call.}

\item{MaxResults}{The maximum number of objects returned per page.}

\item{DomainName}{[required] The unique identifier of the domain.}

\item{ObjectTypeName}{[required] The name of the profile object type.}
}
\description{
Fetch the possible attribute values given the attribute name.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_object_type_attributes/} for full documentation.
}
\keyword{internal}
