% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_profile_attribute_values}
\alias{customerprofiles_list_profile_attribute_values}
\title{Fetch the possible attribute values given the attribute name}
\usage{
customerprofiles_list_profile_attribute_values(DomainName, AttributeName)
}
\arguments{
\item{DomainName}{[required] The unique identifier of the domain.}

\item{AttributeName}{[required] The attribute name.}
}
\description{
Fetch the possible attribute values given the attribute name.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_profile_attribute_values/} for full documentation.
}
\keyword{internal}
