% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_profile_history_records}
\alias{customerprofiles_list_profile_history_records}
\title{Returns a list of history records for a specific profile, for a specific
domain}
\usage{
customerprofiles_list_profile_history_records(
  DomainName,
  ProfileId,
  ObjectTypeName = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  ActionType = NULL,
  PerformedBy = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain for which to return profile history
records.}

\item{ProfileId}{[required] The identifier of the profile to be taken.}

\item{ObjectTypeName}{Applies a filter to include profile history records only with the
specified \code{ObjectTypeName} value in the response.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{ActionType}{Applies a filter to include profile history records only with the
specified \code{ActionType} value in the response.}

\item{PerformedBy}{Applies a filter to include profile history records only with the
specified \code{PerformedBy} value in the response. The \code{PerformedBy} value
can be the Amazon Resource Name (ARN) of the person or service principal
who performed the action.}
}
\description{
Returns a list of history records for a specific profile, for a specific domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_profile_history_records/} for full documentation.
}
\keyword{internal}
