% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_profile_object_types}
\alias{customerprofiles_list_profile_object_types}
\title{Lists all of the templates available within the service}
\usage{
customerprofiles_list_profile_object_types(
  DomainName,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{The maximum number of objects returned per page.}
}
\description{
Lists all of the templates available within the service.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_profile_object_types/} for full documentation.
}
\keyword{internal}
