% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_recommender_filters}
\alias{customerprofiles_list_recommender_filters}
\title{Returns a list of recommender filters in the specified domain}
\usage{
customerprofiles_list_recommender_filters(
  DomainName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{MaxResults}{The maximum number of recommender filters to return in the response. The
default value is 100.}

\item{NextToken}{A token received from a previous ListRecommenderFilters call to retrieve
the next page of results.}
}
\description{
Returns a list of recommender filters in the specified domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_recommender_filters/} for full documentation.
}
\keyword{internal}
