% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_recommender_recipes}
\alias{customerprofiles_list_recommender_recipes}
\title{Returns a list of available recommender recipes that can be used to
create recommenders}
\usage{
customerprofiles_list_recommender_recipes(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of recommender recipes to return in the response. The
default value is 100.}

\item{NextToken}{A token received from a previous ListRecommenderRecipes call to retrieve
the next page of results.}
}
\description{
Returns a list of available recommender recipes that can be used to create recommenders.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_recommender_recipes/} for full documentation.
}
\keyword{internal}
