% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_recommenders}
\alias{customerprofiles_list_recommenders}
\title{Returns a list of recommenders in the specified domain}
\usage{
customerprofiles_list_recommenders(
  DomainName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{MaxResults}{The maximum number of recommenders to return in the response. The
default value is 100.}

\item{NextToken}{A token received from a previous ListRecommenders call to retrieve the
next page of results.}
}
\description{
Returns a list of recommenders in the specified domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_recommenders/} for full documentation.
}
\keyword{internal}
