% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_rule_based_matches}
\alias{customerprofiles_list_rule_based_matches}
\title{Returns a set of MatchIds that belong to the given domain}
\usage{
customerprofiles_list_rule_based_matches(
  NextToken = NULL,
  MaxResults = NULL,
  DomainName
)
}
\arguments{
\item{NextToken}{The pagination token from the previous
\code{\link[=customerprofiles_list_rule_based_matches]{list_rule_based_matches}}
API call.}

\item{MaxResults}{The maximum number of \code{MatchIds} returned per page.}

\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Returns a set of \code{MatchIds} that belong to the given domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_rule_based_matches/} for full documentation.
}
\keyword{internal}
