% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_segment_definitions}
\alias{customerprofiles_list_segment_definitions}
\title{Lists all segment definitions under a domain}
\usage{
customerprofiles_list_segment_definitions(
  DomainName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique identifier of the domain.}

\item{MaxResults}{The maximum number of objects returned per page.}

\item{NextToken}{The pagination token from the previous call.}
}
\description{
Lists all segment definitions under a domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_segment_definitions/} for full documentation.
}
\keyword{internal}
