% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_segment_subscription_events}
\alias{customerprofiles_list_segment_subscription_events}
\title{Returns the most recent membership events for a segment}
\usage{
customerprofiles_list_segment_subscription_events(
  DomainName,
  SegmentDefinitionName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{SegmentDefinitionName}{[required] The unique name of the segment definition.}

\item{MaxResults}{The maximum number of events to return per page.}

\item{NextToken}{The pagination token from the previous call to retrieve the next page of results.}
}
\description{
Returns the most recent membership events for a segment. Each event represents a profile that entered or exited the segment.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_segment_subscription_events/} for full documentation.
}
\keyword{internal}
