% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_upload_jobs}
\alias{customerprofiles_list_upload_jobs}
\title{This API retrieves a list of upload jobs for the specified domain}
\usage{
customerprofiles_list_upload_jobs(
  DomainName,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain to list upload jobs for.}

\item{MaxResults}{The maximum number of upload jobs to return per page.}

\item{NextToken}{The pagination token from the previous call to retrieve the next page of
results.}
}
\description{
This API retrieves a list of upload jobs for the specified domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_list_upload_jobs/} for full documentation.
}
\keyword{internal}
