% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_workflows}
\alias{customerprofiles_list_workflows}
\title{Query to list all workflows}
\usage{
customerprofiles_list_workflows(
  DomainName,
  WorkflowType = NULL,
  Status = NULL,
  QueryStartDate = NULL,
  QueryEndDate = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{WorkflowType}{The type of workflow. The only supported value is APPFLOW_INTEGRATION.}

\item{Status}{Status of workflow execution.}

\item{QueryStartDate}{Retrieve workflows started after timestamp.}

\item{QueryEndDate}{Retrieve workflows ended after timestamp.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Query to list all workflows.

See \url{https://paws-r.github.io/docs/customerprofiles/list_workflows.html} for full documentation.
}
\keyword{internal}
