% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_merge_profiles}
\alias{customerprofiles_merge_profiles}
\title{Runs an AWS Lambda job that does the following:}
\usage{
customerprofiles_merge_profiles(
  DomainName,
  MainProfileId,
  ProfileIdsToBeMerged,
  FieldSourceProfileIds = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{MainProfileId}{[required] The identifier of the profile to be taken.}

\item{ProfileIdsToBeMerged}{[required] The identifier of the profile to be merged into MainProfileId.}

\item{FieldSourceProfileIds}{The identifiers of the fields in the profile that has the information
you want to apply to the merge. For example, say you want to merge
EmailAddress from Profile1 into MainProfile. This would be the
identifier of the EmailAddress field in Profile1.}
}
\description{
Runs an AWS Lambda job that does the following:

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_merge_profiles/} for full documentation.
}
\keyword{internal}
