% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_put_domain_object_type}
\alias{customerprofiles_put_domain_object_type}
\title{Create/Update a DomainObjectType in a Customer Profiles domain}
\usage{
customerprofiles_put_domain_object_type(
  DomainName,
  ObjectTypeName,
  Description = NULL,
  EncryptionKey = NULL,
  Fields,
  Tags = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ObjectTypeName}{[required] The unique name of the domain object type.}

\item{Description}{The description of the domain object type.}

\item{EncryptionKey}{The customer provided KMS key used to encrypt this type of domain
object.}

\item{Fields}{[required] A map of field names to their corresponding domain object type field
definitions.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Create/Update a DomainObjectType in a Customer Profiles domain. To create a new DomainObjectType, Data Store needs to be enabled on the Domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_put_domain_object_type/} for full documentation.
}
\keyword{internal}
