% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_put_profile_object}
\alias{customerprofiles_put_profile_object}
\title{Adds additional objects to customer profiles of a given ObjectType}
\usage{
customerprofiles_put_profile_object(ObjectTypeName, Object, DomainName)
}
\arguments{
\item{ObjectTypeName}{[required] The name of the profile object type.}

\item{Object}{[required] A string that is serialized from a JSON object.}

\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Adds additional objects to customer profiles of a given ObjectType.

See \url{https://paws-r.github.io/docs/customerprofiles/put_profile_object.html} for full documentation.
}
\keyword{internal}
