% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_put_segment_subscription}
\alias{customerprofiles_put_segment_subscription}
\title{Creates or updates a segment subscription for membership events}
\usage{
customerprofiles_put_segment_subscription(
  DomainName,
  SegmentDefinitionName,
  ScheduleConfiguration = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{SegmentDefinitionName}{[required] The unique name of the segment definition.}

\item{ScheduleConfiguration}{The optional schedule configuration that controls how often membership snapshots are run. If not provided, the subscription defaults to a 24-hour interval.}
}
\description{
Creates or updates a segment subscription for membership events. When a subscription is created, an initial snapshot is taken and the system begins monitoring for membership changes.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_put_segment_subscription/} for full documentation.
}
\keyword{internal}
