% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_start_upload_job}
\alias{customerprofiles_start_upload_job}
\title{This API starts the processing of an upload job to ingest profile data}
\usage{
customerprofiles_start_upload_job(DomainName, JobId)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain containing the upload job to start.}

\item{JobId}{[required] The unique identifier of the upload job to start.}
}
\description{
This API starts the processing of an upload job to ingest profile data.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_start_upload_job/} for full documentation.
}
\keyword{internal}
