% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_stop_recommender}
\alias{customerprofiles_stop_recommender}
\title{Stops a recommender, suspending its ability to generate recommendations}
\usage{
customerprofiles_stop_recommender(DomainName, RecommenderName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderName}{[required] The name of the recommender to stop.}
}
\description{
Stops a recommender, suspending its ability to generate recommendations. The recommender can be restarted later using StartRecommender.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_stop_recommender/} for full documentation.
}
\keyword{internal}
