% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_stop_upload_job}
\alias{customerprofiles_stop_upload_job}
\title{This API stops the processing of an upload job}
\usage{
customerprofiles_stop_upload_job(DomainName, JobId)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain containing the upload job to stop.}

\item{JobId}{[required] The unique identifier of the upload job to stop.}
}
\description{
This API stops the processing of an upload job.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_stop_upload_job/} for full documentation.
}
\keyword{internal}
