% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_tag_resource}
\alias{customerprofiles_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified Amazon
Connect Customer Profiles resource}
\usage{
customerprofiles_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource that you're adding tags to.}

\item{tags}{[required] The tags used to organize, track, or control access for this resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified Amazon Connect Customer Profiles resource. Tags can help you organize and categorize your resources. You can also use them to scope user permissions by granting a user permission to access or change only resources with certain tag values. In Connect Customer Profiles, domains, profile object types, and integrations can be tagged.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_tag_resource/} for full documentation.
}
\keyword{internal}
