% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_update_calculated_attribute_definition}
\alias{customerprofiles_update_calculated_attribute_definition}
\title{Updates an existing calculated attribute definition}
\usage{
customerprofiles_update_calculated_attribute_definition(
  DomainName,
  CalculatedAttributeName,
  DisplayName = NULL,
  Description = NULL,
  Conditions = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{CalculatedAttributeName}{[required] The unique name of the calculated attribute.}

\item{DisplayName}{The display name of the calculated attribute.}

\item{Description}{The description of the calculated attribute.}

\item{Conditions}{The conditions including range, object count, and threshold for the
calculated attribute.}
}
\description{
Updates an existing calculated attribute definition. When updating the Conditions, note that increasing the date range of a calculated attribute will not trigger inclusion of historical data greater than the current date range.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_update_calculated_attribute_definition/} for full documentation.
}
\keyword{internal}
