% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_update_domain_layout}
\alias{customerprofiles_update_domain_layout}
\title{Updates the layout used to view data for a specific domain}
\usage{
customerprofiles_update_domain_layout(
  DomainName,
  LayoutDefinitionName,
  Description = NULL,
  DisplayName = NULL,
  IsDefault = NULL,
  LayoutType = NULL,
  Layout = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{LayoutDefinitionName}{[required] The unique name of the layout.}

\item{Description}{The description of the layout}

\item{DisplayName}{The display name of the layout}

\item{IsDefault}{If set to true for a layout, this layout will be used by default to view
data. If set to false, then the layout will not be used by default, but
it can be used to view data by explicitly selecting it in the console.}

\item{LayoutType}{The type of layout that can be used to view data under a Customer
Profiles domain.}

\item{Layout}{A customizable layout that can be used to view data under a Customer
Profiles domain.}
}
\description{
Updates the layout used to view data for a specific domain. This API can only be invoked from the Amazon Connect admin website.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_update_domain_layout/} for full documentation.
}
\keyword{internal}
