% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_update_event_trigger}
\alias{customerprofiles_update_event_trigger}
\title{Update the properties of an Event Trigger}
\usage{
customerprofiles_update_event_trigger(
  DomainName,
  EventTriggerName,
  ObjectTypeName = NULL,
  Description = NULL,
  EventTriggerConditions = NULL,
  SegmentFilter = NULL,
  EventTriggerLimits = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{EventTriggerName}{[required] The unique name of the event trigger.}

\item{ObjectTypeName}{The unique name of the object type.}

\item{Description}{The description of the event trigger.}

\item{EventTriggerConditions}{A list of conditions that determine when an event should trigger the
destination.}

\item{SegmentFilter}{The destination is triggered only for profiles that meet the criteria of
a segment definition.}

\item{EventTriggerLimits}{Defines limits controlling whether an event triggers the destination,
based on ingestion latency and the number of invocations per profile
over specific time periods.}
}
\description{
Update the properties of an Event Trigger.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_update_event_trigger/} for full documentation.
}
\keyword{internal}
