% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_update_recommender}
\alias{customerprofiles_update_recommender}
\title{Updates the properties of an existing recommender, allowing you to
modify its configuration and description}
\usage{
customerprofiles_update_recommender(
  DomainName,
  RecommenderName,
  Description = NULL,
  RecommenderConfig = NULL,
  RecommenderVersionName = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{RecommenderName}{[required] The name of the recommender to update.}

\item{Description}{The new description to assign to the recommender.}

\item{RecommenderConfig}{The new configuration settings to apply to the recommender, including updated parameters and settings that define its behavior.}

\item{RecommenderVersionName}{The name of a specific recommender version to activate as part of this update (for example, to roll back to a previously trained version).}
}
\description{
Updates the properties of an existing recommender, allowing you to modify its configuration and description.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_update_recommender/} for full documentation.
}
\keyword{internal}
