% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_app}
\alias{pinpoint_create_app}
\title{Creates an application}
\usage{
pinpoint_create_app(CreateApplicationRequest)
}
\arguments{
\item{CreateApplicationRequest}{[required]}
}
\description{
Creates an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_create_app/} for full documentation.
}
\keyword{internal}
