% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_campaign}
\alias{pinpoint_create_campaign}
\title{Creates a new campaign for an application or updates the settings of an
existing campaign for an application}
\usage{
pinpoint_create_campaign(ApplicationId, WriteCampaignRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{WriteCampaignRequest}{[required]}
}
\description{
Creates a new campaign for an application or updates the settings of an existing campaign for an application.

See \url{https://paws-r.github.io/docs/pinpoint/create_campaign.html} for full documentation.
}
\keyword{internal}
