% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_export_job}
\alias{pinpoint_create_export_job}
\title{Creates an export job for an application}
\usage{
pinpoint_create_export_job(ApplicationId, ExportJobRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{ExportJobRequest}{[required] Specifies the settings for a job that exports endpoint definitions to an Amazon Simple Storage Service (Amazon S3) bucket.}
}
\description{
Creates an export job for an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_create_export_job/} for full documentation.
}
\keyword{internal}
