% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_in_app_template}
\alias{pinpoint_create_in_app_template}
\title{Creates a new message template for messages using the in-app message
channel}
\usage{
pinpoint_create_in_app_template(InAppTemplateRequest, TemplateName)
}
\arguments{
\item{InAppTemplateRequest}{[required]}

\item{TemplateName}{[required] The name of the message template. A template name must start with an
alphanumeric character and can contain a maximum of 128 characters. The
characters can be alphanumeric characters, underscores (_), or hyphens
(-). Template names are case sensitive.}
}
\description{
Creates a new message template for messages using the in-app message channel.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_create_in_app_template/} for full documentation.
}
\keyword{internal}
