% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_journey}
\alias{pinpoint_create_journey}
\title{Creates a journey for an application}
\usage{
pinpoint_create_journey(ApplicationId, WriteJourneyRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{WriteJourneyRequest}{[required]}
}
\description{
Creates a journey for an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_create_journey/} for full documentation.
}
\keyword{internal}
