% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_recommender_configuration}
\alias{pinpoint_create_recommender_configuration}
\title{Creates an Amazon Pinpoint configuration for a recommender model}
\usage{
pinpoint_create_recommender_configuration(CreateRecommenderConfiguration)
}
\arguments{
\item{CreateRecommenderConfiguration}{[required]}
}
\description{
Creates an Amazon Pinpoint configuration for a recommender model.

See \url{https://paws-r.github.io/docs/pinpoint/create_recommender_configuration.html} for full documentation.
}
\keyword{internal}
