% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_apns_voip_channel}
\alias{pinpoint_delete_apns_voip_channel}
\title{Disables the APNs VoIP channel for an application and deletes any
existing settings for the channel}
\usage{
pinpoint_delete_apns_voip_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\description{
Disables the APNs VoIP channel for an application and deletes any
existing settings for the channel.
}
\section{Request syntax}{
\preformatted{svc$delete_apns_voip_channel(
  ApplicationId = "string"
)
}
}

\keyword{internal}
