% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_app}
\alias{pinpoint_delete_app}
\title{Deletes an app}
\usage{
pinpoint_delete_app(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Deletes an app.
}
\section{Request syntax}{
\preformatted{svc$delete_app(
  ApplicationId = "string"
)
}
}

\keyword{internal}
