% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_endpoint}
\alias{pinpoint_delete_endpoint}
\title{Deletes an endpoint from an application}
\usage{
pinpoint_delete_endpoint(ApplicationId, EndpointId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{EndpointId}{[required] The unique identifier for the endpoint.}
}
\description{
Deletes an endpoint from an application.

See \url{https://paws-r.github.io/docs/pinpoint/delete_endpoint.html} for full documentation.
}
\keyword{internal}
