% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_event_stream}
\alias{pinpoint_delete_event_stream}
\title{Deletes the event stream for an application}
\usage{
pinpoint_delete_event_stream(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EventStream = list(
    ApplicationId = "string",
    DestinationStreamArn = "string",
    ExternalId = "string",
    LastModifiedDate = "string",
    LastUpdatedBy = "string",
    RoleArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the event stream for an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_event_stream(
  ApplicationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
