% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_gcm_channel}
\alias{pinpoint_delete_gcm_channel}
\title{Disables the GCM channel for an application and deletes any existing
settings for the channel}
\usage{
pinpoint_delete_gcm_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\description{
Disables the GCM channel for an application and deletes any existing settings for the channel.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_delete_gcm_channel/} for full documentation.
}
\keyword{internal}
