% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_journey}
\alias{pinpoint_delete_journey}
\title{Deletes a journey from an application}
\usage{
pinpoint_delete_journey(ApplicationId, JourneyId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{JourneyId}{[required] The unique identifier for the journey.}
}
\description{
Deletes a journey from an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_delete_journey/} for full documentation.
}
\keyword{internal}
