% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_push_template}
\alias{pinpoint_delete_push_template}
\title{Deletes a message template that was designed for use in messages that
were sent through a push notification channel}
\usage{
pinpoint_delete_push_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the message template. A template name must start with an
alphanumeric character and can contain a maximum of 128 characters. The
characters can be alphanumeric characters, underscores (\\_), or hyphens
(-). Template names are case sensitive.}
}
\description{
Deletes a message template that was designed for use in messages that
were sent through a push notification channel.
}
\section{Request syntax}{
\preformatted{svc$delete_push_template(
  TemplateName = "string"
)
}
}

\keyword{internal}
