% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_user_endpoints}
\alias{pinpoint_delete_user_endpoints}
\title{Deletes all the endpoints that are associated with a specific user ID}
\usage{
pinpoint_delete_user_endpoints(ApplicationId, UserId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{UserId}{[required] The unique identifier for the user.}
}
\description{
Deletes all the endpoints that are associated with a specific user ID.
}
\section{Request syntax}{
\preformatted{svc$delete_user_endpoints(
  ApplicationId = "string",
  UserId = "string"
)
}
}

\keyword{internal}
