% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_apps}
\alias{pinpoint_get_apps}
\title{Retrieves information about all of your applications}
\usage{
pinpoint_get_apps(PageSize, Token)
}
\arguments{
\item{PageSize}{The maximum number of items to include on each page in a paginated
response.}

\item{Token}{The NextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about all of your applications.
}
\section{Request syntax}{
\preformatted{svc$get_apps(
  PageSize = "string",
  Token = "string"
)
}
}

\keyword{internal}
