% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_campaign}
\alias{pinpoint_get_campaign}
\title{Retrieves information about the status, configuration, and other
settings for a campaign}
\usage{
pinpoint_get_campaign(ApplicationId, CampaignId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{CampaignId}{[required] The unique identifier for the campaign.}
}
\description{
Retrieves information about the status, configuration, and other settings for a campaign.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_get_campaign/} for full documentation.
}
\keyword{internal}
