% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_campaign_versions}
\alias{pinpoint_get_campaign_versions}
\title{Retrieves information about the status, configuration, and other
settings for all versions of a specific campaign}
\usage{
pinpoint_get_campaign_versions(ApplicationId, CampaignId, PageSize,
  Token)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{CampaignId}{[required] The unique identifier for the campaign.}

\item{PageSize}{The maximum number of items to include on each page in a paginated
response.}

\item{Token}{The NextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about the status, configuration, and other
settings for all versions of a specific campaign.
}
\section{Request syntax}{
\preformatted{svc$get_campaign_versions(
  ApplicationId = "string",
  CampaignId = "string",
  PageSize = "string",
  Token = "string"
)
}
}

\keyword{internal}
