% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_export_job}
\alias{pinpoint_get_export_job}
\title{Returns information about an export job}
\usage{
pinpoint_get_export_job(ApplicationId, JobId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{JobId}{[required] The unique ID of the job.}
}
\description{
Returns information about an export job.
}
\section{Request syntax}{
\preformatted{svc$get_export_job(
  ApplicationId = "string",
  JobId = "string"
)
}
}

\keyword{internal}
