% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_export_jobs}
\alias{pinpoint_get_export_jobs}
\title{Returns information about your export jobs}
\usage{
pinpoint_get_export_jobs(ApplicationId, PageSize, Token)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{PageSize}{The number of entries you want on each page in the response.}

\item{Token}{The NextToken string returned on a previous page that you use to get the next page of results in a paginated response.}
}
\description{
Returns information about your export jobs.
}
\section{Request syntax}{
\preformatted{svc$get_export_jobs(
  ApplicationId = "string",
  PageSize = "string",
  Token = "string"
)
}
}

\keyword{internal}
