% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_import_job}
\alias{pinpoint_get_import_job}
\title{Retrieves information about the status and settings of a specific import
job for an application}
\usage{
pinpoint_get_import_job(ApplicationId, JobId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{JobId}{[required] The unique identifier for the job.}
}
\description{
Retrieves information about the status and settings of a specific import job for an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_get_import_job/} for full documentation.
}
\keyword{internal}
