% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_import_jobs}
\alias{pinpoint_get_import_jobs}
\title{Retrieves information about the status and settings of all the import
jobs for an application}
\usage{
pinpoint_get_import_jobs(ApplicationId, PageSize, Token)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{PageSize}{The maximum number of items to include in each page of a paginated
response. This parameter is not supported for application, campaign, and
journey metrics.}

\item{Token}{The NextToken string that specifies which page of results to return in a
paginated response.}
}
\value{
A list with the following syntax:\preformatted{list(
  ImportJobsResponse = list(
    Item = list(
      list(
        ApplicationId = "string",
        CompletedPieces = 123,
        CompletionDate = "string",
        CreationDate = "string",
        Definition = list(
          DefineSegment = TRUE|FALSE,
          ExternalId = "string",
          Format = "CSV"|"JSON",
          RegisterEndpoints = TRUE|FALSE,
          RoleArn = "string",
          S3Url = "string",
          SegmentId = "string",
          SegmentName = "string"
        ),
        FailedPieces = 123,
        Failures = list(
          "string"
        ),
        Id = "string",
        JobStatus = "CREATED"|"PREPARING_FOR_INITIALIZATION"|"INITIALIZING"|"PROCESSING"|"PENDING_JOB"|"COMPLETING"|"COMPLETED"|"FAILING"|"FAILED",
        TotalFailures = 123,
        TotalPieces = 123,
        TotalProcessed = 123,
        Type = "string"
      )
    ),
    NextToken = "string"
  )
)
}
}
\description{
Retrieves information about the status and settings of all the import
jobs for an application.
}
\section{Request syntax}{
\preformatted{svc$get_import_jobs(
  ApplicationId = "string",
  PageSize = "string",
  Token = "string"
)
}
}

\keyword{internal}
