% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_in_app_messages}
\alias{pinpoint_get_in_app_messages}
\title{Retrieves the in-app messages targeted for the provided endpoint ID}
\usage{
pinpoint_get_in_app_messages(ApplicationId, EndpointId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{EndpointId}{[required] The unique identifier for the endpoint.}
}
\description{
Retrieves the in-app messages targeted for the provided endpoint ID.

See \url{https://paws-r.github.io/docs/pinpoint/get_in_app_messages.html} for full documentation.
}
\keyword{internal}
