% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_journey}
\alias{pinpoint_get_journey}
\title{Retrieves information about the status, configuration, and other
settings for a journey}
\usage{
pinpoint_get_journey(ApplicationId, JourneyId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{JourneyId}{[required] The unique identifier for the journey.}
}
\description{
Retrieves information about the status, configuration, and other settings for a journey.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_get_journey/} for full documentation.
}
\keyword{internal}
