% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_journey_runs}
\alias{pinpoint_get_journey_runs}
\title{Provides information about the runs of a journey}
\usage{
pinpoint_get_journey_runs(
  ApplicationId,
  JourneyId,
  PageSize = NULL,
  Token = NULL
)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{JourneyId}{[required] The unique identifier for the journey.}

\item{PageSize}{The maximum number of items to include in each page of a paginated
response. This parameter is not supported for application, campaign, and
journey metrics.}

\item{Token}{The NextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Provides information about the runs of a journey.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_get_journey_runs/} for full documentation.
}
\keyword{internal}
