% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_recommender_configuration}
\alias{pinpoint_get_recommender_configuration}
\title{Retrieves information about an Amazon Pinpoint configuration for a
recommender model}
\usage{
pinpoint_get_recommender_configuration(RecommenderId)
}
\arguments{
\item{RecommenderId}{[required] The unique identifier for the recommender model configuration. This
identifier is displayed as the \strong{Recommender ID} on the Amazon Pinpoint
console.}
}
\description{
Retrieves information about an Amazon Pinpoint configuration for a
recommender model.
}
\section{Request syntax}{
\preformatted{svc$get_recommender_configuration(
  RecommenderId = "string"
)
}
}

\keyword{internal}
