% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_segment}
\alias{pinpoint_get_segment}
\title{Retrieves information about the configuration, dimension, and other
settings for a specific segment that's associated with an application}
\usage{
pinpoint_get_segment(ApplicationId, SegmentId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{SegmentId}{[required] The unique identifier for the segment.}
}
\description{
Retrieves information about the configuration, dimension, and other
settings for a specific segment that's associated with an application.
}
\section{Request syntax}{
\preformatted{svc$get_segment(
  ApplicationId = "string",
  SegmentId = "string"
)
}
}

\keyword{internal}
