% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_segment_version}
\alias{pinpoint_get_segment_version}
\title{Retrieves information about the configuration, dimension, and other
settings for a specific version of a segment that's associated with an
application}
\usage{
pinpoint_get_segment_version(ApplicationId, SegmentId, Version)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{SegmentId}{[required] The unique identifier for the segment.}

\item{Version}{[required] The unique version number (Version property) for the campaign version.}
}
\description{
Retrieves information about the configuration, dimension, and other settings for a specific version of a segment that's associated with an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_get_segment_version/} for full documentation.
}
\keyword{internal}
