% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_sms_template}
\alias{pinpoint_get_sms_template}
\title{Retrieves the content and settings for a message template that you can
use in messages that are sent through the SMS channel}
\usage{
pinpoint_get_sms_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the message template. A template name must start with an
alphanumeric character and can contain a maximum of 128 characters. The
characters can be alphanumeric characters, underscores (\\_), or hyphens
(-). Template names are case sensitive.}
}
\description{
Retrieves the content and settings for a message template that you can
use in messages that are sent through the SMS channel.
}
\section{Request syntax}{
\preformatted{svc$get_sms_template(
  TemplateName = "string"
)
}
}

\keyword{internal}
