% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_voice_channel}
\alias{pinpoint_get_voice_channel}
\title{Retrieves information about the status and settings of the voice channel
for an application}
\usage{
pinpoint_get_voice_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  VoiceChannelResponse = list(
    ApplicationId = "string",
    CreationDate = "string",
    Enabled = TRUE|FALSE,
    HasCredential = TRUE|FALSE,
    Id = "string",
    IsArchived = TRUE|FALSE,
    LastModifiedBy = "string",
    LastModifiedDate = "string",
    Platform = "string",
    Version = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information about the status and settings of the voice channel
for an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_voice_channel(
  ApplicationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
