% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_list_journeys}
\alias{pinpoint_list_journeys}
\title{Retrieves information about the status, configuration, and other
settings for all the journeys that are associated with an application}
\usage{
pinpoint_list_journeys(ApplicationId, PageSize, Token)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{PageSize}{The maximum number of items to include in each page of a paginated
response. This parameter is currently not supported for application,
campaign, and journey metrics.}

\item{Token}{The NextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about the status, configuration, and other
settings for all the journeys that are associated with an application.
}
\section{Request syntax}{
\preformatted{svc$list_journeys(
  ApplicationId = "string",
  PageSize = "string",
  Token = "string"
)
}
}

\keyword{internal}
