% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_list_tags_for_resource}
\alias{pinpoint_list_tags_for_resource}
\title{Retrieves all the tags (keys and values) that are associated with an
application, campaign, message template, or segment}
\usage{
pinpoint_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TagsModel = list(
    tags = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves all the tags (keys and values) that are associated with an
application, campaign, message template, or segment.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
