% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_list_template_versions}
\alias{pinpoint_list_template_versions}
\title{Retrieves information about all the versions of a specific message
template}
\usage{
pinpoint_list_template_versions(NextToken, PageSize, TemplateName,
  TemplateType)
}
\arguments{
\item{NextToken}{The string that specifies which page of results to return in a paginated
response. This parameter is not supported for application, campaign, and
journey metrics.}

\item{PageSize}{The maximum number of items to include in each page of a paginated
response. This parameter is not supported for application, campaign, and
journey metrics.}

\item{TemplateName}{[required] The name of the message template. A template name must start with an
alphanumeric character and can contain a maximum of 128 characters. The
characters can be alphanumeric characters, underscores (_), or hyphens
(-). Template names are case sensitive.}

\item{TemplateType}{[required] The type of channel that the message template is designed for. Valid
values are: EMAIL, PUSH, SMS, and VOICE.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TemplateVersionsResponse = list(
    Item = list(
      list(
        CreationDate = "string",
        DefaultSubstitutions = "string",
        LastModifiedDate = "string",
        TemplateDescription = "string",
        TemplateName = "string",
        TemplateType = "string",
        Version = "string"
      )
    ),
    Message = "string",
    NextToken = "string",
    RequestID = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information about all the versions of a specific message
template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_template_versions(
  NextToken = "string",
  PageSize = "string",
  TemplateName = "string",
  TemplateType = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
