% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_phone_number_validate}
\alias{pinpoint_phone_number_validate}
\title{Retrieves information about a phone number}
\usage{
pinpoint_phone_number_validate(NumberValidateRequest)
}
\arguments{
\item{NumberValidateRequest}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NumberValidateResponse = list(
    Carrier = "string",
    City = "string",
    CleansedPhoneNumberE164 = "string",
    CleansedPhoneNumberNational = "string",
    Country = "string",
    CountryCodeIso2 = "string",
    CountryCodeNumeric = "string",
    County = "string",
    OriginalCountryCodeIso2 = "string",
    OriginalPhoneNumber = "string",
    PhoneType = "string",
    PhoneTypeCode = 123,
    Timezone = "string",
    ZipCode = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information about a phone number.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$phone_number_validate(
  NumberValidateRequest = list(
    IsoCountryCode = "string",
    PhoneNumber = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
