% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_put_event_stream}
\alias{pinpoint_put_event_stream}
\title{Creates a new event stream for an application or updates the settings of
an existing event stream for an application}
\usage{
pinpoint_put_event_stream(ApplicationId, WriteEventStream)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{WriteEventStream}{[required]}
}
\description{
Creates a new event stream for an application or updates the settings of an existing event stream for an application.

See \url{https://paws-r.github.io/docs/pinpoint/put_event_stream.html} for full documentation.
}
\keyword{internal}
