% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_put_events}
\alias{pinpoint_put_events}
\title{Creates a new event to record for endpoints, or creates or updates
endpoint data that existing events are associated with}
\usage{
pinpoint_put_events(ApplicationId, EventsRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{EventsRequest}{[required]}
}
\description{
Creates a new event to record for endpoints, or creates or updates endpoint data that existing events are associated with.

See \url{https://paws-r.github.io/docs/pinpoint/put_events.html} for full documentation.
}
\keyword{internal}
