% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_send_messages}
\alias{pinpoint_send_messages}
\title{Creates and sends a direct message}
\usage{
pinpoint_send_messages(ApplicationId, MessageRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{MessageRequest}{[required]}
}
\description{
Creates and sends a direct message.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_send_messages/} for full documentation.
}
\keyword{internal}
