% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_send_otp_message}
\alias{pinpoint_send_otp_message}
\title{Send an OTP message}
\usage{
pinpoint_send_otp_message(ApplicationId, SendOTPMessageRequestParameters)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{SendOTPMessageRequestParameters}{[required]}
}
\description{
Send an OTP message

See \url{https://www.paws-r-sdk.com/docs/pinpoint_send_otp_message/} for full documentation.
}
\keyword{internal}
