% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_tag_resource}
\alias{pinpoint_tag_resource}
\title{Adds one or more tags (keys and values) to an application, campaign,
message template, or segment}
\usage{
pinpoint_tag_resource(ResourceArn, TagsModel)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{TagsModel}{[required]}
}
\description{
Adds one or more tags (keys and values) to an application, campaign, message template, or segment.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_tag_resource/} for full documentation.
}
\keyword{internal}
