% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_untag_resource}
\alias{pinpoint_untag_resource}
\title{Removes one or more tags (keys and values) from an application,
campaign, message template, or segment}
\usage{
pinpoint_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{TagKeys}{[required] The key of the tag to remove from the resource. To remove multiple tags,
append the tagKeys parameter and argument for each additional tag to
remove, separated by an ampersand (&).}
}
\description{
Removes one or more tags (keys and values) from an application, campaign, message template, or segment.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_untag_resource/} for full documentation.
}
\keyword{internal}
