% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_adm_channel}
\alias{pinpoint_update_adm_channel}
\title{Enables the ADM channel for an application or updates the status and
settings of the ADM channel for an application}
\usage{
pinpoint_update_adm_channel(ADMChannelRequest, ApplicationId)
}
\arguments{
\item{ADMChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\description{
Enables the ADM channel for an application or updates the status and settings of the ADM channel for an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_update_adm_channel/} for full documentation.
}
\keyword{internal}
