% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_apns_channel}
\alias{pinpoint_update_apns_channel}
\title{Use to update the APNs channel for an app}
\usage{
pinpoint_update_apns_channel(APNSChannelRequest, ApplicationId)
}
\arguments{
\item{APNSChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}
}
\description{
Use to update the APNs channel for an app.
}
\section{Request syntax}{
\preformatted{svc$update_apns_channel(
  APNSChannelRequest = list(
    BundleId = "string",
    Certificate = "string",
    DefaultAuthenticationMethod = "string",
    Enabled = TRUE|FALSE,
    PrivateKey = "string",
    TeamId = "string",
    TokenKey = "string",
    TokenKeyId = "string"
  ),
  ApplicationId = "string"
)
}
}

\keyword{internal}
