% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_application_settings}
\alias{pinpoint_update_application_settings}
\title{Updates the settings for an application}
\usage{
pinpoint_update_application_settings(
  ApplicationId,
  WriteApplicationSettingsRequest
)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{WriteApplicationSettingsRequest}{[required]}
}
\description{
Updates the settings for an application.

See \url{https://paws-r.github.io/docs/pinpoint/update_application_settings.html} for full documentation.
}
\keyword{internal}
