% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_campaign}
\alias{pinpoint_update_campaign}
\title{Updates the configuration and other settings for a campaign}
\usage{
pinpoint_update_campaign(ApplicationId, CampaignId, WriteCampaignRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{CampaignId}{[required] The unique identifier for the campaign.}

\item{WriteCampaignRequest}{[required]}
}
\description{
Updates the configuration and other settings for a campaign.

See \url{https://paws-r.github.io/docs/pinpoint/update_campaign.html} for full documentation.
}
\keyword{internal}
