% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_endpoints_batch}
\alias{pinpoint_update_endpoints_batch}
\title{Creates a new batch of endpoints for an application or updates the
settings and attributes of a batch of existing endpoints for an
application}
\usage{
pinpoint_update_endpoints_batch(ApplicationId, EndpointBatchRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{EndpointBatchRequest}{[required]}
}
\description{
Creates a new batch of endpoints for an application or updates the settings and attributes of a batch of existing endpoints for an application. You can also use this operation to define custom attributes for a batch of endpoints. If an update includes one or more values for a custom attribute, Amazon Pinpoint replaces (overwrites) any existing values with the new values.

See \url{https://paws-r.github.io/docs/pinpoint/update_endpoints_batch.html} for full documentation.
}
\keyword{internal}
