% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_journey}
\alias{pinpoint_update_journey}
\title{Updates the configuration and other settings for a journey}
\usage{
pinpoint_update_journey(ApplicationId, JourneyId, WriteJourneyRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{JourneyId}{[required] The unique identifier for the journey.}

\item{WriteJourneyRequest}{[required]}
}
\description{
Updates the configuration and other settings for a journey.

See \url{https://paws-r.github.io/docs/pinpoint/update_journey.html} for full documentation.
}
\keyword{internal}
