% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_journey_state}
\alias{pinpoint_update_journey_state}
\title{Cancels (stops) an active journey}
\usage{
pinpoint_update_journey_state(ApplicationId, JourneyId, JourneyStateRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{JourneyId}{[required] The unique identifier for the journey.}

\item{JourneyStateRequest}{[required]}
}
\description{
Cancels (stops) an active journey.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_update_journey_state/} for full documentation.
}
\keyword{internal}
