% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_segment}
\alias{pinpoint_update_segment}
\title{Creates a new segment for an application or updates the configuration,
dimension, and other settings for an existing segment that's associated
with an application}
\usage{
pinpoint_update_segment(ApplicationId, SegmentId, WriteSegmentRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{SegmentId}{[required] The unique identifier for the segment.}

\item{WriteSegmentRequest}{[required]}
}
\description{
Creates a new segment for an application or updates the configuration, dimension, and other settings for an existing segment that's associated with an application.

See \url{https://paws-r.github.io/docs/pinpoint/update_segment.html} for full documentation.
}
\keyword{internal}
