% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_template_active_version}
\alias{pinpoint_update_template_active_version}
\title{Changes the status of a specific version of a message template to active}
\usage{
pinpoint_update_template_active_version(TemplateActiveVersionRequest,
  TemplateName, TemplateType)
}
\arguments{
\item{TemplateActiveVersionRequest}{[required]}

\item{TemplateName}{[required] The name of the message template. A template name must start with an
alphanumeric character and can contain a maximum of 128 characters. The
characters can be alphanumeric characters, underscores (_), or hyphens
(-). Template names are case sensitive.}

\item{TemplateType}{[required] The type of channel that the message template is designed for. Valid
values are: EMAIL, PUSH, SMS, and VOICE.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  MessageBody = list(
    Message = "string",
    RequestID = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Changes the status of a specific version of a message template to
\emph{active}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_template_active_version(
  TemplateActiveVersionRequest = list(
    Version = "string"
  ),
  TemplateName = "string",
  TemplateType = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
