% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_voice_channel}
\alias{pinpoint_update_voice_channel}
\title{Enables the voice channel for an application or updates the status and
settings of the voice channel for an application}
\usage{
pinpoint_update_voice_channel(ApplicationId, VoiceChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{VoiceChannelRequest}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  VoiceChannelResponse = list(
    ApplicationId = "string",
    CreationDate = "string",
    Enabled = TRUE|FALSE,
    HasCredential = TRUE|FALSE,
    Id = "string",
    IsArchived = TRUE|FALSE,
    LastModifiedBy = "string",
    LastModifiedDate = "string",
    Platform = "string",
    Version = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Enables the voice channel for an application or updates the status and
settings of the voice channel for an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_voice_channel(
  ApplicationId = "string",
  VoiceChannelRequest = list(
    Enabled = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
