% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_create_configuration_set_event_destination}
\alias{pinpointemail_create_configuration_set_event_destination}
\title{Create an event destination}
\usage{
pinpointemail_create_configuration_set_event_destination(
  ConfigurationSetName,
  EventDestinationName,
  EventDestination
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to add an event
destination to.}

\item{EventDestinationName}{[required] A name that identifies the event destination within the configuration
set.}

\item{EventDestination}{[required] An object that defines the event destination.}
}
\description{
Create an event destination. In Amazon Pinpoint, \emph{events} include message sends, deliveries, opens, clicks, bounces, and complaints. \emph{Event destinations} are places that you can send information about these events to. For example, you can send event data to Amazon SNS to receive notifications when you receive bounces or complaints, or you can use Amazon Kinesis Data Firehose to stream data to Amazon S3 for long-term storage.

See \url{https://www.paws-r-sdk.com/docs/pinpointemail_create_configuration_set_event_destination/} for full documentation.
}
\keyword{internal}
