% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_create_dedicated_ip_pool}
\alias{pinpointemail_create_dedicated_ip_pool}
\title{Create a new pool of dedicated IP addresses}
\usage{
pinpointemail_create_dedicated_ip_pool(PoolName, Tags)
}
\arguments{
\item{PoolName}{[required] The name of the dedicated IP pool.}

\item{Tags}{An object that defines the tags (keys and values) that you want to
associate with the pool.}
}
\value{
An empty list.
}
\description{
Create a new pool of dedicated IP addresses. A pool can include one or
more dedicated IP addresses that are associated with your Amazon
Pinpoint account. You can associate a pool with a configuration set.
When you send an email that uses that configuration set, Amazon Pinpoint
sends it using only the IP addresses in the associated pool.
}
\section{Request syntax}{
\preformatted{svc$create_dedicated_ip_pool(
  PoolName = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
