% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_delete_configuration_set}
\alias{pinpointemail_delete_configuration_set}
\title{Delete an existing configuration set}
\usage{
pinpointemail_delete_configuration_set(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to delete.}
}
\description{
Delete an existing configuration set.

See \url{https://www.paws-r-sdk.com/docs/pinpointemail_delete_configuration_set/} for full documentation.
}
\keyword{internal}
