% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_delete_dedicated_ip_pool}
\alias{pinpointemail_delete_dedicated_ip_pool}
\title{Delete a dedicated IP pool}
\usage{
pinpointemail_delete_dedicated_ip_pool(PoolName)
}
\arguments{
\item{PoolName}{[required] The name of the dedicated IP pool that you want to delete.}
}
\description{
Delete a dedicated IP pool.

See \url{https://paws-r.github.io/docs/pinpointemail/delete_dedicated_ip_pool.html} for full documentation.
}
\keyword{internal}
