% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_get_account}
\alias{pinpointemail_get_account}
\title{Obtain information about the email-sending status and capabilities of
your Amazon Pinpoint account in the current AWS Region}
\usage{
pinpointemail_get_account()
}
\description{
Obtain information about the email-sending status and capabilities of your Amazon Pinpoint account in the current AWS Region.

See \url{https://www.paws-r-sdk.com/docs/pinpointemail_get_account/} for full documentation.
}
\keyword{internal}
