% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_get_blacklist_reports}
\alias{pinpointemail_get_blacklist_reports}
\title{Retrieve a list of the blacklists that your dedicated IP addresses
appear on}
\usage{
pinpointemail_get_blacklist_reports(BlacklistItemNames)
}
\arguments{
\item{BlacklistItemNames}{[required] A list of IP addresses that you want to retrieve blacklist information
about. You can only specify the dedicated IP addresses that you use to
send email using Amazon Pinpoint or Amazon SES.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BlacklistReport = list(
    list(
      list(
        RblName = "string",
        ListingTime = as.POSIXct(
          "2015-01-01"
        ),
        Description = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieve a list of the blacklists that your dedicated IP addresses
appear on.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_blacklist_reports(
  BlacklistItemNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
