% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_get_dedicated_ip}
\alias{pinpointemail_get_dedicated_ip}
\title{Get information about a dedicated IP address, including the name of the
dedicated IP pool that it's associated with, as well information about
the automatic warm-up process for the address}
\usage{
pinpointemail_get_dedicated_ip(Ip)
}
\arguments{
\item{Ip}{[required] The IP address that you want to obtain more information about. The value
you specify has to be a dedicated IP address that's assocaited with your
Amazon Pinpoint account.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DedicatedIp = list(
    Ip = "string",
    WarmupStatus = "IN_PROGRESS"|"DONE",
    WarmupPercentage = 123,
    PoolName = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Get information about a dedicated IP address, including the name of the
dedicated IP pool that it's associated with, as well information about
the automatic warm-up process for the address.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_dedicated_ip(
  Ip = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
