% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_list_configuration_sets}
\alias{pinpointemail_list_configuration_sets}
\title{List all of the configuration sets associated with your Amazon Pinpoint
account in the current region}
\usage{
pinpointemail_list_configuration_sets(NextToken = NULL, PageSize = NULL)
}
\arguments{
\item{NextToken}{A token returned from a previous call to
\code{\link[=pinpointemail_list_configuration_sets]{list_configuration_sets}} to
indicate the position in the list of configuration sets.}

\item{PageSize}{The number of results to show in a single call to
\code{\link[=pinpointemail_list_configuration_sets]{list_configuration_sets}}. If
the number of results is larger than the number you specified in this
parameter, then the response includes a \code{NextToken} element, which you
can use to obtain additional results.}
}
\description{
List all of the configuration sets associated with your Amazon Pinpoint account in the current region.

See \url{https://paws-r.github.io/docs/pinpointemail/list_configuration_sets.html} for full documentation.
}
\keyword{internal}
