% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_list_dedicated_ip_pools}
\alias{pinpointemail_list_dedicated_ip_pools}
\title{List all of the dedicated IP pools that exist in your Amazon Pinpoint
account in the current AWS Region}
\usage{
pinpointemail_list_dedicated_ip_pools(NextToken = NULL, PageSize = NULL)
}
\arguments{
\item{NextToken}{A token returned from a previous call to
\code{\link[=pinpointemail_list_dedicated_ip_pools]{list_dedicated_ip_pools}} to
indicate the position in the list of dedicated IP pools.}

\item{PageSize}{The number of results to show in a single call to
\code{\link[=pinpointemail_list_dedicated_ip_pools]{list_dedicated_ip_pools}}. If
the number of results is larger than the number you specified in this
parameter, then the response includes a \code{NextToken} element, which you
can use to obtain additional results.}
}
\description{
List all of the dedicated IP pools that exist in your Amazon Pinpoint account in the current AWS Region.

See \url{https://paws-r.github.io/docs/pinpointemail/list_dedicated_ip_pools.html} for full documentation.
}
\keyword{internal}
