% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_list_tags_for_resource}
\alias{pinpointemail_list_tags_for_resource}
\title{Retrieve a list of the tags (keys and values) that are associated with a
specified resource}
\usage{
pinpointemail_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to retrieve
tag information for.}
}
\description{
Retrieve a list of the tags (keys and values) that are associated with a specified resource. A \emph{tag} is a label that you optionally define and associate with a resource in Amazon Pinpoint. Each tag consists of a required \emph{tag key} and an optional associated \emph{tag value}. A tag key is a general label that acts as a category for more specific tag values. A tag value acts as a descriptor within a tag key.

See \url{https://paws-r.github.io/docs/pinpointemail/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
