% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_put_configuration_set_sending_options}
\alias{pinpointemail_put_configuration_set_sending_options}
\title{Enable or disable email sending for messages that use a particular
configuration set in a specific AWS Region}
\usage{
pinpointemail_put_configuration_set_sending_options(
  ConfigurationSetName,
  SendingEnabled = NULL
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to enable or disable
email sending for.}

\item{SendingEnabled}{If \code{true}, email sending is enabled for the configuration set. If
\code{false}, email sending is disabled for the configuration set.}
}
\description{
Enable or disable email sending for messages that use a particular configuration set in a specific AWS Region.

See \url{https://www.paws-r-sdk.com/docs/pinpointemail_put_configuration_set_sending_options/} for full documentation.
}
\keyword{internal}
