% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_put_configuration_set_tracking_options}
\alias{pinpointemail_put_configuration_set_tracking_options}
\title{Specify a custom domain to use for open and click tracking elements in
email that you send using Amazon Pinpoint}
\usage{
pinpointemail_put_configuration_set_tracking_options(
  ConfigurationSetName, CustomRedirectDomain)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to add a custom tracking
domain to.}

\item{CustomRedirectDomain}{The domain that you want to use to track open and click events.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Specify a custom domain to use for open and click tracking elements in
email that you send using Amazon Pinpoint.
}
\section{Request syntax}{
\preformatted{svc$put_configuration_set_tracking_options(
  ConfigurationSetName = "string",
  CustomRedirectDomain = "string"
)
}
}

\keyword{internal}
