% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_put_dedicated_ip_in_pool}
\alias{pinpointemail_put_dedicated_ip_in_pool}
\title{Move a dedicated IP address to an existing dedicated IP pool}
\usage{
pinpointemail_put_dedicated_ip_in_pool(Ip, DestinationPoolName)
}
\arguments{
\item{Ip}{[required] The IP address that you want to move to the dedicated IP pool. The value
you specify has to be a dedicated IP address that's associated with your
Amazon Pinpoint account.}

\item{DestinationPoolName}{[required] The name of the IP pool that you want to add the dedicated IP address
to. You have to specify an IP pool that already exists.}
}
\description{
Move a dedicated IP address to an existing dedicated IP pool.

See \url{https://www.paws-r-sdk.com/docs/pinpointemail_put_dedicated_ip_in_pool/} for full documentation.
}
\keyword{internal}
