% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_put_dedicated_ip_warmup_attributes}
\alias{pinpointemail_put_dedicated_ip_warmup_attributes}
\title{Put dedicated ip warmup attributes}
\usage{
pinpointemail_put_dedicated_ip_warmup_attributes(Ip, WarmupPercentage)
}
\arguments{
\item{Ip}{[required] The dedicated IP address that you want to update the warm-up attributes
for.}

\item{WarmupPercentage}{[required] The warm-up percentage that you want to associate with the dedicated IP
address.}
}
\description{
Put dedicated ip warmup attributes

See \url{https://paws-r.github.io/docs/pinpointemail/put_dedicated_ip_warmup_attributes.html} for full documentation.
}
\keyword{internal}
