% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_tag_resource}
\alias{pinpointemail_tag_resource}
\title{Add one or more tags (keys and values) to a specified resource}
\usage{
pinpointemail_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to add one
or more tags to.}

\item{Tags}{[required] A list of the tags that you want to add to the resource. A tag consists
of a required tag key (\code{Key}) and an associated tag value (\code{Value}). The
maximum length of a tag key is 128 characters. The maximum length of a
tag value is 256 characters.}
}
\description{
Add one or more tags (keys and values) to a specified resource. A \emph{tag} is a label that you optionally define and associate with a resource in Amazon Pinpoint. Tags can help you categorize and manage resources in different ways, such as by purpose, owner, environment, or other criteria. A resource can have as many as 50 tags.

See \url{https://www.paws-r-sdk.com/docs/pinpointemail_tag_resource/} for full documentation.
}
\keyword{internal}
