% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_update_configuration_set_event_destination}
\alias{pinpointemail_update_configuration_set_event_destination}
\title{Update the configuration of an event destination for a configuration set}
\usage{
pinpointemail_update_configuration_set_event_destination(
  ConfigurationSetName,
  EventDestinationName,
  EventDestination
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that contains the event destination
that you want to modify.}

\item{EventDestinationName}{[required] The name of the event destination that you want to modify.}

\item{EventDestination}{[required] An object that defines the event destination.}
}
\description{
Update the configuration of an event destination for a configuration set.

See \url{https://paws-r.github.io/docs/pinpointemail/update_configuration_set_event_destination.html} for full documentation.
}
\keyword{internal}
