% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoice_operations.R
\name{pinpointsmsvoice_create_configuration_set}
\alias{pinpointsmsvoice_create_configuration_set}
\title{Create a new configuration set}
\usage{
pinpointsmsvoice_create_configuration_set(ConfigurationSetName = NULL)
}
\arguments{
\item{ConfigurationSetName}{The name that you want to give the configuration set.}
}
\description{
Create a new configuration set. After you create the configuration set, you can add one or more event destinations to it.

See \url{https://paws-r.github.io/docs/pinpointsmsvoice/create_configuration_set.html} for full documentation.
}
\keyword{internal}
