% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoice_operations.R
\name{pinpointsmsvoice_create_configuration_set_event_destination}
\alias{pinpointsmsvoice_create_configuration_set_event_destination}
\title{Create a new event destination in a configuration set}
\usage{
pinpointsmsvoice_create_configuration_set_event_destination(
  ConfigurationSetName,
  EventDestination = NULL,
  EventDestinationName = NULL
)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}

\item{EventDestination}{}

\item{EventDestinationName}{A name that identifies the event destination.}
}
\description{
Create a new event destination in a configuration set.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoice_create_configuration_set_event_destination/} for full documentation.
}
\keyword{internal}
