% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoice_operations.R
\name{pinpointsmsvoice_delete_configuration_set}
\alias{pinpointsmsvoice_delete_configuration_set}
\title{Deletes an existing configuration set}
\usage{
pinpointsmsvoice_delete_configuration_set(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes an existing configuration set.
}
\section{Request syntax}{
\preformatted{svc$delete_configuration_set(
  ConfigurationSetName = "string"
)
}
}

\keyword{internal}
