% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoice_operations.R
\name{pinpointsmsvoice_delete_configuration_set_event_destination}
\alias{pinpointsmsvoice_delete_configuration_set_event_destination}
\title{Deletes an event destination in a configuration set}
\usage{
pinpointsmsvoice_delete_configuration_set_event_destination(
  ConfigurationSetName, EventDestinationName)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}

\item{EventDestinationName}{[required] EventDestinationName}
}
\value{
An empty list.
}
\description{
Deletes an event destination in a configuration set.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_configuration_set_event_destination(
  ConfigurationSetName = "string",
  EventDestinationName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
