% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoice_operations.R
\name{pinpointsmsvoice_update_configuration_set_event_destination}
\alias{pinpointsmsvoice_update_configuration_set_event_destination}
\title{Update an event destination in a configuration set}
\usage{
pinpointsmsvoice_update_configuration_set_event_destination(
  ConfigurationSetName,
  EventDestination = NULL,
  EventDestinationName
)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}

\item{EventDestination}{}

\item{EventDestinationName}{[required] EventDestinationName}
}
\description{
Update an event destination in a configuration set. An event destination is a location that you publish information about your voice calls to. For example, you can log an event to an Amazon CloudWatch destination when a call fails.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoice_update_configuration_set_event_destination/} for full documentation.
}
\keyword{internal}
