% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_associate_protect_configuration}
\alias{pinpointsmsvoicev2_associate_protect_configuration}
\title{Associate a protect configuration with a configuration set}
\usage{
pinpointsmsvoicev2_associate_protect_configuration(
  ProtectConfigurationId,
  ConfigurationSetName
)
}
\arguments{
\item{ProtectConfigurationId}{[required] The unique identifier for the protect configuration.}

\item{ConfigurationSetName}{[required] The name of the ConfigurationSet.}
}
\description{
Associate a protect configuration with a configuration set. This replaces the configuration sets current protect configuration. A configuration set can only be associated with one protect configuration at a time. A protect configuration can be associated with multiple configuration sets.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_associate_protect_configuration/} for full documentation.
}
\keyword{internal}
