% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_create_notify_configuration}
\alias{pinpointsmsvoicev2_create_notify_configuration}
\title{Creates a new notify configuration for managed messaging}
\usage{
pinpointsmsvoicev2_create_notify_configuration(
  DisplayName,
  UseCase,
  DefaultTemplateId = NULL,
  PoolId = NULL,
  EnabledCountries = NULL,
  EnabledChannels,
  DeletionProtectionEnabled = NULL,
  ClientToken = NULL,
  Tags = NULL
)
}
\arguments{
\item{DisplayName}{[required] The display name to associate with the notify configuration.}

\item{UseCase}{[required] The use case for the notify configuration.}

\item{DefaultTemplateId}{The default template identifier to associate with the notify
configuration. If specified, this template is used when sending messages
without an explicit template identifier.}

\item{PoolId}{The identifier of the pool to associate with the notify configuration.}

\item{EnabledCountries}{An array of two-character ISO country codes, in ISO 3166-1 alpha-2
format, that are enabled for the notify configuration.}

\item{EnabledChannels}{[required] An array of channels to enable for the notify configuration. Supported
values include \code{SMS} and \code{VOICE}.}

\item{DeletionProtectionEnabled}{By default this is set to false. When set to true the notify
configuration can't be deleted. You can change this value using the
\code{\link[=pinpointsmsvoicev2_update_notify_configuration]{update_notify_configuration}}
action.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don't specify a client token, a
randomly generated token is used for the request to ensure idempotency.}

\item{Tags}{An array of tags (key and value pairs) associated with the notify
configuration.}
}
\description{
Creates a new notify configuration for managed messaging. A notify configuration defines the settings for sending templated messages, including the display name, use case, enabled channels, and enabled countries.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_create_notify_configuration/} for full documentation.
}
\keyword{internal}
