% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_create_opt_out_list}
\alias{pinpointsmsvoicev2_create_opt_out_list}
\title{Creates a new opt-out list}
\usage{
pinpointsmsvoicev2_create_opt_out_list(
  OptOutListName,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{OptOutListName}{[required] The name of the new OptOutList.}

\item{Tags}{An array of tags (key and value pairs) to associate with the new
OptOutList.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don't specify a client token, a
randomly generated token is used for the request to ensure idempotency.}
}
\description{
Creates a new opt-out list.

See \url{https://paws-r.github.io/docs/pinpointsmsvoicev2/create_opt_out_list.html} for full documentation.
}
\keyword{internal}
