% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_create_registration}
\alias{pinpointsmsvoicev2_create_registration}
\title{Creates a new registration based on the RegistrationType field}
\usage{
pinpointsmsvoicev2_create_registration(
  RegistrationType,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{RegistrationType}{[required] The type of registration form to create. The list of
\strong{RegistrationTypes} can be found using the
\code{\link[=pinpointsmsvoicev2_describe_registration_type_definitions]{describe_registration_type_definitions}}
action.}

\item{Tags}{An array of tags (key and value pairs) to associate with the
registration.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don't specify a client token, a
randomly generated token is used for the request to ensure idempotency.}
}
\description{
Creates a new registration based on the \strong{RegistrationType} field.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_create_registration/} for full documentation.
}
\keyword{internal}
