% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_create_registration_association}
\alias{pinpointsmsvoicev2_create_registration_association}
\title{Associate the registration with an origination identity such as a phone
number or sender ID}
\usage{
pinpointsmsvoicev2_create_registration_association(RegistrationId, ResourceId)
}
\arguments{
\item{RegistrationId}{[required] The unique identifier for the registration.}

\item{ResourceId}{[required] The unique identifier for the origination identity. For example this
could be a \strong{PhoneNumberId} or \strong{SenderId}.}
}
\description{
Associate the registration with an origination identity such as a phone number or sender ID.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_create_registration_association/} for full documentation.
}
\keyword{internal}
