% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_create_registration_version}
\alias{pinpointsmsvoicev2_create_registration_version}
\title{Create a new version of the registration and increase the VersionNumber}
\usage{
pinpointsmsvoicev2_create_registration_version(RegistrationId)
}
\arguments{
\item{RegistrationId}{[required] The unique identifier for the registration.}
}
\description{
Create a new version of the registration and increase the \strong{VersionNumber}. The previous version of the registration becomes read-only.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_create_registration_version/} for full documentation.
}
\keyword{internal}
