% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_create_verified_destination_number}
\alias{pinpointsmsvoicev2_create_verified_destination_number}
\title{You can only send messages to verified destination numbers when your
account is in the sandbox}
\usage{
pinpointsmsvoicev2_create_verified_destination_number(
  DestinationPhoneNumber,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{DestinationPhoneNumber}{[required] The verified destination phone number, in E.164 format.}

\item{Tags}{An array of tags (key and value pairs) to associate with the destination
number.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If you don't specify a client token, a
randomly generated token is used for the request to ensure idempotency.}
}
\description{
You can only send messages to verified destination numbers when your account is in the sandbox. You can add up to 10 verified destination numbers.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_create_verified_destination_number/} for full documentation.
}
\keyword{internal}
