% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_protect_configuration_rule_set_number_override}
\alias{pinpointsmsvoicev2_delete_protect_configuration_rule_set_number_override}
\title{Permanently delete the protect configuration rule set number override}
\usage{
pinpointsmsvoicev2_delete_protect_configuration_rule_set_number_override(
  ProtectConfigurationId,
  DestinationPhoneNumber
)
}
\arguments{
\item{ProtectConfigurationId}{[required] The unique identifier for the protect configuration.}

\item{DestinationPhoneNumber}{[required] The destination phone number in E.164 format.}
}
\description{
Permanently delete the protect configuration rule set number override.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_delete_protect_configuration_rule_set_number_override/} for full documentation.
}
\keyword{internal}
