% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_configuration_set}
\alias{pinpointsmsvoicev2_delete_configuration_set}
\title{Deletes an existing configuration set}
\usage{
pinpointsmsvoicev2_delete_configuration_set(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set or the configuration set ARN that you
want to delete. The ConfigurationSetName and ConfigurationSetArn can be
found using the
\code{\link[=pinpointsmsvoicev2_describe_configuration_sets]{describe_configuration_sets}}
action.}
}
\description{
Deletes an existing configuration set.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_delete_configuration_set/} for full documentation.
}
\keyword{internal}
