% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_default_message_type}
\alias{pinpointsmsvoicev2_delete_default_message_type}
\title{Deletes an existing default message type on a configuration set}
\usage{
pinpointsmsvoicev2_delete_default_message_type(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set or the configuration set Amazon
Resource Name (ARN) to delete the default message type from. The
ConfigurationSetName and ConfigurationSetArn can be found using the
\code{\link[=pinpointsmsvoicev2_describe_configuration_sets]{describe_configuration_sets}}
action.}
}
\description{
Deletes an existing default message type on a configuration set.

See \url{https://paws-r.github.io/docs/pinpointsmsvoicev2/delete_default_message_type.html} for full documentation.
}
\keyword{internal}
