% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_default_sender_id}
\alias{pinpointsmsvoicev2_delete_default_sender_id}
\title{Deletes an existing default sender ID on a configuration set}
\usage{
pinpointsmsvoicev2_delete_default_sender_id(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set or the configuration set Amazon
Resource Name (ARN) to delete the default sender ID from. The
ConfigurationSetName and ConfigurationSetArn can be found using the
\code{\link[=pinpointsmsvoicev2_describe_configuration_sets]{describe_configuration_sets}}
action.}
}
\description{
Deletes an existing default sender ID on a configuration set.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_delete_default_sender_id/} for full documentation.
}
\keyword{internal}
