% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_notify_configuration}
\alias{pinpointsmsvoicev2_delete_notify_configuration}
\title{Deletes an existing notify configuration}
\usage{
pinpointsmsvoicev2_delete_notify_configuration(NotifyConfigurationId)
}
\arguments{
\item{NotifyConfigurationId}{[required] The identifier of the notify configuration to delete. The
NotifyConfigurationId can be found using the
\code{\link[=pinpointsmsvoicev2_describe_notify_configurations]{describe_notify_configurations}}
operation.}
}
\description{
Deletes an existing notify configuration.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_delete_notify_configuration/} for full documentation.
}
\keyword{internal}
