% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_opted_out_number}
\alias{pinpointsmsvoicev2_delete_opted_out_number}
\title{Deletes an existing opted out destination phone number from the
specified opt-out list}
\usage{
pinpointsmsvoicev2_delete_opted_out_number(OptOutListName, OptedOutNumber)
}
\arguments{
\item{OptOutListName}{[required] The OptOutListName or OptOutListArn to remove the phone number from.}

\item{OptedOutNumber}{[required] The phone number, in E.164 format, to remove from the OptOutList.}
}
\description{
Deletes an existing opted out destination phone number from the specified opt-out list.

See \url{https://paws-r.github.io/docs/pinpointsmsvoicev2/delete_opted_out_number.html} for full documentation.
}
\keyword{internal}
