% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_pool}
\alias{pinpointsmsvoicev2_delete_pool}
\title{Deletes an existing pool}
\usage{
pinpointsmsvoicev2_delete_pool(PoolId)
}
\arguments{
\item{PoolId}{[required] The PoolId or PoolArn of the pool to delete. You can use
\code{\link[=pinpointsmsvoicev2_describe_pools]{describe_pools}} to find the values
for PoolId and PoolArn .}
}
\description{
Deletes an existing pool. Deleting a pool disassociates all origination identities from that pool.

See \url{https://paws-r.github.io/docs/pinpointsmsvoicev2/delete_pool.html} for full documentation.
}
\keyword{internal}
