% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_protect_configuration}
\alias{pinpointsmsvoicev2_delete_protect_configuration}
\title{Permanently delete the protect configuration}
\usage{
pinpointsmsvoicev2_delete_protect_configuration(ProtectConfigurationId)
}
\arguments{
\item{ProtectConfigurationId}{[required] The unique identifier for the protect configuration.}
}
\description{
Permanently delete the protect configuration. The protect configuration must have deletion protection disabled and must not be associated as the account default protect configuration or associated with a configuration set.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_delete_protect_configuration/} for full documentation.
}
\keyword{internal}
