% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_rcs_agent}
\alias{pinpointsmsvoicev2_delete_rcs_agent}
\title{Deletes an existing RCS agent}
\usage{
pinpointsmsvoicev2_delete_rcs_agent(RcsAgentId)
}
\arguments{
\item{RcsAgentId}{[required] The unique identifier of the RCS agent to delete. You can use either the
RcsAgentId or RcsAgentArn.}
}
\description{
Deletes an existing RCS agent. If deletion protection is enabled, an error is returned.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_delete_rcs_agent/} for full documentation.
}
\keyword{internal}
