% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_registration}
\alias{pinpointsmsvoicev2_delete_registration}
\title{Permanently delete an existing registration from your account}
\usage{
pinpointsmsvoicev2_delete_registration(RegistrationId)
}
\arguments{
\item{RegistrationId}{[required] The unique identifier for the registration.}
}
\description{
Permanently delete an existing registration from your account.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_delete_registration/} for full documentation.
}
\keyword{internal}
