% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_registration_attachment}
\alias{pinpointsmsvoicev2_delete_registration_attachment}
\title{Permanently delete the specified registration attachment}
\usage{
pinpointsmsvoicev2_delete_registration_attachment(RegistrationAttachmentId)
}
\arguments{
\item{RegistrationAttachmentId}{[required] The unique identifier for the registration attachment.}
}
\description{
Permanently delete the specified registration attachment.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_delete_registration_attachment/} for full documentation.
}
\keyword{internal}
