% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_verified_destination_number}
\alias{pinpointsmsvoicev2_delete_verified_destination_number}
\title{Delete a verified destination phone number}
\usage{
pinpointsmsvoicev2_delete_verified_destination_number(
  VerifiedDestinationNumberId
)
}
\arguments{
\item{VerifiedDestinationNumberId}{[required] The unique identifier for the verified destination phone number.}
}
\description{
Delete a verified destination phone number.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_delete_verified_destination_number/} for full documentation.
}
\keyword{internal}
