% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_account_attributes}
\alias{pinpointsmsvoicev2_describe_account_attributes}
\title{Describes attributes of your Amazon Web Services account}
\usage{
pinpointsmsvoicev2_describe_account_attributes(
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Describes attributes of your Amazon Web Services account. The supported account attributes include account tier, which indicates whether your account is in the sandbox or production environment. When you're ready to move your account out of the sandbox, create an Amazon Web Services Support case for a service limit increase request.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_account_attributes/} for full documentation.
}
\keyword{internal}
