% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_configuration_sets}
\alias{pinpointsmsvoicev2_describe_configuration_sets}
\title{Describes the specified configuration sets or all in your account}
\usage{
pinpointsmsvoicev2_describe_configuration_sets(
  ConfigurationSetNames = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ConfigurationSetNames}{An array of strings. Each element can be either a ConfigurationSetName
or ConfigurationSetArn.}

\item{Filters}{An array of filters to apply to the results that are returned.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Describes the specified configuration sets or all in your account.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_configuration_sets/} for full documentation.
}
\keyword{internal}
