% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_notify_configurations}
\alias{pinpointsmsvoicev2_describe_notify_configurations}
\title{Describes the specified notify configurations or all notify
configurations in your account}
\usage{
pinpointsmsvoicev2_describe_notify_configurations(
  NotifyConfigurationIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{NotifyConfigurationIds}{An array of notify configuration IDs to describe.}

\item{Filters}{An array of NotifyConfigurationFilter objects to filter the results on.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Describes the specified notify configurations or all notify configurations in your account.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_notify_configurations/} for full documentation.
}
\keyword{internal}
