% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_notify_templates}
\alias{pinpointsmsvoicev2_describe_notify_templates}
\title{Describes the specified notify templates or all notify templates in your
account}
\usage{
pinpointsmsvoicev2_describe_notify_templates(
  TemplateIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{TemplateIds}{An array of template IDs to describe.}

\item{Filters}{An array of NotifyTemplateFilter objects to filter the results on.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Describes the specified notify templates or all notify templates in your account.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_notify_templates/} for full documentation.
}
\keyword{internal}
