% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_opt_out_lists}
\alias{pinpointsmsvoicev2_describe_opt_out_lists}
\title{Describes the specified opt-out list or all opt-out lists in your
account}
\usage{
pinpointsmsvoicev2_describe_opt_out_lists(
  OptOutListNames = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{OptOutListNames}{The OptOutLists to show the details of. This is an array of strings that
can be either the OptOutListName or OptOutListArn.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Describes the specified opt-out list or all opt-out lists in your account.

See \url{https://paws-r.github.io/docs/pinpointsmsvoicev2/describe_opt_out_lists.html} for full documentation.
}
\keyword{internal}
