% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_phone_numbers}
\alias{pinpointsmsvoicev2_describe_phone_numbers}
\title{Describes the specified origination phone number, or all the phone
numbers in your account}
\usage{
pinpointsmsvoicev2_describe_phone_numbers(
  PhoneNumberIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{PhoneNumberIds}{The unique identifier of phone numbers to find information about. This
is an array of strings that can be either the PhoneNumberId or
PhoneNumberArn.}

\item{Filters}{An array of PhoneNumberFilter objects to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Describes the specified origination phone number, or all the phone numbers in your account.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_phone_numbers/} for full documentation.
}
\keyword{internal}
