% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_pools}
\alias{pinpointsmsvoicev2_describe_pools}
\title{Retrieves the specified pools or all pools associated with your Amazon
Web Services account}
\usage{
pinpointsmsvoicev2_describe_pools(
  PoolIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{PoolIds}{The unique identifier of pools to find. This is an array of strings that
can be either the PoolId or PoolArn.}

\item{Filters}{An array of PoolFilter objects to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Retrieves the specified pools or all pools associated with your Amazon Web Services account.

See \url{https://paws-r.github.io/docs/pinpointsmsvoicev2/describe_pools.html} for full documentation.
}
\keyword{internal}
