% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_protect_configurations}
\alias{pinpointsmsvoicev2_describe_protect_configurations}
\title{Retrieves the protect configurations that match any of filters}
\usage{
pinpointsmsvoicev2_describe_protect_configurations(
  ProtectConfigurationIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ProtectConfigurationIds}{An array of protect configuration identifiers to search for.}

\item{Filters}{An array of ProtectConfigurationFilter objects to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Retrieves the protect configurations that match any of filters. If a filter isn’t provided then all protect configurations are returned.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_protect_configurations/} for full documentation.
}
\keyword{internal}
