% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_rcs_agent_country_launch_status}
\alias{pinpointsmsvoicev2_describe_rcs_agent_country_launch_status}
\title{Retrieves the per-country launch status of an RCS agent, including
carrier-level details for each country}
\usage{
pinpointsmsvoicev2_describe_rcs_agent_country_launch_status(
  RcsAgentId,
  IsoCountryCodes = NULL,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{RcsAgentId}{[required] The unique identifier of the RCS agent. You can use either the
RcsAgentId or RcsAgentArn.}

\item{IsoCountryCodes}{An array of two-character ISO country codes, in ISO 3166-1 alpha-2
format, to filter the results.}

\item{Filters}{An array of CountryLaunchStatusFilter objects to filter the results.}

\item{MaxResults}{The maximum number of results to return per each request.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}
}
\description{
Retrieves the per-country launch status of an RCS agent, including carrier-level details for each country.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_rcs_agent_country_launch_status/} for full documentation.
}
\keyword{internal}
