% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_rcs_agents}
\alias{pinpointsmsvoicev2_describe_rcs_agents}
\title{Retrieves the specified RCS agents or all RCS agents associated with
your Amazon Web Services account}
\usage{
pinpointsmsvoicev2_describe_rcs_agents(
  RcsAgentIds = NULL,
  Owner = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{RcsAgentIds}{An array of unique identifiers for the RCS agents. This is an array of
strings that can be either the RcsAgentId or RcsAgentArn.}

\item{Owner}{Use \code{SELF} to filter the list of RCS agents to ones your account owns or
use \code{SHARED} to filter on RCS agents shared with your account. The
\code{Owner} and \code{RcsAgentIds} parameters can't be used at the same time.}

\item{Filters}{An array of RcsAgentFilter objects to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Retrieves the specified RCS agents or all RCS agents associated with your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_rcs_agents/} for full documentation.
}
\keyword{internal}
