% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_registration_attachments}
\alias{pinpointsmsvoicev2_describe_registration_attachments}
\title{Retrieves the specified registration attachments or all registration
attachments associated with your Amazon Web Services account}
\usage{
pinpointsmsvoicev2_describe_registration_attachments(
  RegistrationAttachmentIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{RegistrationAttachmentIds}{The unique identifier of registration attachments to find. This is an
array of \strong{RegistrationAttachmentId}.}

\item{Filters}{An array of RegistrationAttachmentFilter objects to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Retrieves the specified registration attachments or all registration attachments associated with your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_registration_attachments/} for full documentation.
}
\keyword{internal}
