% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_registration_type_definitions}
\alias{pinpointsmsvoicev2_describe_registration_type_definitions}
\title{Retrieves the specified registration type definitions}
\usage{
pinpointsmsvoicev2_describe_registration_type_definitions(
  RegistrationTypes = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{RegistrationTypes}{The type of registration form. The list of \strong{RegistrationTypes} can be
found using the
\code{\link[=pinpointsmsvoicev2_describe_registration_type_definitions]{describe_registration_type_definitions}}
action.}

\item{Filters}{An array of RegistrationFilter objects to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Retrieves the specified registration type definitions. You can use DescribeRegistrationTypeDefinitions to view the requirements for creating, filling out, and submitting each registration type.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_registration_type_definitions/} for full documentation.
}
\keyword{internal}
