% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_registration_versions}
\alias{pinpointsmsvoicev2_describe_registration_versions}
\title{Retrieves the specified registration version}
\usage{
pinpointsmsvoicev2_describe_registration_versions(
  RegistrationId,
  VersionNumbers = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{RegistrationId}{[required] The unique identifier for the registration.}

\item{VersionNumbers}{An array of registration version numbers.}

\item{Filters}{An array of RegistrationVersionFilter objects to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Retrieves the specified registration version.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_registration_versions/} for full documentation.
}
\keyword{internal}
