% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_registrations}
\alias{pinpointsmsvoicev2_describe_registrations}
\title{Retrieves the specified registrations}
\usage{
pinpointsmsvoicev2_describe_registrations(
  RegistrationIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{RegistrationIds}{An array of unique identifiers for each registration.}

\item{Filters}{An array of RegistrationFilter objects to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Retrieves the specified registrations.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_registrations/} for full documentation.
}
\keyword{internal}
