% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_spend_limits}
\alias{pinpointsmsvoicev2_describe_spend_limits}
\title{Describes the current Amazon Pinpoint monthly spend limits for sending
voice and text messages}
\usage{
pinpointsmsvoicev2_describe_spend_limits(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Describes the current Amazon Pinpoint monthly spend limits for sending voice and text messages.

See \url{https://paws-r.github.io/docs/pinpointsmsvoicev2/describe_spend_limits.html} for full documentation.
}
\keyword{internal}
