% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_verified_destination_numbers}
\alias{pinpointsmsvoicev2_describe_verified_destination_numbers}
\title{Retrieves the specified verified destination numbers}
\usage{
pinpointsmsvoicev2_describe_verified_destination_numbers(
  VerifiedDestinationNumberIds = NULL,
  DestinationPhoneNumbers = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{VerifiedDestinationNumberIds}{An array of VerifiedDestinationNumberid to retrieve.}

\item{DestinationPhoneNumbers}{An array of verified destination phone number, in E.164 format.}

\item{Filters}{An array of VerifiedDestinationNumberFilter objects to filter the
results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Retrieves the specified verified destination numbers.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_verified_destination_numbers/} for full documentation.
}
\keyword{internal}
