% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_disassociate_protect_configuration}
\alias{pinpointsmsvoicev2_disassociate_protect_configuration}
\title{Disassociate a protect configuration from a configuration set}
\usage{
pinpointsmsvoicev2_disassociate_protect_configuration(
  ProtectConfigurationId,
  ConfigurationSetName
)
}
\arguments{
\item{ProtectConfigurationId}{[required] The unique identifier for the protect configuration.}

\item{ConfigurationSetName}{[required] The name of the ConfigurationSet.}
}
\description{
Disassociate a protect configuration from a configuration set.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_disassociate_protect_configuration/} for full documentation.
}
\keyword{internal}
