% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_discard_registration_version}
\alias{pinpointsmsvoicev2_discard_registration_version}
\title{Discard the current version of the registration}
\usage{
pinpointsmsvoicev2_discard_registration_version(RegistrationId)
}
\arguments{
\item{RegistrationId}{[required] The unique identifier for the registration.}
}
\description{
Discard the current version of the registration.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_discard_registration_version/} for full documentation.
}
\keyword{internal}
