% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_get_protect_configuration_country_rule_set}
\alias{pinpointsmsvoicev2_get_protect_configuration_country_rule_set}
\title{Retrieve the CountryRuleSet for the specified NumberCapability from a
protect configuration}
\usage{
pinpointsmsvoicev2_get_protect_configuration_country_rule_set(
  ProtectConfigurationId,
  NumberCapability
)
}
\arguments{
\item{ProtectConfigurationId}{[required] The unique identifier for the protect configuration.}

\item{NumberCapability}{[required] The capability type to return the CountryRuleSet for. Valid values are
\code{SMS}, \code{VOICE}, or \code{MMS}.}
}
\description{
Retrieve the CountryRuleSet for the specified NumberCapability from a protect configuration.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_get_protect_configuration_country_rule_set/} for full documentation.
}
\keyword{internal}
