% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_list_notify_countries}
\alias{pinpointsmsvoicev2_list_notify_countries}
\title{Lists countries that support notify messaging}
\usage{
pinpointsmsvoicev2_list_notify_countries(
  Channels = NULL,
  UseCases = NULL,
  Tier = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Channels}{An array of channels to filter the results by.}

\item{UseCases}{An array of use cases to filter the results by.}

\item{Tier}{The tier to filter the results by.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Lists countries that support notify messaging. You can optionally filter by channel, use case, or tier.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_list_notify_countries/} for full documentation.
}
\keyword{internal}
