% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_list_pool_origination_identities}
\alias{pinpointsmsvoicev2_list_pool_origination_identities}
\title{Lists all associated origination identities in your pool}
\usage{
pinpointsmsvoicev2_list_pool_origination_identities(
  PoolId,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{PoolId}{[required] The unique identifier for the pool. This value can be either the PoolId
or PoolArn.}

\item{Filters}{An array of PoolOriginationIdentitiesFilter objects to filter the
results..}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Lists all associated origination identities in your pool.

See \url{https://paws-r.github.io/docs/pinpointsmsvoicev2/list_pool_origination_identities.html} for full documentation.
}
\keyword{internal}
