% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_list_protect_configuration_rule_set_number_overrides}
\alias{pinpointsmsvoicev2_list_protect_configuration_rule_set_number_overrides}
\title{Retrieve all of the protect configuration rule set number overrides that
match the filters}
\usage{
pinpointsmsvoicev2_list_protect_configuration_rule_set_number_overrides(
  ProtectConfigurationId,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ProtectConfigurationId}{[required] The unique identifier for the protect configuration.}

\item{Filters}{An array of ProtectConfigurationRuleSetNumberOverrideFilterItem objects
to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Retrieve all of the protect configuration rule set number overrides that match the filters.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_list_protect_configuration_rule_set_number_overrides/} for full documentation.
}
\keyword{internal}
