% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_list_registration_associations}
\alias{pinpointsmsvoicev2_list_registration_associations}
\title{Retreive all of the origination identies that are associated with a
registration}
\usage{
pinpointsmsvoicev2_list_registration_associations(
  RegistrationId,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{RegistrationId}{[required] The unique identifier for the registration.}

\item{Filters}{An array of RegistrationAssociationFilter to apply to the results that
are returned.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}
}
\description{
Retreive all of the origination identies that are associated with a registration.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_list_registration_associations/} for full documentation.
}
\keyword{internal}
