% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_list_tags_for_resource}
\alias{pinpointsmsvoicev2_list_tags_for_resource}
\title{List all tags associated with a resource}
\usage{
pinpointsmsvoicev2_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to query for.}
}
\description{
List all tags associated with a resource.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
