% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_put_message_feedback}
\alias{pinpointsmsvoicev2_put_message_feedback}
\title{Set the MessageFeedbackStatus as RECEIVED or FAILED for the passed in
MessageId}
\usage{
pinpointsmsvoicev2_put_message_feedback(MessageId, MessageFeedbackStatus)
}
\arguments{
\item{MessageId}{[required] The unique identifier for the message.}

\item{MessageFeedbackStatus}{[required] Set the message feedback to be either \code{RECEIVED} or \code{FAILED}.}
}
\description{
Set the MessageFeedbackStatus as \code{RECEIVED} or \code{FAILED} for the passed in MessageId.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_put_message_feedback/} for full documentation.
}
\keyword{internal}
