% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_put_opted_out_number}
\alias{pinpointsmsvoicev2_put_opted_out_number}
\title{Creates an opted out destination phone number in the opt-out list}
\usage{
pinpointsmsvoicev2_put_opted_out_number(OptOutListName, OptedOutNumber)
}
\arguments{
\item{OptOutListName}{[required] The OptOutListName or OptOutListArn to add the phone number to.}

\item{OptedOutNumber}{[required] The phone number to add to the OptOutList in E.164 format.}
}
\description{
Creates an opted out destination phone number in the opt-out list.

See \url{https://paws-r.github.io/docs/pinpointsmsvoicev2/put_opted_out_number.html} for full documentation.
}
\keyword{internal}
