% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_put_registration_field_value}
\alias{pinpointsmsvoicev2_put_registration_field_value}
\title{Creates or updates a field value for a registration}
\usage{
pinpointsmsvoicev2_put_registration_field_value(
  RegistrationId,
  FieldPath,
  SelectChoices = NULL,
  TextValue = NULL,
  RegistrationAttachmentId = NULL
)
}
\arguments{
\item{RegistrationId}{[required] The unique identifier for the registration.}

\item{FieldPath}{[required] The path to the registration form field. You can use
\code{\link[=pinpointsmsvoicev2_describe_registration_field_definitions]{describe_registration_field_definitions}}
for a list of \strong{FieldPaths}.}

\item{SelectChoices}{An array of values for the form field.}

\item{TextValue}{The text data for a free form field.}

\item{RegistrationAttachmentId}{The unique identifier for the registration attachment.}
}
\description{
Creates or updates a field value for a registration.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_put_registration_field_value/} for full documentation.
}
\keyword{internal}
