% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_release_phone_number}
\alias{pinpointsmsvoicev2_release_phone_number}
\title{Releases an existing origination phone number in your account}
\usage{
pinpointsmsvoicev2_release_phone_number(PhoneNumberId)
}
\arguments{
\item{PhoneNumberId}{[required] The PhoneNumberId or PhoneNumberArn of the phone number to release. You
can use
\code{\link[=pinpointsmsvoicev2_describe_phone_numbers]{describe_phone_numbers}} to
get the values for PhoneNumberId and PhoneNumberArn.}
}
\description{
Releases an existing origination phone number in your account. Once released, a phone number is no longer available for sending messages.

See \url{https://paws-r.github.io/docs/pinpointsmsvoicev2/release_phone_number.html} for full documentation.
}
\keyword{internal}
