% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_release_sender_id}
\alias{pinpointsmsvoicev2_release_sender_id}
\title{Releases an existing sender ID in your account}
\usage{
pinpointsmsvoicev2_release_sender_id(SenderId, IsoCountryCode)
}
\arguments{
\item{SenderId}{[required] The sender ID to release.}

\item{IsoCountryCode}{[required] The two-character code, in ISO 3166-1 alpha-2 format, for the country or
region.}
}
\description{
Releases an existing sender ID in your account.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_release_sender_id/} for full documentation.
}
\keyword{internal}
