% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_request_phone_number}
\alias{pinpointsmsvoicev2_request_phone_number}
\title{Request an origination phone number for use in your account}
\usage{
pinpointsmsvoicev2_request_phone_number(
  IsoCountryCode,
  MessageType,
  NumberCapabilities,
  NumberType,
  OptOutListName = NULL,
  PoolId = NULL,
  RegistrationId = NULL,
  NumberPreference = NULL,
  InternationalSendingEnabled = NULL,
  DeletionProtectionEnabled = NULL,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{IsoCountryCode}{[required] The two-character code, in ISO 3166-1 alpha-2 format, for the country or region.}

\item{MessageType}{[required] The type of message. Valid values are \code{TRANSACTIONAL} for messages that are critical or time-sensitive and \code{PROMOTIONAL} for messages that aren't critical or time-sensitive.}

\item{NumberCapabilities}{[required] Indicates if the phone number will be used for text messages, voice messages, or both.}

\item{NumberType}{[required] The type of phone number to request.

When you request a \code{SIMULATOR} phone number, you must set \strong{MessageType} as \code{TRANSACTIONAL}.}

\item{OptOutListName}{The name of the OptOutList to associate with the phone number. You can use the OptOutListName or OptOutListArn.

If you are using a shared End User Messaging SMS resource then you must use the full Amazon Resource Name(ARN).}

\item{PoolId}{The pool to associated with the phone number. You can use the PoolId or PoolArn.

If you are using a shared End User Messaging SMS resource then you must use the full Amazon Resource Name(ARN).}

\item{RegistrationId}{Use this field to attach your phone number for an external registration process.}

\item{NumberPreference}{An optional selection preference used to request a specific phone number, such as a number that starts with, ends with, or contains a particular digit pattern. You can specify at most one preference. Number preferences apply only to \code{TEN_DLC} requests in the \code{US}.}

\item{InternationalSendingEnabled}{By default this is set to false. When set to true the international sending of phone number is Enabled.}

\item{DeletionProtectionEnabled}{By default this is set to false. When set to true the phone number can't be deleted.}

\item{Tags}{An array of tags (key and value pairs) to associate with the requested phone number.}

\item{ClientToken}{Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If you don't specify a client token, a randomly generated token is used for the request to ensure idempotency.}
}
\description{
Request an origination phone number for use in your account. For more information on phone number request see \href{https://docs.aws.amazon.com/sms-voice/latest/userguide/phone-numbers-request.html}{Request a phone number} in the \emph{End User Messaging SMS User Guide}.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_request_phone_number/} for full documentation.
}
\keyword{internal}
