% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_set_account_default_protect_configuration}
\alias{pinpointsmsvoicev2_set_account_default_protect_configuration}
\title{Set a protect configuration as your account default}
\usage{
pinpointsmsvoicev2_set_account_default_protect_configuration(
  ProtectConfigurationId
)
}
\arguments{
\item{ProtectConfigurationId}{[required] The unique identifier for the protect configuration.}
}
\description{
Set a protect configuration as your account default. You can only have one account default protect configuration at a time. The current account default protect configuration is replaced with the provided protect configuration.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_set_account_default_protect_configuration/} for full documentation.
}
\keyword{internal}
