% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_set_default_message_feedback_enabled}
\alias{pinpointsmsvoicev2_set_default_message_feedback_enabled}
\title{Sets a configuration set's default for message feedback}
\usage{
pinpointsmsvoicev2_set_default_message_feedback_enabled(
  ConfigurationSetName,
  MessageFeedbackEnabled
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set to use. This can be either the
ConfigurationSetName or ConfigurationSetArn.}

\item{MessageFeedbackEnabled}{[required] Set to true to enable message feedback.}
}
\description{
Sets a configuration set's default for message feedback.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_set_default_message_feedback_enabled/} for full documentation.
}
\keyword{internal}
