% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_set_default_message_type}
\alias{pinpointsmsvoicev2_set_default_message_type}
\title{Sets the default message type on a configuration set}
\usage{
pinpointsmsvoicev2_set_default_message_type(ConfigurationSetName, MessageType)
}
\arguments{
\item{ConfigurationSetName}{[required] The configuration set to update with a new default message type. This
field can be the ConsigurationSetName or ConfigurationSetArn.}

\item{MessageType}{[required] The type of message. Valid values are TRANSACTIONAL for messages that
are critical or time-sensitive and PROMOTIONAL for messages that aren't
critical or time-sensitive.}
}
\description{
Sets the default message type on a configuration set.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_set_default_message_type/} for full documentation.
}
\keyword{internal}
