% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_submit_registration_version}
\alias{pinpointsmsvoicev2_submit_registration_version}
\title{Submit the specified registration for review and approval}
\usage{
pinpointsmsvoicev2_submit_registration_version(RegistrationId)
}
\arguments{
\item{RegistrationId}{[required] The unique identifier for the registration.}
}
\description{
Submit the specified registration for review and approval.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_submit_registration_version/} for full documentation.
}
\keyword{internal}
