% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_tag_resource}
\alias{pinpointsmsvoicev2_tag_resource}
\title{Adds or overwrites only the specified tags for the specified resource}
\usage{
pinpointsmsvoicev2_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{Tags}{[required] An array of key and value pair tags that are associated with the
resource.}
}
\description{
Adds or overwrites only the specified tags for the specified resource. When you specify an existing tag key, the value is overwritten with the new value. Each tag consists of a key and an optional value. Tag keys must be unique per resource. For more information about tags, see \href{https://docs.aws.amazon.com/sms-voice/latest/userguide/phone-numbers-tags.html}{Tags} in the \emph{End User Messaging SMS User Guide}.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_tag_resource/} for full documentation.
}
\keyword{internal}
