% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_untag_resource}
\alias{pinpointsmsvoicev2_untag_resource}
\title{Removes the association of the specified tags from an Amazon Pinpoint
SMS Voice V2 resource}
\usage{
pinpointsmsvoicev2_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{TagKeys}{[required] An array of tag key values to unassociate with the resource.}
}
\description{
Removes the association of the specified tags from an Amazon Pinpoint SMS Voice V2 resource. For more information on tags see \href{https://docs.aws.amazon.com/pinpoint/latest/developerguide/tagging-resources.html}{Tagging Amazon Pinpoint resources} in the \emph{Amazon Pinpoint Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_untag_resource/} for full documentation.
}
\keyword{internal}
