% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_update_protect_configuration_country_rule_set}
\alias{pinpointsmsvoicev2_update_protect_configuration_country_rule_set}
\title{Update a country rule set to ALLOW, BLOCK, MONITOR, or FILTER messages
to be sent to the specified destination counties}
\usage{
pinpointsmsvoicev2_update_protect_configuration_country_rule_set(
  ProtectConfigurationId,
  NumberCapability,
  CountryRuleSetUpdates
)
}
\arguments{
\item{ProtectConfigurationId}{[required] The unique identifier for the protect configuration.}

\item{NumberCapability}{[required] The number capability to apply the CountryRuleSetUpdates updates to.}

\item{CountryRuleSetUpdates}{[required] A map of ProtectConfigurationCountryRuleSetInformation objects that
contain the details for the requested NumberCapability. The Key is the
two-letter ISO country code. For a list of supported ISO country codes,
see \href{https://docs.aws.amazon.com/sms-voice/latest/userguide/phone-numbers-sms-by-country.html}{Supported countries and regions (SMS channel)}
in the End User Messaging SMS User Guide.

For example, to set the United States as allowed and Canada as blocked,
the \code{CountryRuleSetUpdates} would be formatted as:
\verb{"CountryRuleSetUpdates": \{ "US" : \{ "ProtectStatus": "ALLOW" \} "CA" : \{ "ProtectStatus": "BLOCK" \} \}}}
}
\description{
Update a country rule set to \code{ALLOW}, \code{BLOCK}, \code{MONITOR}, or \code{FILTER} messages to be sent to the specified destination counties. You can update one or multiple countries at a time. The updates are only applied to the specified NumberCapability type.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_update_protect_configuration_country_rule_set/} for full documentation.
}
\keyword{internal}
