% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_update_protect_configuration}
\alias{pinpointsmsvoicev2_update_protect_configuration}
\title{Update the setting for an existing protect configuration}
\usage{
pinpointsmsvoicev2_update_protect_configuration(
  ProtectConfigurationId,
  DeletionProtectionEnabled = NULL
)
}
\arguments{
\item{ProtectConfigurationId}{[required] The unique identifier for the protect configuration.}

\item{DeletionProtectionEnabled}{When set to true deletion protection is enabled. By default this is set
to false.}
}
\description{
Update the setting for an existing protect configuration.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_update_protect_configuration/} for full documentation.
}
\keyword{internal}
