% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_update_sender_id}
\alias{pinpointsmsvoicev2_update_sender_id}
\title{Updates the configuration of an existing sender ID}
\usage{
pinpointsmsvoicev2_update_sender_id(
  SenderId,
  IsoCountryCode,
  DeletionProtectionEnabled = NULL
)
}
\arguments{
\item{SenderId}{[required] The sender ID to update.}

\item{IsoCountryCode}{[required] The two-character code, in ISO 3166-1 alpha-2 format, for the country or
region.}

\item{DeletionProtectionEnabled}{By default this is set to false. When set to true the sender ID can't be
deleted.}
}
\description{
Updates the configuration of an existing sender ID.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_update_sender_id/} for full documentation.
}
\keyword{internal}
