% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_verify_destination_number}
\alias{pinpointsmsvoicev2_verify_destination_number}
\title{Use the verification code that was received by the verified destination
phone number to opt-in the verified destination phone number to receive
more messages}
\usage{
pinpointsmsvoicev2_verify_destination_number(
  VerifiedDestinationNumberId,
  VerificationCode
)
}
\arguments{
\item{VerifiedDestinationNumberId}{[required] The unique identifier for the verififed destination phone number.}

\item{VerificationCode}{[required] The verification code that was received by the verified destination
phone number.}
}
\description{
Use the verification code that was received by the verified destination phone number to opt-in the verified destination phone number to receive more messages.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_verify_destination_number/} for full documentation.
}
\keyword{internal}
