% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_clone_receipt_rule_set}
\alias{ses_clone_receipt_rule_set}
\title{Creates a receipt rule set by cloning an existing one}
\usage{
ses_clone_receipt_rule_set(RuleSetName, OriginalRuleSetName)
}
\arguments{
\item{RuleSetName}{[required] The name of the rule set to create. The name must meet the following
requirements:
\itemize{
\item Contain only ASCII letters (a-z, A-Z), numbers (0-9), underscores
(_), or dashes (-).
\item Start and end with a letter or number.
\item Contain 64 characters or fewer.
}}

\item{OriginalRuleSetName}{[required] The name of the rule set to clone.}
}
\description{
Creates a receipt rule set by cloning an existing one. All receipt rules and configurations are copied to the new receipt rule set and are completely independent of the source rule set.

See \url{https://www.paws-r-sdk.com/docs/ses_clone_receipt_rule_set/} for full documentation.
}
\keyword{internal}
