% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_create_configuration_set}
\alias{ses_create_configuration_set}
\title{Creates a configuration set}
\usage{
ses_create_configuration_set(ConfigurationSet)
}
\arguments{
\item{ConfigurationSet}{[required] A data structure that contains the name of the configuration set.}
}
\description{
Creates a configuration set.

Configuration sets enable you to publish email sending events. For
information about using configuration sets, see the \href{https://docs.aws.amazon.com/ses/latest/DeveloperGuide/monitor-sending-activity.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$create_configuration_set(
  ConfigurationSet = list(
    Name = "string"
  )
)
}
}

\keyword{internal}
