% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_create_configuration_set_tracking_options}
\alias{ses_create_configuration_set_tracking_options}
\title{Creates an association between a configuration set and a custom domain
for open and click event tracking}
\usage{
ses_create_configuration_set_tracking_options(
  ConfigurationSetName,
  TrackingOptions
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that the tracking options should be associated with.}

\item{TrackingOptions}{[required] A domain that is used to redirect email recipients to an Amazon SES-operated domain. This domain captures open and click events generated by Amazon SES emails.

For more information, see \href{https://docs.aws.amazon.com/ses/latest/dg/configure-custom-open-click-domains.html}{Configuring Custom Domains to Handle Open and Click Tracking} in the \emph{Amazon SES Developer Guide}.}
}
\description{
Creates an association between a configuration set and a custom domain for open and click event tracking.

See \url{https://www.paws-r-sdk.com/docs/ses_create_configuration_set_tracking_options/} for full documentation.
}
\keyword{internal}
