% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_create_receipt_filter}
\alias{ses_create_receipt_filter}
\title{Creates a new IP address filter}
\usage{
ses_create_receipt_filter(Filter)
}
\arguments{
\item{Filter}{[required] A data structure that describes the IP address filter to create, which
consists of a name, an IP address range, and whether to allow or block
mail from it.}
}
\description{
Creates a new IP address filter.

See \url{https://paws-r.github.io/docs/ses/create_receipt_filter.html} for full documentation.
}
\keyword{internal}
