% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_create_receipt_rule_set}
\alias{ses_create_receipt_rule_set}
\title{Creates an empty receipt rule set}
\usage{
ses_create_receipt_rule_set(RuleSetName)
}
\arguments{
\item{RuleSetName}{[required] The name of the rule set to create. The name must meet the following
requirements:
\itemize{
\item Contain only ASCII letters (a-z, A-Z), numbers (0-9), underscores
(_), or dashes (-).
\item Start and end with a letter or number.
\item Contain 64 characters or fewer.
}}
}
\description{
Creates an empty receipt rule set.

See \url{https://www.paws-r-sdk.com/docs/ses_create_receipt_rule_set/} for full documentation.
}
\keyword{internal}
