% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_delete_configuration_set}
\alias{ses_delete_configuration_set}
\title{Deletes a configuration set}
\usage{
ses_delete_configuration_set(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set to delete.}
}
\description{
Deletes a configuration set. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the \href{https://docs.aws.amazon.com/ses/latest/dg/monitor-sending-activity.html}{Amazon SES Developer Guide}.

See \url{https://paws-r.github.io/docs/ses/delete_configuration_set.html} for full documentation.
}
\keyword{internal}
