% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_delete_configuration_set_tracking_options}
\alias{ses_delete_configuration_set_tracking_options}
\title{Deletes an association between a configuration set and a custom domain
for open and click event tracking}
\usage{
ses_delete_configuration_set_tracking_options(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set from which you want to delete the
tracking options.}
}
\description{
Deletes an association between a configuration set and a custom domain for open and click event tracking.

See \url{https://paws-r.github.io/docs/ses/delete_configuration_set_tracking_options.html} for full documentation.
}
\keyword{internal}
