% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_delete_identity}
\alias{ses_delete_identity}
\title{Deletes the specified identity (an email address or a domain) from the
list of verified identities}
\usage{
ses_delete_identity(Identity)
}
\arguments{
\item{Identity}{[required] The identity to be removed from the list of identities for the AWS
Account.}
}
\description{
Deletes the specified identity (an email address or a domain) from the list of verified identities.

See \url{https://www.paws-r-sdk.com/docs/ses_delete_identity/} for full documentation.
}
\keyword{internal}
