% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_delete_receipt_rule}
\alias{ses_delete_receipt_rule}
\title{Deletes the specified receipt rule}
\usage{
ses_delete_receipt_rule(RuleSetName, RuleName)
}
\arguments{
\item{RuleSetName}{[required] The name of the receipt rule set that contains the receipt rule to
delete.}

\item{RuleName}{[required] The name of the receipt rule to delete.}
}
\description{
Deletes the specified receipt rule.

See \url{https://www.paws-r-sdk.com/docs/ses_delete_receipt_rule/} for full documentation.
}
\keyword{internal}
