% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_delete_receipt_rule_set}
\alias{ses_delete_receipt_rule_set}
\title{Deletes the specified receipt rule set and all of the receipt rules it
contains}
\usage{
ses_delete_receipt_rule_set(RuleSetName)
}
\arguments{
\item{RuleSetName}{[required] The name of the receipt rule set to delete.}
}
\description{
Deletes the specified receipt rule set and all of the receipt rules it contains.

See \url{https://paws-r.github.io/docs/ses/delete_receipt_rule_set.html} for full documentation.
}
\keyword{internal}
