% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_describe_active_receipt_rule_set}
\alias{ses_describe_active_receipt_rule_set}
\title{Returns the metadata and receipt rules for the receipt rule set that is
currently active}
\usage{
ses_describe_active_receipt_rule_set()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Metadata = list(
    Name = "string",
    CreatedTimestamp = as.POSIXct(
      "2015-01-01"
    )
  ),
  Rules = list(
    list(
      Name = "string",
      Enabled = TRUE|FALSE,
      TlsPolicy = "Require"|"Optional",
      Recipients = list(
        "string"
      ),
      Actions = list(
        list(
          S3Action = list(
            TopicArn = "string",
            BucketName = "string",
            ObjectKeyPrefix = "string",
            KmsKeyArn = "string"
          ),
          BounceAction = list(
            TopicArn = "string",
            SmtpReplyCode = "string",
            StatusCode = "string",
            Message = "string",
            Sender = "string"
          ),
          WorkmailAction = list(
            TopicArn = "string",
            OrganizationArn = "string"
          ),
          LambdaAction = list(
            TopicArn = "string",
            FunctionArn = "string",
            InvocationType = "Event"|"RequestResponse"
          ),
          StopAction = list(
            Scope = "RuleSet",
            TopicArn = "string"
          ),
          AddHeaderAction = list(
            HeaderName = "string",
            HeaderValue = "string"
          ),
          SNSAction = list(
            TopicArn = "string",
            Encoding = "UTF-8"|"Base64"
          )
        )
      ),
      ScanEnabled = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the metadata and receipt rules for the receipt rule set that is
currently active.

For information about setting up receipt rule sets, see the \href{https://docs.aws.amazon.com/ses/latest/dg/receiving-email-receipt-rules-console-walkthrough.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_active_receipt_rule_set()
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example returns the metadata and receipt rules for the
# receipt rule set that is currently active:
svc$describe_active_receipt_rule_set()
}

}
\keyword{internal}
