% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_describe_receipt_rule}
\alias{ses_describe_receipt_rule}
\title{Returns the details of the specified receipt rule}
\usage{
ses_describe_receipt_rule(RuleSetName, RuleName)
}
\arguments{
\item{RuleSetName}{[required] The name of the receipt rule set that the receipt rule belongs to.}

\item{RuleName}{[required] The name of the receipt rule.}
}
\description{
Returns the details of the specified receipt rule.

See \url{https://www.paws-r-sdk.com/docs/ses_describe_receipt_rule/} for full documentation.
}
\keyword{internal}
