% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_describe_receipt_rule_set}
\alias{ses_describe_receipt_rule_set}
\title{Returns the details of the specified receipt rule set}
\usage{
ses_describe_receipt_rule_set(RuleSetName)
}
\arguments{
\item{RuleSetName}{[required] The name of the receipt rule set to describe.}
}
\description{
Returns the details of the specified receipt rule set.

See \url{https://paws-r.github.io/docs/ses/describe_receipt_rule_set.html} for full documentation.
}
\keyword{internal}
