% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_get_custom_verification_email_template}
\alias{ses_get_custom_verification_email_template}
\title{Returns the custom email verification template for the template name you
specify}
\usage{
ses_get_custom_verification_email_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the custom verification email template to retrieve.}
}
\description{
Returns the custom email verification template for the template name you specify.

See \url{https://www.paws-r-sdk.com/docs/ses_get_custom_verification_email_template/} for full documentation.
}
\keyword{internal}
