% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_get_identity_mail_from_domain_attributes}
\alias{ses_get_identity_mail_from_domain_attributes}
\title{Returns the custom MAIL FROM attributes for a list of identities (email
addresses : domains)}
\usage{
ses_get_identity_mail_from_domain_attributes(Identities)
}
\arguments{
\item{Identities}{[required] A list of one or more identities.}
}
\description{
Returns the custom MAIL FROM attributes for a list of identities (email addresses : domains).

See \url{https://paws-r.github.io/docs/ses/get_identity_mail_from_domain_attributes.html} for full documentation.
}
\keyword{internal}
