% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_get_identity_notification_attributes}
\alias{ses_get_identity_notification_attributes}
\title{Given a list of verified identities (email addresses and/or domains),
returns a structure describing identity notification attributes}
\usage{
ses_get_identity_notification_attributes(Identities)
}
\arguments{
\item{Identities}{[required] A list of one or more identities. You can specify an identity by using
its name or by using its Amazon Resource Name (ARN). Examples:
\code{user@example.com}, \code{example.com},
\code{arn:aws:ses:us-east-1:123456789012:identity/example.com}.}
}
\description{
Given a list of verified identities (email addresses and/or domains), returns a structure describing identity notification attributes.

See \url{https://paws-r.github.io/docs/ses/get_identity_notification_attributes.html} for full documentation.
}
\keyword{internal}
