% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_get_identity_verification_attributes}
\alias{ses_get_identity_verification_attributes}
\title{Given a list of identities (email addresses and/or domains), returns the
verification status and (for domain identities) the verification token
for each identity}
\usage{
ses_get_identity_verification_attributes(Identities)
}
\arguments{
\item{Identities}{[required] A list of identities.}
}
\description{
Given a list of identities (email addresses and/or domains), returns the verification status and (for domain identities) the verification token for each identity.

See \url{https://www.paws-r-sdk.com/docs/ses_get_identity_verification_attributes/} for full documentation.
}
\keyword{internal}
