% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_get_send_statistics}
\alias{ses_get_send_statistics}
\title{Provides sending statistics for the current AWS Region}
\usage{
ses_get_send_statistics()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SendDataPoints = list(
    list(
      Timestamp = as.POSIXct(
        "2015-01-01"
      ),
      DeliveryAttempts = 123,
      Bounces = 123,
      Complaints = 123,
      Rejects = 123
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Provides sending statistics for the current AWS Region. The result is a
list of data points, representing the last two weeks of sending
activity. Each data point in the list contains statistics for a
15-minute period of time.

You can execute this operation no more than once per second.
}
\examples{
\dontrun{
# The following example returns Amazon SES sending statistics:
svc$get_send_statistics()
}

}
\keyword{internal}
