% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_list_configuration_sets}
\alias{ses_list_configuration_sets}
\title{Provides a list of the configuration sets associated with your Amazon
SES account in the current Amazon Web Services Region}
\usage{
ses_list_configuration_sets(NextToken = NULL, MaxItems = NULL)
}
\arguments{
\item{NextToken}{A token returned from a previous call to
\code{\link[=ses_list_configuration_sets]{list_configuration_sets}} to indicate the
position of the configuration set in the configuration set list.}

\item{MaxItems}{The number of configuration sets to return.}
}
\description{
Provides a list of the configuration sets associated with your Amazon SES account in the current Amazon Web Services Region. For information about using configuration sets, see \href{https://docs.aws.amazon.com/ses/latest/dg/monitor-sending-activity.html}{Monitoring Your Amazon SES Sending Activity} in the \emph{Amazon SES Developer Guide.}

See \url{https://www.paws-r-sdk.com/docs/ses_list_configuration_sets/} for full documentation.
}
\keyword{internal}
