% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_list_identity_policies}
\alias{ses_list_identity_policies}
\title{Returns a list of sending authorization policies that are attached to
the given identity (an email address or a domain)}
\usage{
ses_list_identity_policies(Identity)
}
\arguments{
\item{Identity}{[required] The identity that is associated with the policy for which the policies
are listed. You can specify an identity by using its name or by using
its Amazon Resource Name (ARN). Examples: \code{user@example.com},
\code{example.com},
\code{arn:aws:ses:us-east-1:123456789012:identity/example.com}.

To successfully call this operation, you must own the identity.}
}
\description{
Returns a list of sending authorization policies that are attached to the given identity (an email address or a domain). This operation returns only a list. To get the actual policy content, use \code{\link[=ses_get_identity_policies]{get_identity_policies}}.

See \url{https://www.paws-r-sdk.com/docs/ses_list_identity_policies/} for full documentation.
}
\keyword{internal}
