% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_list_receipt_filters}
\alias{ses_list_receipt_filters}
\title{Lists the IP address filters associated with your AWS account in the
current AWS Region}
\usage{
ses_list_receipt_filters()
}
\description{
Lists the IP address filters associated with your AWS account in the current AWS Region.

See \url{https://paws-r.github.io/docs/ses/list_receipt_filters.html} for full documentation.
}
\keyword{internal}
