% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_list_verified_email_addresses}
\alias{ses_list_verified_email_addresses}
\title{Deprecated}
\usage{
ses_list_verified_email_addresses()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  VerifiedEmailAddresses = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deprecated. Use the \code{\link[=ses_list_identities]{list_identities}} operation
to list the email addresses and domains associated with your account.
}
\examples{
\dontrun{
# The following example lists all email addresses that have been submitted
# for verification with Amazon SES:
svc$list_verified_email_addresses()
}

}
\keyword{internal}
