% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_put_configuration_set_delivery_options}
\alias{ses_put_configuration_set_delivery_options}
\title{Adds or updates the delivery options for a configuration set}
\usage{
ses_put_configuration_set_delivery_options(
  ConfigurationSetName,
  DeliveryOptions = NULL
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to specify the delivery
options for.}

\item{DeliveryOptions}{Specifies whether messages that use the configuration set are required
to use Transport Layer Security (TLS).}
}
\description{
Adds or updates the delivery options for a configuration set.

See \url{https://paws-r.github.io/docs/ses/put_configuration_set_delivery_options.html} for full documentation.
}
\keyword{internal}
