% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_send_custom_verification_email}
\alias{ses_send_custom_verification_email}
\title{Adds an email address to the list of identities for your Amazon SES
account in the current AWS Region and attempts to verify it}
\usage{
ses_send_custom_verification_email(
  EmailAddress,
  TemplateName,
  ConfigurationSetName = NULL
)
}
\arguments{
\item{EmailAddress}{[required] The email address to verify.}

\item{TemplateName}{[required] The name of the custom verification email template to use when sending
the verification email.}

\item{ConfigurationSetName}{Name of a configuration set to use when sending the verification email.}
}
\description{
Adds an email address to the list of identities for your Amazon SES account in the current AWS Region and attempts to verify it. As a result of executing this operation, a customized verification email is sent to the specified address.

See \url{https://paws-r.github.io/docs/ses/send_custom_verification_email.html} for full documentation.
}
\keyword{internal}
