% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_set_active_receipt_rule_set}
\alias{ses_set_active_receipt_rule_set}
\title{Sets the specified receipt rule set as the active receipt rule set}
\usage{
ses_set_active_receipt_rule_set(RuleSetName = NULL)
}
\arguments{
\item{RuleSetName}{The name of the receipt rule set to make active. Setting this value to
null disables all email receiving.}
}
\description{
Sets the specified receipt rule set as the active receipt rule set.

See \url{https://www.paws-r-sdk.com/docs/ses_set_active_receipt_rule_set/} for full documentation.
}
\keyword{internal}
