% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_set_receipt_rule_position}
\alias{ses_set_receipt_rule_position}
\title{Sets the position of the specified receipt rule in the receipt rule set}
\usage{
ses_set_receipt_rule_position(RuleSetName, RuleName, After = NULL)
}
\arguments{
\item{RuleSetName}{[required] The name of the receipt rule set that contains the receipt rule to
reposition.}

\item{RuleName}{[required] The name of the receipt rule to reposition.}

\item{After}{The name of the receipt rule after which to place the specified receipt
rule.}
}
\description{
Sets the position of the specified receipt rule in the receipt rule set.

See \url{https://www.paws-r-sdk.com/docs/ses_set_receipt_rule_position/} for full documentation.
}
\keyword{internal}
