% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_test_render_template}
\alias{ses_test_render_template}
\title{Creates a preview of the MIME content of an email when provided with a
template and a set of replacement data}
\usage{
ses_test_render_template(TemplateName, TemplateData)
}
\arguments{
\item{TemplateName}{[required] The name of the template that you want to render.}

\item{TemplateData}{[required] A list of replacement values to apply to the template. This parameter is
a JSON object, typically consisting of key-value pairs in which the keys
correspond to replacement tags in the email template.}
}
\description{
Creates a preview of the MIME content of an email when provided with a
template and a set of replacement data.
}
\details{
You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$test_render_template(
  TemplateName = "string",
  TemplateData = "string"
)
}
}

\keyword{internal}
