% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_update_receipt_rule}
\alias{ses_update_receipt_rule}
\title{Updates a receipt rule}
\usage{
ses_update_receipt_rule(RuleSetName, Rule)
}
\arguments{
\item{RuleSetName}{[required] The name of the receipt rule set that the receipt rule belongs to.}

\item{Rule}{[required] A data structure that contains the updated receipt rule information.}
}
\description{
Updates a receipt rule.

See \url{https://paws-r.github.io/docs/ses/update_receipt_rule.html} for full documentation.
}
\keyword{internal}
