% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_update_template}
\alias{ses_update_template}
\title{Updates an email template}
\usage{
ses_update_template(Template)
}
\arguments{
\item{Template}{[required]}
}
\description{
Updates an email template. Email templates enable you to send personalized email to one or more destinations in a single API operation. For more information, see the \href{https://docs.aws.amazon.com/ses/latest/dg/send-personalized-email-api.html}{Amazon SES Developer Guide}.

See \url{https://paws-r.github.io/docs/ses/update_template.html} for full documentation.
}
\keyword{internal}
