% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_verify_domain_dkim}
\alias{ses_verify_domain_dkim}
\title{Returns a set of DKIM tokens for a domain identity}
\usage{
ses_verify_domain_dkim(Domain)
}
\arguments{
\item{Domain}{[required] The name of the domain to be verified for Easy DKIM signing.}
}
\description{
Returns a set of DKIM tokens for a domain identity.

See \url{https://paws-r.github.io/docs/ses/verify_domain_dkim.html} for full documentation.
}
\keyword{internal}
