% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_verify_email_address}
\alias{ses_verify_email_address}
\title{Deprecated}
\usage{
ses_verify_email_address(EmailAddress)
}
\arguments{
\item{EmailAddress}{[required] The email address to be verified.}
}
\description{
Deprecated. Use the \code{\link[=ses_verify_email_identity]{verify_email_identity}} operation to verify a new email address.

See \url{https://paws-r.github.io/docs/ses/verify_email_address.html} for full documentation.
}
\keyword{internal}
