% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_associate_email_identity_certificate}
\alias{sesv2_associate_email_identity_certificate}
\title{Associates an S/MIME certificate with an email identity}
\usage{
sesv2_associate_email_identity_certificate(
  EmailIdentity,
  FromAddress = NULL,
  CertificateArn
)
}
\arguments{
\item{EmailIdentity}{[required] The email identity, either an email address or a domain, to associate the certificate with.}

\item{FromAddress}{The email address that the certificate applies to. This value is required when the email identity is a domain, and the address must belong to that domain or one of its subdomains. When the email identity is an email address, this value is optional. If you specify it, it must exactly match the email identity.}

\item{CertificateArn}{[required] The Amazon Resource Name (ARN) of the Certificate Manager (ACM) certificate to associate with the email identity.}
}
\description{
Associates an S/MIME certificate with an email identity. After the certificate is active, Amazon SES API v2 can add an S/MIME signature to messages that you send from the associated address when signing is enabled on the configuration set used to send the message.

See \url{https://www.paws-r-sdk.com/docs/sesv2_associate_email_identity_certificate/} for full documentation.
}
\keyword{internal}
