% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_batch_get_metric_data}
\alias{sesv2_batch_get_metric_data}
\title{Retrieves batches of metric data collected based on your sending
activity}
\usage{
sesv2_batch_get_metric_data(Queries)
}
\arguments{
\item{Queries}{[required] A list of queries for metrics to be retrieved.}
}
\description{
Retrieves batches of metric data collected based on your sending activity.

See \url{https://paws-r.github.io/docs/sesv2/batch_get_metric_data.html} for full documentation.
}
\keyword{internal}
