% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_cancel_export_job}
\alias{sesv2_cancel_export_job}
\title{Cancels an export job}
\usage{
sesv2_cancel_export_job(JobId)
}
\arguments{
\item{JobId}{[required] The export job ID.}
}
\description{
Cancels an export job.

See \url{https://www.paws-r-sdk.com/docs/sesv2_cancel_export_job/} for full documentation.
}
\keyword{internal}
