% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_create_configuration_set_event_destination}
\alias{sesv2_create_configuration_set_event_destination}
\title{Create an event destination}
\usage{
sesv2_create_configuration_set_event_destination(
  ConfigurationSetName,
  EventDestinationName,
  EventDestination
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set .}

\item{EventDestinationName}{[required] A name that identifies the event destination within the configuration
set.}

\item{EventDestination}{[required] An object that defines the event destination.}
}
\description{
Create an event destination. \emph{Events} include message sends, deliveries, opens, clicks, bounces, and complaints. \emph{Event destinations} are places that you can send information about these events to. For example, you can send event data to Amazon EventBridge and associate a rule to send the event to the specified target.

See \url{https://www.paws-r-sdk.com/docs/sesv2_create_configuration_set_event_destination/} for full documentation.
}
\keyword{internal}
