% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_create_contact}
\alias{sesv2_create_contact}
\title{Creates a contact, which is an end-user who is receiving the email, and
adds them to a contact list}
\usage{
sesv2_create_contact(
  ContactListName,
  EmailAddress,
  TopicPreferences = NULL,
  UnsubscribeAll = NULL,
  AttributesData = NULL
)
}
\arguments{
\item{ContactListName}{[required] The name of the contact list to which the contact should be added.}

\item{EmailAddress}{[required] The contact's email address.}

\item{TopicPreferences}{The contact's preferences for being opted-in to or opted-out of topics.}

\item{UnsubscribeAll}{A boolean value status noting if the contact is unsubscribed from all
contact list topics.}

\item{AttributesData}{The attribute data attached to a contact.}
}
\description{
Creates a contact, which is an end-user who is receiving the email, and adds them to a contact list.

See \url{https://paws-r.github.io/docs/sesv2/create_contact.html} for full documentation.
}
\keyword{internal}
