% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_create_export_job}
\alias{sesv2_create_export_job}
\title{Creates an export job for a data source and destination}
\usage{
sesv2_create_export_job(ExportDataSource, ExportDestination)
}
\arguments{
\item{ExportDataSource}{[required] The data source for the export job.}

\item{ExportDestination}{[required] The destination for the export job.}
}
\description{
Creates an export job for a data source and destination.

See \url{https://www.paws-r-sdk.com/docs/sesv2_create_export_job/} for full documentation.
}
\keyword{internal}
