% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_create_import_job}
\alias{sesv2_create_import_job}
\title{Creates an import job for a data destination}
\usage{
sesv2_create_import_job(ImportDestination, ImportDataSource)
}
\arguments{
\item{ImportDestination}{[required] The destination for the import job.}

\item{ImportDataSource}{[required] The data source for the import job.}
}
\description{
Creates an import job for a data destination.

See \url{https://www.paws-r-sdk.com/docs/sesv2_create_import_job/} for full documentation.
}
\keyword{internal}
