% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_create_multi_region_endpoint}
\alias{sesv2_create_multi_region_endpoint}
\title{Creates a multi-region endpoint (global-endpoint)}
\usage{
sesv2_create_multi_region_endpoint(EndpointName, Details, Tags = NULL)
}
\arguments{
\item{EndpointName}{[required] The name of the multi-region endpoint (global-endpoint).}

\item{Details}{[required] Contains details of a multi-region endpoint (global-endpoint) being
created.}

\item{Tags}{An array of objects that define the tags (keys and values) to associate
with the multi-region endpoint (global-endpoint).}
}
\description{
Creates a multi-region endpoint (global-endpoint).

See \url{https://www.paws-r-sdk.com/docs/sesv2_create_multi_region_endpoint/} for full documentation.
}
\keyword{internal}
