% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_create_tenant_resource_association}
\alias{sesv2_create_tenant_resource_association}
\title{Associate a resource with a tenant}
\usage{
sesv2_create_tenant_resource_association(TenantName, ResourceArn)
}
\arguments{
\item{TenantName}{[required] The name of the tenant to associate the resource with.}

\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to associate with the
tenant.}
}
\description{
Associate a resource with a tenant.

See \url{https://www.paws-r-sdk.com/docs/sesv2_create_tenant_resource_association/} for full documentation.
}
\keyword{internal}
