% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_configuration_set}
\alias{sesv2_delete_configuration_set}
\title{Delete an existing configuration set}
\usage{
sesv2_delete_configuration_set(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set.}
}
\description{
Delete an existing configuration set.

See \url{https://www.paws-r-sdk.com/docs/sesv2_delete_configuration_set/} for full documentation.
}
\keyword{internal}
