% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_email_identity_policy}
\alias{sesv2_delete_email_identity_policy}
\title{Deletes the specified sending authorization policy for the given
identity (an email address or a domain)}
\usage{
sesv2_delete_email_identity_policy(EmailIdentity, PolicyName)
}
\arguments{
\item{EmailIdentity}{[required] The email identity.}

\item{PolicyName}{[required] The name of the policy.

The policy name cannot exceed 64 characters and can only include
alphanumeric characters, dashes, and underscores.}
}
\description{
Deletes the specified sending authorization policy for the given identity (an email address or a domain). This API returns successfully even if a policy with the specified name does not exist.

See \url{https://paws-r.github.io/docs/sesv2/delete_email_identity_policy.html} for full documentation.
}
\keyword{internal}
