% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_multi_region_endpoint}
\alias{sesv2_delete_multi_region_endpoint}
\title{Deletes a multi-region endpoint (global-endpoint)}
\usage{
sesv2_delete_multi_region_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{[required] The name of the multi-region endpoint (global-endpoint) to be deleted.}
}
\description{
Deletes a multi-region endpoint (global-endpoint).

See \url{https://www.paws-r-sdk.com/docs/sesv2_delete_multi_region_endpoint/} for full documentation.
}
\keyword{internal}
