% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_tenant}
\alias{sesv2_delete_tenant}
\title{Delete an existing tenant}
\usage{
sesv2_delete_tenant(TenantName)
}
\arguments{
\item{TenantName}{[required] The name of the tenant to delete.}
}
\description{
Delete an existing tenant.

See \url{https://www.paws-r-sdk.com/docs/sesv2_delete_tenant/} for full documentation.
}
\keyword{internal}
