% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_tenant_resource_association}
\alias{sesv2_delete_tenant_resource_association}
\title{Delete an association between a tenant and a resource}
\usage{
sesv2_delete_tenant_resource_association(TenantName, ResourceArn)
}
\arguments{
\item{TenantName}{[required] The name of the tenant to remove the resource association from.}

\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to remove from the tenant
association.}
}
\description{
Delete an association between a tenant and a resource.

See \url{https://www.paws-r-sdk.com/docs/sesv2_delete_tenant_resource_association/} for full documentation.
}
\keyword{internal}
