% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_disassociate_email_identity_certificate}
\alias{sesv2_disassociate_email_identity_certificate}
\title{Removes the association between an S/MIME certificate and an email
identity}
\usage{
sesv2_disassociate_email_identity_certificate(
  EmailIdentity,
  FromAddress = NULL
)
}
\arguments{
\item{EmailIdentity}{[required] The email identity whose certificate association you want to remove.}

\item{FromAddress}{The email address whose certificate association you want to remove. This value is required when the email identity is a domain. When the email identity is an email address, this value is optional.}
}
\description{
Removes the association between an S/MIME certificate and an email identity. After the association is removed, Amazon SES API v2 stops adding an S/MIME signature to messages sent from that address.

See \url{https://www.paws-r-sdk.com/docs/sesv2_disassociate_email_identity_certificate/} for full documentation.
}
\keyword{internal}
