% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_account}
\alias{sesv2_get_account}
\title{Obtain information about the email-sending status and capabilities of
your Amazon SES account in the current Amazon Web Services Region}
\usage{
sesv2_get_account()
}
\description{
Obtain information about the email-sending status and capabilities of your Amazon SES account in the current Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_account/} for full documentation.
}
\keyword{internal}
