% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_configuration_set_event_destinations}
\alias{sesv2_get_configuration_set_event_destinations}
\title{Retrieve a list of event destinations that are associated with a
configuration set}
\usage{
sesv2_get_configuration_set_event_destinations(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that contains the event destination.}
}
\description{
Retrieve a list of event destinations that are associated with a configuration set.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_configuration_set_event_destinations/} for full documentation.
}
\keyword{internal}
