% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_contact}
\alias{sesv2_get_contact}
\title{Returns a contact from a contact list}
\usage{
sesv2_get_contact(ContactListName, EmailAddress)
}
\arguments{
\item{ContactListName}{[required] The name of the contact list to which the contact belongs.}

\item{EmailAddress}{[required] The contact's email addres.}
}
\description{
Returns a contact from a contact list.

See \url{https://paws-r.github.io/docs/sesv2/get_contact.html} for full documentation.
}
\keyword{internal}
