% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_contact_list}
\alias{sesv2_get_contact_list}
\title{Returns contact list metadata}
\usage{
sesv2_get_contact_list(ContactListName)
}
\arguments{
\item{ContactListName}{[required] The name of the contact list.}
}
\description{
Returns contact list metadata. It does not return any information about the contacts present in the list.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_contact_list/} for full documentation.
}
\keyword{internal}
