% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_dedicated_ip_pool}
\alias{sesv2_get_dedicated_ip_pool}
\title{Retrieve information about the dedicated pool}
\usage{
sesv2_get_dedicated_ip_pool(PoolName)
}
\arguments{
\item{PoolName}{[required] The name of the dedicated IP pool to retrieve.}
}
\description{
Retrieve information about the dedicated pool.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_dedicated_ip_pool/} for full documentation.
}
\keyword{internal}
