% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_deliverability_dashboard_options}
\alias{sesv2_get_deliverability_dashboard_options}
\title{Retrieve information about the status of the Deliverability dashboard
for your account}
\usage{
sesv2_get_deliverability_dashboard_options()
}
\description{
Retrieve information about the status of the Deliverability dashboard for your account. When the Deliverability dashboard is enabled, you gain access to reputation, deliverability, and other metrics for the domains that you use to send email. You also gain the ability to perform predictive inbox placement tests.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_deliverability_dashboard_options/} for full documentation.
}
\keyword{internal}
