% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_deliverability_test_report}
\alias{sesv2_get_deliverability_test_report}
\title{Retrieve the results of a predictive inbox placement test}
\usage{
sesv2_get_deliverability_test_report(ReportId)
}
\arguments{
\item{ReportId}{[required] A unique string that identifies the predictive inbox placement test.}
}
\description{
Retrieve the results of a predictive inbox placement test.

See \url{https://paws-r.github.io/docs/sesv2/get_deliverability_test_report.html} for full documentation.
}
\keyword{internal}
