% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_email_address_insights}
\alias{sesv2_get_email_address_insights}
\title{Provides validation insights about a specific email address, including
syntax validation, DNS record checks, mailbox existence, and other
deliverability factors}
\usage{
sesv2_get_email_address_insights(EmailAddress)
}
\arguments{
\item{EmailAddress}{[required] The email address to analyze for validation insights.}
}
\description{
Provides validation insights about a specific email address, including syntax validation, DNS record checks, mailbox existence, and other deliverability factors.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_email_address_insights/} for full documentation.
}
\keyword{internal}
