% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_email_identity}
\alias{sesv2_get_email_identity}
\title{Provides information about a specific identity, including the identity's
verification status, sending authorization policies, its DKIM
authentication status, and its custom Mail-From settings}
\usage{
sesv2_get_email_identity(EmailIdentity)
}
\arguments{
\item{EmailIdentity}{[required] The email identity.}
}
\description{
Provides information about a specific identity, including the identity's verification status, sending authorization policies, its DKIM authentication status, and its custom Mail-From settings.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_email_identity/} for full documentation.
}
\keyword{internal}
