% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_email_identity_policies}
\alias{sesv2_get_email_identity_policies}
\title{Returns the requested sending authorization policies for the given
identity (an email address or a domain)}
\usage{
sesv2_get_email_identity_policies(EmailIdentity)
}
\arguments{
\item{EmailIdentity}{[required] The email identity.}
}
\description{
Returns the requested sending authorization policies for the given identity (an email address or a domain). The policies are returned as a map of policy names to policy contents. You can retrieve a maximum of 20 policies at a time.

See \url{https://paws-r.github.io/docs/sesv2/get_email_identity_policies.html} for full documentation.
}
\keyword{internal}
