% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_message_insights}
\alias{sesv2_get_message_insights}
\title{Provides information about a specific message, including the from
address, the subject, the recipient address, email tags, as well as
events associated with the message}
\usage{
sesv2_get_message_insights(MessageId)
}
\arguments{
\item{MessageId}{[required] A \code{MessageId} is a unique identifier for a message, and is returned when
sending emails through Amazon SES.}
}
\description{
Provides information about a specific message, including the from address, the subject, the recipient address, email tags, as well as events associated with the message.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_message_insights/} for full documentation.
}
\keyword{internal}
