% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_multi_region_endpoint}
\alias{sesv2_get_multi_region_endpoint}
\title{Displays the multi-region endpoint (global-endpoint) configuration}
\usage{
sesv2_get_multi_region_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{[required] The name of the multi-region endpoint (global-endpoint).}
}
\description{
Displays the multi-region endpoint (global-endpoint) configuration.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_multi_region_endpoint/} for full documentation.
}
\keyword{internal}
