% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_reputation_entity}
\alias{sesv2_get_reputation_entity}
\title{Retrieve information about a specific reputation entity, including its
reputation management policy, customer-managed status, Amazon Web
Services Amazon SES-managed status, and aggregate sending status}
\usage{
sesv2_get_reputation_entity(ReputationEntityReference, ReputationEntityType)
}
\arguments{
\item{ReputationEntityReference}{[required] The unique identifier for the reputation entity. For resource-type
entities, this is the Amazon Resource Name (ARN) of the resource.}

\item{ReputationEntityType}{[required] The type of reputation entity. Currently, only \code{RESOURCE} type entities
are supported.}
}
\description{
Retrieve information about a specific reputation entity, including its reputation management policy, customer-managed status, Amazon Web Services Amazon SES-managed status, and aggregate sending status.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_reputation_entity/} for full documentation.
}
\keyword{internal}
