% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_tenant}
\alias{sesv2_get_tenant}
\title{Get information about a specific tenant, including the tenant's name,
ID, ARN, creation timestamp, tags, sending status, and suppression
attributes}
\usage{
sesv2_get_tenant(TenantName)
}
\arguments{
\item{TenantName}{[required] The name of the tenant to retrieve information about.}
}
\description{
Get information about a specific tenant, including the tenant's name, ID, ARN, creation timestamp, tags, sending status, and suppression attributes.

See \url{https://www.paws-r-sdk.com/docs/sesv2_get_tenant/} for full documentation.
}
\keyword{internal}
