% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_email_identities}
\alias{sesv2_list_email_identities}
\title{Returns a list of all of the email identities that are associated with
your Amazon Web Services account}
\usage{
sesv2_list_email_identities(Filter = NULL, NextToken = NULL, PageSize = NULL)
}
\arguments{
\item{Filter}{An object that contains filters to apply when listing email identities. You can filter by identity name, identity type, or verification status.}

\item{NextToken}{A token returned from a previous call to \code{\link[=sesv2_list_email_identities]{list_email_identities}} to indicate the position in the list of identities.}

\item{PageSize}{The number of results to show in a single call to \code{\link[=sesv2_list_email_identities]{list_email_identities}}. If the number of results is larger than the number you specified in this parameter, then the response includes a \code{NextToken} element, which you can use to obtain additional results.

The value you specify has to be at least 0, and can be no more than 1000.}
}
\description{
Returns a list of all of the email identities that are associated with your Amazon Web Services account. An identity can be either an email address or a domain. This operation returns identities that are verified as well as those that aren't. This operation returns identities that are associated with Amazon SES and Amazon Pinpoint.

See \url{https://www.paws-r-sdk.com/docs/sesv2_list_email_identities/} for full documentation.
}
\keyword{internal}
