% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_email_identity_certificates}
\alias{sesv2_list_email_identity_certificates}
\title{Lists the S/MIME certificates that are associated with the specified
email identity}
\usage{
sesv2_list_email_identity_certificates(
  EmailIdentity,
  NextToken = NULL,
  PageSize = NULL
)
}
\arguments{
\item{EmailIdentity}{[required] The email identity whose certificate associations you want to list.}

\item{NextToken}{A token returned from a previous call to \code{\link[=sesv2_list_email_identity_certificates]{list_email_identity_certificates}} to indicate the position in the list of certificates.}

\item{PageSize}{The number of results to show in a single call to \code{\link[=sesv2_list_email_identity_certificates]{list_email_identity_certificates}}. If the number of results is larger than the number you specified in this parameter, then the response includes a \code{NextToken} element, which you can use to obtain additional results.}
}
\description{
Lists the S/MIME certificates that are associated with the specified email identity. The results include certificates in all states, such as \code{PROVISIONING}, \code{ACTIVE}, \code{INACTIVE}, \code{DEPROVISIONING}, and \code{FAILED}.

See \url{https://www.paws-r-sdk.com/docs/sesv2_list_email_identity_certificates/} for full documentation.
}
\keyword{internal}
