% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_export_jobs}
\alias{sesv2_list_export_jobs}
\title{Lists all of the export jobs}
\usage{
sesv2_list_export_jobs(
  NextToken = NULL,
  PageSize = NULL,
  ExportSourceType = NULL,
  JobStatus = NULL
)
}
\arguments{
\item{NextToken}{The pagination token returned from a previous call to
\code{\link[=sesv2_list_export_jobs]{list_export_jobs}} to indicate the position in
the list of export jobs.}

\item{PageSize}{Maximum number of export jobs to return at once. Use this parameter to
paginate results. If additional export jobs exist beyond the specified
limit, the \code{NextToken} element is sent in the response. Use the
\code{NextToken} value in subsequent calls to
\code{\link[=sesv2_list_export_jobs]{list_export_jobs}} to retrieve additional
export jobs.}

\item{ExportSourceType}{A value used to list export jobs that have a certain \code{ExportSourceType}.}

\item{JobStatus}{A value used to list export jobs that have a certain \code{JobStatus}.}
}
\description{
Lists all of the export jobs.

See \url{https://www.paws-r-sdk.com/docs/sesv2_list_export_jobs/} for full documentation.
}
\keyword{internal}
