% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_import_jobs}
\alias{sesv2_list_import_jobs}
\title{Lists all of the import jobs}
\usage{
sesv2_list_import_jobs(
  ImportDestinationType = NULL,
  NextToken = NULL,
  PageSize = NULL
)
}
\arguments{
\item{ImportDestinationType}{The destination of the import job, which can be used to list import jobs
that have a certain \code{ImportDestinationType}.}

\item{NextToken}{A string token indicating that there might be additional import jobs
available to be listed. Copy this token to a subsequent call to
\code{\link[=sesv2_list_import_jobs]{list_import_jobs}} with the same parameters to
retrieve the next page of import jobs.}

\item{PageSize}{Maximum number of import jobs to return at once. Use this parameter to
paginate results. If additional import jobs exist beyond the specified
limit, the \code{NextToken} element is sent in the response. Use the
\code{NextToken} value in subsequent requests to retrieve additional
addresses.}
}
\description{
Lists all of the import jobs.

See \url{https://www.paws-r-sdk.com/docs/sesv2_list_import_jobs/} for full documentation.
}
\keyword{internal}
