% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_multi_region_endpoints}
\alias{sesv2_list_multi_region_endpoints}
\title{List the multi-region endpoints (global-endpoints)}
\usage{
sesv2_list_multi_region_endpoints(NextToken = NULL, PageSize = NULL)
}
\arguments{
\item{NextToken}{A token returned from a previous call to
\code{\link[=sesv2_list_multi_region_endpoints]{list_multi_region_endpoints}} to
indicate the position in the list of multi-region endpoints
(global-endpoints).}

\item{PageSize}{The number of results to show in a single call to
\code{\link[=sesv2_list_multi_region_endpoints]{list_multi_region_endpoints}}. If
the number of results is larger than the number you specified in this
parameter, the response includes a \code{NextToken} element that you can use
to retrieve the next page of results.}
}
\description{
List the multi-region endpoints (global-endpoints).

See \url{https://www.paws-r-sdk.com/docs/sesv2_list_multi_region_endpoints/} for full documentation.
}
\keyword{internal}
