% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_recommendations}
\alias{sesv2_list_recommendations}
\title{Lists the recommendations present in your Amazon SES account in the
current Amazon Web Services Region}
\usage{
sesv2_list_recommendations(Filter = NULL, NextToken = NULL, PageSize = NULL)
}
\arguments{
\item{Filter}{Filters applied when retrieving recommendations. Can eiter be an
individual filter, or combinations of \code{STATUS} and \code{IMPACT} or \code{STATUS}
and \code{TYPE}}

\item{NextToken}{A token returned from a previous call to
\code{\link[=sesv2_list_recommendations]{list_recommendations}} to indicate the
position in the list of recommendations.}

\item{PageSize}{The number of results to show in a single call to
\code{\link[=sesv2_list_recommendations]{list_recommendations}}. If the number of
results is larger than the number you specified in this parameter, then
the response includes a \code{NextToken} element, which you can use to obtain
additional results.

The value you specify has to be at least 1, and can be no more than 100.}
}
\description{
Lists the recommendations present in your Amazon SES account in the current Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/sesv2_list_recommendations/} for full documentation.
}
\keyword{internal}
