% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_reputation_entities}
\alias{sesv2_list_reputation_entities}
\title{List reputation entities in your Amazon SES account in the current
Amazon Web Services Region}
\usage{
sesv2_list_reputation_entities(
  Filter = NULL,
  NextToken = NULL,
  PageSize = NULL
)
}
\arguments{
\item{Filter}{An object that contains filters to apply when listing reputation
entities. You can filter by entity type, reputation impact, sending
status, or entity reference prefix.}

\item{NextToken}{A token returned from a previous call to
\code{\link[=sesv2_list_reputation_entities]{list_reputation_entities}} to indicate
the position in the list of reputation entities.}

\item{PageSize}{The number of results to show in a single call to
\code{\link[=sesv2_list_reputation_entities]{list_reputation_entities}}. If the
number of results is larger than the number you specified in this
parameter, then the response includes a \code{NextToken} element, which you
can use to obtain additional results.}
}
\description{
List reputation entities in your Amazon SES account in the current Amazon Web Services Region. You can filter the results by entity type, reputation impact, sending status, or entity reference prefix.

See \url{https://www.paws-r-sdk.com/docs/sesv2_list_reputation_entities/} for full documentation.
}
\keyword{internal}
