% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_resource_tenants}
\alias{sesv2_list_resource_tenants}
\title{List all tenants associated with a specific resource}
\usage{
sesv2_list_resource_tenants(ResourceArn, PageSize = NULL, NextToken = NULL)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to list associated
tenants for.}

\item{PageSize}{The number of results to show in a single call to
\code{\link[=sesv2_list_resource_tenants]{list_resource_tenants}}. If the number of
results is larger than the number you specified in this parameter, then
the response includes a \code{NextToken} element, which you can use to obtain
additional results.}

\item{NextToken}{A token returned from a previous call to
\code{\link[=sesv2_list_resource_tenants]{list_resource_tenants}} to indicate the
position in the list of resource tenants.}
}
\description{
List all tenants associated with a specific resource.

See \url{https://www.paws-r-sdk.com/docs/sesv2_list_resource_tenants/} for full documentation.
}
\keyword{internal}
