% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_tenant_resources}
\alias{sesv2_list_tenant_resources}
\title{List all resources associated with a specific tenant}
\usage{
sesv2_list_tenant_resources(
  TenantName,
  Filter = NULL,
  PageSize = NULL,
  NextToken = NULL
)
}
\arguments{
\item{TenantName}{[required] The name of the tenant to list resources for.}

\item{Filter}{A map of filter keys and values for filtering the list of tenant
resources. Currently, the only supported filter key is \code{RESOURCE_TYPE}.}

\item{PageSize}{The number of results to show in a single call to
\code{\link[=sesv2_list_tenant_resources]{list_tenant_resources}}. If the number of
results is larger than the number you specified in this parameter, then
the response includes a \code{NextToken} element, which you can use to obtain
additional results.}

\item{NextToken}{A token returned from a previous call to
\code{\link[=sesv2_list_tenant_resources]{list_tenant_resources}} to indicate the
position in the list of tenant resources.}
}
\description{
List all resources associated with a specific tenant.

See \url{https://www.paws-r-sdk.com/docs/sesv2_list_tenant_resources/} for full documentation.
}
\keyword{internal}
