% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_list_tenants}
\alias{sesv2_list_tenants}
\title{List all tenants associated with your account in the current Amazon Web
Services Region}
\usage{
sesv2_list_tenants(Filter = NULL, NextToken = NULL, PageSize = NULL)
}
\arguments{
\item{Filter}{An object that contains filters to apply when listing tenants. You can filter by tenant name or sending status.}

\item{NextToken}{A token returned from a previous call to \code{\link[=sesv2_list_tenants]{list_tenants}} to indicate the position in the list of tenants.}

\item{PageSize}{The number of results to show in a single call to \code{\link[=sesv2_list_tenants]{list_tenants}}. If the number of results is larger than the number you specified in this parameter, then the response includes a \code{NextToken} element, which you can use to obtain additional results.}
}
\description{
List all tenants associated with your account in the current Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/sesv2_list_tenants/} for full documentation.
}
\keyword{internal}
