% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_account_pricing_attributes}
\alias{sesv2_put_account_pricing_attributes}
\title{Set the pricing plan for your Amazon SES account}
\usage{
sesv2_put_account_pricing_attributes(Plan)
}
\arguments{
\item{Plan}{[required] The pricing plan to apply to your Amazon SES account. For details about each plan, see \href{https://aws.amazon.com/ses/pricing/}{Amazon SES Pricing}. Can be one of the following:
\itemize{
\item \code{NONE}
\item \code{ESSENTIALS}
\item \code{PRO}
\item \code{ENTERPRISE}
}}
}
\description{
Set the pricing plan for your Amazon SES account.

See \url{https://www.paws-r-sdk.com/docs/sesv2_put_account_pricing_attributes/} for full documentation.
}
\keyword{internal}
