% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_configuration_set_reputation_options}
\alias{sesv2_put_configuration_set_reputation_options}
\title{Enable or disable collection of reputation metrics for emails that you
send using a particular configuration set in a specific Amazon Web
Services Region}
\usage{
sesv2_put_configuration_set_reputation_options(
  ConfigurationSetName,
  ReputationMetricsEnabled = NULL
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set.}

\item{ReputationMetricsEnabled}{If \code{true}, tracking of reputation metrics is enabled for the
configuration set. If \code{false}, tracking of reputation metrics is
disabled for the configuration set.}
}
\description{
Enable or disable collection of reputation metrics for emails that you send using a particular configuration set in a specific Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/sesv2_put_configuration_set_reputation_options/} for full documentation.
}
\keyword{internal}
