% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_configuration_set_sending_options}
\alias{sesv2_put_configuration_set_sending_options}
\title{Enable or disable email sending for messages that use a particular
configuration set in a specific Amazon Web Services Region}
\usage{
sesv2_put_configuration_set_sending_options(
  ConfigurationSetName,
  SendingEnabled = NULL
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set to enable or disable email sending
for.}

\item{SendingEnabled}{If \code{true}, email sending is enabled for the configuration set. If
\code{false}, email sending is disabled for the configuration set.}
}
\description{
Enable or disable email sending for messages that use a particular configuration set in a specific Amazon Web Services Region.

See \url{https://paws-r.github.io/docs/sesv2/put_configuration_set_sending_options.html} for full documentation.
}
\keyword{internal}
