% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_configuration_set_tracking_options}
\alias{sesv2_put_configuration_set_tracking_options}
\title{Specify a custom domain to use for open and click tracking elements in
email that you send}
\usage{
sesv2_put_configuration_set_tracking_options(
  ConfigurationSetName,
  CustomRedirectDomain = NULL
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set.}

\item{CustomRedirectDomain}{The domain to use to track open and click events.}
}
\description{
Specify a custom domain to use for open and click tracking elements in email that you send.

See \url{https://paws-r.github.io/docs/sesv2/put_configuration_set_tracking_options.html} for full documentation.
}
\keyword{internal}
