% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_configuration_set_vdm_options}
\alias{sesv2_put_configuration_set_vdm_options}
\title{Specify VDM preferences for email that you send using the configuration
set}
\usage{
sesv2_put_configuration_set_vdm_options(
  ConfigurationSetName,
  VdmOptions = NULL
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set.}

\item{VdmOptions}{The VDM options to apply to the configuration set.}
}
\description{
Specify VDM preferences for email that you send using the configuration set.

See \url{https://www.paws-r-sdk.com/docs/sesv2_put_configuration_set_vdm_options/} for full documentation.
}
\keyword{internal}
