% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_dedicated_ip_in_pool}
\alias{sesv2_put_dedicated_ip_in_pool}
\title{Move a dedicated IP address to an existing dedicated IP pool}
\usage{
sesv2_put_dedicated_ip_in_pool(Ip, DestinationPoolName)
}
\arguments{
\item{Ip}{[required] The IP address that you want to move to the dedicated IP pool. The value
you specify has to be a dedicated IP address that's associated with your
Amazon Web Services account.}

\item{DestinationPoolName}{[required] The name of the IP pool that you want to add the dedicated IP address
to. You have to specify an IP pool that already exists.}
}
\description{
Move a dedicated IP address to an existing dedicated IP pool.

See \url{https://paws-r.github.io/docs/sesv2/put_dedicated_ip_in_pool.html} for full documentation.
}
\keyword{internal}
