% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_dedicated_ip_pool_scaling_attributes}
\alias{sesv2_put_dedicated_ip_pool_scaling_attributes}
\title{Used to convert a dedicated IP pool to a different scaling mode}
\usage{
sesv2_put_dedicated_ip_pool_scaling_attributes(PoolName, ScalingMode)
}
\arguments{
\item{PoolName}{[required] The name of the dedicated IP pool.}

\item{ScalingMode}{[required] The scaling mode to apply to the dedicated IP pool.

Changing the scaling mode from \code{MANAGED} to \code{STANDARD} is not supported.}
}
\description{
Used to convert a dedicated IP pool to a different scaling mode.

See \url{https://www.paws-r-sdk.com/docs/sesv2_put_dedicated_ip_pool_scaling_attributes/} for full documentation.
}
\keyword{internal}
