% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_deliverability_dashboard_option}
\alias{sesv2_put_deliverability_dashboard_option}
\title{Enable or disable the Deliverability dashboard}
\usage{
sesv2_put_deliverability_dashboard_option(
  DashboardEnabled,
  SubscribedDomains = NULL
)
}
\arguments{
\item{DashboardEnabled}{[required] Specifies whether to enable the Deliverability dashboard. To enable the
dashboard, set this value to \code{true}.}

\item{SubscribedDomains}{An array of objects, one for each verified domain that you use to send
email and enabled the Deliverability dashboard for.}
}
\description{
Enable or disable the Deliverability dashboard. When you enable the Deliverability dashboard, you gain access to reputation, deliverability, and other metrics for the domains that you use to send email. You also gain the ability to perform predictive inbox placement tests.

See \url{https://paws-r.github.io/docs/sesv2/put_deliverability_dashboard_option.html} for full documentation.
}
\keyword{internal}
