% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_email_identity_configuration_set_attributes}
\alias{sesv2_put_email_identity_configuration_set_attributes}
\title{Used to associate a configuration set with an email identity}
\usage{
sesv2_put_email_identity_configuration_set_attributes(
  EmailIdentity,
  ConfigurationSetName = NULL
)
}
\arguments{
\item{EmailIdentity}{[required] The email address or domain to associate with a configuration set.}

\item{ConfigurationSetName}{The configuration set to associate with an email identity.}
}
\description{
Used to associate a configuration set with an email identity.

See \url{https://www.paws-r-sdk.com/docs/sesv2_put_email_identity_configuration_set_attributes/} for full documentation.
}
\keyword{internal}
