% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_send_custom_verification_email}
\alias{sesv2_send_custom_verification_email}
\title{Adds an email address to the list of identities for your Amazon SES
account in the current Amazon Web Services Region and attempts to verify
it}
\usage{
sesv2_send_custom_verification_email(
  EmailAddress,
  TemplateName,
  ConfigurationSetName = NULL
)
}
\arguments{
\item{EmailAddress}{[required] The email address to verify.}

\item{TemplateName}{[required] The name of the custom verification email template to use when sending
the verification email.}

\item{ConfigurationSetName}{Name of a configuration set to use when sending the verification email.}
}
\description{
Adds an email address to the list of identities for your Amazon SES account in the current Amazon Web Services Region and attempts to verify it. As a result of executing this operation, a customized verification email is sent to the specified address.

See \url{https://www.paws-r-sdk.com/docs/sesv2_send_custom_verification_email/} for full documentation.
}
\keyword{internal}
