% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_test_render_email_template}
\alias{sesv2_test_render_email_template}
\title{Creates a preview of the MIME content of an email when provided with a
template and a set of replacement data}
\usage{
sesv2_test_render_email_template(TemplateName, TemplateData)
}
\arguments{
\item{TemplateName}{[required] The name of the template.}

\item{TemplateData}{[required] A list of replacement values to apply to the template. This parameter is
a JSON object, typically consisting of key-value pairs in which the keys
correspond to replacement tags in the email template.}
}
\description{
Creates a preview of the MIME content of an email when provided with a template and a set of replacement data.

See \url{https://paws-r.github.io/docs/sesv2/test_render_email_template.html} for full documentation.
}
\keyword{internal}
