% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_untag_resource}
\alias{sesv2_untag_resource}
\title{Remove one or more tags (keys and values) from a specified resource}
\usage{
sesv2_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to remove
one or more tags from.}

\item{TagKeys}{[required] The tags (tag keys) that you want to remove from the resource. When you
specify a tag key, the action removes both that key and its associated
tag value.

To remove more than one tag from the resource, append the \code{TagKeys}
parameter and argument for each additional tag to remove, separated by
an ampersand. For example:
\verb{/v2/email/tags?ResourceArn=ResourceArn&TagKeys=Key1&TagKeys=Key2}}
}
\description{
Remove one or more tags (keys and values) from a specified resource.

See \url{https://www.paws-r-sdk.com/docs/sesv2_untag_resource/} for full documentation.
}
\keyword{internal}
