% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_update_configuration_set}
\alias{sesv2_update_configuration_set}
\title{Updates an existing configuration set}
\usage{
sesv2_update_configuration_set(
  ConfigurationSetName,
  MessageSecurityOptions = NULL
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set to update.}

\item{MessageSecurityOptions}{The security options that apply to the MIME message itself for messages sent with the configuration set.}
}
\description{
Updates an existing configuration set.

See \url{https://www.paws-r-sdk.com/docs/sesv2_update_configuration_set/} for full documentation.
}
\keyword{internal}
