% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_update_configuration_set_event_destination}
\alias{sesv2_update_configuration_set_event_destination}
\title{Update the configuration of an event destination for a configuration set}
\usage{
sesv2_update_configuration_set_event_destination(
  ConfigurationSetName,
  EventDestinationName,
  EventDestination
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that contains the event destination to
modify.}

\item{EventDestinationName}{[required] The name of the event destination.}

\item{EventDestination}{[required] An object that defines the event destination.}
}
\description{
Update the configuration of an event destination for a configuration set.

See \url{https://www.paws-r-sdk.com/docs/sesv2_update_configuration_set_event_destination/} for full documentation.
}
\keyword{internal}
