% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_update_reputation_entity_customer_managed_status}
\alias{sesv2_update_reputation_entity_customer_managed_status}
\title{Update the customer-managed sending status for a reputation entity}
\usage{
sesv2_update_reputation_entity_customer_managed_status(
  ReputationEntityType,
  ReputationEntityReference,
  SendingStatus
)
}
\arguments{
\item{ReputationEntityType}{[required] The type of reputation entity. Currently, only \code{RESOURCE} type entities
are supported.}

\item{ReputationEntityReference}{[required] The unique identifier for the reputation entity. For resource-type
entities, this is the Amazon Resource Name (ARN) of the resource.}

\item{SendingStatus}{[required] The new customer-managed sending status for the reputation entity. This
can be one of the following:
\itemize{
\item \code{ENABLED} – Allow sending for this entity.
\item \code{DISABLED} – Prevent sending for this entity.
\item \code{REINSTATED} – Allow sending even if there are active reputation
findings.
}}
}
\description{
Update the customer-managed sending status for a reputation entity. This allows you to enable, disable, or reinstate sending for the entity.

See \url{https://www.paws-r-sdk.com/docs/sesv2_update_reputation_entity_customer_managed_status/} for full documentation.
}
\keyword{internal}
