% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_approved_origin}
\alias{connect_associate_approved_origin}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_associate_approved_origin(InstanceId, Origin)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{Origin}{[required] The domain to add to your allow list.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://www.paws-r-sdk.com/docs/connect_associate_approved_origin/} for full documentation.
}
\keyword{internal}
