% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_lambda_function}
\alias{connect_associate_lambda_function}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_associate_lambda_function(InstanceId, FunctionArn)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{FunctionArn}{[required] The Amazon Resource Name (ARN) for the Lambda function being associated.
Maximum number of characters allowed is 140.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Allows the specified Amazon Connect instance to access the specified
Lambda function.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_lambda_function(
  InstanceId = "string",
  FunctionArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
