% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_lex_bot}
\alias{connect_associate_lex_bot}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_associate_lex_bot(InstanceId, LexBot)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{LexBot}{[required] The Amazon Lex bot to associate with the instance.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/associate_lex_bot.html} for full documentation.
}
\keyword{internal}
