% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_phone_number_contact_flow}
\alias{connect_associate_phone_number_contact_flow}
\title{Associates a contact flow with a phone number claimed to your Amazon
Connect instance}
\usage{
connect_associate_phone_number_contact_flow(
  PhoneNumberId,
  InstanceId,
  ContactFlowId
)
}
\arguments{
\item{PhoneNumberId}{[required] A unique identifier for the phone number.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ContactFlowId}{[required] The identifier of the contact flow.}
}
\description{
Associates a contact flow with a phone number claimed to your Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/associate_phone_number_contact_flow.html} for full documentation.
}
\keyword{internal}
