% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_queue_quick_connects}
\alias{connect_associate_queue_quick_connects}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_associate_queue_quick_connects(InstanceId, QueueId, QuickConnectIds)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{QuickConnectIds}{[required] The quick connects to associate with this queue.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/associate_queue_quick_connects.html} for full documentation.
}
\keyword{internal}
