% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_routing_profile_queues}
\alias{connect_associate_routing_profile_queues}
\title{Associates a set of queues with a routing profile}
\usage{
connect_associate_routing_profile_queues(
  InstanceId,
  RoutingProfileId,
  QueueConfigs
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{QueueConfigs}{[required] The queues to associate with this routing profile.}
}
\description{
Associates a set of queues with a routing profile.

See \url{https://paws-r.github.io/docs/connect/associate_routing_profile_queues.html} for full documentation.
}
\keyword{internal}
