% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_associate_security_key}
\alias{connect_associate_security_key}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_associate_security_key(InstanceId, Key)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{Key}{[required] A valid security key in PEM format.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Associates a security key to the instance.
}
\section{Request syntax}{
\preformatted{svc$associate_security_key(
  InstanceId = "string",
  Key = "string"
)
}
}

\keyword{internal}
