% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_batch_get_attached_file_metadata}
\alias{connect_batch_get_attached_file_metadata}
\title{Allows you to retrieve metadata about multiple attached files on an
associated resource}
\usage{
connect_batch_get_attached_file_metadata(
  FileIds,
  InstanceId,
  AssociatedResourceArn
)
}
\arguments{
\item{FileIds}{[required] The unique identifiers of the attached file resource.}

\item{InstanceId}{[required] The unique identifier of the Connect instance.}

\item{AssociatedResourceArn}{[required] The resource to which the attached file is (being) uploaded to. The
supported resources are
\href{https://docs.aws.amazon.com/connect/latest/adminguide/cases.html}{Cases}
and
\href{https://docs.aws.amazon.com/connect/latest/adminguide/setup-email-channel.html}{Email}.

This value must be a valid ARN.}
}
\description{
Allows you to retrieve metadata about multiple attached files on an associated resource. Each attached file provided in the input list must be associated with the input AssociatedResourceArn.

See \url{https://www.paws-r-sdk.com/docs/connect_batch_get_attached_file_metadata/} for full documentation.
}
\keyword{internal}
