% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_claim_phone_number}
\alias{connect_claim_phone_number}
\title{Claims an available phone number to your Amazon Connect instance}
\usage{
connect_claim_phone_number(
  TargetArn,
  PhoneNumber,
  PhoneNumberDescription = NULL,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{TargetArn}{[required] The Amazon Resource Name (ARN) for Amazon Connect instances that phone
numbers are claimed to.}

\item{PhoneNumber}{[required] The phone number you want to claim. Phone numbers are formatted
\verb{[+] [country code] [subscriber number including area code]}.}

\item{PhoneNumberDescription}{The description of the phone number.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Claims an available phone number to your Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/claim_phone_number.html} for full documentation.
}
\keyword{internal}
