% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_complete_attached_file_upload}
\alias{connect_complete_attached_file_upload}
\title{Allows you to confirm that the attached file has been uploaded using the
pre-signed URL provided in the StartAttachedFileUpload API}
\usage{
connect_complete_attached_file_upload(
  InstanceId,
  FileId,
  AssociatedResourceArn
)
}
\arguments{
\item{InstanceId}{[required] The unique identifier of the Amazon Connect instance.}

\item{FileId}{[required] The unique identifier of the attached file resource.}

\item{AssociatedResourceArn}{[required] The resource to which the attached file is (being) uploaded to. The
supported resources are
\href{https://docs.aws.amazon.com/connect/latest/adminguide/cases.html}{Cases}
and
\href{https://docs.aws.amazon.com/connect/latest/adminguide/setup-email-channel.html}{Email}.

This value must be a valid ARN.}
}
\description{
Allows you to confirm that the attached file has been uploaded using the pre-signed URL provided in the StartAttachedFileUpload API.

See \url{https://www.paws-r-sdk.com/docs/connect_complete_attached_file_upload/} for full documentation.
}
\keyword{internal}
