% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_agent_status}
\alias{connect_create_agent_status}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_create_agent_status(
  InstanceId,
  Name,
  Description = NULL,
  State,
  DisplayOrder = NULL,
  Tags = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{Name}{[required] The name of the status.}

\item{Description}{The description of the status.}

\item{State}{[required] The state of the status.}

\item{DisplayOrder}{The display order of the status.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/create_agent_status.html} for full documentation.
}
\keyword{internal}
