% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_hours_of_operation}
\alias{connect_create_hours_of_operation}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_create_hours_of_operation(
  InstanceId,
  Name,
  Description = NULL,
  TimeZone,
  Config,
  Tags = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{Name}{[required] The name of the hours of operation.}

\item{Description}{The description of the hours of operation.}

\item{TimeZone}{[required] The time zone of the hours of operation.}

\item{Config}{[required] Configuration information for the hours of operation: day, start time,
and end time.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/create_hours_of_operation.html} for full documentation.
}
\keyword{internal}
