% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_hours_of_operation}
\alias{connect_create_hours_of_operation}
\title{Creates hours of operation}
\usage{
connect_create_hours_of_operation(
  InstanceId,
  Name,
  Description = NULL,
  TimeZone,
  Config,
  Tags = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{Name}{[required] The name of the hours of operation.}

\item{Description}{The description of the hours of operation.}

\item{TimeZone}{[required] The time zone of the hours of operation.}

\item{Config}{[required] Configuration information for the hours of operation: day, start time,
and end time.}

\item{Tags}{The tags used to organize, track, or control access for this resource.
For example, \{ "Tags": \{"key1":"value1", "key2":"value2"\} \}.}
}
\description{
Creates hours of operation.

See \url{https://www.paws-r-sdk.com/docs/connect_create_hours_of_operation/} for full documentation.
}
\keyword{internal}
