% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_instance}
\alias{connect_create_instance}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_create_instance(
  ClientToken = NULL,
  IdentityManagementType,
  InstanceAlias = NULL,
  DirectoryId = NULL,
  InboundCallsEnabled,
  OutboundCallsEnabled
)
}
\arguments{
\item{ClientToken}{The idempotency token.}

\item{IdentityManagementType}{[required] The type of identity management for your Amazon Connect users.}

\item{InstanceAlias}{The name for your instance.}

\item{DirectoryId}{The identifier for the directory.}

\item{InboundCallsEnabled}{[required] Your contact center handles incoming contacts.}

\item{OutboundCallsEnabled}{[required] Your contact center allows outbound calls.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://www.paws-r-sdk.com/docs/connect_create_instance/} for full documentation.
}
\keyword{internal}
