% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_integration_association}
\alias{connect_create_integration_association}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_create_integration_association(InstanceId, IntegrationType,
  IntegrationArn, SourceApplicationUrl, SourceApplicationName, SourceType)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{IntegrationType}{[required] The type of information to be ingested.}

\item{IntegrationArn}{[required] The Amazon Resource Name (ARN) of the integration.}

\item{SourceApplicationUrl}{[required] The URL for the external application.}

\item{SourceApplicationName}{[required] The name of the external application.}

\item{SourceType}{[required] The type of the data source.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Create an AppIntegration association with an Amazon Connect instance.
}
\section{Request syntax}{
\preformatted{svc$create_integration_association(
  InstanceId = "string",
  IntegrationType = "EVENT",
  IntegrationArn = "string",
  SourceApplicationUrl = "string",
  SourceApplicationName = "string",
  SourceType = "SALESFORCE"|"ZENDESK"
)
}
}

\keyword{internal}
