% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_integration_association}
\alias{connect_create_integration_association}
\title{Create an AppIntegration association with an Amazon Connect instance}
\usage{
connect_create_integration_association(InstanceId, IntegrationType,
  IntegrationArn, SourceApplicationUrl, SourceApplicationName, SourceType,
  Tags)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{IntegrationType}{[required] The type of information to be ingested.}

\item{IntegrationArn}{[required] The Amazon Resource Name (ARN) of the integration.}

\item{SourceApplicationUrl}{[required] The URL for the external application.}

\item{SourceApplicationName}{[required] The name of the external application.}

\item{SourceType}{[required] The type of the data source.}

\item{Tags}{One or more tags.}
}
\value{
A list with the following syntax:\preformatted{list(
  IntegrationAssociationId = "string",
  IntegrationAssociationArn = "string"
)
}
}
\description{
Create an AppIntegration association with an Amazon Connect instance.
}
\section{Request syntax}{
\preformatted{svc$create_integration_association(
  InstanceId = "string",
  IntegrationType = "EVENT",
  IntegrationArn = "string",
  SourceApplicationUrl = "string",
  SourceApplicationName = "string",
  SourceType = "SALESFORCE"|"ZENDESK",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
