% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_quick_connect}
\alias{connect_create_quick_connect}
\title{Creates a quick connect for the specified Amazon Connect instance}
\usage{
connect_create_quick_connect(
  InstanceId,
  Name,
  Description = NULL,
  QuickConnectConfig,
  Tags = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{Name}{[required] A unique name of the quick connect.}

\item{Description}{The description of the quick connect.}

\item{QuickConnectConfig}{[required] Configuration settings for the quick connect.}

\item{Tags}{The tags used to organize, track, or control access for this resource.
For example, \{ "Tags": \{"key1":"value1", "key2":"value2"\} \}.}
}
\description{
Creates a quick connect for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_create_quick_connect/} for full documentation.
}
\keyword{internal}
