% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_quick_connect}
\alias{connect_create_quick_connect}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_create_quick_connect(InstanceId, Name, Description,
  QuickConnectConfig, Tags)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{Name}{[required] The name of the quick connect.}

\item{Description}{The description of the quick connect.}

\item{QuickConnectConfig}{[required] Configuration settings for the quick connect.}

\item{Tags}{One or more tags.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.
}
\details{
Creates a quick connect for the specified Amazon Connect instance.
}
\section{Request syntax}{
\preformatted{svc$create_quick_connect(
  InstanceId = "string",
  Name = "string",
  Description = "string",
  QuickConnectConfig = list(
    QuickConnectType = "USER"|"QUEUE"|"PHONE_NUMBER",
    UserConfig = list(
      UserId = "string",
      ContactFlowId = "string"
    ),
    QueueConfig = list(
      QueueId = "string",
      ContactFlowId = "string"
    ),
    PhoneConfig = list(
      PhoneNumber = "string"
    )
  ),
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
