% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_routing_profile}
\alias{connect_create_routing_profile}
\title{Creates a new routing profile}
\usage{
connect_create_routing_profile(InstanceId, Name, Description,
  DefaultOutboundQueueId, QueueConfigs, MediaConcurrencies, Tags)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{Name}{[required] The name of the routing profile. Must not be more than 127 characters.}

\item{Description}{[required] Description of the routing profile. Must not be more than 250
characters.}

\item{DefaultOutboundQueueId}{[required] The default outbound queue for the routing profile.}

\item{QueueConfigs}{The inbound queues associated with the routing profile. If no queue is
added, the agent can only make outbound calls.}

\item{MediaConcurrencies}{[required] The channels agents can handle in the Contact Control Panel (CCP) for
this routing profile.}

\item{Tags}{One or more tags.}
}
\description{
Creates a new routing profile.
}
\section{Request syntax}{
\preformatted{svc$create_routing_profile(
  InstanceId = "string",
  Name = "string",
  Description = "string",
  DefaultOutboundQueueId = "string",
  QueueConfigs = list(
    list(
      QueueReference = list(
        QueueId = "string",
        Channel = "VOICE"|"CHAT"|"TASK"
      ),
      Priority = 123,
      Delay = 123
    )
  ),
  MediaConcurrencies = list(
    list(
      Channel = "VOICE"|"CHAT"|"TASK",
      Concurrency = 123
    )
  ),
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
