% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_security_profile}
\alias{connect_create_security_profile}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_create_security_profile(
  SecurityProfileName,
  Description = NULL,
  Permissions = NULL,
  InstanceId,
  Tags = NULL
)
}
\arguments{
\item{SecurityProfileName}{[required] The name of the security profile.}

\item{Description}{The description of the security profile.}

\item{Permissions}{Permissions assigned to the security profile.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{Tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/create_security_profile.html} for full documentation.
}
\keyword{internal}
