% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_use_case}
\alias{connect_create_use_case}
\title{Creates a use case for an AppIntegration association}
\usage{
connect_create_use_case(InstanceId, IntegrationAssociationId,
  UseCaseType, Tags)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{IntegrationAssociationId}{[required] The identifier for the AppIntegration association.}

\item{UseCaseType}{[required] The type of use case to associate to the AppIntegration association.
Each AppIntegration association can have only one of each use case type.}

\item{Tags}{One or more tags.}
}
\value{
A list with the following syntax:\preformatted{list(
  UseCaseId = "string",
  UseCaseArn = "string"
)
}
}
\description{
Creates a use case for an AppIntegration association.
}
\section{Request syntax}{
\preformatted{svc$create_use_case(
  InstanceId = "string",
  IntegrationAssociationId = "string",
  UseCaseType = "RULES_EVALUATION",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
