% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_create_user_hierarchy_group}
\alias{connect_create_user_hierarchy_group}
\title{Creates a new user hierarchy group}
\usage{
connect_create_user_hierarchy_group(Name, ParentGroupId, InstanceId)
}
\arguments{
\item{Name}{[required] The name of the user hierarchy group. Must not be more than 100
characters.}

\item{ParentGroupId}{The identifier for the parent hierarchy group. The user hierarchy is
created at level one if the parent group ID is null.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\value{
A list with the following syntax:\preformatted{list(
  HierarchyGroupId = "string",
  HierarchyGroupArn = "string"
)
}
}
\description{
Creates a new user hierarchy group.
}
\section{Request syntax}{
\preformatted{svc$create_user_hierarchy_group(
  Name = "string",
  ParentGroupId = "string",
  InstanceId = "string"
)
}
}

\keyword{internal}
