% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_attached_file}
\alias{connect_delete_attached_file}
\title{Deletes an attached file along with the underlying S3 Object}
\usage{
connect_delete_attached_file(InstanceId, FileId, AssociatedResourceArn)
}
\arguments{
\item{InstanceId}{[required] The unique identifier of the Connect instance.}

\item{FileId}{[required] The unique identifier of the attached file resource.}

\item{AssociatedResourceArn}{[required] The resource to which the attached file is (being) uploaded to.
\href{https://docs.aws.amazon.com/connect/latest/APIReference/API_connect-cases_CreateCase.html}{Cases}
are the only current supported resource.

This value must be a valid ARN.}
}
\description{
Deletes an attached file along with the underlying S3 Object.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_attached_file/} for full documentation.
}
\keyword{internal}
