% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_contact_flow}
\alias{connect_delete_contact_flow}
\title{Deletes a flow for the specified Amazon Connect instance}
\usage{
connect_delete_contact_flow(InstanceId, ContactFlowId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowId}{[required] The identifier of the flow.}
}
\description{
Deletes a flow for the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/delete_contact_flow.html} for full documentation.
}
\keyword{internal}
