% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_hours_of_operation}
\alias{connect_delete_hours_of_operation}
\title{Deletes an hours of operation}
\usage{
connect_delete_hours_of_operation(InstanceId, HoursOfOperationId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{HoursOfOperationId}{[required] The identifier for the hours of operation.}
}
\description{
Deletes an hours of operation.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_hours_of_operation/} for full documentation.
}
\keyword{internal}
