% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_instance}
\alias{connect_delete_instance}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_delete_instance(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Deletes the Amazon Connect instance.
}
\section{Request syntax}{
\preformatted{svc$delete_instance(
  InstanceId = "string"
)
}
}

\keyword{internal}
