% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_instance}
\alias{connect_delete_instance}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_delete_instance(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Deletes the Amazon Connect instance.

Amazon Connect enforces a limit on the total number of instances that
you can create or delete in 30 days. If you exceed this limit, you will
get an error message indicating there has been an excessive number of
attempts at creating or deleting instances. You must wait 30 days before
you can restart creating and deleting instances in your account.
}
\section{Request syntax}{
\preformatted{svc$delete_instance(
  InstanceId = "string"
)
}
}

\keyword{internal}
