% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_integration_association}
\alias{connect_delete_integration_association}
\title{Deletes an Amazon Web Services resource association from an Amazon
Connect instance}
\usage{
connect_delete_integration_association(InstanceId, IntegrationAssociationId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{IntegrationAssociationId}{[required] The identifier for the integration association.}
}
\description{
Deletes an Amazon Web Services resource association from an Amazon Connect instance. The association must not have any use cases associated with it.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_integration_association/} for full documentation.
}
\keyword{internal}
