% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_integration_association}
\alias{connect_delete_integration_association}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_delete_integration_association(InstanceId,
  IntegrationAssociationId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{IntegrationAssociationId}{[required] The identifier for the AppIntegration association.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Deletes an AppIntegration association from an Amazon Connect instance.
The association must not have any use cases associated with it.
}
\section{Request syntax}{
\preformatted{svc$delete_integration_association(
  InstanceId = "string",
  IntegrationAssociationId = "string"
)
}
}

\keyword{internal}
