% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_quick_connect}
\alias{connect_delete_quick_connect}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_delete_quick_connect(InstanceId, QuickConnectId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QuickConnectId}{[required] The identifier for the quick connect.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Deletes a quick connect.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_quick_connect(
  InstanceId = "string",
  QuickConnectId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
