% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_quick_connect}
\alias{connect_delete_quick_connect}
\title{Deletes a quick connect}
\usage{
connect_delete_quick_connect(InstanceId, QuickConnectId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QuickConnectId}{[required] The identifier for the quick connect.}
}
\description{
Deletes a quick connect.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_quick_connect/} for full documentation.
}
\keyword{internal}
