% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_use_case}
\alias{connect_delete_use_case}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_delete_use_case(InstanceId, IntegrationAssociationId, UseCaseId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{IntegrationAssociationId}{[required] The identifier for the AppIntegration association.}

\item{UseCaseId}{[required] The identifier for the use case.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Deletes a use case from an AppIntegration association.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_use_case(
  InstanceId = "string",
  IntegrationAssociationId = "string",
  UseCaseId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
