% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_use_case}
\alias{connect_delete_use_case}
\title{Deletes a use case from an integration association}
\usage{
connect_delete_use_case(InstanceId, IntegrationAssociationId, UseCaseId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{IntegrationAssociationId}{[required] The identifier for the integration association.}

\item{UseCaseId}{[required] The identifier for the use case.}
}
\description{
Deletes a use case from an integration association.

See \url{https://www.paws-r-sdk.com/docs/connect_delete_use_case/} for full documentation.
}
\keyword{internal}
