% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_delete_user}
\alias{connect_delete_user}
\title{Deletes a user account from the specified Amazon Connect instance}
\usage{
connect_delete_user(InstanceId, UserId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{UserId}{[required] The identifier of the user.}
}
\description{
Deletes a user account from the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/delete_user.html} for full documentation.
}
\keyword{internal}
