% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_agent_status}
\alias{connect_describe_agent_status}
\title{Describes an agent status}
\usage{
connect_describe_agent_status(InstanceId, AgentStatusId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{AgentStatusId}{[required] The identifier for the agent status.}
}
\description{
Describes an agent status.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_agent_status/} for full documentation.
}
\keyword{internal}
