% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_contact_flow}
\alias{connect_describe_contact_flow}
\title{Describes the specified flow}
\usage{
connect_describe_contact_flow(InstanceId, ContactFlowId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{ContactFlowId}{[required] The identifier of the flow.}
}
\description{
Describes the specified flow.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_contact_flow/} for full documentation.
}
\keyword{internal}
