% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_contact_flow}
\alias{connect_describe_contact_flow}
\title{Describes the specified contact flow}
\usage{
connect_describe_contact_flow(InstanceId, ContactFlowId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{ContactFlowId}{[required] The identifier of the contact flow.}
}
\description{
Describes the specified contact flow.

See \url{https://paws-r.github.io/docs/connect/describe_contact_flow.html} for full documentation.
}
\keyword{internal}
