% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_contact_flow_module}
\alias{connect_describe_contact_flow_module}
\title{Describes the specified contact flow module}
\usage{
connect_describe_contact_flow_module(InstanceId, ContactFlowModuleId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ContactFlowModuleId}{[required] The identifier of the contact flow module.}
}
\description{
Describes the specified contact flow module.

See \url{https://paws-r.github.io/docs/connect/describe_contact_flow_module.html} for full documentation.
}
\keyword{internal}
