% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_instance_attribute}
\alias{connect_describe_instance_attribute}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_instance_attribute(InstanceId, AttributeType)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{AttributeType}{[required] The type of attribute.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_instance_attribute/} for full documentation.
}
\keyword{internal}
