% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_instance_storage_config}
\alias{connect_describe_instance_storage_config}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_instance_storage_config(InstanceId, AssociationId,
  ResourceType)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{AssociationId}{[required] The existing association identifier that uniquely identifies the
resource type and storage config for the given instance ID.}

\item{ResourceType}{[required] A valid resource type.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  StorageConfig = list(
    AssociationId = "string",
    StorageType = "S3"|"KINESIS_VIDEO_STREAM"|"KINESIS_STREAM"|"KINESIS_FIREHOSE",
    S3Config = list(
      BucketName = "string",
      BucketPrefix = "string",
      EncryptionConfig = list(
        EncryptionType = "KMS",
        KeyId = "string"
      )
    ),
    KinesisVideoStreamConfig = list(
      Prefix = "string",
      RetentionPeriodHours = 123,
      EncryptionConfig = list(
        EncryptionType = "KMS",
        KeyId = "string"
      )
    ),
    KinesisStreamConfig = list(
      StreamArn = "string"
    ),
    KinesisFirehoseConfig = list(
      FirehoseArn = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Retrieves the current storage configurations for the specified resource
type, association ID, and instance ID.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_instance_storage_config(
  InstanceId = "string",
  AssociationId = "string",
  ResourceType = "CHAT_TRANSCRIPTS"|"CALL_RECORDINGS"|"SCHEDULED_REPORTS"|"MEDIA_STREAMS"|"CONTACT_TRACE_RECORDS"|"AGENT_EVENTS"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
