% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_predefined_attribute}
\alias{connect_describe_predefined_attribute}
\title{Describes a predefined attribute for the specified Amazon Connect
instance}
\usage{
connect_describe_predefined_attribute(InstanceId, Name)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the instance
ID in the Amazon Resource Name (ARN) of the instance.}

\item{Name}{[required] The name of the predefined attribute.}
}
\description{
Describes a predefined attribute for the specified Amazon Connect instance. A \emph{predefined attribute} is made up of a name and a value. You can use predefined attributes for:

See \url{https://www.paws-r-sdk.com/docs/connect_describe_predefined_attribute/} for full documentation.
}
\keyword{internal}
