% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_queue}
\alias{connect_describe_queue}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_queue(InstanceId, QueueId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QueueId}{[required] The identifier for the queue.}
}
\value{
A list with the following syntax:\preformatted{list(
  Queue = list(
    Name = "string",
    QueueArn = "string",
    QueueId = "string",
    Description = "string",
    OutboundCallerConfig = list(
      OutboundCallerIdName = "string",
      OutboundCallerIdNumberId = "string",
      OutboundFlowId = "string"
    ),
    HoursOfOperationId = "string",
    MaxContacts = 123,
    Status = "ENABLED"|"DISABLED",
    Tags = list(
      "string"
    )
  )
)
}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Describes the specified queue.
}
\section{Request syntax}{
\preformatted{svc$describe_queue(
  InstanceId = "string",
  QueueId = "string"
)
}
}

\keyword{internal}
