% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_quick_connect}
\alias{connect_describe_quick_connect}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_describe_quick_connect(InstanceId, QuickConnectId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QuickConnectId}{[required] The identifier for the quick connect.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Describes the quick connect.
}
\section{Request syntax}{
\preformatted{svc$describe_quick_connect(
  InstanceId = "string",
  QuickConnectId = "string"
)
}
}

\keyword{internal}
