% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_quick_connect}
\alias{connect_describe_quick_connect}
\title{Describes the quick connect}
\usage{
connect_describe_quick_connect(InstanceId, QuickConnectId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{QuickConnectId}{[required] The identifier for the quick connect.}
}
\description{
Describes the quick connect.

See \url{https://paws-r.github.io/docs/connect/describe_quick_connect.html} for full documentation.
}
\keyword{internal}
