% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_routing_profile}
\alias{connect_describe_routing_profile}
\title{Describes the specified routing profile}
\usage{
connect_describe_routing_profile(InstanceId, RoutingProfileId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}
}
\description{
Describes the specified routing profile.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_routing_profile/} for full documentation.
}
\keyword{internal}
