% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_rule}
\alias{connect_describe_rule}
\title{Describes a rule for the specified Amazon Connect instance}
\usage{
connect_describe_rule(InstanceId, RuleId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{RuleId}{[required] A unique identifier for the rule.}
}
\description{
Describes a rule for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_rule/} for full documentation.
}
\keyword{internal}
