% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_user}
\alias{connect_describe_user}
\title{Describes the specified user}
\usage{
connect_describe_user(UserId, InstanceId)
}
\arguments{
\item{UserId}{[required] The identifier of the user account.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}
}
\description{
Describes the specified user. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID in the Amazon Connect console} (it’s the final part of the ARN). The console does not display the user IDs. Instead, list the users and note the IDs provided in the output.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_user/} for full documentation.
}
\keyword{internal}
