% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_user}
\alias{connect_describe_user}
\title{Describes the specified user account}
\usage{
connect_describe_user(UserId, InstanceId)
}
\arguments{
\item{UserId}{[required] The identifier of the user account.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  User = list(
    Id = "string",
    Arn = "string",
    Username = "string",
    IdentityInfo = list(
      FirstName = "string",
      LastName = "string",
      Email = "string"
    ),
    PhoneConfig = list(
      PhoneType = "SOFT_PHONE"|"DESK_PHONE",
      AutoAccept = TRUE|FALSE,
      AfterContactWorkTimeLimit = 123,
      DeskPhoneNumber = "string"
    ),
    DirectoryUserId = "string",
    SecurityProfileIds = list(
      "string"
    ),
    RoutingProfileId = "string",
    HierarchyGroupId = "string",
    Tags = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the specified user account. You can find the instance ID in
the console (it’s the final part of the ARN). The console does not
display the user IDs. Instead, list the users and note the IDs provided
in the output.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_user(
  UserId = "string",
  InstanceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
