% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_user_hierarchy_group}
\alias{connect_describe_user_hierarchy_group}
\title{Describes the specified hierarchy group}
\usage{
connect_describe_user_hierarchy_group(HierarchyGroupId, InstanceId)
}
\arguments{
\item{HierarchyGroupId}{[required] The identifier of the hierarchy group.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}
}
\description{
Describes the specified hierarchy group.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_user_hierarchy_group/} for full documentation.
}
\keyword{internal}
