% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_user_hierarchy_structure}
\alias{connect_describe_user_hierarchy_structure}
\title{Describes the hierarchy structure of the specified Amazon Connect
instance}
\usage{
connect_describe_user_hierarchy_structure(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}
}
\description{
Describes the hierarchy structure of the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/describe_user_hierarchy_structure.html} for full documentation.
}
\keyword{internal}
