% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_user_hierarchy_structure}
\alias{connect_describe_user_hierarchy_structure}
\title{Describes the hierarchy structure of the specified Amazon Connect
instance}
\usage{
connect_describe_user_hierarchy_structure(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\value{
A list with the following syntax:\preformatted{list(
  HierarchyStructure = list(
    LevelOne = list(
      Id = "string",
      Arn = "string",
      Name = "string"
    ),
    LevelTwo = list(
      Id = "string",
      Arn = "string",
      Name = "string"
    ),
    LevelThree = list(
      Id = "string",
      Arn = "string",
      Name = "string"
    ),
    LevelFour = list(
      Id = "string",
      Arn = "string",
      Name = "string"
    ),
    LevelFive = list(
      Id = "string",
      Arn = "string",
      Name = "string"
    )
  )
)
}
}
\description{
Describes the hierarchy structure of the specified Amazon Connect
instance.
}
\section{Request syntax}{
\preformatted{svc$describe_user_hierarchy_structure(
  InstanceId = "string"
)
}
}

\keyword{internal}
