% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_user_hierarchy_structure}
\alias{connect_describe_user_hierarchy_structure}
\title{Describes the hierarchy structure of the specified Amazon Connect
instance}
\usage{
connect_describe_user_hierarchy_structure(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\description{
Describes the hierarchy structure of the specified Amazon Connect
instance.
}
\section{Request syntax}{
\preformatted{svc$describe_user_hierarchy_structure(
  InstanceId = "string"
)
}
}

\keyword{internal}
