% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_describe_vocabulary}
\alias{connect_describe_vocabulary}
\title{Describes the specified vocabulary}
\usage{
connect_describe_vocabulary(InstanceId, VocabularyId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{VocabularyId}{[required] The identifier of the custom vocabulary.}
}
\description{
Describes the specified vocabulary.

See \url{https://www.paws-r-sdk.com/docs/connect_describe_vocabulary/} for full documentation.
}
\keyword{internal}
