% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_lambda_function}
\alias{connect_disassociate_lambda_function}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_disassociate_lambda_function(InstanceId, FunctionArn)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance..}

\item{FunctionArn}{[required] The Amazon Resource Name (ARN) of the Lambda function being
disassociated.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Remove the Lambda function from the dropdown options available in the
relevant contact flow blocks.
}
\section{Request syntax}{
\preformatted{svc$disassociate_lambda_function(
  InstanceId = "string",
  FunctionArn = "string"
)
}
}

\keyword{internal}
