% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_lex_bot}
\alias{connect_disassociate_lex_bot}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_disassociate_lex_bot(InstanceId, BotName, LexRegion)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{BotName}{[required] The name of the Amazon Lex bot. Maximum character limit of 50.}

\item{LexRegion}{[required] The Region in which the Amazon Lex bot has been created.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/disassociate_lex_bot.html} for full documentation.
}
\keyword{internal}
