% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_lex_bot}
\alias{connect_disassociate_lex_bot}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_disassociate_lex_bot(InstanceId, BotName, LexRegion)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{BotName}{[required] The name of the Amazon Lex bot. Maximum character limit of 50.}

\item{LexRegion}{[required] The Region in which the Amazon Lex bot has been created.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Revokes authorization from the specified instance to access the
specified Amazon Lex bot.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disassociate_lex_bot(
  InstanceId = "string",
  BotName = "string",
  LexRegion = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
