% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_queue_quick_connects}
\alias{connect_disassociate_queue_quick_connects}
\title{Disassociates a set of quick connects from a queue}
\usage{
connect_disassociate_queue_quick_connects(InstanceId, QueueId, QuickConnectIds)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{QuickConnectIds}{[required] The quick connects to disassociate from the queue.}
}
\description{
Disassociates a set of quick connects from a queue.

See \url{https://www.paws-r-sdk.com/docs/connect_disassociate_queue_quick_connects/} for full documentation.
}
\keyword{internal}
