% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_routing_profile_queues}
\alias{connect_disassociate_routing_profile_queues}
\title{Disassociates a set of queues from a routing profile}
\usage{
connect_disassociate_routing_profile_queues(
  InstanceId,
  RoutingProfileId,
  QueueReferences
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{QueueReferences}{[required] The queues to disassociate from this routing profile.}
}
\description{
Disassociates a set of queues from a routing profile.

See \url{https://paws-r.github.io/docs/connect/disassociate_routing_profile_queues.html} for full documentation.
}
\keyword{internal}
