% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_routing_profile_queues}
\alias{connect_disassociate_routing_profile_queues}
\title{Disassociates a set of queues from a routing profile}
\usage{
connect_disassociate_routing_profile_queues(InstanceId,
  RoutingProfileId, QueueReferences)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{QueueReferences}{[required] The queues to disassociate from this routing profile.}
}
\value{
An empty list.
}
\description{
Disassociates a set of queues from a routing profile.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disassociate_routing_profile_queues(
  InstanceId = "string",
  RoutingProfileId = "string",
  QueueReferences = list(
    list(
      QueueId = "string",
      Channel = "VOICE"|"CHAT"|"TASK"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
