% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_routing_profile_queues}
\alias{connect_disassociate_routing_profile_queues}
\title{Disassociates a set of queues from a routing profile}
\usage{
connect_disassociate_routing_profile_queues(
  InstanceId,
  RoutingProfileId,
  QueueReferences = NULL,
  ManualAssignmentQueueReferences = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{QueueReferences}{The queues to disassociate from this routing profile.}

\item{ManualAssignmentQueueReferences}{The manual assignment queues to disassociate with this routing profile.}
}
\description{
Disassociates a set of queues from a routing profile.

See \url{https://www.paws-r-sdk.com/docs/connect_disassociate_routing_profile_queues/} for full documentation.
}
\keyword{internal}
