% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_disassociate_security_key}
\alias{connect_disassociate_security_key}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_disassociate_security_key(InstanceId, AssociationId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{AssociationId}{[required] The existing association identifier that uniquely identifies the
resource type and storage config for the given instance ID.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Deletes the specified security key.
}
\section{Request syntax}{
\preformatted{svc$disassociate_security_key(
  InstanceId = "string",
  AssociationId = "string"
)
}
}

\keyword{internal}
