% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_contact_attributes}
\alias{connect_get_contact_attributes}
\title{Retrieves the contact attributes for the specified contact}
\usage{
connect_get_contact_attributes(InstanceId, InitialContactId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{InitialContactId}{[required] The identifier of the initial contact.}
}
\description{
Retrieves the contact attributes for the specified contact.

See \url{https://paws-r.github.io/docs/connect/get_contact_attributes.html} for full documentation.
}
\keyword{internal}
