% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_contact_metrics}
\alias{connect_get_contact_metrics}
\title{Retrieves the position of the contact in the queue}
\usage{
connect_get_contact_metrics(InstanceId, ContactId, Metrics)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact in this instance of Amazon Connect.}

\item{Metrics}{[required] A list of contact-level metrics to retrieve.}
}
\description{
Retrieves the position of the contact in the queue.

See \url{https://www.paws-r-sdk.com/docs/connect_get_contact_metrics/} for full documentation.
}
\keyword{internal}
