% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_federation_token}
\alias{connect_get_federation_token}
\title{Retrieves a token for federation}
\usage{
connect_get_federation_token(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}
}
\description{
Retrieves a token for federation.

See \url{https://paws-r.github.io/docs/connect/get_federation_token.html} for full documentation.
}
\keyword{internal}
