% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_federation_token}
\alias{connect_get_federation_token}
\title{Retrieves a token for federation}
\usage{
connect_get_federation_token(InstanceId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\description{
Retrieves a token for federation.
}
\section{Request syntax}{
\preformatted{svc$get_federation_token(
  InstanceId = "string"
)
}
}

\keyword{internal}
