% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_get_metric_data_v2}
\alias{connect_get_metric_data_v2}
\title{Gets metric data from the specified Amazon Connect instance}
\usage{
connect_get_metric_data_v2(
  ResourceArn,
  StartTime,
  EndTime,
  Interval = NULL,
  Filters,
  Groupings = NULL,
  Metrics,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource. This includes the
\code{instanceId} an Amazon Connect instance.}

\item{StartTime}{[required] The timestamp, in UNIX Epoch time format, at which to start the
reporting interval for the retrieval of historical metrics data. The
time must be before the end time timestamp. The start and end time
depends on the \code{IntervalPeriod} selected. By default the time range
between start and end time is 35 days. Historical metrics are available
for 3 months.}

\item{EndTime}{[required] The timestamp, in UNIX Epoch time format, at which to end the reporting
interval for the retrieval of historical metrics data. The time must be
later than the start time timestamp. It cannot be later than the current
timestamp.}

\item{Interval}{The interval period and timezone to apply to returned metrics.
\itemize{
\item \code{IntervalPeriod}: An aggregated grouping applied to request metrics.
Valid \code{IntervalPeriod} values are: \code{FIFTEEN_MIN} | \code{THIRTY_MIN} |
\code{HOUR} | \code{DAY} | \code{WEEK} | \code{TOTAL}.

For example, if \code{IntervalPeriod} is selected \code{THIRTY_MIN},
\code{StartTime} and \code{EndTime} differs by 1 day, then Amazon Connect
returns 48 results in the response. Each result is aggregated by the
THIRTY_MIN period. By default Amazon Connect aggregates results
based on the \code{TOTAL} interval period.

The following list describes restrictions on \code{StartTime} and
\code{EndTime} based on which \code{IntervalPeriod} is requested.
\itemize{
\item \code{FIFTEEN_MIN}: The difference between \code{StartTime} and \code{EndTime}
must be less than 3 days.
\item \code{THIRTY_MIN}: The difference between \code{StartTime} and \code{EndTime}
must be less than 3 days.
\item \code{HOUR}: The difference between \code{StartTime} and \code{EndTime} must be
less than 3 days.
\item \code{DAY}: The difference between \code{StartTime} and \code{EndTime} must be
less than 35 days.
\item \code{WEEK}: The difference between \code{StartTime} and \code{EndTime} must be
less than 35 days.
\item \code{TOTAL}: The difference between \code{StartTime} and \code{EndTime} must
be less than 35 days.
}
\item \code{TimeZone}: The timezone applied to requested metrics.
}}

\item{Filters}{[required] The filters to apply to returned metrics. You can filter on the
following resources:
\itemize{
\item Agents
\item Campaigns
\item Channels
\item Feature
\item Queues
\item Routing profiles
\item Routing step expression
\item User hierarchy groups
}

At least one filter must be passed from queues, routing profiles,
agents, or user hierarchy groups.

For metrics for outbound campaigns analytics, you can also use campaigns
to satisfy at least one filter requirement.

To filter by phone number, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/create-historical-metrics-report.html}{Create a historical metrics report}
in the \emph{Amazon Connect Administrator Guide}.

Note the following limits:
\itemize{
\item \strong{Filter keys}: A maximum of 5 filter keys are supported in a
single request. Valid filter keys: \code{AGENT} |
\code{AGENT_HIERARCHY_LEVEL_FIVE} | \code{AGENT_HIERARCHY_LEVEL_FOUR} |
\verb{AGENT_ HIERARCHY_LEVEL_ONE} | \code{AGENT_HIERARCHY_LEVEL_THREE} |
\code{AGENT_HIERARCHY_LEVEL_TWO} | \code{ ANSWERING_MACHINE_DETECTION_STATUS}
| \code{BOT_ALIAS} | \code{BOT_ID} | \code{BOT_INTENT_NAME} | \code{BOT_LOCALE} |
\code{BOT_VERSION} | \code{CAMPAIGN} | \code{CAMPAIGN_DELIVERY_EVENT_TYPE} |
\code{ CAMPAIGN_EXCLUDED_EVENT_TYPE} | \code{CASE_STATUS} |
\code{CASE_TEMPLATE_ARN} | \code{CHANNEL} |
\code{ contact/segmentAttributes/connect:Subtype} |
\code{contact/segmentAttributes/connect:ValidationTestType} |
\code{ DISCONNECT_REASON} | \code{EVALUATION_FORM} | \code{EVALUATION_QUESTION} |
\code{EVALUATION_SECTION} | \code{EVALUATION_SOURCE} | \code{EVALUATOR_ID} |
\code{FEATURE} | \code{FLOW_ACTION_ID} | \code{FLOW_TYPE} |
\code{FLOWS_MODULE_RESOURCE_ID} | \code{FLOWS_NEXT_RESOURCE_ID} |
\code{FLOWS_NEXT_RESOURCE_QUEUE_ID} | \code{FLOWS_OUTCOME_TYPE} |
\code{FLOWS_RESOURCE_ID} | \code{FORM_VERSION} | \code{INITIATING_FLOW} |
\code{INITIATION_METHOD} | \code{INVOKING_RESOURCE_PUBLISHED_TIMESTAMP} |
\code{INVOKING_RESOURCE_TYPE} | \code{PARENT_FLOWS_RESOURCE_ID} |
\code{Q_CONNECT_ENABLED} | \code{QUEUE} | \verb{RESOURCE_PUBLISHED_ TIMESTAMP} |
\code{ROUTING_PROFILE} | \code{ROUTING_STEP_EXPRESSION} | \code{TEST_CASE} |
\verb{TEST_ CASE_EXECUTION_FAILURE_REASON} | \code{TEST_CASE_EXECUTION_RESULT}
| \code{TEST_CASE_EXECUTION_STATE}
\item \strong{Filter values}: A maximum of 100 filter values are supported in a
single request. VOICE, CHAT, and TASK are valid \code{filterValue} for
the CHANNEL filter key. They do not count towards limitation of 100
filter values. For example, a GetMetricDataV2 request can filter by
50 queues, 35 agents, and 15 routing profiles for a total of 100
filter values, along with 3 channel filters.

\code{contact_lens_conversational_analytics} is a valid filterValue for
the \code{FEATURE} filter key. It is available only to contacts analyzed
by Contact Lens conversational analytics.

\code{connect:Chat}, \code{connect:SMS}, \code{connect:Telephony}, and
\code{connect:WebRTC} are valid \code{filterValue} examples (not exhaustive)
for the \verb{contact/segmentAttributes/connect:Subtype filter} key.

\code{ROUTING_STEP_EXPRESSION} is a valid filter key with a filter value
up to 3000 length. This filter is case and order sensitive. JSON
string fields must be sorted in ascending order and JSON array order
should be kept as is.

\code{Q_CONNECT_ENABLED}. TRUE and FALSE are the only valid filterValues
for the \code{Q_CONNECT_ENABLED} filter key.
\itemize{
\item TRUE includes all contacts that had Connect AI Agents enabled as
part of the flow.
\item FALSE includes all contacts that did not have Connect AI Agents
enabled as part of the flow
\item EXPERIENCE_VALIDATION and FLOW_VALIDATION are the only valid
filterValues for the
contact/segmentAttributes/connect:ValidationTestType filter key
}

This filter is available only for contact record-driven metrics.

\href{https://docs.aws.amazon.com/connect/latest/APIReference/API_connect-outbound-campaigns_Campaign.html}{Campaign}
ARNs are valid \code{filterValues} for the \code{CAMPAIGN} filter key.
}}

\item{Groupings}{The grouping applied to the metrics that are returned. For example, when
results are grouped by queue, the metrics returned are grouped by queue.
The values that are returned apply to the metrics for each queue. They
are not aggregated for all queues.

If no grouping is specified, a summary of all metrics is returned.

Valid grouping keys: \code{AGENT} | \code{AGENT_HIERARCHY_LEVEL_ONE} |
\code{AGENT_HIERARCHY_LEVEL_TWO} | \code{AGENT_HIERARCHY_LEVEL_THREE} |
\code{AGENT_HIERARCHY_LEVEL_FOUR} | \code{AGENT_HIERARCHY_LEVEL_FIVE} |
\code{ANSWERING_MACHINE_DETECTION_STATUS} | \code{BOT_ID} | \code{BOT_ALIAS} |
\code{BOT_VERSION} | \code{BOT_LOCALE} | \code{BOT_INTENT_NAME} | \code{CAMPAIGN} |
\code{CAMPAIGN_DELIVERY_EVENT_TYPE} | \code{CAMPAIGN_EXCLUDED_EVENT_TYPE} |
\code{CAMPAIGN_EXECUTION_TIMESTAMP} | \code{CASE_TEMPLATE_ARN} | \code{CASE_STATUS} |
\code{CHANNEL} | \code{contact/segmentAttributes/connect:Subtype} |
\code{DISCONNECT_REASON} | \code{EVALUATION_FORM} | \code{EVALUATION_SECTION} |
\code{EVALUATION_QUESTION} | \code{EVALUATION_SOURCE} | \code{EVALUATOR_ID} |
\code{FLOWS_RESOURCE_ID} | \code{FLOWS_MODULE_RESOURCE_ID} | \code{FLOW_ACTION_ID} |
\code{FLOW_TYPE} | \code{FLOWS_OUTCOME_TYPE} | \code{FORM_VERSION} |
\code{INITIATION_METHOD} | \code{INVOKING_RESOURCE_PUBLISHED_TIMESTAMP} |
\code{INVOKING_RESOURCE_TYPE} | \code{PARENT_FLOWS_RESOURCE_ID} |
\code{Q_CONNECT_ENABLED} | \code{QUEUE} | \code{RESOURCE_PUBLISHED_TIMESTAMP} |
\code{ROUTING_PROFILE} | \code{ROUTING_STEP_EXPRESSION} | \code{TEST_CASE} |
\code{TEST_CASE_EXECUTION_FAILURE_REASON} | \code{TEST_CASE_INVOCATION_METHOD}

API, SCHEDULE, and EVENT are the only valid filterValues for
TEST_CASE_INVOCATION_METHOD.

OBSERVE_EVENT, SEND_INSTRUCTION, ASSERT_DATA, and
OVERRIDE_SYSTEM_BEHAVIOR are the only valid filterValues for
TEST_CASE_EXECUTION_FAILURE_REASON

Type: Array of strings

Array Members: Maximum number of 4 items

Required: No}

\item{Metrics}{[required] The metrics to retrieve. Specify the name or metricId, groupings, and
filters for each metric. The following historical metrics are available.
For a description of each metric, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html}{Metrics definition}
in the \emph{Amazon Connect Administrator Guide}.

MetricId should be used to reference custom metrics or out of the box
metrics as Arn. If using MetricId, the limit is 20 MetricId per request.

\strong{ABANDONMENT_RATE}

Unit: Percent

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Feature, contact/segmentAttributes/connect:Subtype, Q
in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#abandonment-rate}{Abandonment rate}

\strong{AGENT_ADHERENT_TIME}

This metric is available only in Amazon Web Services Regions where
\href{https://docs.aws.amazon.com/connect/latest/adminguide/regions.html#optimization_region}{Forecasting, capacity planning, and scheduling}
is available.

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#adherent-time}{Adherent time}

\strong{AGENT_ANSWER_RATE}

Unit: Percent

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-answer-rate}{Agent answer rate}

\strong{AGENT_NON_ADHERENT_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#non-adherent-time}{Non-adherent time}

\strong{AGENT_NON_RESPONSE}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-non-response}{Agent non-response}

\strong{AGENT_NON_RESPONSE_WITHOUT_CUSTOMER_ABANDONS}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy

Data for this metric is available starting from October 1, 2023 0:00:00
GMT.

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-non-response-without-customer-abandons}{Agent non-response without customer abandons}

\strong{AGENT_OCCUPANCY}

Unit: Percentage

Valid groupings and filters: Routing Profile, Agent, Agent Hierarchy

UI name:
\href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#occupancy}{Occupancy}

\strong{AGENT_SCHEDULE_ADHERENCE}

This metric is available only in Amazon Web Services Regions where
\href{https://docs.aws.amazon.com/connect/latest/adminguide/regions.html#optimization_region}{Forecasting, capacity planning, and scheduling}
is available.

Unit: Percent

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy

UI name:
\href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#adherence}{Adherence}

\strong{AGENT_SCHEDULED_TIME}

This metric is available only in Amazon Web Services Regions where
\href{https://docs.aws.amazon.com/connect/latest/adminguide/regions.html#optimization_region}{Forecasting, capacity planning, and scheduling}
is available.

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#scheduled-time}{Scheduled time}

\strong{AVG_ABANDON_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Feature, contact/segmentAttributes/connect:Subtype, Q
in Connect

Valid metric filter key: \code{INITIATION_METHOD}

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-queue-abandon-time}{Average queue abandon time}

\strong{AVG_ACTIVE_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-active-time}{Average active time}

\strong{AVG_AFTER_CONTACT_WORK_TIME}

Unit: Seconds

Valid metric filter key: \code{INITIATION_METHOD}

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Feature, contact/segmentAttributes/connect:Subtype, Q
in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#after-contact-work-time}{Average after contact work time}

Feature is a valid filter but not a valid grouping.

\strong{AVG_AGENT_CONCURRENCY}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-concurrency}{Average agent concurrency}

\strong{AVG_AGENT_CONNECTING_TIME}

Unit: Seconds

Valid metric filter key: \code{INITIATION_METHOD}. For now, this metric only
supports the following as \code{INITIATION_METHOD}: \code{INBOUND} | \code{OUTBOUND} |
\code{CALLBACK} | \code{API}

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-api-connecting-time}{Average agent API connecting time}

The \code{Negate} key in metric-level filters is not applicable for this
metric.

\strong{AVG_AGENT_PAUSE_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-pause-time}{Average agent pause time}

\strong{AVG_BOT_CONVERSATION_TIME}

Unit: Seconds

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Bot ID, Bot alias, Bot
version, Bot locale, Flows resource ID, Flows module resource ID, Flow
type, Flow action ID, Invoking resource published timestamp, Initiation
method, Invoking resource type, Parent flows resource ID

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/bot-metrics.html#average-bot-conversation-time}{Average bot conversation time}

\strong{AVG_BOT_CONVERSATION_TURNS}

Unit: Count

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Bot ID, Bot alias, Bot
version, Bot locale, Flows resource ID, Flows module resource ID, Flow
type, Flow action ID, Invoking resource published timestamp, Initiation
method, Invoking resource type, Parent flows resource ID

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/bot-metrics.html#average-bot-conversation-turns}{Average bot conversation turns}

\strong{AVG_CASE_RELATED_CONTACTS}

Unit: Count

Required filter key: CASE_TEMPLATE_ARN

Valid groupings and filters: CASE_TEMPLATE_ARN, CASE_STATUS

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-contacts-per-case}{Average contacts per case}

\strong{AVG_CASE_RESOLUTION_TIME}

Unit: Seconds

Required filter key: CASE_TEMPLATE_ARN

Valid groupings and filters: CASE_TEMPLATE_ARN, CASE_STATUS

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-case-resolution-time}{Average case resolution time}

\strong{AVG_CONTACT_DURATION}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Feature, contact/segmentAttributes/connect:Subtype, Q
in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-contact-duration}{Average contact duration}

Feature is a valid filter but not a valid grouping.

\strong{AVG_CONTACT_FIRST_RESPONSE_TIME_AGENT}

Unit: Seconds

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-average-contact-first-response-wait-time}{Agent average contact first response wait time}

\strong{AVG_CONVERSATION_CLOSE_TIME}

Unit: Seconds

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-conversation-close-time}{Average conversation close time}

\strong{AVG_CONVERSATION_DURATION}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Feature, contact/segmentAttributes/connect:Subtype, Q
in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-conversation-duration}{Average conversation duration}

\strong{AVG_DIALS_PER_MINUTE}

This metric is available only for outbound campaigns that use the agent
assisted voice and automated voice delivery modes.

Unit: Count

Valid groupings and filters: Agent, Campaign, Queue, Routing Profile

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-dials-per-minute}{Average dials per minute}

\strong{AVG_EVALUATION_SCORE}

Unit: Percent

Valid groupings and filters: Agent, Agent Hierarchy, Channel, Evaluation
Form ID, Evaluation Section ID, Evaluation Question ID, Evaluation
Source, Form Version, Queue, Routing Profile

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-evaluation-score}{Average evaluation score}

\strong{AVG_FIRST_RESPONSE_TIME_AGENT}

Unit: Seconds

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-first-response-time}{Average agent first response time}

\strong{AVG_FLOW_TIME}

Unit: Seconds

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Flow type, Flows module
resource ID, Flows next resource ID, Flows next resource queue ID, Flows
outcome type, Flows resource ID, Initiation method, Resource published
timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-flow-time}{Average flow time}

\strong{AVG_GREETING_TIME_AGENT}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-greeting-time}{Average agent greeting time}

\strong{AVG_HANDLE_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Feature, contact/segmentAttributes/connect:Subtype,
RoutingStepExpression

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-handle-time}{Average handle time}

Feature is a valid filter but not a valid grouping.

\strong{ACTIVE_AI_AGENTS}

Unit: Count

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Type, AI
Use Case, Channel, Queue, Routing Profile

UI name: Active AI Agents

\strong{AI_HANDOFF_RATE}

Unit: Percent

Valid groupings and filters: AI Use Case, Channel, Queue, Routing
Profile

UI name: Handoff Rate

\strong{AI_HANDOFFS}

Unit: Count

Valid groupings and filters: AI Use Case, Channel, Queue, Routing
Profile

UI name: AI Handoff Count

\strong{AI_AGENT_INVOCATION_SUCCESS}

Unit: Count

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Name
Version, AI Agent Type, AI Use Case, Channel, Queue, Routing Profile

UI name: AI Agent Invocation Success Count

AI Agent Name Version is not a valid filter but a valid grouping.

\strong{AI_AGENT_INVOCATION_SUCCESS_RATE}

Unit: Percent

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Name
Version, AI Agent Type, AI Use Case, Channel, Queue, Routing Profile

UI name: AI Agent Invocation Success Rate

AI Agent Name Version is not a valid filter but a valid grouping.

\strong{AI_AGENT_INVOCATIONS}

Unit: Count

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Type, AI
Agent Name Version, AI Use Case, Channel, Queue, Routing Profile

UI name: AI Agent Invocation Count

AI Agent Name Version is not a valid filter but a valid grouping.

\strong{AI_RESPONSE_COMPLETION_RATE}

Unit: Percent

Valid groupings and filters: AI Use Case, Channel, Queue, Routing
Profile

UI name: AI Response Completion Rate

\strong{AI_INVOLVED_CONTACTS}

Unit: Count

Valid groupings and filters: AI Use Case, Channel, Queue, Routing
Profile

UI name: AI Contacts

\strong{AI_PROMPT_INVOCATION_SUCCESS}

Unit: Count

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Name
Version, AI Agent Type, AI Prompt, AI Prompt ID, AI Prompt Name, AI
Prompt Type, AI Use Case, Channel, Queue, Routing Profile

UI name: AI Prompt Invocation Success Count

AI Agent Name Version is not a valid filter but a valid grouping.

\strong{AI_PROMPT_INVOCATION_SUCCESS_RATE}

Unit: Percent

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Name
Version, AI Agent Type, AI Prompt, AI Prompt ID, AI Prompt Name, AI
Prompt Type, AI Use Case, Channel, Queue, Routing Profile

UI name: AI Prompt Invocation Success Rate

AI Agent Name Version is not a valid filter but a valid grouping.

\strong{AI_TOOL_INVOCATIONS}

Unit: Count

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Name
Version, AI Agent Type, AI Tool ID, AI Tool Name, AI Tool Type, AI Use
Case, Channel, Queue, Routing Profile

UI name: AI Tool Invocation Count

AI Agent Name Version is not a valid filter but a valid grouping.

\strong{AVG_AI_AGENT_CONVERSATION_TURNS}

Unit: Count

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Name
Version, AI Agent Type, AI Use Case, Channel, Queue, Routing Profile

UI name: Average AI Agent Conversation Turns

AI Agent Name Version is not a valid filter but a valid grouping.

\strong{AVG_AI_CONVERSATION_TURNS}

Unit: Count

Valid groupings and filters: AI Use Case, Channel, Queue, Routing
Profile

UI name: AI Conversation Turns

\strong{AVG_AI_PROMPT_INVOCATION_LATENCY}

Unit: Milliseconds

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Name
Version, AI Agent Type, AI Prompt, AI Prompt ID, AI Prompt Name, AI
Prompt Type, AI Use Case, Channel, Queue, Routing Profile

UI name: Average AI Prompt Invocation Latency

AI Agent Name Version is not a valid filter but a valid grouping.

\strong{AVG_AI_TOOL_INVOCATION_LATENCY}

Unit: Milliseconds

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Name
Version, AI Agent Type, AI Tool ID, AI Tool Name, AI Tool Type, AI Use
Case, Channel, Queue, Routing Profile

UI name: Average AI Tool Invocation Latency

AI Agent Name Version is not a valid filter but a valid grouping.

\strong{KNOWLEDGE_CONTENT_REFERENCES}

Unit: Count

Valid groupings and filters: AI Agent, AI Agent Name, AI Agent Type, AI
Use Case, Channel, Knowledge Base Name, Queue, Routing Profile

UI name: KnowledgeBase Reference Count

\strong{PROACTIVE_INTENT_ENGAGEMENT_RATE}

Unit: Percent

Valid groupings and filters: AI Use Case, Channel, Queue, Routing
Profile

UI name: Proactive Intent Engagement Rate

\strong{PROACTIVE_INTENT_RESPONSE_RATE}

Unit: Percent

Valid groupings and filters: AI Use Case, Channel, Queue, Routing
Profile

UI name: Proactive Intent Response Rate

\strong{PROACTIVE_INTENTS_ANSWERED}

Unit: Count

Valid groupings and filters: AI Use Case, Channel, Queue, Routing
Profile

UI name: Proactive Intents Answered

\strong{PROACTIVE_INTENTS_DETECTED}

Unit: Count

Valid groupings and filters: AI Use Case, Channel, Queue, Routing
Profile

UI name: Proactive Intents Detected

Unit:

Valid groupings and filters:

UI name:

Unit:

Valid groupings and filters:

UI name:

\strong{PROACTIVE_INTENTS_ENGAGED}

Unit: Count

Valid groupings and filters: AI Use Case, Channel, Queue, Routing
Profile

UI name: UI name:

\strong{AVG_HOLD_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Feature, contact/segmentAttributes/connect:Subtype, Q
in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-customer-hold-time}{Average customer hold time}

Feature is a valid filter but not a valid grouping.

\strong{AVG_HOLD_TIME_ALL_CONTACTS}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-customer-hold-time-all-contacts}{Average customer hold time all contacts}

\strong{AVG_HOLDS}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Feature, contact/segmentAttributes/connect:Subtype, Q
in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-holds}{Average holds}

Feature is a valid filter but not a valid grouping.

\strong{AVG_INTERACTION_AND_HOLD_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-interaction-and-customer-hold-time}{Average agent interaction and customer hold time}

\strong{AVG_INTERACTION_TIME}

Unit: Seconds

Valid metric filter key: \code{INITIATION_METHOD}

Valid groupings and filters: Queue, Channel, Routing Profile, Feature,
contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-interaction-time}{Average agent interaction time}

Feature is a valid filter but not a valid grouping.

\strong{AVG_INTERRUPTIONS_AGENT}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-interruptions}{Average agent interruptions}

\strong{AVG_INTERRUPTION_TIME_AGENT}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-interruption-time}{Average agent interruption time}

\strong{AVG_MESSAGE_LENGTH_AGENT}

Unit: Count

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-message-length}{Average agent message length}

\strong{AVG_MESSAGE_LENGTH_CUSTOMER}

Unit: Count

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-customer-message-length}{Average customer message length}

\strong{AVG_MESSAGES}

Unit: Count

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-messages}{Average messages}

\strong{AVG_MESSAGES_AGENT}

Unit: Count

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-messages}{Average agent messages}

\strong{AVG_MESSAGES_BOT}

Unit: Count

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-bot-messages}{Average bot messages}

\strong{AVG_MESSAGES_CUSTOMER}

Unit: Count

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-customer-messages}{Average customer messages}

\strong{AVG_NON_TALK_TIME}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-non-talk-time}{Average non-talk time}

\strong{AVG_QUEUE_ANSWER_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Feature,
contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-queue-answer-time}{Average queue answer time}

Valid metric level filters: \code{INITIATION_METHOD}, \code{FEATURE},
\code{DISCONNECT_REASON}

Feature is a valid filter but not a valid grouping.

\strong{AVG_QUEUE_ANSWER_TIME_CUSTOMER_FIRST_CALLBACK}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Feature, contact/segmentAttributes/connect:Subtype, Q in Connect, Agent
Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-queue-answer-time-customer-first-callback}{Avg. queue answer time - customer first callback}

\strong{AVG_RESPONSE_TIME_AGENT}

Unit: Seconds

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-response-time-agent}{Average agent response time}

\strong{AVG_RESPONSE_TIME_CUSTOMER}

Unit: Seconds

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-customer-time-agent}{Average customer response time}

\strong{AVG_RESOLUTION_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile,
contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-resolution-time}{Average resolution time}

\strong{AVG_TALK_TIME}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-talk-time}{Average talk time}

\strong{AVG_TALK_TIME_AGENT}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-agent-talk-time}{Average agent talk time}

\strong{AVG_TALK_TIME_CUSTOMER}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-customer-talk-time}{Average customer talk time}

\strong{AVG_WAIT_TIME_AFTER_CUSTOMER_CONNECTION}

This metric is available only for outbound campaigns that use the agent
assisted voice and automated voice delivery modes.

Unit: Seconds

Valid groupings and filters: Campaign

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-wait-time-after-customer-connection}{Average wait time after customer connection}

\strong{AVG_WAIT_TIME_AFTER_CUSTOMER_FIRST_CALLBACK_CONNECTION}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Feature, contact/segmentAttributes/connect:Subtype, Q in Connect, Agent
Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-wait-time-after-customer-connection-customer-first-callback}{Avg. wait time after customer connection - customer first callback}

\strong{AVG_WEIGHTED_EVALUATION_SCORE}

Unit: Percent

Valid groupings and filters: Agent, Agent Hierarchy, Channel, Evaluation
Form Id, Evaluation Section ID, Evaluation Question ID, Evaluation
Source, Form Version, Queue, Routing Profile

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#average-weighted-evaluation-score}{Average weighted evaluation score}

\strong{BOT_CONVERSATIONS_COMPLETED}

Unit: Count

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Bot ID, Bot alias, Bot
version, Bot locale, Flows resource ID, Flows module resource ID, Flow
type, Flow action ID, Invoking resource published timestamp, Initiation
method, Invoking resource type, Parent flows resource ID

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/bot-metrics.html#bot-conversations-completed}{Bot conversations completed}

\strong{BOT_INTENTS_COMPLETED}

Unit: Count

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Bot ID, Bot alias, Bot
version, Bot locale, Bot intent name, Flows resource ID, Flows module
resource ID, Flow type, Flow action ID, Invoking resource published
timestamp, Initiation method, Invoking resource type, Parent flows
resource ID

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/bot-metrics.html#bot-intents-completed}{Bot intents completed}

\strong{CAMPAIGN_CONTACTS_ABANDONED_AFTER_X}

This metric is available only for outbound campaigns using the agent
assisted voice and automated voice delivery modes.

Unit: Count

Valid groupings and filters: Agent, Campaign

Threshold: For \code{ThresholdValue}, enter any whole number from 1 to 604800
(inclusive), in seconds. For \code{Comparison}, you must enter \code{GT} (for
\emph{Greater than}).

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#campaign-contacts-abandoned-after-x}{Campaign contacts abandoned after X}

\strong{CAMPAIGN_CONTACTS_ABANDONED_AFTER_X_RATE}

This metric is available only for outbound campaigns using the agent
assisted voice and automated voice delivery modes.

Unit: Percent

Valid groupings and filters: Agent, Campaign

Threshold: For \code{ThresholdValue}, enter any whole number from 1 to 604800
(inclusive), in seconds. For \code{Comparison}, you must enter \code{GT} (for
\emph{Greater than}).

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#campaign-contacts-abandoned-after-x-rate}{Campaign contacts abandoned after X rate}

\strong{CAMPAIGN_INTERACTIONS}

This metric is available only for outbound campaigns using the email
delivery mode.

Unit: Count

Valid metric filter key: CAMPAIGN_INTERACTION_EVENT_TYPE

Valid groupings and filters: Campaign

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#campaign-interactions}{Campaign interactions}

\strong{CAMPAIGN_PROGRESS_RATE}

This metric is only available for outbound campaigns initiated using a
customer segment. It is not available for event triggered campaigns.

Unit: Percent

Valid groupings and filters: Campaign, Campaign Execution Timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#campaign-progress-rate}{Campaign progress rate}

\strong{CAMPAIGN_SEND_ATTEMPTS}

This metric is available only for outbound campaigns.

Unit: Count

Valid groupings and filters: Campaign, Channel,
contact/segmentAttributes/connect:Subtype

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#campaign-send-attempts}{Campaign send attempts}

\strong{CAMPAIGN_SEND_EXCLUSIONS}

This metric is available only for outbound campaigns.

Valid metric filter key: CAMPAIGN_EXCLUDED_EVENT_TYPE

Unit: Count

Valid groupings and filters: Campaign, Campaign Excluded Event Type,
Campaign Execution Timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#campaign-send-exclusions}{Campaign send exclusions}

\strong{CASES_CREATED}

Unit: Count

Required filter key: CASE_TEMPLATE_ARN

Valid groupings and filters: CASE_TEMPLATE_ARN, CASE_STATUS

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#cases-created}{Cases created}

\strong{CONTACTS_CREATED}

Unit: Count

Valid metric filter key: \code{INITIATION_METHOD}

Valid groupings and filters: Queue, Channel, Routing Profile, Feature,
contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-created}{Contacts created}

Feature is a valid filter but not a valid grouping.

\strong{CONTACTS_HANDLED}

Unit: Count

Valid metric filter key: \code{INITIATION_METHOD}, \code{DISCONNECT_REASON}

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Feature, contact/segmentAttributes/connect:Subtype,
RoutingStepExpression, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-handled}{Contacts handled}

Feature is a valid filter but not a valid grouping.

\strong{CONTACTS_HANDLED_BY_CONNECTED_TO_AGENT}

Unit: Count

Valid metric filter key: \code{INITIATION_METHOD}

Valid groupings and filters: Queue, Channel, Agent, Agent Hierarchy,
contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-handled-by-connected-to-agent-timestamp}{Contacts handled (connected to agent timestamp)}

\strong{CONTACTS_HOLD_ABANDONS}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-hold-disconnect}{Contacts hold disconnect}

\strong{CONTACTS_ON_HOLD_AGENT_DISCONNECT}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-hold-agent-disconnect}{Contacts hold agent disconnect}

\strong{CONTACTS_ON_HOLD_CUSTOMER_DISCONNECT}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-hold-customer-disconnect}{Contacts hold customer disconnect}

\strong{CONTACTS_PUT_ON_HOLD}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-put-on-hold}{Contacts put on hold}

\strong{CONTACTS_TRANSFERRED_OUT_EXTERNAL}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-transferred-out-external}{Contacts transferred out external}

\strong{CONTACTS_TRANSFERRED_OUT_INTERNAL}

Unit: Percent

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-transferred-out-internal}{Contacts transferred out internal}

\strong{CONTACTS_QUEUED}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-queued}{Contacts queued}

\strong{CONTACTS_QUEUED_BY_ENQUEUE}

Unit: Count

Valid groupings and filters: Queue, Channel, Agent, Agent Hierarchy,
contact/segmentAttributes/connect:Subtype

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-queued-by-enqueue}{Contacts queued (enqueue timestamp)}

\strong{CONTACTS_REMOVED_FROM_QUEUE_IN_X}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Q in
Connect

Threshold: For \code{ThresholdValue}, enter any whole number from 1 to 604800
(inclusive), in seconds. For \code{Comparison}, you can use \code{LT} (for "Less
than") or \code{LTE} (for "Less than equal").

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-removed-from-queue}{Contacts removed from queue in X seconds}

\strong{CONTACTS_RESOLVED_IN_X}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile,
contact/segmentAttributes/connect:Subtype, Q in Connect

Threshold: For \code{ThresholdValue}, enter any whole number from 1 to 604800
(inclusive), in seconds. For \code{Comparison}, you can use \code{LT} (for "Less
than") or \code{LTE} (for "Less than equal").

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-resolved}{Contacts resolved in X}

\strong{CONTACTS_TRANSFERRED_OUT}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Feature, contact/segmentAttributes/connect:Subtype, Q
in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-transferred-out}{Contacts transferred out}

Feature is a valid filter but not a valid grouping.

\strong{CONTACTS_TRANSFERRED_OUT_BY_AGENT}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-transferred-out-by-agent}{Contacts transferred out by agent}

\strong{CONTACTS_TRANSFERRED_OUT_FROM_QUEUE}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-transferred-out-queue}{Contacts transferred out queue}

\strong{CURRENT_CASES}

Unit: Count

Required filter key: CASE_TEMPLATE_ARN

Valid groupings and filters: CASE_TEMPLATE_ARN, CASE_STATUS

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#current-cases}{Current cases}

\strong{CONVERSATIONS_ABANDONED}

Unit: Count

Valid groupings and filters: Agent, Agent Hierarchy, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Feature,
RoutingStepExpression, Initiation method, Routing Profile, Queue, Q in
Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#conversations-abandoned}{Conversations abandoned}

\strong{DELIVERY_ATTEMPTS}

This metric is available only for outbound campaigns.

Unit: Count

Valid metric filter key: \code{ANSWERING_MACHINE_DETECTION_STATUS},
\code{CAMPAIGN_DELIVERY_EVENT_TYPE}, \code{DISCONNECT_REASON}

Valid groupings and filters: Agent, Answering Machine Detection Status,
Campaign, Campaign Delivery EventType, Channel,
contact/segmentAttributes/connect:Subtype, Disconnect Reason, Queue,
Routing Profile

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#delivery-attempts}{Delivery attempts}

Campaign Delivery EventType filter and grouping are only available for
SMS and Email campaign delivery modes. Agent, Queue, Routing Profile,
Answering Machine Detection Status and Disconnect Reason are only
available for agent assisted voice and automated voice delivery modes.

\strong{DELIVERY_ATTEMPT_DISPOSITION_RATE}

This metric is available only for outbound campaigns. Dispositions for
the agent assisted voice and automated voice delivery modes are only
available with answering machine detection enabled.

Unit: Percent

Valid metric filter key: \code{ANSWERING_MACHINE_DETECTION_STATUS},
\code{CAMPAIGN_DELIVERY_EVENT_TYPE}, \code{DISCONNECT_REASON}

Valid groupings and filters: Agent, Answering Machine Detection Status,
Campaign, Channel, contact/segmentAttributes/connect:Subtype, Disconnect
Reason, Queue, Routing Profile

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#delivery-attempt-disposition-rate}{Delivery attempt disposition rate}

Campaign Delivery Event Type filter and grouping are only available for
SMS and Email campaign delivery modes. Agent, Queue, Routing Profile,
Answering Machine Detection Status and Disconnect Reason are only
available for agent assisted voice and automated voice delivery modes.

\strong{EVALUATIONS_PERFORMED}

Unit: Count

Valid groupings and filters: Agent, Agent Hierarchy, Channel, Evaluation
Form ID, Evaluation Source, Form Version, Queue, Routing Profile

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#evaluations-performed}{Evaluations performed}

\strong{FLOWS_OUTCOME}

Unit: Count

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Flow type, Flows module
resource ID, Flows next resource ID, Flows next resource queue ID, Flows
outcome type, Flows resource ID, Initiation method, Resource published
timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#flows-outcome}{Flows outcome}

\strong{FLOWS_STARTED}

Unit: Count

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Flow type, Flows module
resource ID, Flows resource ID, Initiation method, Resource published
timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#flows-started}{Flows started}

\strong{HUMAN_ANSWERED_CALLS}

This metric is available only for outbound campaigns. Dispositions for
the agent assisted voice and automated voice delivery modes are only
available with answering machine detection enabled.

Unit: Count

Valid groupings and filters: Agent, Campaign

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#human-answered}{Human answered}

\strong{MAX_FLOW_TIME}

Unit: Seconds

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Flow type, Flows module
resource ID, Flows next resource ID, Flows next resource queue ID, Flows
outcome type, Flows resource ID, Initiation method, Resource published
timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#maximum-flow-time}{Maximum flow time}

\strong{MAX_QUEUED_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#maximum-queued-time}{Maximum queued time}

\strong{MIN_FLOW_TIME}

Unit: Seconds

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Flow type, Flows module
resource ID, Flows next resource ID, Flows next resource queue ID, Flows
outcome type, Flows resource ID, Initiation method, Resource published
timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#minimum-flow-time}{Minimum flow time}

\strong{PERCENT_AUTOMATIC_FAILS}

Unit: Percent

Valid groupings and filters: Agent, Agent Hierarchy, Channel, Evaluation
Form ID, Evaluation Source, Form Version, Queue, Routing Profile

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#automatic-fails-percent}{Automatic fails percent}

\strong{PERCENT_BOT_CONVERSATIONS_OUTCOME}

Unit: Percent

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Bot ID, Bot alias, Bot
version, Bot locale, Flows resource ID, Flows module resource ID, Flow
type, Flow action ID, Invoking resource published timestamp, Initiation
method, Invoking resource type, Parent flows resource ID

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/bot-metrics.html#percent-bot-conversations-outcome}{Percent bot conversations outcome}

\strong{PERCENT_BOT_INTENTS_OUTCOME}

Unit: Percent

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Bot ID, Bot alias, Bot
version, Bot locale, Bot intent name, Flows resource ID, Flows module
resource ID, Flow type, Flow action ID, Invoking resource published
timestamp, Initiation method, Invoking resource type, Parent flows
resource ID

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/bot-metrics.html#percent-bot-intents-outcome}{Percent bot intents outcome}

\strong{PERCENT_CASES_FIRST_CONTACT_RESOLVED}

Unit: Percent

Required filter key: CASE_TEMPLATE_ARN

Valid groupings and filters: CASE_TEMPLATE_ARN, CASE_STATUS

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#cases-resolved-on-first-contact}{Cases resolved on first contact}

\strong{PERCENT_CONTACTS_STEP_EXPIRED}

Unit: Percent

Valid groupings and filters: Queue, RoutingStepExpression

UI name: This metric is available in Real-time Metrics UI but not on the
Historical Metrics UI.

\strong{PERCENT_CONTACTS_STEP_JOINED}

Unit: Percent

Valid groupings and filters: Queue, RoutingStepExpression

UI name: This metric is available in Real-time Metrics UI but not on the
Historical Metrics UI.

\strong{PERCENT_FLOWS_OUTCOME}

Unit: Percent

Valid metric filter key: \code{FLOWS_OUTCOME_TYPE}

Valid groupings and filters: Channel,
contact/segmentAttributes/connect:Subtype, Flow type, Flows module
resource ID, Flows next resource ID, Flows next resource queue ID, Flows
outcome type, Flows resource ID, Initiation method, Resource published
timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#flows-outcome-percentage}{Flows outcome percentage}.

The \code{FLOWS_OUTCOME_TYPE} is not a valid grouping.

\strong{PERCENT_NON_TALK_TIME}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Percentage

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#non-talk-time-percent}{Non-talk time percent}

\strong{PERCENT_TALK_TIME}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Percentage

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#talk-time-percent}{Talk time percent}

\strong{PERCENT_TALK_TIME_AGENT}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Percentage

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-talk-time-percent}{Agent talk time percent}

\strong{PERCENT_TALK_TIME_CUSTOMER}

This metric is available only for contacts analyzed by Contact Lens
conversational analytics.

Unit: Percentage

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#customer-talk-time-percent}{Customer talk time percent}

\strong{RECIPIENTS_ATTEMPTED}

This metric is only available for outbound campaigns initiated using a
customer segment. It is not available for event triggered campaigns.

Unit: Count

Valid groupings and filters: Campaign, Campaign Execution Timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#recipients-attempted}{Recipients attempted}

\strong{RECIPIENTS_INTERACTED}

This metric is only available for outbound campaigns initiated using a
customer segment. It is not available for event triggered campaigns.

Valid metric filter key: CAMPAIGN_INTERACTION_EVENT_TYPE

Unit: Count

Valid groupings and filters: Campaign, Channel,
contact/segmentAttributes/connect:Subtype, Campaign Execution Timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#recipients-interacted}{Recipients interacted}

\strong{RECIPIENTS_TARGETED}

This metric is only available for outbound campaigns initiated using a
customer segment. It is not available for event triggered campaigns.

Unit: Count

Valid groupings and filters: Campaign, Campaign Execution Timestamp

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#recipients-targeted}{Recipients targeted}

\strong{REOPENED_CASE_ACTIONS}

Unit: Count

Required filter key: CASE_TEMPLATE_ARN

Valid groupings and filters: CASE_TEMPLATE_ARN, CASE_STATUS

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#cases-reopened}{Cases reopened}

\strong{RESOLVED_CASE_ACTIONS}

Unit: Count

Required filter key: CASE_TEMPLATE_ARN

Valid groupings and filters: CASE_TEMPLATE_ARN, CASE_STATUS

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#cases-resolved}{Cases resolved}

\strong{SERVICE_LEVEL}

You can include up to 20 SERVICE_LEVEL metrics in a request.

Unit: Percent

Valid groupings and filters: Queue, Channel, Routing Profile, Q in
Connect

Threshold: For \code{ThresholdValue}, enter any whole number from 1 to 604800
(inclusive), in seconds. For \code{Comparison}, you can use \code{LT} (for "Less
than") or \code{LTE} (for "Less than equal").

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#service-level}{Service level X}

\strong{STEP_CONTACTS_QUEUED}

Unit: Count

Valid groupings and filters: Queue, RoutingStepExpression

UI name: This metric is available in Real-time Metrics UI but not on the
Historical Metrics UI.

\strong{SUM_AFTER_CONTACT_WORK_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#after-contact-work-time}{After contact work time}

\strong{SUM_CONNECTING_TIME_AGENT}

Unit: Seconds

Valid metric filter key: \code{INITIATION_METHOD}. This metric only supports
the following filter keys as \code{INITIATION_METHOD}: \code{INBOUND} | \code{OUTBOUND}
| \code{CALLBACK} | \code{API} | \code{CALLBACK_CUSTOMER_FIRST_DIALED}

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-api-connecting-time}{Agent API connecting time}

The \code{Negate} key in metric-level filters is not applicable for this
metric.

\strong{CONTACTS_ABANDONED}

Unit: Count

Metric filter:
\itemize{
\item Valid values: \code{API}| \code{INCOMING} | \code{OUTBOUND} | \code{TRANSFER} |
\code{CALLBACK} | \code{QUEUE_TRANSFER}| \code{Disconnect} |
\code{CALLBACK_CUSTOMER_FIRST_DIALED}
}

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype,
RoutingStepExpression, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-abandoned}{Contact abandoned}

\strong{SUM_CONTACTS_ABANDONED_IN_X}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile,
contact/segmentAttributes/connect:Subtype, Q in Connect

Threshold: For \code{ThresholdValue}, enter any whole number from 1 to 604800
(inclusive), in seconds. For \code{Comparison}, you can use \code{LT} (for "Less
than") or \code{LTE} (for "Less than equal").

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-abandoned-in-x-seconds}{Contacts abandoned in X seconds}

\strong{SUM_CONTACTS_ANSWERED_IN_X}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile,
contact/segmentAttributes/connect:Subtype, Q in Connect

Threshold: For \code{ThresholdValue}, enter any whole number from 1 to 604800
(inclusive), in seconds. For \code{Comparison}, you can use \code{LT} (for "Less
than") or \code{LTE} (for "Less than equal").

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contacts-answered-in-x-seconds}{Contacts answered in X seconds}

\strong{SUM_CONTACT_FLOW_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contact-flow-time}{Contact flow time}

\strong{SUM_CONTACT_TIME_AGENT}

Unit: Seconds

Valid groupings and filters: Routing Profile, Agent, Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-on-contact-time}{Agent on contact time}

\strong{SUM_CONTACTS_DISCONNECTED}

Valid metric filter key: \code{DISCONNECT_REASON}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contact-disconnected}{Contact disconnected}

\strong{SUM_ERROR_STATUS_TIME_AGENT}

Unit: Seconds

Valid groupings and filters: Routing Profile, Agent, Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#error-status-time}{Error status time}

\strong{SUM_HANDLE_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#contact-handle-time}{Contact handle time}

\strong{SUM_HOLD_TIME}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#customer-hold-time}{Customer hold time}

\strong{SUM_IDLE_TIME_AGENT}

Unit: Seconds

Valid groupings and filters: Routing Profile, Agent, Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-idle-time}{Agent idle time}

\strong{SUM_INTERACTION_AND_HOLD_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-interaction-and-hold-time}{Agent interaction and hold time}

\strong{SUM_INTERACTION_TIME}

Unit: Seconds

Valid groupings and filters: Queue, Channel, Routing Profile, Agent,
Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-interaction-time}{Agent interaction time}

\strong{SUM_NON_PRODUCTIVE_TIME_AGENT}

Unit: Seconds

Valid groupings and filters: Routing Profile, Agent, Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#agent-non-productive-time}{Agent non-productive time}

\strong{SUM_ONLINE_TIME_AGENT}

Unit: Seconds

Valid groupings and filters: Routing Profile, Agent, Agent Hierarchy

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#online-time}{Online time}

\strong{SUM_RETRY_CALLBACK_ATTEMPTS}

Unit: Count

Valid groupings and filters: Queue, Channel, Routing Profile,
contact/segmentAttributes/connect:Subtype, Q in Connect

UI name: \href{https://docs.aws.amazon.com/connect/latest/adminguide/metrics-definitions.html#callback-attempts}{Callback attempts}}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Gets metric data from the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_get_metric_data_v2/} for full documentation.
}
\keyword{internal}
