% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_contact_evaluations}
\alias{connect_list_contact_evaluations}
\title{Lists contact evaluations in the specified Amazon Connect instance}
\usage{
connect_list_contact_evaluations(InstanceId, ContactId, NextToken = NULL)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact in this instance of Amazon Connect.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.

This is not expected to be set because the value returned in the
previous response is always null.}
}
\description{
Lists contact evaluations in the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/list_contact_evaluations.html} for full documentation.
}
\keyword{internal}
