% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_queues}
\alias{connect_list_queues}
\title{Provides information about the queues for the specified Amazon Connect
instance}
\usage{
connect_list_queues(
  InstanceId,
  QueueTypes = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{QueueTypes}{The type of queue.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}
}
\description{
Provides information about the queues for the specified Amazon Connect instance.

See \url{https://paws-r.github.io/docs/connect/list_queues.html} for full documentation.
}
\keyword{internal}
