% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_tags_for_resource}
\alias{connect_list_tags_for_resource}
\title{Lists the tags for the specified resource}
\usage{
connect_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource. All Amazon Connect
resources (instances, queues, flows, routing profiles, etc) have an ARN.
To locate the ARN for an instance, for example, see \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{Find your Amazon Connect instance ID/ARN}.}
}
\description{
Lists the tags for the specified resource.

See \url{https://www.paws-r-sdk.com/docs/connect_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
