% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_list_use_cases}
\alias{connect_list_use_cases}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_list_use_cases(InstanceId, IntegrationAssociationId, NextToken,
  MaxResults)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{IntegrationAssociationId}{[required] The identifier for the integration association.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximimum number of results to return per page.}
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

List the use cases.
}
\section{Request syntax}{
\preformatted{svc$list_use_cases(
  InstanceId = "string",
  IntegrationAssociationId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
