% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_resume_contact_recording}
\alias{connect_resume_contact_recording}
\title{When a contact is being recorded, and the recording has been suspended
using SuspendContactRecording, this API resumes recording the call}
\usage{
connect_resume_contact_recording(InstanceId, ContactId,
  InitialContactId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{ContactId}{[required] The identifier of the contact.}

\item{InitialContactId}{[required] The identifier of the contact. This is the identifier of the contact
associated with the first interaction with the contact center.}
}
\value{
An empty list.
}
\description{
When a contact is being recorded, and the recording has been suspended
using SuspendContactRecording, this API resumes recording the call.

Only voice recordings are supported at this time.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$resume_contact_recording(
  InstanceId = "string",
  ContactId = "string",
  InitialContactId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
