% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_resume_contact_recording}
\alias{connect_resume_contact_recording}
\title{When a contact is being recorded, and the recording has been suspended
using SuspendContactRecording, this API resumes recording the call}
\usage{
connect_resume_contact_recording(InstanceId, ContactId, InitialContactId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact.}

\item{InitialContactId}{[required] The identifier of the contact. This is the identifier of the contact
associated with the first interaction with the contact center.}
}
\description{
When a contact is being recorded, and the recording has been suspended using SuspendContactRecording, this API resumes recording the call.

See \url{https://paws-r.github.io/docs/connect/resume_contact_recording.html} for full documentation.
}
\keyword{internal}
