% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_search_users}
\alias{connect_search_users}
\title{Searches users in an Amazon Connect instance, with optional filtering}
\usage{
connect_search_users(
  InstanceId = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  SearchFilter = NULL,
  SearchCriteria = NULL
)
}
\arguments{
\item{InstanceId}{The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{NextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{MaxResults}{The maximum number of results to return per page.}

\item{SearchFilter}{Filters to be applied to search results.}

\item{SearchCriteria}{}
}
\description{
Searches users in an Amazon Connect instance, with optional filtering.

See \url{https://paws-r.github.io/docs/connect/search_users.html} for full documentation.
}
\keyword{internal}
