% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_start_contact_streaming}
\alias{connect_start_contact_streaming}
\title{Initiates real-time message streaming for a new chat contact}
\usage{
connect_start_contact_streaming(
  InstanceId,
  ContactId,
  ChatStreamingConfiguration,
  ClientToken
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}

\item{ContactId}{[required] The identifier of the contact. This is the identifier of the contact
associated with the first interaction with the contact center.}

\item{ChatStreamingConfiguration}{[required] The streaming configuration, such as the Amazon SNS streaming endpoint.}

\item{ClientToken}{[required] A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\description{
Initiates real-time message streaming for a new chat contact.

See \url{https://paws-r.github.io/docs/connect/start_contact_streaming.html} for full documentation.
}
\keyword{internal}
