% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_tag_resource}
\alias{connect_tag_resource}
\title{Adds the specified tags to the specified resource}
\usage{
connect_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tags}{[required] One or more tags. For example, \{ "tags": \{"key1":"value1",
"key2":"value2"\} \}.}
}
\description{
Adds the specified tags to the specified resource.
}
\details{
The supported resource type is users.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
