% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_flow_content}
\alias{connect_update_contact_flow_content}
\title{Updates the specified flow}
\usage{
connect_update_contact_flow_content(InstanceId, ContactFlowId, Content)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{ContactFlowId}{[required] The identifier of the flow.}

\item{Content}{[required] The JSON string that represents the content of the flow. For an example,
see \href{https://docs.aws.amazon.com/connect/latest/APIReference/flow-language-example.html}{Example flow in Amazon Connect Flow language}.

Length Constraints: Minimum length of 1. Maximum length of 256000.}
}
\description{
Updates the specified flow.

See \url{https://www.paws-r-sdk.com/docs/connect_update_contact_flow_content/} for full documentation.
}
\keyword{internal}
