% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_flow_metadata}
\alias{connect_update_contact_flow_metadata}
\title{Updates metadata about specified flow}
\usage{
connect_update_contact_flow_metadata(
  InstanceId,
  ContactFlowId,
  Name = NULL,
  Description = NULL,
  ContactFlowState = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowId}{[required] The identifier of the flow.}

\item{Name}{The name of the flow.}

\item{Description}{The description of the flow.}

\item{ContactFlowState}{The state of flow.}
}
\description{
Updates metadata about specified flow.

See \url{https://www.paws-r-sdk.com/docs/connect_update_contact_flow_metadata/} for full documentation.
}
\keyword{internal}
