% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_flow_module_content}
\alias{connect_update_contact_flow_module_content}
\title{Updates specified flow module for the specified Amazon Connect instance}
\usage{
connect_update_contact_flow_module_content(
  InstanceId,
  ContactFlowModuleId,
  Content
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactFlowModuleId}{[required] The identifier of the flow module.}

\item{Content}{[required] The content of the flow module.}
}
\description{
Updates specified flow module for the specified Amazon Connect instance.

See \url{https://www.paws-r-sdk.com/docs/connect_update_contact_flow_module_content/} for full documentation.
}
\keyword{internal}
