% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_flow_name}
\alias{connect_update_contact_flow_name}
\title{The name of the flow}
\usage{
connect_update_contact_flow_name(
  InstanceId,
  ContactFlowId,
  Name = NULL,
  Description = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{ContactFlowId}{[required] The identifier of the flow.}

\item{Name}{The name of the flow.}

\item{Description}{The description of the flow.}
}
\description{
The name of the flow.

See \url{https://www.paws-r-sdk.com/docs/connect_update_contact_flow_name/} for full documentation.
}
\keyword{internal}
