% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_contact_routing_data}
\alias{connect_update_contact_routing_data}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_contact_routing_data(
  InstanceId,
  ContactId,
  QueueTimeAdjustmentSeconds = NULL,
  QueuePriority = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{ContactId}{[required] The identifier of the contact in this instance of Amazon Connect.}

\item{QueueTimeAdjustmentSeconds}{The number of seconds to add or subtract from the contact's routing age.
Contacts are routed to agents on a first-come, first-serve basis. This
means that changing their amount of time in queue compared to others
also changes their position in queue.}

\item{QueuePriority}{Priority of the contact in the queue. The default priority for new
contacts is 5. You can raise the priority of a contact compared to other
contacts in the queue by assigning them a higher priority, such as 1 or
2.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://www.paws-r-sdk.com/docs/connect_update_contact_routing_data/} for full documentation.
}
\keyword{internal}
