% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_hours_of_operation}
\alias{connect_update_hours_of_operation}
\title{Updates the hours of operation}
\usage{
connect_update_hours_of_operation(
  InstanceId,
  HoursOfOperationId,
  Name = NULL,
  Description = NULL,
  TimeZone = NULL,
  Config = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{HoursOfOperationId}{[required] The identifier of the hours of operation.}

\item{Name}{The name of the hours of operation.}

\item{Description}{The description of the hours of operation.}

\item{TimeZone}{The time zone of the hours of operation.}

\item{Config}{Configuration information of the hours of operation.}
}
\description{
Updates the hours of operation.

See \url{https://www.paws-r-sdk.com/docs/connect_update_hours_of_operation/} for full documentation.
}
\keyword{internal}
