% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_hours_of_operation_override}
\alias{connect_update_hours_of_operation_override}
\title{Update the hours of operation override}
\usage{
connect_update_hours_of_operation_override(
  InstanceId,
  HoursOfOperationId,
  HoursOfOperationOverrideId,
  Name = NULL,
  Description = NULL,
  Config = NULL,
  EffectiveFrom = NULL,
  EffectiveTill = NULL,
  RecurrenceConfig = NULL,
  OverrideType = NULL
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{HoursOfOperationId}{[required] The identifier for the hours of operation.}

\item{HoursOfOperationOverrideId}{[required] The identifier for the hours of operation override.}

\item{Name}{The name of the hours of operation override.}

\item{Description}{The description of the hours of operation override.}

\item{Config}{Configuration information for the hours of operation override: day,
start time, and end time.}

\item{EffectiveFrom}{The date from when the hours of operation override would be effective.}

\item{EffectiveTill}{The date until the hours of operation override is effective.}

\item{RecurrenceConfig}{Configuration for a recurring event.}

\item{OverrideType}{Whether the override will be defined as a \emph{standard} or as a \emph{recurring
event}.

For more information about how override types are applied, see Build
your list of overrides in the \emph{Administrator Guide}.}
}
\description{
Update the hours of operation override.

See \url{https://www.paws-r-sdk.com/docs/connect_update_hours_of_operation_override/} for full documentation.
}
\keyword{internal}
