% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_phone_number}
\alias{connect_update_phone_number}
\title{Updates your claimed phone number from its current Amazon Connect
instance or traffic distribution group to another Amazon Connect
instance or traffic distribution group in the same Amazon Web Services
Region}
\usage{
connect_update_phone_number(PhoneNumberId, TargetArn, ClientToken = NULL)
}
\arguments{
\item{PhoneNumberId}{[required] A unique identifier for the phone number.}

\item{TargetArn}{[required] The Amazon Resource Name (ARN) for Amazon Connect instances or traffic
distribution groups that phone numbers are claimed to.}

\item{ClientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}
}
\description{
Updates your claimed phone number from its current Amazon Connect instance or traffic distribution group to another Amazon Connect instance or traffic distribution group in the same Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/connect_update_phone_number/} for full documentation.
}
\keyword{internal}
