% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_queue_hours_of_operation}
\alias{connect_update_queue_hours_of_operation}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_queue_hours_of_operation(InstanceId, QueueId,
  HoursOfOperationId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{HoursOfOperationId}{[required] The identifier for the hours of operation.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Updates the hours of operation for the specified queue.
}
\section{Request syntax}{
\preformatted{svc$update_queue_hours_of_operation(
  InstanceId = "string",
  QueueId = "string",
  HoursOfOperationId = "string"
)
}
}

\keyword{internal}
