% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_queue_hours_of_operation}
\alias{connect_update_queue_hours_of_operation}
\title{Updates the hours of operation for the specified queue}
\usage{
connect_update_queue_hours_of_operation(
  InstanceId,
  QueueId,
  HoursOfOperationId
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{HoursOfOperationId}{[required] The identifier for the hours of operation.}
}
\description{
Updates the hours of operation for the specified queue.

See \url{https://www.paws-r-sdk.com/docs/connect_update_queue_hours_of_operation/} for full documentation.
}
\keyword{internal}
