% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_queue_name}
\alias{connect_update_queue_name}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_queue_name(InstanceId, QueueId, Name, Description)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{Name}{The name of the queue.}

\item{Description}{The description of the queue.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Updates the name and description of a queue. At least \code{Name} or
\code{Description} must be provided.
}
\section{Request syntax}{
\preformatted{svc$update_queue_name(
  InstanceId = "string",
  QueueId = "string",
  Name = "string",
  Description = "string"
)
}
}

\keyword{internal}
