% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_queue_name}
\alias{connect_update_queue_name}
\title{Updates the name and description of a queue}
\usage{
connect_update_queue_name(InstanceId, QueueId, Name = NULL, Description = NULL)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{Name}{The name of the queue.}

\item{Description}{The description of the queue.}
}
\description{
Updates the name and description of a queue. At least \code{Name} or \code{Description} must be provided.

See \url{https://www.paws-r-sdk.com/docs/connect_update_queue_name/} for full documentation.
}
\keyword{internal}
