% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_queue_outbound_caller_config}
\alias{connect_update_queue_outbound_caller_config}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_queue_outbound_caller_config(InstanceId, QueueId,
  OutboundCallerConfig)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{OutboundCallerConfig}{[required] The outbound caller ID name, number, and outbound whisper flow.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Updates the outbound caller ID name, number, and outbound whisper flow
for a specified queue.
}
\section{Request syntax}{
\preformatted{svc$update_queue_outbound_caller_config(
  InstanceId = "string",
  QueueId = "string",
  OutboundCallerConfig = list(
    OutboundCallerIdName = "string",
    OutboundCallerIdNumberId = "string",
    OutboundFlowId = "string"
  )
)
}
}

\keyword{internal}
