% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_queue_status}
\alias{connect_update_queue_status}
\title{Updates the status of the queue}
\usage{
connect_update_queue_status(InstanceId, QueueId, Status)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{QueueId}{[required] The identifier for the queue.}

\item{Status}{[required] The status of the queue.}
}
\description{
Updates the status of the queue.

See \url{https://www.paws-r-sdk.com/docs/connect_update_queue_status/} for full documentation.
}
\keyword{internal}
