% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_quick_connect_name}
\alias{connect_update_quick_connect_name}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_quick_connect_name(InstanceId, QuickConnectId, Name,
  Description)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{QuickConnectId}{[required] The identifier for the quick connect.}

\item{Name}{The name of the quick connect.}

\item{Description}{The description of the quick connect.}
}
\value{
An empty list.
}
\description{
This API is in preview release for Amazon Connect and is subject to
change.

Updates the name and description of a quick connect. The request accepts
the following data in JSON format. At least \code{Name} or \code{Description} must
be provided.
}
\section{Request syntax}{
\preformatted{svc$update_quick_connect_name(
  InstanceId = "string",
  QuickConnectId = "string",
  Name = "string",
  Description = "string"
)
}
}

\keyword{internal}
