% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_routing_profile_concurrency}
\alias{connect_update_routing_profile_concurrency}
\title{Updates the channels that agents can handle in the Contact Control Panel
(CCP) for a routing profile}
\usage{
connect_update_routing_profile_concurrency(
  InstanceId,
  RoutingProfileId,
  MediaConcurrencies
)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{MediaConcurrencies}{[required] The channels that agents can handle in the Contact Control Panel (CCP).}
}
\description{
Updates the channels that agents can handle in the Contact Control Panel (CCP) for a routing profile.

See \url{https://www.paws-r-sdk.com/docs/connect_update_routing_profile_concurrency/} for full documentation.
}
\keyword{internal}
