% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_routing_profile_concurrency}
\alias{connect_update_routing_profile_concurrency}
\title{Updates the channels that agents can handle in the Contact Control Panel
(CCP) for a routing profile}
\usage{
connect_update_routing_profile_concurrency(InstanceId, RoutingProfileId,
  MediaConcurrencies)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{MediaConcurrencies}{[required] The channels agents can handle in the Contact Control Panel (CCP).}
}
\description{
Updates the channels that agents can handle in the Contact Control Panel
(CCP) for a routing profile.
}
\section{Request syntax}{
\preformatted{svc$update_routing_profile_concurrency(
  InstanceId = "string",
  RoutingProfileId = "string",
  MediaConcurrencies = list(
    list(
      Channel = "VOICE"|"CHAT"|"TASK",
      Concurrency = 123
    )
  )
)
}
}

\keyword{internal}
