% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_routing_profile_default_outbound_queue}
\alias{connect_update_routing_profile_default_outbound_queue}
\title{Updates the default outbound queue of a routing profile}
\usage{
connect_update_routing_profile_default_outbound_queue(InstanceId,
  RoutingProfileId, DefaultOutboundQueueId)
}
\arguments{
\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}

\item{RoutingProfileId}{[required] The identifier of the routing profile.}

\item{DefaultOutboundQueueId}{[required] The identifier for the default outbound queue.}
}
\value{
An empty list.
}
\description{
Updates the default outbound queue of a routing profile.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_routing_profile_default_outbound_queue(
  InstanceId = "string",
  RoutingProfileId = "string",
  DefaultOutboundQueueId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
