% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_security_profile}
\alias{connect_update_security_profile}
\title{This API is in preview release for Amazon Connect and is subject to
change}
\usage{
connect_update_security_profile(
  Description = NULL,
  Permissions = NULL,
  SecurityProfileId,
  InstanceId
)
}
\arguments{
\item{Description}{The description of the security profile.}

\item{Permissions}{The permissions granted to a security profile.}

\item{SecurityProfileId}{[required] The identifier for the security profle.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}
}
\description{
This API is in preview release for Amazon Connect and is subject to change.

See \url{https://paws-r.github.io/docs/connect/update_security_profile.html} for full documentation.
}
\keyword{internal}
