% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_user_hierarchy_structure}
\alias{connect_update_user_hierarchy_structure}
\title{Updates the user hierarchy structure: add, remove, and rename user
hierarchy levels}
\usage{
connect_update_user_hierarchy_structure(HierarchyStructure, InstanceId)
}
\arguments{
\item{HierarchyStructure}{[required] The hierarchy levels to update.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}
}
\description{
Updates the user hierarchy structure: add, remove, and rename user hierarchy levels.

See \url{https://www.paws-r-sdk.com/docs/connect_update_user_hierarchy_structure/} for full documentation.
}
\keyword{internal}
