% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_user_hierarchy_structure}
\alias{connect_update_user_hierarchy_structure}
\title{Updates the user hierarchy structure: add, remove, and rename user
hierarchy levels}
\usage{
connect_update_user_hierarchy_structure(HierarchyStructure, InstanceId)
}
\arguments{
\item{HierarchyStructure}{[required] The hierarchy levels to update.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance.}
}
\value{
An empty list.
}
\description{
Updates the user hierarchy structure: add, remove, and rename user
hierarchy levels.
}
\section{Request syntax}{
\preformatted{svc$update_user_hierarchy_structure(
  HierarchyStructure = list(
    LevelOne = list(
      Name = "string"
    ),
    LevelTwo = list(
      Name = "string"
    ),
    LevelThree = list(
      Name = "string"
    ),
    LevelFour = list(
      Name = "string"
    ),
    LevelFive = list(
      Name = "string"
    )
  ),
  InstanceId = "string"
)
}
}

\keyword{internal}
