% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_user_phone_config}
\alias{connect_update_user_phone_config}
\title{Updates the phone configuration settings for the specified user}
\usage{
connect_update_user_phone_config(PhoneConfig, UserId, InstanceId)
}
\arguments{
\item{PhoneConfig}{[required] Information about phone configuration settings for the user.}

\item{UserId}{[required] The identifier of the user account.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can find the
instanceId in the ARN of the instance.}
}
\description{
Updates the phone configuration settings for the specified user.

See \url{https://paws-r.github.io/docs/connect/update_user_phone_config.html} for full documentation.
}
\keyword{internal}
