% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_operations.R
\name{connect_update_user_security_profiles}
\alias{connect_update_user_security_profiles}
\title{Assigns the specified security profiles to the specified user}
\usage{
connect_update_user_security_profiles(SecurityProfileIds, UserId, InstanceId)
}
\arguments{
\item{SecurityProfileIds}{[required] The identifiers of the security profiles for the user.}

\item{UserId}{[required] The identifier of the user account.}

\item{InstanceId}{[required] The identifier of the Amazon Connect instance. You can \href{https://docs.aws.amazon.com/connect/latest/adminguide/find-instance-arn.html}{find the instance ID}
in the Amazon Resource Name (ARN) of the instance.}
}
\description{
Assigns the specified security profiles to the specified user.

See \url{https://paws-r.github.io/docs/connect/update_user_security_profiles.html} for full documentation.
}
\keyword{internal}
