% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_get_campaign_state}
\alias{connectcampaignservice_get_campaign_state}
\title{Get state of a campaign for the specified Amazon Connect account}
\usage{
connectcampaignservice_get_campaign_state(id)
}
\arguments{
\item{id}{[required]}
}
\description{
Get state of a campaign for the specified Amazon Connect account.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_get_campaign_state/} for full documentation.
}
\keyword{internal}
