% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_get_campaign_state_batch}
\alias{connectcampaignservice_get_campaign_state_batch}
\title{Get state of campaigns for the specified Amazon Connect account}
\usage{
connectcampaignservice_get_campaign_state_batch(campaignIds)
}
\arguments{
\item{campaignIds}{[required]}
}
\description{
Get state of campaigns for the specified Amazon Connect account.

See \url{https://paws-r.github.io/docs/connectcampaignservice/get_campaign_state_batch.html} for full documentation.
}
\keyword{internal}
