% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_get_connect_instance_config}
\alias{connectcampaignservice_get_connect_instance_config}
\title{Get the specific Connect instance config}
\usage{
connectcampaignservice_get_connect_instance_config(connectInstanceId)
}
\arguments{
\item{connectInstanceId}{[required]}
}
\description{
Get the specific Connect instance config.

See \url{https://paws-r.github.io/docs/connectcampaignservice/get_connect_instance_config.html} for full documentation.
}
\keyword{internal}
