% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_list_campaigns}
\alias{connectcampaignservice_list_campaigns}
\title{Provides summary information about the campaigns under the specified
Amazon Connect account}
\usage{
connectcampaignservice_list_campaigns(
  maxResults = NULL,
  nextToken = NULL,
  filters = NULL
)
}
\arguments{
\item{maxResults}{}

\item{nextToken}{}

\item{filters}{}
}
\description{
Provides summary information about the campaigns under the specified Amazon Connect account.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_list_campaigns/} for full documentation.
}
\keyword{internal}
