% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_list_tags_for_resource}
\alias{connectcampaignservice_list_tags_for_resource}
\title{List tags for a resource}
\usage{
connectcampaignservice_list_tags_for_resource(arn)
}
\arguments{
\item{arn}{[required]}
}
\description{
List tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
