% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_start_instance_onboarding_job}
\alias{connectcampaignservice_start_instance_onboarding_job}
\title{Onboard the specific Amazon Connect instance to Connect Campaigns}
\usage{
connectcampaignservice_start_instance_onboarding_job(
  connectInstanceId,
  encryptionConfig
)
}
\arguments{
\item{connectInstanceId}{[required]}

\item{encryptionConfig}{[required]}
}
\description{
Onboard the specific Amazon Connect instance to Connect Campaigns.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_start_instance_onboarding_job/} for full documentation.
}
\keyword{internal}
