% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_untag_resource}
\alias{connectcampaignservice_untag_resource}
\title{Untag a resource}
\usage{
connectcampaignservice_untag_resource(arn, tagKeys)
}
\arguments{
\item{arn}{[required]}

\item{tagKeys}{[required]}
}
\description{
Untag a resource.

See \url{https://paws-r.github.io/docs/connectcampaignservice/untag_resource.html} for full documentation.
}
\keyword{internal}
