% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_update_campaign_dialer_config}
\alias{connectcampaignservice_update_campaign_dialer_config}
\title{Updates the dialer config of a campaign}
\usage{
connectcampaignservice_update_campaign_dialer_config(dialerConfig, id)
}
\arguments{
\item{dialerConfig}{[required]}

\item{id}{[required]}
}
\description{
Updates the dialer config of a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_update_campaign_dialer_config/} for full documentation.
}
\keyword{internal}
