% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservice_operations.R
\name{connectcampaignservice_update_campaign_outbound_call_config}
\alias{connectcampaignservice_update_campaign_outbound_call_config}
\title{Updates the outbound call config of a campaign}
\usage{
connectcampaignservice_update_campaign_outbound_call_config(
  id,
  connectContactFlowId = NULL,
  connectSourcePhoneNumber = NULL,
  answerMachineDetectionConfig = NULL
)
}
\arguments{
\item{id}{[required]}

\item{connectContactFlowId}{}

\item{connectSourcePhoneNumber}{}

\item{answerMachineDetectionConfig}{}
}
\description{
Updates the outbound call config of a campaign. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservice_update_campaign_outbound_call_config/} for full documentation.
}
\keyword{internal}
