% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_create_campaign}
\alias{connectcampaignservicev2_create_campaign}
\title{Creates a campaign for the specified Amazon Connect account}
\usage{
connectcampaignservicev2_create_campaign(
  name,
  connectInstanceId,
  channelSubtypeConfig = NULL,
  type = NULL,
  source = NULL,
  connectCampaignFlowArn = NULL,
  schedule = NULL,
  communicationTimeConfig = NULL,
  communicationLimitsOverride = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required]}

\item{connectInstanceId}{[required]}

\item{channelSubtypeConfig}{}

\item{type}{}

\item{source}{}

\item{connectCampaignFlowArn}{}

\item{schedule}{}

\item{communicationTimeConfig}{}

\item{communicationLimitsOverride}{}

\item{tags}{}
}
\description{
Creates a campaign for the specified Amazon Connect account. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_create_campaign/} for full documentation.
}
\keyword{internal}
