% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcampaignservicev2_operations.R
\name{connectcampaignservicev2_put_instance_communication_limits}
\alias{connectcampaignservicev2_put_instance_communication_limits}
\title{Put the instance communication limits}
\usage{
connectcampaignservicev2_put_instance_communication_limits(
  connectInstanceId,
  communicationLimitsConfig
)
}
\arguments{
\item{connectInstanceId}{[required]}

\item{communicationLimitsConfig}{[required]}
}
\description{
Put the instance communication limits. This API is idempotent.

See \url{https://www.paws-r-sdk.com/docs/connectcampaignservicev2_put_instance_communication_limits/} for full documentation.
}
\keyword{internal}
