% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_create_case}
\alias{connectcases_create_case}
\title{Creates a case in the specified Cases domain}
\usage{
connectcases_create_case(clientToken = NULL, domainId, fields, templateId)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{fields}{[required] An array of objects with field ID (matching ListFields/DescribeField)
and value union data.}

\item{templateId}{[required] A unique identifier of a template.}
}
\description{
Creates a case in the specified Cases domain. Case system and custom fields are taken as an array id/value pairs with a declared data types.

See \url{https://www.paws-r-sdk.com/docs/connectcases_create_case/} for full documentation.
}
\keyword{internal}
