% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_create_related_item}
\alias{connectcases_create_related_item}
\title{Creates a related item (comments, tasks, and contacts) and associates it
with a case}
\usage{
connectcases_create_related_item(
  domainId,
  caseId,
  type,
  content,
  performedBy = NULL
)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{caseId}{[required] A unique identifier of the case.}

\item{type}{[required] The type of a related item.}

\item{content}{[required] The content of a related item to be created.}

\item{performedBy}{Represents the creator of the related item.}
}
\description{
Creates a related item (comments, tasks, and contacts) and associates it with a case.

See \url{https://www.paws-r-sdk.com/docs/connectcases_create_related_item/} for full documentation.
}
\keyword{internal}
