% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_search_related_items}
\alias{connectcases_search_related_items}
\title{Searches for related items that are associated with a case}
\usage{
connectcases_search_related_items(
  domainId,
  caseId,
  maxResults = NULL,
  nextToken = NULL,
  filters = NULL
)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{caseId}{[required] A unique identifier of the case.}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{filters}{The list of types of related items and their parameters to use for
filtering.}
}
\description{
Searches for related items that are associated with a case.

See \url{https://www.paws-r-sdk.com/docs/connectcases_search_related_items/} for full documentation.
}
\keyword{internal}
