% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectcases_operations.R
\name{connectcases_update_case}
\alias{connectcases_update_case}
\title{If you provide a value for PerformedBy}
\usage{
connectcases_update_case(domainId, caseId, fields, performedBy = NULL)
}
\arguments{
\item{domainId}{[required] The unique identifier of the Cases domain.}

\item{caseId}{[required] A unique identifier of the case.}

\item{fields}{[required] An array of objects with \code{fieldId} (matching ListFields/DescribeField)
and value union data, structured identical to
\code{\link[=connectcases_create_case]{create_case}}.}

\item{performedBy}{}
}
\description{
If you provide a value for \code{PerformedBy.UserArn} you must also have \href{https://docs.aws.amazon.com/connect/latest/APIReference/API_DescribeUser.html}{connect:DescribeUser} permission on the User ARN resource that you provide

See \url{https://www.paws-r-sdk.com/docs/connectcases_update_case/} for full documentation.
}
\keyword{internal}
