% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectparticipant_operations.R
\name{connectparticipant_get_attachment}
\alias{connectparticipant_get_attachment}
\title{Provides a pre-signed URL for download of a completed attachment}
\usage{
connectparticipant_get_attachment(AttachmentId, ConnectionToken)
}
\arguments{
\item{AttachmentId}{[required] A unique identifier for the attachment.}

\item{ConnectionToken}{[required] The authentication token associated with the participant's connection.}
}
\description{
Provides a pre-signed URL for download of a completed attachment. This is an asynchronous API for use with active contacts.

See \url{https://www.paws-r-sdk.com/docs/connectparticipant_get_attachment/} for full documentation.
}
\keyword{internal}
