% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_create_session}
\alias{connectwisdomservice_create_session}
\title{Creates a session}
\usage{
connectwisdomservice_create_session(
  assistantId,
  clientToken = NULL,
  description = NULL,
  name,
  tags = NULL
)
}
\arguments{
\item{assistantId}{[required] The identifier of the Wisdom assistant. Can be either the ID or the ARN.
URLs cannot contain the ARN.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. If not provided, the Amazon Web Services SDK
populates this field. For more information about idempotency, see
\href{https://aws.amazon.com/builders-library/making-retries-safe-with-idempotent-APIs/}{Making retries safe with idempotent APIs}.}

\item{description}{The description.}

\item{name}{[required] The name of the session.}

\item{tags}{The tags used to organize, track, or control access for this resource.}
}
\description{
Creates a session. A session is a contextual container used for generating recommendations. Amazon Connect creates a new Wisdom session for each contact on which Wisdom is enabled.

See \url{https://paws-r.github.io/docs/connectwisdomservice/create_session.html} for full documentation.
}
\keyword{internal}
