% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_get_assistant}
\alias{connectwisdomservice_get_assistant}
\title{Retrieves information about an assistant}
\usage{
connectwisdomservice_get_assistant(assistantId)
}
\arguments{
\item{assistantId}{[required] The identifier of the Wisdom assistant. Can be either the ID or the ARN.
URLs cannot contain the ARN.}
}
\description{
Retrieves information about an assistant.

See \url{https://paws-r.github.io/docs/connectwisdomservice/get_assistant.html} for full documentation.
}
\keyword{internal}
