% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_get_knowledge_base}
\alias{connectwisdomservice_get_knowledge_base}
\title{Retrieves information about the knowledge base}
\usage{
connectwisdomservice_get_knowledge_base(knowledgeBaseId)
}
\arguments{
\item{knowledgeBaseId}{[required] The identifier of the knowledge base. This should not be a
QUICK_RESPONSES type knowledge base if you're storing Wisdom Content
resource to it. Can be either the ID or the ARN. URLs cannot contain the
ARN.}
}
\description{
Retrieves information about the knowledge base.

See \url{https://www.paws-r-sdk.com/docs/connectwisdomservice_get_knowledge_base/} for full documentation.
}
\keyword{internal}
