% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwisdomservice_operations.R
\name{connectwisdomservice_search_content}
\alias{connectwisdomservice_search_content}
\title{Searches for content in a specified knowledge base}
\usage{
connectwisdomservice_search_content(
  knowledgeBaseId,
  maxResults = NULL,
  nextToken = NULL,
  searchExpression
)
}
\arguments{
\item{knowledgeBaseId}{[required] The identifier of the knowledge base. Can be either the ID or the ARN.
URLs cannot contain the ARN.}

\item{maxResults}{The maximum number of results to return per page.}

\item{nextToken}{The token for the next set of results. Use the value returned in the
previous response in the next request to retrieve the next set of
results.}

\item{searchExpression}{[required] The search expression to filter results.}
}
\description{
Searches for content in a specified knowledge base. Can be used to get a specific content resource by its name.

See \url{https://paws-r.github.io/docs/connectwisdomservice/search_content.html} for full documentation.
}
\keyword{internal}
