% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_domain}
\alias{customerprofiles_get_domain}
\title{Returns information about a specific domain}
\usage{
customerprofiles_get_domain(DomainName)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}
}
\description{
Returns information about a specific domain.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_domain/} for full documentation.
}
\keyword{internal}
