% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_get_profile_recommendations}
\alias{customerprofiles_get_profile_recommendations}
\title{Fetches the recommendations for a profile in the input Customer Profiles
domain}
\usage{
customerprofiles_get_profile_recommendations(
  DomainName,
  ProfileId,
  RecommenderName,
  Context = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DomainName}{[required] The unique name of the domain.}

\item{ProfileId}{[required] The unique identifier of the profile for which to retrieve
recommendations.}

\item{RecommenderName}{[required] The unique name of the recommender.}

\item{Context}{The contextual metadata used to provide dynamic runtime information to
tailor recommendations.}

\item{MaxResults}{The maximum number of recommendations to return. The default value is
10.}
}
\description{
Fetches the recommendations for a profile in the input Customer Profiles domain. Fetches all the profile recommendations

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_get_profile_recommendations/} for full documentation.
}
\keyword{internal}
