% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_list_profile_objects}
\alias{customerprofiles_list_profile_objects}
\title{Returns a list of objects associated with a profile of a given
ProfileObjectType}
\usage{
customerprofiles_list_profile_objects(
  NextToken = NULL,
  MaxResults = NULL,
  DomainName,
  ObjectTypeName,
  ProfileId,
  ObjectFilter = NULL
)
}
\arguments{
\item{NextToken}{The pagination token from the previous call to ListProfileObjects.}

\item{MaxResults}{The maximum number of objects returned per page.}

\item{DomainName}{[required] The unique name of the domain.}

\item{ObjectTypeName}{[required] The name of the profile object type.}

\item{ProfileId}{[required] The unique identifier of a customer profile.}

\item{ObjectFilter}{Applies a filter to the response to include profile objects with the
specified index values. This filter is only supported for ObjectTypeName
_asset, _case and _order.}
}
\description{
Returns a list of objects associated with a profile of a given ProfileObjectType.

See \url{https://paws-r.github.io/docs/customerprofiles/list_profile_objects.html} for full documentation.
}
\keyword{internal}
