% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customerprofiles_operations.R
\name{customerprofiles_untag_resource}
\alias{customerprofiles_untag_resource}
\title{Removes one or more tags from the specified Amazon Connect Customer
Profiles resource}
\usage{
customerprofiles_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource from which you are removing tags.}

\item{tagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified Amazon Connect Customer Profiles resource. In Connect Customer Profiles, domains, profile object types, and integrations can be tagged.

See \url{https://www.paws-r-sdk.com/docs/customerprofiles_untag_resource/} for full documentation.
}
\keyword{internal}
