% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_app}
\alias{pinpoint_create_app}
\title{Creates an application}
\usage{
pinpoint_create_app(CreateApplicationRequest)
}
\arguments{
\item{CreateApplicationRequest}{[required]}
}
\value{
A list with the following syntax:\preformatted{list(
  ApplicationResponse = list(
    Arn = "string",
    Id = "string",
    Name = "string",
    tags = list(
      "string"
    )
  )
)
}
}
\description{
Creates an application.
}
\section{Request syntax}{
\preformatted{svc$create_app(
  CreateApplicationRequest = list(
    Name = "string",
    tags = list(
      "string"
    )
  )
)
}
}

\keyword{internal}
