% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_email_template}
\alias{pinpoint_create_email_template}
\title{Creates a message template for messages that are sent through the email
channel}
\usage{
pinpoint_create_email_template(EmailTemplateRequest, TemplateName)
}
\arguments{
\item{EmailTemplateRequest}{[required]}

\item{TemplateName}{[required] The name of the message template. A template name must start with an
alphanumeric character and can contain a maximum of 128 characters. The
characters can be alphanumeric characters, underscores (_), or hyphens
(-). Template names are case sensitive.}
}
\description{
Creates a message template for messages that are sent through the email channel.

See \url{https://paws-r.github.io/docs/pinpoint/create_email_template.html} for full documentation.
}
\keyword{internal}
