% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_import_job}
\alias{pinpoint_create_import_job}
\title{Creates an import job for an application}
\usage{
pinpoint_create_import_job(ApplicationId, ImportJobRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{ImportJobRequest}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ImportJobResponse = list(
    ApplicationId = "string",
    CompletedPieces = 123,
    CompletionDate = "string",
    CreationDate = "string",
    Definition = list(
      DefineSegment = TRUE|FALSE,
      ExternalId = "string",
      Format = "CSV"|"JSON",
      RegisterEndpoints = TRUE|FALSE,
      RoleArn = "string",
      S3Url = "string",
      SegmentId = "string",
      SegmentName = "string"
    ),
    FailedPieces = 123,
    Failures = list(
      "string"
    ),
    Id = "string",
    JobStatus = "CREATED"|"PREPARING_FOR_INITIALIZATION"|"INITIALIZING"|"PROCESSING"|"PENDING_JOB"|"COMPLETING"|"COMPLETED"|"FAILING"|"FAILED",
    TotalFailures = 123,
    TotalPieces = 123,
    TotalProcessed = 123,
    Type = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates an import job for an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_import_job(
  ApplicationId = "string",
  ImportJobRequest = list(
    DefineSegment = TRUE|FALSE,
    ExternalId = "string",
    Format = "CSV"|"JSON",
    RegisterEndpoints = TRUE|FALSE,
    RoleArn = "string",
    S3Url = "string",
    SegmentId = "string",
    SegmentName = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
