% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_import_job}
\alias{pinpoint_create_import_job}
\title{Creates an import job for an application}
\usage{
pinpoint_create_import_job(ApplicationId, ImportJobRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{ImportJobRequest}{[required]}
}
\description{
Creates an import job for an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_create_import_job/} for full documentation.
}
\keyword{internal}
