% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_import_job}
\alias{pinpoint_create_import_job}
\title{Creates a new import job for an application}
\usage{
pinpoint_create_import_job(ApplicationId, ImportJobRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{ImportJobRequest}{[required]}
}
\description{
Creates a new import job for an application.
}
\section{Request syntax}{
\preformatted{svc$create_import_job(
  ApplicationId = "string",
  ImportJobRequest = list(
    DefineSegment = TRUE|FALSE,
    ExternalId = "string",
    Format = "CSV"|"JSON",
    RegisterEndpoints = TRUE|FALSE,
    RoleArn = "string",
    S3Url = "string",
    SegmentId = "string",
    SegmentName = "string"
  )
)
}
}

\keyword{internal}
