% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_recommender_configuration}
\alias{pinpoint_create_recommender_configuration}
\title{Creates an Amazon Pinpoint configuration for a recommender model}
\usage{
pinpoint_create_recommender_configuration(
  CreateRecommenderConfiguration)
}
\arguments{
\item{CreateRecommenderConfiguration}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RecommenderConfigurationResponse = list(
    Attributes = list(
      "string"
    ),
    CreationDate = "string",
    Description = "string",
    Id = "string",
    LastModifiedDate = "string",
    Name = "string",
    RecommendationProviderIdType = "string",
    RecommendationProviderRoleArn = "string",
    RecommendationProviderUri = "string",
    RecommendationTransformerUri = "string",
    RecommendationsDisplayName = "string",
    RecommendationsPerMessage = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates an Amazon Pinpoint configuration for a recommender model.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_recommender_configuration(
  CreateRecommenderConfiguration = list(
    Attributes = list(
      "string"
    ),
    Description = "string",
    Name = "string",
    RecommendationProviderIdType = "string",
    RecommendationProviderRoleArn = "string",
    RecommendationProviderUri = "string",
    RecommendationTransformerUri = "string",
    RecommendationsDisplayName = "string",
    RecommendationsPerMessage = 123
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
