% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_create_voice_template}
\alias{pinpoint_create_voice_template}
\title{Creates a message template for messages that are sent through the voice
channel}
\usage{
pinpoint_create_voice_template(TemplateName, VoiceTemplateRequest)
}
\arguments{
\item{TemplateName}{[required] The name of the message template. A template name must start with an
alphanumeric character and can contain a maximum of 128 characters. The
characters can be alphanumeric characters, underscores (_), or hyphens
(-). Template names are case sensitive.}

\item{VoiceTemplateRequest}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CreateTemplateMessageBody = list(
    Arn = "string",
    Message = "string",
    RequestID = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a message template for messages that are sent through the voice
channel.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_voice_template(
  TemplateName = "string",
  VoiceTemplateRequest = list(
    Body = "string",
    DefaultSubstitutions = "string",
    LanguageCode = "string",
    tags = list(
      "string"
    ),
    TemplateDescription = "string",
    VoiceId = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
