% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_app}
\alias{pinpoint_delete_app}
\title{Deletes an application}
\usage{
pinpoint_delete_app(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ApplicationResponse = list(
    Arn = "string",
    Id = "string",
    Name = "string",
    tags = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_app(
  ApplicationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
