% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_app}
\alias{pinpoint_delete_app}
\title{Deletes an application}
\usage{
pinpoint_delete_app(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\description{
Deletes an application.

See \url{https://paws-r.github.io/docs/pinpoint/delete_app.html} for full documentation.
}
\keyword{internal}
