% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_campaign}
\alias{pinpoint_delete_campaign}
\title{Deletes a campaign from an application}
\usage{
pinpoint_delete_campaign(ApplicationId, CampaignId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{CampaignId}{[required] The unique identifier for the campaign.}
}
\description{
Deletes a campaign from an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_delete_campaign/} for full documentation.
}
\keyword{internal}
