% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_email_channel}
\alias{pinpoint_delete_email_channel}
\title{Disables the email channel for an application and deletes any existing
settings for the channel}
\usage{
pinpoint_delete_email_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EmailChannelResponse = list(
    ApplicationId = "string",
    ConfigurationSet = "string",
    CreationDate = "string",
    Enabled = TRUE|FALSE,
    FromAddress = "string",
    HasCredential = TRUE|FALSE,
    Id = "string",
    Identity = "string",
    IsArchived = TRUE|FALSE,
    LastModifiedBy = "string",
    LastModifiedDate = "string",
    MessagesPerSecond = 123,
    Platform = "string",
    RoleArn = "string",
    Version = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Disables the email channel for an application and deletes any existing
settings for the channel.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_email_channel(
  ApplicationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
