% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_recommender_configuration}
\alias{pinpoint_delete_recommender_configuration}
\title{Deletes an Amazon Pinpoint configuration for a recommender model}
\usage{
pinpoint_delete_recommender_configuration(RecommenderId)
}
\arguments{
\item{RecommenderId}{[required] The unique identifier for the recommender model configuration. This
identifier is displayed as the \strong{Recommender ID} on the Amazon Pinpoint
console.}
}
\description{
Deletes an Amazon Pinpoint configuration for a recommender model.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_delete_recommender_configuration/} for full documentation.
}
\keyword{internal}
