% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_delete_voice_channel}
\alias{pinpoint_delete_voice_channel}
\title{Disables the voice channel for an application and deletes any existing
settings for the channel}
\usage{
pinpoint_delete_voice_channel(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\description{
Disables the voice channel for an application and deletes any existing settings for the channel.

See \url{https://paws-r.github.io/docs/pinpoint/delete_voice_channel.html} for full documentation.
}
\keyword{internal}
