% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_application_settings}
\alias{pinpoint_get_application_settings}
\title{Retrieves information about the settings for an application}
\usage{
pinpoint_get_application_settings(ApplicationId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\description{
Retrieves information about the settings for an application.

See \url{https://paws-r.github.io/docs/pinpoint/get_application_settings.html} for full documentation.
}
\keyword{internal}
