% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_campaign}
\alias{pinpoint_get_campaign}
\title{Returns information about a campaign}
\usage{
pinpoint_get_campaign(ApplicationId, CampaignId)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{CampaignId}{[required] The unique ID of the campaign.}
}
\description{
Returns information about a campaign.
}
\section{Request syntax}{
\preformatted{svc$get_campaign(
  ApplicationId = "string",
  CampaignId = "string"
)
}
}

\keyword{internal}
