% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_campaign_activities}
\alias{pinpoint_get_campaign_activities}
\title{Retrieves information about all the activities for a campaign}
\usage{
pinpoint_get_campaign_activities(
  ApplicationId,
  CampaignId,
  PageSize = NULL,
  Token = NULL
)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{CampaignId}{[required] The unique identifier for the campaign.}

\item{PageSize}{The maximum number of items to include in each page of a paginated
response. This parameter is not supported for application, campaign, and
journey metrics.}

\item{Token}{The NextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about all the activities for a campaign.

See \url{https://paws-r.github.io/docs/pinpoint/get_campaign_activities.html} for full documentation.
}
\keyword{internal}
