% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_campaign_version}
\alias{pinpoint_get_campaign_version}
\title{Retrieves information about the status, configuration, and other
settings for a specific version of a campaign}
\usage{
pinpoint_get_campaign_version(ApplicationId, CampaignId, Version)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{CampaignId}{[required] The unique identifier for the campaign.}

\item{Version}{[required] The unique version number (Version property) for the campaign version.}
}
\description{
Retrieves information about the status, configuration, and other settings for a specific version of a campaign.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_get_campaign_version/} for full documentation.
}
\keyword{internal}
