% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_campaigns}
\alias{pinpoint_get_campaigns}
\title{Retrieves information about the status, configuration, and other
settings for all the campaigns that are associated with an application}
\usage{
pinpoint_get_campaigns(ApplicationId, PageSize = NULL, Token = NULL)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{PageSize}{The maximum number of items to include in each page of a paginated
response. This parameter is not supported for application, campaign, and
journey metrics.}

\item{Token}{The NextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about the status, configuration, and other settings for all the campaigns that are associated with an application.

See \url{https://paws-r.github.io/docs/pinpoint/get_campaigns.html} for full documentation.
}
\keyword{internal}
