% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_endpoint}
\alias{pinpoint_get_endpoint}
\title{Retrieves information about the settings and attributes of a specific
endpoint for an application}
\usage{
pinpoint_get_endpoint(ApplicationId, EndpointId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{EndpointId}{[required] The case insensitive unique identifier for the endpoint. The identifier
can't contain \code{$}, \verb{\\\{} or \verb{\\\}}.}
}
\description{
Retrieves information about the settings and attributes of a specific endpoint for an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_get_endpoint/} for full documentation.
}
\keyword{internal}
