% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_export_job}
\alias{pinpoint_get_export_job}
\title{Retrieves information about the status and settings of a specific export
job for an application}
\usage{
pinpoint_get_export_job(ApplicationId, JobId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{JobId}{[required] The unique identifier for the job.}
}
\description{
Retrieves information about the status and settings of a specific export job for an application.

See \url{https://paws-r.github.io/docs/pinpoint/get_export_job.html} for full documentation.
}
\keyword{internal}
