% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_recommender_configurations}
\alias{pinpoint_get_recommender_configurations}
\title{Retrieves information about all the recommender model configurations
that are associated with your Amazon Pinpoint account}
\usage{
pinpoint_get_recommender_configurations(PageSize = NULL, Token = NULL)
}
\arguments{
\item{PageSize}{The maximum number of items to include in each page of a paginated
response. This parameter is not supported for application, campaign, and
journey metrics.}

\item{Token}{The NextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about all the recommender model configurations that are associated with your Amazon Pinpoint account.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_get_recommender_configurations/} for full documentation.
}
\keyword{internal}
