% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_segment_version}
\alias{pinpoint_get_segment_version}
\title{Retrieves information about the configuration, dimension, and other
settings for a specific version of a segment that's associated with an
application}
\usage{
pinpoint_get_segment_version(ApplicationId, SegmentId, Version)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{SegmentId}{[required] The unique identifier for the segment.}

\item{Version}{[required] The unique version number (Version property) for the campaign version.}
}
\value{
A list with the following syntax:\preformatted{list(
  SegmentResponse = list(
    ApplicationId = "string",
    Arn = "string",
    CreationDate = "string",
    Dimensions = list(
      Attributes = list(
        list(
          AttributeType = "INCLUSIVE"|"EXCLUSIVE"|"CONTAINS"|"BEFORE"|"AFTER"|"ON"|"BETWEEN",
          Values = list(
            "string"
          )
        )
      ),
      Behavior = list(
        Recency = list(
          Duration = "HR_24"|"DAY_7"|"DAY_14"|"DAY_30",
          RecencyType = "ACTIVE"|"INACTIVE"
        )
      ),
      Demographic = list(
        AppVersion = list(
          DimensionType = "INCLUSIVE"|"EXCLUSIVE",
          Values = list(
            "string"
          )
        ),
        Channel = list(
          DimensionType = "INCLUSIVE"|"EXCLUSIVE",
          Values = list(
            "string"
          )
        ),
        DeviceType = list(
          DimensionType = "INCLUSIVE"|"EXCLUSIVE",
          Values = list(
            "string"
          )
        ),
        Make = list(
          DimensionType = "INCLUSIVE"|"EXCLUSIVE",
          Values = list(
            "string"
          )
        ),
        Model = list(
          DimensionType = "INCLUSIVE"|"EXCLUSIVE",
          Values = list(
            "string"
          )
        ),
        Platform = list(
          DimensionType = "INCLUSIVE"|"EXCLUSIVE",
          Values = list(
            "string"
          )
        )
      ),
      Location = list(
        Country = list(
          DimensionType = "INCLUSIVE"|"EXCLUSIVE",
          Values = list(
            "string"
          )
        ),
        GPSPoint = list(
          Coordinates = list(
            Latitude = 123.0,
            Longitude = 123.0
          ),
          RangeInKilometers = 123.0
        )
      ),
      Metrics = list(
        list(
          ComparisonOperator = "string",
          Value = 123.0
        )
      ),
      UserAttributes = list(
        list(
          AttributeType = "INCLUSIVE"|"EXCLUSIVE"|"CONTAINS"|"BEFORE"|"AFTER"|"ON"|"BETWEEN",
          Values = list(
            "string"
          )
        )
      )
    ),
    Id = "string",
    ImportDefinition = list(
      ChannelCounts = list(
        123
      ),
      ExternalId = "string",
      Format = "CSV"|"JSON",
      RoleArn = "string",
      S3Url = "string",
      Size = 123
    ),
    LastModifiedDate = "string",
    Name = "string",
    SegmentGroups = list(
      Groups = list(
        list(
          Dimensions = list(
            list(
              Attributes = list(
                list(
                  AttributeType = "INCLUSIVE"|"EXCLUSIVE"|"CONTAINS"|"BEFORE"|"AFTER"|"ON"|"BETWEEN",
                  Values = list(
                    "string"
                  )
                )
              ),
              Behavior = list(
                Recency = list(
                  Duration = "HR_24"|"DAY_7"|"DAY_14"|"DAY_30",
                  RecencyType = "ACTIVE"|"INACTIVE"
                )
              ),
              Demographic = list(
                AppVersion = list(
                  DimensionType = "INCLUSIVE"|"EXCLUSIVE",
                  Values = list(
                    "string"
                  )
                ),
                Channel = list(
                  DimensionType = "INCLUSIVE"|"EXCLUSIVE",
                  Values = list(
                    "string"
                  )
                ),
                DeviceType = list(
                  DimensionType = "INCLUSIVE"|"EXCLUSIVE",
                  Values = list(
                    "string"
                  )
                ),
                Make = list(
                  DimensionType = "INCLUSIVE"|"EXCLUSIVE",
                  Values = list(
                    "string"
                  )
                ),
                Model = list(
                  DimensionType = "INCLUSIVE"|"EXCLUSIVE",
                  Values = list(
                    "string"
                  )
                ),
                Platform = list(
                  DimensionType = "INCLUSIVE"|"EXCLUSIVE",
                  Values = list(
                    "string"
                  )
                )
              ),
              Location = list(
                Country = list(
                  DimensionType = "INCLUSIVE"|"EXCLUSIVE",
                  Values = list(
                    "string"
                  )
                ),
                GPSPoint = list(
                  Coordinates = list(
                    Latitude = 123.0,
                    Longitude = 123.0
                  ),
                  RangeInKilometers = 123.0
                )
              ),
              Metrics = list(
                list(
                  ComparisonOperator = "string",
                  Value = 123.0
                )
              ),
              UserAttributes = list(
                list(
                  AttributeType = "INCLUSIVE"|"EXCLUSIVE"|"CONTAINS"|"BEFORE"|"AFTER"|"ON"|"BETWEEN",
                  Values = list(
                    "string"
                  )
                )
              )
            )
          ),
          SourceSegments = list(
            list(
              Id = "string",
              Version = 123
            )
          ),
          SourceType = "ALL"|"ANY"|"NONE",
          Type = "ALL"|"ANY"|"NONE"
        )
      ),
      Include = "ALL"|"ANY"|"NONE"
    ),
    SegmentType = "DIMENSIONAL"|"IMPORT",
    tags = list(
      "string"
    ),
    Version = 123
  )
)
}
}
\description{
Retrieves information about the configuration, dimension, and other
settings for a specific version of a segment that's associated with an
application.
}
\section{Request syntax}{
\preformatted{svc$get_segment_version(
  ApplicationId = "string",
  SegmentId = "string",
  Version = "string"
)
}
}

\keyword{internal}
