% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_get_user_endpoints}
\alias{pinpoint_get_user_endpoints}
\title{Retrieves information about all the endpoints that are associated with a
specific user ID}
\usage{
pinpoint_get_user_endpoints(ApplicationId, UserId)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{UserId}{[required] The unique identifier for the user.}
}
\description{
Retrieves information about all the endpoints that are associated with a specific user ID.

See \url{https://paws-r.github.io/docs/pinpoint/get_user_endpoints.html} for full documentation.
}
\keyword{internal}
