% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_phone_number_validate}
\alias{pinpoint_phone_number_validate}
\title{Retrieves information about a phone number}
\usage{
pinpoint_phone_number_validate(NumberValidateRequest)
}
\arguments{
\item{NumberValidateRequest}{[required]}
}
\description{
Retrieves information about a phone number.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_phone_number_validate/} for full documentation.
}
\keyword{internal}
