% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_send_users_messages}
\alias{pinpoint_send_users_messages}
\title{Creates and sends a message to a list of users}
\usage{
pinpoint_send_users_messages(ApplicationId, SendUsersMessageRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{SendUsersMessageRequest}{[required]}
}
\description{
Creates and sends a message to a list of users.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_send_users_messages/} for full documentation.
}
\keyword{internal}
