% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_tag_resource}
\alias{pinpoint_tag_resource}
\title{Adds one or more tags (keys and values) to an application, campaign,
message template, or segment}
\usage{
pinpoint_tag_resource(ResourceArn, TagsModel)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{TagsModel}{[required]}
}
\value{
An empty list.
}
\description{
Adds one or more tags (keys and values) to an application, campaign,
message template, or segment.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  TagsModel = list(
    tags = list(
      "string"
    )
  )
)
}
}

\keyword{internal}
