% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_adm_channel}
\alias{pinpoint_update_adm_channel}
\title{Enables the ADM channel for an application or updates the status and
settings of the ADM channel for an application}
\usage{
pinpoint_update_adm_channel(ADMChannelRequest, ApplicationId)
}
\arguments{
\item{ADMChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ADMChannelResponse = list(
    ApplicationId = "string",
    CreationDate = "string",
    Enabled = TRUE|FALSE,
    HasCredential = TRUE|FALSE,
    Id = "string",
    IsArchived = TRUE|FALSE,
    LastModifiedBy = "string",
    LastModifiedDate = "string",
    Platform = "string",
    Version = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Enables the ADM channel for an application or updates the status and
settings of the ADM channel for an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_adm_channel(
  ADMChannelRequest = list(
    ClientId = "string",
    ClientSecret = "string",
    Enabled = TRUE|FALSE
  ),
  ApplicationId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
