% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_apns_sandbox_channel}
\alias{pinpoint_update_apns_sandbox_channel}
\title{Enables the APNs sandbox channel for an application or updates the
status and settings of the APNs sandbox channel for an application}
\usage{
pinpoint_update_apns_sandbox_channel(APNSSandboxChannelRequest, ApplicationId)
}
\arguments{
\item{APNSSandboxChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\description{
Enables the APNs sandbox channel for an application or updates the status and settings of the APNs sandbox channel for an application.

See \url{https://paws-r.github.io/docs/pinpoint/update_apns_sandbox_channel.html} for full documentation.
}
\keyword{internal}
