% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_apns_voip_channel}
\alias{pinpoint_update_apns_voip_channel}
\title{Enables the APNs VoIP channel for an application or updates the status
and settings of the APNs VoIP channel for an application}
\usage{
pinpoint_update_apns_voip_channel(APNSVoipChannelRequest, ApplicationId)
}
\arguments{
\item{APNSVoipChannelRequest}{[required]}

\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}
}
\description{
Enables the APNs VoIP channel for an application or updates the status and settings of the APNs VoIP channel for an application.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_update_apns_voip_channel/} for full documentation.
}
\keyword{internal}
