% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_baidu_channel}
\alias{pinpoint_update_baidu_channel}
\title{Enables the Baidu channel for an application or updates the status and
settings of the Baidu channel for an application}
\usage{
pinpoint_update_baidu_channel(ApplicationId, BaiduChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{BaiduChannelRequest}{[required]}
}
\value{
A list with the following syntax:\preformatted{list(
  BaiduChannelResponse = list(
    ApplicationId = "string",
    CreationDate = "string",
    Credential = "string",
    Enabled = TRUE|FALSE,
    HasCredential = TRUE|FALSE,
    Id = "string",
    IsArchived = TRUE|FALSE,
    LastModifiedBy = "string",
    LastModifiedDate = "string",
    Platform = "string",
    Version = 123
  )
)
}
}
\description{
Enables the Baidu channel for an application or updates the status and
settings of the Baidu channel for an application.
}
\section{Request syntax}{
\preformatted{svc$update_baidu_channel(
  ApplicationId = "string",
  BaiduChannelRequest = list(
    ApiKey = "string",
    Enabled = TRUE|FALSE,
    SecretKey = "string"
  )
)
}
}

\keyword{internal}
