% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_baidu_channel}
\alias{pinpoint_update_baidu_channel}
\title{Update a BAIDU GCM channel}
\usage{
pinpoint_update_baidu_channel(ApplicationId, BaiduChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{BaiduChannelRequest}{[required]}
}
\description{
Update a BAIDU GCM channel
}
\section{Request syntax}{
\preformatted{svc$update_baidu_channel(
  ApplicationId = "string",
  BaiduChannelRequest = list(
    ApiKey = "string",
    Enabled = TRUE|FALSE,
    SecretKey = "string"
  )
)
}
}

\keyword{internal}
