% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_email_channel}
\alias{pinpoint_update_email_channel}
\title{Enables the email channel for an application or updates the status and
settings of the email channel for an application}
\usage{
pinpoint_update_email_channel(ApplicationId, EmailChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{EmailChannelRequest}{[required]}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  EmailChannelResponse = list(
    ApplicationId = "string",
    ConfigurationSet = "string",
    CreationDate = "string",
    Enabled = TRUE|FALSE,
    FromAddress = "string",
    HasCredential = TRUE|FALSE,
    Id = "string",
    Identity = "string",
    IsArchived = TRUE|FALSE,
    LastModifiedBy = "string",
    LastModifiedDate = "string",
    MessagesPerSecond = 123,
    Platform = "string",
    RoleArn = "string",
    Version = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Enables the email channel for an application or updates the status and
settings of the email channel for an application.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_email_channel(
  ApplicationId = "string",
  EmailChannelRequest = list(
    ConfigurationSet = "string",
    Enabled = TRUE|FALSE,
    FromAddress = "string",
    Identity = "string",
    RoleArn = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
