% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_recommender_configuration}
\alias{pinpoint_update_recommender_configuration}
\title{Updates an Amazon Pinpoint configuration for a recommender model}
\usage{
pinpoint_update_recommender_configuration(
  RecommenderId,
  UpdateRecommenderConfiguration
)
}
\arguments{
\item{RecommenderId}{[required] The unique identifier for the recommender model configuration. This
identifier is displayed as the \strong{Recommender ID} on the Amazon Pinpoint
console.}

\item{UpdateRecommenderConfiguration}{[required]}
}
\description{
Updates an Amazon Pinpoint configuration for a recommender model.

See \url{https://www.paws-r-sdk.com/docs/pinpoint_update_recommender_configuration/} for full documentation.
}
\keyword{internal}
