% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_sms_channel}
\alias{pinpoint_update_sms_channel}
\title{Enables the SMS channel for an application or updates the status and
settings of the SMS channel for an application}
\usage{
pinpoint_update_sms_channel(ApplicationId, SMSChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{SMSChannelRequest}{[required]}
}
\value{
A list with the following syntax:\preformatted{list(
  SMSChannelResponse = list(
    ApplicationId = "string",
    CreationDate = "string",
    Enabled = TRUE|FALSE,
    HasCredential = TRUE|FALSE,
    Id = "string",
    IsArchived = TRUE|FALSE,
    LastModifiedBy = "string",
    LastModifiedDate = "string",
    Platform = "string",
    PromotionalMessagesPerSecond = 123,
    SenderId = "string",
    ShortCode = "string",
    TransactionalMessagesPerSecond = 123,
    Version = 123
  )
)
}
}
\description{
Enables the SMS channel for an application or updates the status and
settings of the SMS channel for an application.
}
\section{Request syntax}{
\preformatted{svc$update_sms_channel(
  ApplicationId = "string",
  SMSChannelRequest = list(
    Enabled = TRUE|FALSE,
    SenderId = "string",
    ShortCode = "string"
  )
)
}
}

\keyword{internal}
