% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_update_sms_channel}
\alias{pinpoint_update_sms_channel}
\title{Enables the SMS channel for an application or updates the status and
settings of the SMS channel for an application}
\usage{
pinpoint_update_sms_channel(ApplicationId, SMSChannelRequest)
}
\arguments{
\item{ApplicationId}{[required] The unique identifier for the application. This identifier is displayed
as the \strong{Project ID} on the Amazon Pinpoint console.}

\item{SMSChannelRequest}{[required]}
}
\description{
Enables the SMS channel for an application or updates the status and settings of the SMS channel for an application.

See \url{https://paws-r.github.io/docs/pinpoint/update_sms_channel.html} for full documentation.
}
\keyword{internal}
