% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpoint_operations.R
\name{pinpoint_verify_otp_message}
\alias{pinpoint_verify_otp_message}
\title{Verify an OTP}
\usage{
pinpoint_verify_otp_message(ApplicationId, VerifyOTPMessageRequestParameters)
}
\arguments{
\item{ApplicationId}{[required] The unique ID of your Amazon Pinpoint application.}

\item{VerifyOTPMessageRequestParameters}{[required]}
}
\description{
Verify an OTP

See \url{https://www.paws-r-sdk.com/docs/pinpoint_verify_otp_message/} for full documentation.
}
\keyword{internal}
