% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_delete_configuration_set_event_destination}
\alias{pinpointemail_delete_configuration_set_event_destination}
\title{Delete an event destination}
\usage{
pinpointemail_delete_configuration_set_event_destination(
  ConfigurationSetName,
  EventDestinationName
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that contains the event destination
that you want to delete.}

\item{EventDestinationName}{[required] The name of the event destination that you want to delete.}
}
\description{
Delete an event destination.

See \url{https://paws-r.github.io/docs/pinpointemail/delete_configuration_set_event_destination.html} for full documentation.
}
\keyword{internal}
