% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_get_account}
\alias{pinpointemail_get_account}
\title{Obtain information about the email-sending status and capabilities of
your Amazon Pinpoint account in the current AWS Region}
\usage{
pinpointemail_get_account()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SendQuota = list(
    Max24HourSend = 123.0,
    MaxSendRate = 123.0,
    SentLast24Hours = 123.0
  ),
  SendingEnabled = TRUE|FALSE,
  DedicatedIpAutoWarmupEnabled = TRUE|FALSE,
  EnforcementStatus = "string",
  ProductionAccessEnabled = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
Obtain information about the email-sending status and capabilities of
your Amazon Pinpoint account in the current AWS Region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_account()
}\if{html}{\out{</div>}}
}

\keyword{internal}
