% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_get_blacklist_reports}
\alias{pinpointemail_get_blacklist_reports}
\title{Retrieve a list of the blacklists that your dedicated IP addresses
appear on}
\usage{
pinpointemail_get_blacklist_reports(BlacklistItemNames)
}
\arguments{
\item{BlacklistItemNames}{[required] A list of IP addresses that you want to retrieve blacklist information
about. You can only specify the dedicated IP addresses that you use to
send email using Amazon Pinpoint or Amazon SES.}
}
\description{
Retrieve a list of the blacklists that your dedicated IP addresses appear on.

See \url{https://paws-r.github.io/docs/pinpointemail/get_blacklist_reports.html} for full documentation.
}
\keyword{internal}
