% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_get_configuration_set}
\alias{pinpointemail_get_configuration_set}
\title{Get information about an existing configuration set, including the
dedicated IP pool that it's associated with, whether or not it's enabled
for sending email, and more}
\usage{
pinpointemail_get_configuration_set(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to obtain more
information about.}
}
\description{
Get information about an existing configuration set, including the dedicated IP pool that it's associated with, whether or not it's enabled for sending email, and more.

See \url{https://www.paws-r-sdk.com/docs/pinpointemail_get_configuration_set/} for full documentation.
}
\keyword{internal}
