% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_get_deliverability_dashboard_options}
\alias{pinpointemail_get_deliverability_dashboard_options}
\title{Retrieve information about the status of the Deliverability dashboard
for your Amazon Pinpoint account}
\usage{
pinpointemail_get_deliverability_dashboard_options()
}
\description{
Retrieve information about the status of the Deliverability dashboard for your Amazon Pinpoint account. When the Deliverability dashboard is enabled, you gain access to reputation, deliverability, and other metrics for the domains that you use to send email using Amazon Pinpoint. You also gain the ability to perform predictive inbox placement tests.

See \url{https://paws-r.github.io/docs/pinpointemail/get_deliverability_dashboard_options.html} for full documentation.
}
\keyword{internal}
