% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_get_email_identity}
\alias{pinpointemail_get_email_identity}
\title{Provides information about a specific identity associated with your
Amazon Pinpoint account, including the identity's verification status,
its DKIM authentication status, and its custom Mail-From settings}
\usage{
pinpointemail_get_email_identity(EmailIdentity)
}
\arguments{
\item{EmailIdentity}{[required] The email identity that you want to retrieve details for.}
}
\description{
Provides information about a specific identity associated with your Amazon Pinpoint account, including the identity's verification status, its DKIM authentication status, and its custom Mail-From settings.

See \url{https://paws-r.github.io/docs/pinpointemail/get_email_identity.html} for full documentation.
}
\keyword{internal}
