% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_list_deliverability_test_reports}
\alias{pinpointemail_list_deliverability_test_reports}
\title{Show a list of the predictive inbox placement tests that you've
performed, regardless of their statuses}
\usage{
pinpointemail_list_deliverability_test_reports(NextToken, PageSize)
}
\arguments{
\item{NextToken}{A token returned from a previous call to
\code{\link[=pinpointemail_list_deliverability_test_reports]{list_deliverability_test_reports}}
to indicate the position in the list of predictive inbox placement
tests.}

\item{PageSize}{The number of results to show in a single call to
\code{\link[=pinpointemail_list_deliverability_test_reports]{list_deliverability_test_reports}}.
If the number of results is larger than the number you specified in this
parameter, then the response includes a \code{NextToken} element, which you
can use to obtain additional results.

The value you specify has to be at least 0, and can be no more than
1000.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DeliverabilityTestReports = list(
    list(
      ReportId = "string",
      ReportName = "string",
      Subject = "string",
      FromEmailAddress = "string",
      CreateDate = as.POSIXct(
        "2015-01-01"
      ),
      DeliverabilityTestStatus = "IN_PROGRESS"|"COMPLETED"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Show a list of the predictive inbox placement tests that you've
performed, regardless of their statuses. For predictive inbox placement
tests that are complete, you can use the
\code{\link[=pinpointemail_get_deliverability_test_report]{get_deliverability_test_report}}
operation to view the results.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_deliverability_test_reports(
  NextToken = "string",
  PageSize = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
