% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointemail_operations.R
\name{pinpointemail_put_dedicated_ip_warmup_attributes}
\alias{pinpointemail_put_dedicated_ip_warmup_attributes}
\title{Put dedicated ip warmup attributes}
\usage{
pinpointemail_put_dedicated_ip_warmup_attributes(Ip, WarmupPercentage)
}
\arguments{
\item{Ip}{[required] The dedicated IP address that you want to update the warm-up attributes
for.}

\item{WarmupPercentage}{[required] The warm-up percentage that you want to associate with the dedicated IP
address.}
}
\value{
An empty list.
}
\description{
Put dedicated ip warmup attributes
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$put_dedicated_ip_warmup_attributes(
  Ip = "string",
  WarmupPercentage = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
