% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoice_operations.R
\name{pinpointsmsvoice_delete_configuration_set}
\alias{pinpointsmsvoice_delete_configuration_set}
\title{Deletes an existing configuration set}
\usage{
pinpointsmsvoice_delete_configuration_set(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}
}
\description{
Deletes an existing configuration set.

See \url{https://paws-r.github.io/docs/pinpointsmsvoice/delete_configuration_set.html} for full documentation.
}
\keyword{internal}
