% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoice_operations.R
\name{pinpointsmsvoice_delete_configuration_set}
\alias{pinpointsmsvoice_delete_configuration_set}
\title{Deletes an existing configuration set}
\usage{
pinpointsmsvoice_delete_configuration_set(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}
}
\value{
An empty list.
}
\description{
Deletes an existing configuration set.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_configuration_set(
  ConfigurationSetName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
