% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoice_operations.R
\name{pinpointsmsvoice_delete_configuration_set_event_destination}
\alias{pinpointsmsvoice_delete_configuration_set_event_destination}
\title{Deletes an event destination in a configuration set}
\usage{
pinpointsmsvoice_delete_configuration_set_event_destination(
  ConfigurationSetName,
  EventDestinationName
)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}

\item{EventDestinationName}{[required] EventDestinationName}
}
\description{
Deletes an event destination in a configuration set.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoice_delete_configuration_set_event_destination/} for full documentation.
}
\keyword{internal}
