% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoice_operations.R
\name{pinpointsmsvoice_delete_configuration_set_event_destination}
\alias{pinpointsmsvoice_delete_configuration_set_event_destination}
\title{Deletes an event destination in a configuration set}
\usage{
pinpointsmsvoice_delete_configuration_set_event_destination(
  ConfigurationSetName, EventDestinationName)
}
\arguments{
\item{ConfigurationSetName}{[required] ConfigurationSetName}

\item{EventDestinationName}{[required] EventDestinationName}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes an event destination in a configuration set.
}
\section{Request syntax}{
\preformatted{svc$delete_configuration_set_event_destination(
  ConfigurationSetName = "string",
  EventDestinationName = "string"
)
}
}

\keyword{internal}
