% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoice_operations.R
\name{pinpointsmsvoice_list_configuration_sets}
\alias{pinpointsmsvoice_list_configuration_sets}
\title{List all of the configuration sets associated with your Amazon Pinpoint
account in the current region}
\usage{
pinpointsmsvoice_list_configuration_sets(NextToken, PageSize)
}
\arguments{
\item{NextToken}{A token returned from a previous call to the API that indicates the
position in the list of results.}

\item{PageSize}{Used to specify the number of items that should be returned in the
response.}
}
\value{
A list with the following syntax:\preformatted{list(
  ConfigurationSets = list(
    "string"
  ),
  NextToken = "string"
)
}
}
\description{
List all of the configuration sets associated with your Amazon Pinpoint
account in the current region.
}
\section{Request syntax}{
\preformatted{svc$list_configuration_sets(
  NextToken = "string",
  PageSize = "string"
)
}
}

\keyword{internal}
