% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_carrier_lookup}
\alias{pinpointsmsvoicev2_carrier_lookup}
\title{Returns information about a destination phone number, including whether
the number type and whether it is valid, the carrier, and more}
\usage{
pinpointsmsvoicev2_carrier_lookup(PhoneNumber)
}
\arguments{
\item{PhoneNumber}{[required] The phone number that you want to retrieve information about. You can
provide the phone number in various formats including special characters
such as parentheses, brackets, spaces, hyphens, periods, and commas. The
service automatically converts the input to E164 format for processing.}
}
\description{
Returns information about a destination phone number, including whether the number type and whether it is valid, the carrier, and more.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_carrier_lookup/} for full documentation.
}
\keyword{internal}
