% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_delete_keyword}
\alias{pinpointsmsvoicev2_delete_keyword}
\title{Deletes an existing keyword from an origination phone number or pool}
\usage{
pinpointsmsvoicev2_delete_keyword(OriginationIdentity, Keyword)
}
\arguments{
\item{OriginationIdentity}{[required] The origination identity to use such as a PhoneNumberId, PhoneNumberArn,
PoolId or PoolArn. You can use
\code{\link[=pinpointsmsvoicev2_describe_phone_numbers]{describe_phone_numbers}} to
find the values for PhoneNumberId and PhoneNumberArn and
\code{\link[=pinpointsmsvoicev2_describe_pools]{describe_pools}} to find the values
of PoolId and PoolArn.

If you are using a shared End User Messaging SMS resource then you must
use the full Amazon Resource Name(ARN).}

\item{Keyword}{[required] The keyword to delete.}
}
\description{
Deletes an existing keyword from an origination phone number or pool.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_delete_keyword/} for full documentation.
}
\keyword{internal}
