% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_pools}
\alias{pinpointsmsvoicev2_describe_pools}
\title{Retrieves the specified pools or all pools associated with your Amazon
Web Services account}
\usage{
pinpointsmsvoicev2_describe_pools(
  PoolIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Owner = NULL
)
}
\arguments{
\item{PoolIds}{The unique identifier of pools to find. This is an array of strings that
can be either the PoolId or PoolArn.

If you are using a shared End User Messaging SMS resource then you must
use the full Amazon Resource Name(ARN).}

\item{Filters}{An array of PoolFilter objects to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}

\item{Owner}{Use \code{SELF} to filter the list of Pools to ones your account owns or use
\code{SHARED} to filter on Pools shared with your account. The \code{Owner} and
\code{PoolIds} parameters can't be used at the same time.}
}
\description{
Retrieves the specified pools or all pools associated with your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_pools/} for full documentation.
}
\keyword{internal}
