% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_describe_sender_ids}
\alias{pinpointsmsvoicev2_describe_sender_ids}
\title{Describes the specified SenderIds or all SenderIds associated with your
Amazon Web Services account}
\usage{
pinpointsmsvoicev2_describe_sender_ids(
  SenderIds = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Owner = NULL
)
}
\arguments{
\item{SenderIds}{An array of SenderIdAndCountry objects to search for.

If you are using a shared End User Messaging SMS resource then you must
use the full Amazon Resource Name(ARN).}

\item{Filters}{An array of SenderIdFilter objects to filter the results.}

\item{NextToken}{The token to be used for the next set of paginated results. You don't
need to supply a value for this field in the initial request.}

\item{MaxResults}{The maximum number of results to return per each request.}

\item{Owner}{Use \code{SELF} to filter the list of Sender Ids to ones your account owns or
use \code{SHARED} to filter on Sender Ids shared with your account. The
\code{Owner} and \code{SenderIds} parameters can't be used at the same time.}
}
\description{
Describes the specified SenderIds or all SenderIds associated with your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_describe_sender_ids/} for full documentation.
}
\keyword{internal}
