% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_list_available_phone_numbers}
\alias{pinpointsmsvoicev2_list_available_phone_numbers}
\title{Search available phone numbers from aggregator inventory, optionally
filtered by pattern}
\usage{
pinpointsmsvoicev2_list_available_phone_numbers(
  IsoCountryCode,
  NumberCapabilities,
  NumberType,
  RegistrationId = NULL,
  NumberPreference = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{IsoCountryCode}{[required] The two-character code, in ISO 3166-1 alpha-2 format, for the country or region in which to search for available phone numbers. This operation currently supports only \code{US}.}

\item{NumberCapabilities}{[required] The capabilities to filter by, such as SMS. Only phone numbers that support all of the specified capabilities are returned.}

\item{NumberType}{[required] The type of phone number to search for.}

\item{RegistrationId}{The registration associated with the request. A registration is required for regulated number types. You can specify either:
\itemize{
\item The unique identifier of the registration.
\item The Amazon Resource Name (ARN) of the registration.
}}

\item{NumberPreference}{Optional. If omitted, returns unfiltered available numbers. Max 1 element for List API.}

\item{NextToken}{The token returned from a previous request to retrieve the next page of results.}

\item{MaxResults}{The maximum number of results to return per page. If you don't specify a value, the default is 10.}
}
\description{
Search available phone numbers from aggregator inventory, optionally filtered by pattern. If NumberPreference is omitted, returns unfiltered available numbers. Returns empty list (not an exception) when no numbers match. ResourceNotFoundException is thrown only for invalid RegistrationId (campaign not found).

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_list_available_phone_numbers/} for full documentation.
}
\keyword{internal}
