% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_put_keyword}
\alias{pinpointsmsvoicev2_put_keyword}
\title{Creates or updates a keyword configuration on an origination phone
number or pool}
\usage{
pinpointsmsvoicev2_put_keyword(
  OriginationIdentity,
  Keyword,
  KeywordMessage,
  KeywordAction = NULL
)
}
\arguments{
\item{OriginationIdentity}{[required] The origination identity to use such as a PhoneNumberId, PhoneNumberArn,
SenderId or SenderIdArn. You can use
\code{\link[=pinpointsmsvoicev2_describe_phone_numbers]{describe_phone_numbers}}
get the values for PhoneNumberId and PhoneNumberArn while
\code{\link[=pinpointsmsvoicev2_describe_sender_ids]{describe_sender_ids}} can be
used to get the values for SenderId and SenderIdArn.

If you are using a shared End User Messaging SMS resource then you must
use the full Amazon Resource Name(ARN).}

\item{Keyword}{[required] The new keyword to add.}

\item{KeywordMessage}{[required] The message associated with the keyword.}

\item{KeywordAction}{The action to perform for the new keyword when it is received.
\itemize{
\item AUTOMATIC_RESPONSE: A message is sent to the recipient.
\item OPT_OUT: Keeps the recipient from receiving future messages.
\item OPT_IN: The recipient wants to receive future messages.
}}
}
\description{
Creates or updates a keyword configuration on an origination phone number or pool.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_put_keyword/} for full documentation.
}
\keyword{internal}
