% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinpointsmsvoicev2_operations.R
\name{pinpointsmsvoicev2_update_notify_configuration}
\alias{pinpointsmsvoicev2_update_notify_configuration}
\title{Updates an existing notify configuration}
\usage{
pinpointsmsvoicev2_update_notify_configuration(
  NotifyConfigurationId,
  DefaultTemplateId = NULL,
  PoolId = NULL,
  EnabledCountries = NULL,
  EnabledChannels = NULL,
  DeletionProtectionEnabled = NULL
)
}
\arguments{
\item{NotifyConfigurationId}{[required] The identifier of the notify configuration to update. The
NotifyConfigurationId can be found using the
\code{\link[=pinpointsmsvoicev2_describe_notify_configurations]{describe_notify_configurations}}
operation.}

\item{DefaultTemplateId}{The template ID to set as the default, or the special value
UNSET_DEFAULT_TEMPLATE to clear the current default template.}

\item{PoolId}{The pool ID or ARN to associate, or the special value
UNSET_DEFAULT_POOL_FOR_NOTIFY to clear the current default pool.}

\item{EnabledCountries}{An array of two-character ISO country codes, in ISO 3166-1 alpha-2
format, that are enabled for the notify configuration.}

\item{EnabledChannels}{An array of channels to enable for the notify configuration. Supported
values include \code{SMS} and \code{VOICE}.}

\item{DeletionProtectionEnabled}{When set to true the notify configuration can't be deleted.}
}
\description{
Updates an existing notify configuration. You can update the default template, pool association, enabled channels, enabled countries, and deletion protection settings.

See \url{https://www.paws-r-sdk.com/docs/pinpointsmsvoicev2_update_notify_configuration/} for full documentation.
}
\keyword{internal}
