% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_create_configuration_set}
\alias{ses_create_configuration_set}
\title{Creates a configuration set}
\usage{
ses_create_configuration_set(ConfigurationSet)
}
\arguments{
\item{ConfigurationSet}{[required] A data structure that contains the name of the configuration set.}
}
\description{
Creates a configuration set.

See \url{https://www.paws-r-sdk.com/docs/ses_create_configuration_set/} for full documentation.
}
\keyword{internal}
