% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_create_configuration_set_event_destination}
\alias{ses_create_configuration_set_event_destination}
\title{Creates a configuration set event destination}
\usage{
ses_create_configuration_set_event_destination(
  ConfigurationSetName,
  EventDestination
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that the event destination should be
associated with.}

\item{EventDestination}{[required] An object that describes the AWS service that email sending event
information will be published to.}
}
\description{
Creates a configuration set event destination.

See \url{https://paws-r.github.io/docs/ses/create_configuration_set_event_destination.html} for full documentation.
}
\keyword{internal}
