% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_create_configuration_set_tracking_options}
\alias{ses_create_configuration_set_tracking_options}
\title{Creates an association between a configuration set and a custom domain
for open and click event tracking}
\usage{
ses_create_configuration_set_tracking_options(
  ConfigurationSetName,
  TrackingOptions
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that the tracking options should be
associated with.}

\item{TrackingOptions}{[required]}
}
\description{
Creates an association between a configuration set and a custom domain for open and click event tracking.

See \url{https://www.paws-r-sdk.com/docs/ses_create_configuration_set_tracking_options/} for full documentation.
}
\keyword{internal}
