% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_create_template}
\alias{ses_create_template}
\title{Creates an email template}
\usage{
ses_create_template(Template)
}
\arguments{
\item{Template}{[required] The content of the email, composed of a subject line, an HTML part, and
a text-only part.}
}
\value{
An empty list.
}
\description{
Creates an email template. Email templates enable you to send
personalized email to one or more destinations in a single API
operation. For more information, see the \href{https://docs.aws.amazon.com/ses/latest/dg/send-personalized-email-api.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\section{Request syntax}{
\preformatted{svc$create_template(
  Template = list(
    TemplateName = "string",
    SubjectPart = "string",
    TextPart = "string",
    HtmlPart = "string"
  )
)
}
}

\keyword{internal}
