% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_delete_receipt_filter}
\alias{ses_delete_receipt_filter}
\title{Deletes the specified IP address filter}
\usage{
ses_delete_receipt_filter(FilterName)
}
\arguments{
\item{FilterName}{[required] The name of the IP address filter to delete.}
}
\description{
Deletes the specified IP address filter.

See \url{https://www.paws-r-sdk.com/docs/ses_delete_receipt_filter/} for full documentation.
}
\keyword{internal}
