% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_delete_template}
\alias{ses_delete_template}
\title{Deletes an email template}
\usage{
ses_delete_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the template to be deleted.}
}
\value{
An empty list.
}
\description{
Deletes an email template.

You can execute this operation no more than once per second.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_template(
  TemplateName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
