% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_delete_verified_email_address}
\alias{ses_delete_verified_email_address}
\title{Deprecated}
\usage{
ses_delete_verified_email_address(EmailAddress)
}
\arguments{
\item{EmailAddress}{[required] An email address to be removed from the list of verified addresses.}
}
\description{
Deprecated. Use the \code{\link[=ses_delete_identity]{delete_identity}} operation
to delete email addresses and domains.
}
\section{Request syntax}{
\preformatted{svc$delete_verified_email_address(
  EmailAddress = "string"
)
}
}

\examples{
\dontrun{
# The following example deletes an email address from the list of
# identities that have been submitted for verification with Amazon SES:
svc$delete_verified_email_address(
  EmailAddress = "user@example.com"
)
}

}
\keyword{internal}
