% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_describe_active_receipt_rule_set}
\alias{ses_describe_active_receipt_rule_set}
\title{Returns the metadata and receipt rules for the receipt rule set that is
currently active}
\usage{
ses_describe_active_receipt_rule_set()
}
\description{
Returns the metadata and receipt rules for the receipt rule set that is currently active.

See \url{https://paws-r.github.io/docs/ses/describe_active_receipt_rule_set.html} for full documentation.
}
\keyword{internal}
