% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_describe_configuration_set}
\alias{ses_describe_configuration_set}
\title{Returns the details of the specified configuration set}
\usage{
ses_describe_configuration_set(
  ConfigurationSetName,
  ConfigurationSetAttributeNames = NULL
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set to describe.}

\item{ConfigurationSetAttributeNames}{A list of configuration set attributes to return.}
}
\description{
Returns the details of the specified configuration set. For information about using configuration sets, see the \href{https://docs.aws.amazon.com/ses/latest/dg/monitor-sending-activity.html}{Amazon SES Developer Guide}.

See \url{https://www.paws-r-sdk.com/docs/ses_describe_configuration_set/} for full documentation.
}
\keyword{internal}
