% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_get_account_sending_enabled}
\alias{ses_get_account_sending_enabled}
\title{Returns the email sending status of the Amazon SES account for the
current region}
\usage{
ses_get_account_sending_enabled()
}
\description{
Returns the email sending status of the Amazon SES account for the current region.

See \url{https://www.paws-r-sdk.com/docs/ses_get_account_sending_enabled/} for full documentation.
}
\keyword{internal}
