% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_get_send_quota}
\alias{ses_get_send_quota}
\title{Provides the sending limits for the Amazon SES account}
\usage{
ses_get_send_quota()
}
\description{
Provides the sending limits for the Amazon SES account.

See \url{https://paws-r.github.io/docs/ses/get_send_quota.html} for full documentation.
}
\keyword{internal}
