% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_get_send_quota}
\alias{ses_get_send_quota}
\title{Provides the sending limits for the Amazon SES account}
\usage{
ses_get_send_quota()
}
\value{
A list with the following syntax:\preformatted{list(
  Max24HourSend = 123.0,
  MaxSendRate = 123.0,
  SentLast24Hours = 123.0
)
}
}
\description{
Provides the sending limits for the Amazon SES account.

You can execute this operation no more than once per second.
}
\examples{
\dontrun{
# The following example returns the Amazon SES sending limits for an AWS
# account:
svc$get_send_quota()
}

}
\keyword{internal}
