% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_get_send_statistics}
\alias{ses_get_send_statistics}
\title{Provides sending statistics for the current AWS Region}
\usage{
ses_get_send_statistics()
}
\description{
Provides sending statistics for the current AWS Region. The result is a list of data points, representing the last two weeks of sending activity. Each data point in the list contains statistics for a 15-minute period of time.

See \url{https://www.paws-r-sdk.com/docs/ses_get_send_statistics/} for full documentation.
}
\keyword{internal}
