% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_list_templates}
\alias{ses_list_templates}
\title{Lists the email templates present in your Amazon SES account in the
current Amazon Web Services Region}
\usage{
ses_list_templates(NextToken = NULL, MaxItems = NULL)
}
\arguments{
\item{NextToken}{A token returned from a previous call to
\code{\link[=ses_list_templates]{list_templates}} to indicate the position in the
list of email templates.}

\item{MaxItems}{The maximum number of templates to return. This value must be at least 1
and less than or equal to 100. If more than 100 items are requested, the
page size will automatically set to 100. If you do not specify a value,
10 is the default page size.}
}
\description{
Lists the email templates present in your Amazon SES account in the current Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/ses_list_templates/} for full documentation.
}
\keyword{internal}
