% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_list_verified_email_addresses}
\alias{ses_list_verified_email_addresses}
\title{Deprecated}
\usage{
ses_list_verified_email_addresses()
}
\description{
Deprecated. Use the \code{\link[=ses_list_identities]{list_identities}} operation to list the email addresses and domains associated with your account.

See \url{https://paws-r.github.io/docs/ses/list_verified_email_addresses.html} for full documentation.
}
\keyword{internal}
