% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_reorder_receipt_rule_set}
\alias{ses_reorder_receipt_rule_set}
\title{Reorders the receipt rules within a receipt rule set}
\usage{
ses_reorder_receipt_rule_set(RuleSetName, RuleNames)
}
\arguments{
\item{RuleSetName}{[required] The name of the receipt rule set to reorder.}

\item{RuleNames}{[required] The specified receipt rule set's receipt rules, in order.}
}
\description{
Reorders the receipt rules within a receipt rule set.

See \url{https://www.paws-r-sdk.com/docs/ses_reorder_receipt_rule_set/} for full documentation.
}
\keyword{internal}
