% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_operations.R
\name{ses_send_email}
\alias{ses_send_email}
\title{Composes an email message and immediately queues it for sending}
\usage{
ses_send_email(
  Source,
  Destination,
  Message,
  ReplyToAddresses = NULL,
  ReturnPath = NULL,
  SourceArn = NULL,
  ReturnPathArn = NULL,
  Tags = NULL,
  ConfigurationSetName = NULL
)
}
\arguments{
\item{Source}{[required] The email address that is sending the email. This email address must be
either individually verified with Amazon SES, or from a domain that has
been verified with Amazon SES. For information about verifying
identities, see the \href{https://docs.aws.amazon.com/ses/latest/dg/verify-addresses-and-domains.html}{Amazon SES Developer Guide}.

If you are sending on behalf of another user and have been permitted to
do so by a sending authorization policy, then you must also specify the
\code{SourceArn} parameter. For more information about sending authorization,
see the \href{https://docs.aws.amazon.com/ses/latest/dg/sending-authorization.html}{Amazon SES Developer Guide}.

Amazon SES does not support the SMTPUTF8 extension, as described in
\href{https://datatracker.ietf.org/doc/html/rfc6531}{RFC6531}. For this
reason, the \emph{local part} of a source email address (the part of the
email address that precedes the @ sign) may only contain \href{https://en.wikipedia.org/wiki/Email_address#Local-part}{7-bit ASCII characters}. If
the \emph{domain part} of an address (the part after the @ sign) contains
non-ASCII characters, they must be encoded using Punycode, as described
in \href{https://datatracker.ietf.org/doc/html/rfc3492.html}{RFC3492}. The
sender name (also known as the \emph{friendly name}) may contain non-ASCII
characters. These characters must be encoded using MIME encoded-word
syntax, as described in \href{https://datatracker.ietf.org/doc/html/rfc2047}{RFC 2047}. MIME encoded-word
syntax uses the following form: \verb{=?charset?encoding?encoded-text?=}.}

\item{Destination}{[required] The destination for this email, composed of To:, CC:, and BCC: fields.}

\item{Message}{[required] The message to be sent.}

\item{ReplyToAddresses}{The reply-to email address(es) for the message. If the recipient replies
to the message, each reply-to address will receive the reply.}

\item{ReturnPath}{The email address that bounces and complaints will be forwarded to when
feedback forwarding is enabled. If the message cannot be delivered to
the recipient, then an error message will be returned from the
recipient's ISP; this message will then be forwarded to the email
address specified by the \code{ReturnPath} parameter. The \code{ReturnPath}
parameter is never overwritten. This email address must be either
individually verified with Amazon SES, or from a domain that has been
verified with Amazon SES.}

\item{SourceArn}{This parameter is used only for sending authorization. It is the ARN of
the identity that is associated with the sending authorization policy
that permits you to send for the email address specified in the \code{Source}
parameter.

For example, if the owner of \code{example.com} (which has ARN
\code{arn:aws:ses:us-east-1:123456789012:identity/example.com}) attaches a
policy to it that authorizes you to send from \code{user@example.com}, then
you would specify the \code{SourceArn} to be
\code{arn:aws:ses:us-east-1:123456789012:identity/example.com}, and the
\code{Source} to be \code{user@example.com}.

For more information about sending authorization, see the \href{https://docs.aws.amazon.com/ses/latest/dg/sending-authorization.html}{Amazon SES Developer Guide}.}

\item{ReturnPathArn}{This parameter is used only for sending authorization. It is the ARN of
the identity that is associated with the sending authorization policy
that permits you to use the email address specified in the \code{ReturnPath}
parameter.

For example, if the owner of \code{example.com} (which has ARN
\code{arn:aws:ses:us-east-1:123456789012:identity/example.com}) attaches a
policy to it that authorizes you to use \code{feedback@example.com}, then you
would specify the \code{ReturnPathArn} to be
\code{arn:aws:ses:us-east-1:123456789012:identity/example.com}, and the
\code{ReturnPath} to be \code{feedback@example.com}.

For more information about sending authorization, see the \href{https://docs.aws.amazon.com/ses/latest/dg/sending-authorization.html}{Amazon SES Developer Guide}.}

\item{Tags}{A list of tags, in the form of name/value pairs, to apply to an email
that you send using \code{\link[=ses_send_email]{send_email}}. Tags correspond to
characteristics of the email that you define, so that you can publish
email sending events.}

\item{ConfigurationSetName}{The name of the configuration set to use when you send an email using
\code{\link[=ses_send_email]{send_email}}.}
}
\description{
Composes an email message and immediately queues it for sending. In order to send email using the \code{\link[=ses_send_email]{send_email}} operation, your message must meet the following requirements:

See \url{https://www.paws-r-sdk.com/docs/ses_send_email/} for full documentation.
}
\keyword{internal}
