% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_create_tenant}
\alias{sesv2_create_tenant}
\title{Create a tenant}
\usage{
sesv2_create_tenant(TenantName, Tags = NULL)
}
\arguments{
\item{TenantName}{[required] The name of the tenant to create. The name can contain up to 64
alphanumeric characters, including letters, numbers, hyphens (-) and
underscores (_) only.}

\item{Tags}{An array of objects that define the tags (keys and values) to associate
with the tenant}
}
\description{
Create a tenant.

See \url{https://www.paws-r-sdk.com/docs/sesv2_create_tenant/} for full documentation.
}
\keyword{internal}
