% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_configuration_set_event_destination}
\alias{sesv2_delete_configuration_set_event_destination}
\title{Delete an event destination}
\usage{
sesv2_delete_configuration_set_event_destination(
  ConfigurationSetName,
  EventDestinationName
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that contains the event destination to
delete.}

\item{EventDestinationName}{[required] The name of the event destination to delete.}
}
\description{
Delete an event destination.

See \url{https://paws-r.github.io/docs/sesv2/delete_configuration_set_event_destination.html} for full documentation.
}
\keyword{internal}
