% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_contact_list}
\alias{sesv2_delete_contact_list}
\title{Deletes a contact list and all of the contacts on that list}
\usage{
sesv2_delete_contact_list(ContactListName)
}
\arguments{
\item{ContactListName}{[required] The name of the contact list.}
}
\description{
Deletes a contact list and all of the contacts on that list.

See \url{https://www.paws-r-sdk.com/docs/sesv2_delete_contact_list/} for full documentation.
}
\keyword{internal}
