% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_custom_verification_email_template}
\alias{sesv2_delete_custom_verification_email_template}
\title{Deletes an existing custom verification email template}
\usage{
sesv2_delete_custom_verification_email_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the custom verification email template that you want to
delete.}
}
\description{
Deletes an existing custom verification email template.

See \url{https://www.paws-r-sdk.com/docs/sesv2_delete_custom_verification_email_template/} for full documentation.
}
\keyword{internal}
