% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_email_identity}
\alias{sesv2_delete_email_identity}
\title{Deletes an email identity}
\usage{
sesv2_delete_email_identity(EmailIdentity)
}
\arguments{
\item{EmailIdentity}{[required] The identity (that is, the email address or domain) to delete.}
}
\description{
Deletes an email identity. An identity can be either an email address or a domain name.

See \url{https://paws-r.github.io/docs/sesv2/delete_email_identity.html} for full documentation.
}
\keyword{internal}
