% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_delete_suppressed_destination}
\alias{sesv2_delete_suppressed_destination}
\title{Removes an email address from the suppression list for your account}
\usage{
sesv2_delete_suppressed_destination(EmailAddress)
}
\arguments{
\item{EmailAddress}{[required] The suppressed email destination to remove from the account suppression
list.}
}
\description{
Removes an email address from the suppression list for your account.

See \url{https://paws-r.github.io/docs/sesv2/delete_suppressed_destination.html} for full documentation.
}
\keyword{internal}
