% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_custom_verification_email_template}
\alias{sesv2_get_custom_verification_email_template}
\title{Returns the custom email verification template for the template name you
specify}
\usage{
sesv2_get_custom_verification_email_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the custom verification email template that you want to
retrieve.}
}
\description{
Returns the custom email verification template for the template name you specify.

See \url{https://paws-r.github.io/docs/sesv2/get_custom_verification_email_template.html} for full documentation.
}
\keyword{internal}
