% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_get_import_job}
\alias{sesv2_get_import_job}
\title{Provides information about an import job}
\usage{
sesv2_get_import_job(JobId)
}
\arguments{
\item{JobId}{[required] The ID of the import job.}
}
\description{
Provides information about an import job.

See \url{https://paws-r.github.io/docs/sesv2/get_import_job.html} for full documentation.
}
\keyword{internal}
