% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesv2_operations.R
\name{sesv2_put_account_vdm_attributes}
\alias{sesv2_put_account_vdm_attributes}
\title{Update your Amazon SES account VDM attributes}
\usage{
sesv2_put_account_vdm_attributes(VdmAttributes)
}
\arguments{
\item{VdmAttributes}{[required] The VDM attributes that you wish to apply to your Amazon SES account.}
}
\description{
Update your Amazon SES account VDM attributes.

See \url{https://www.paws-r-sdk.com/docs/sesv2_put_account_vdm_attributes/} for full documentation.
}
\keyword{internal}
